Repeated Spatial Cross Validation Resampling

Spatial Block Cross validation implemented by the blockCV package.

Details

By default blockCV::spatialBlock() does not allow the creation of multiple repetitions. mlr3spatiotempcv adds support for this when using the range argument for fold creation. When supplying a vector of length(repeats) for argument range, these different settings will be used to create folds which differ among the repetitions.

Multiple repetitions are not possible when using the "row & cols" approach because the created folds will always be the same.

References

Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G (2018). “blockCV: an R package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models.” bioRxiv. doi: 10.1101/357798 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVBlock

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVBlock$new()

ResamplingRepeatedSpCVBlock$folds()

ResamplingRepeatedSpCVBlock$repeats()

ResamplingRepeatedSpCVBlock$instantiate()

ResamplingRepeatedSpCVBlock$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "coordinate-based" repeated resampling instance.

Usage

ResamplingRepeatedSpCVBlock$new(id = "repeated-spcv-block")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVBlock$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVBlock$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVBlock$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVBlock$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("diplodia")

# Instantiate Resampling
rrcv = rsmp("repeated-spcv-block",
  folds = 3, repeats = 2,
  range = c(1000, 1500))
rrcv$instantiate(task)

# Individual sets:
rrcv$iters

#> [1] 6

rrcv$folds(1:6)

#> [1] 1 2 1 2 1 2

rrcv$repeats(1:6)

#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)

#>   [1]   1   4   8   9  14  16  17  23  25  26  29  34  37  44  45  54  55  58
#>  [19]  60  63  65  69  71  72  73  75  76  79  81  82  83  84  86  87  88  92
#>  [37]  93  99 101 102 106 107 109 112 118 119 120 123 124 127 148 154 155 156
#>  [55] 157 162 164 165 167 176 177 178 179 183 184 185 189 192 193 194 200 201
#>  [73] 202 203 204 205 211 213 217 219 220 221 223 224 235 237 239 243 244 252
#>  [91] 253 254 257 258 265 266 268 269 274 277 278 287 289 291 292 294 295 299
#> [109] 302 315 316 317 318 322 332 333 335 337 339 341 342 343 345 348 352 354
#> [127] 408 409 410 411 412 413 418 421 431 433 438 439 445 448 449 453 457 458
#> [145] 461 471 472 473 474 476 478 486 487 488 489 490 491 493 494 495 503 505
#> [163] 506 512 514 517 519 520 522 546 547 549 550 552 553 557 559 560 561 583
#> [181] 584 587 589 590 592 593 594 596 598 610 611 614 615 616 620 624 626 627
#> [199] 629 631 638 639 653 654 655 656 659 660 662 663 676 680 687 688 691 692
#> [217] 697 698 699 700 702 705 706 707 708 709 710 711 712 713 714 715 719 724
#> [235] 725 727 732 746 752 754 755 762 765 766 767 768 769 771 772 774 775 780
#> [253] 781 784 786 787 788 795 796 799 800 801 804 808 809 811 821 826 830 832
#> [271] 834 836 838 840 843 854 860 861 864 867 869 870 871 874 885 886 889 900
#> [289] 903 904 911 913 916   5  11  12  18  21  22  24  30  31  32  33  35  36
#> [307]  38  40  41  43  49  50  51  52  56  57  59  61  62  64  67  68  77  80
#> [325]  85  89  91  96  98 103 110 111 116 117 125 126 147 150 153 161 168 174
#> [343] 181 186 187 188 190 195 196 198 199 210 212 214 215 218 226 228 229 230
#> [361] 238 245 246 247 248 249 255 256 259 267 270 271 273 275 276 279 280 281
#> [379] 282 283 290 296 300 303 313 328 330 331 338 340 344 346 353 355 358 359
#> [397] 360 361 362 363 364 365 366 367 368 369 371 372 373 374 375 376 377 378
#> [415] 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396
#> [433] 397 398 399 400 401 402 403 404 405 415 416 422 423 425 434 437 440 441
#> [451] 444 447 451 452 456 459 463 466 467 475 480 482 483 484 485 501 507 511
#> [469] 518 525 526 530 535 536 537 538 539 540 542 544 554 558 562 563 567 574
#> [487] 582 585 588 595 597 599 601 602 603 604 606 607 608 609 618 621 622 623
#> [505] 625 628 632 633 634 635 642 646 647 648 649 650 658 661 666 669 675 678
#> [523] 679 682 683 684 685 689 693 696 703 704 716 717 718 721 722 726 733 737
#> [541] 739 740 741 742 744 745 747 748 749 753 756 757 758 760 761 763 764 770
#> [559] 779 782 785 789 792 793 794 798 802 810 813 814 816 817 818 820 827 828
#> [577] 831 835 837 845 846 847 848 850 853 855 856 857 858 862 875 876 877 879
#> [595] 881 883 890 893 896 897 898 905 906 907 908 910 912 914 917 919 920

rrcv$test_set(1)

#>   [1]   2   3   6   7  10  13  15  19  20  27  28  39  42  46  47  48  53  66
#>  [19]  70  74  78  90  94  95  97 100 104 105 108 113 114 115 121 122 128 129
#>  [37] 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 149
#>  [55] 151 152 158 159 160 163 166 169 170 171 172 173 175 180 182 191 197 206
#>  [73] 207 208 209 216 222 225 227 231 232 233 234 236 240 241 242 250 251 260
#>  [91] 261 262 263 264 272 284 285 286 288 293 297 298 301 304 305 306 307 308
#> [109] 309 310 311 312 314 319 320 321 323 324 325 326 327 329 334 336 347 349
#> [127] 350 351 356 357 370 406 407 414 417 419 420 424 426 427 428 429 430 432
#> [145] 435 436 442 443 446 450 454 455 460 462 464 465 468 469 470 477 479 481
#> [163] 492 496 497 498 499 500 502 504 508 509 510 513 515 516 521 523 524 527
#> [181] 528 529 531 532 533 534 541 543 545 548 551 555 556 564 565 566 568 569
#> [199] 570 571 572 573 575 576 577 578 579 580 581 586 591 600 605 612 613 617
#> [217] 619 630 636 637 640 641 643 644 645 651 652 657 664 665 667 668 670 671
#> [235] 672 673 674 677 681 686 690 694 695 701 720 723 728 729 730 731 734 735
#> [253] 736 738 743 750 751 759 773 776 777 778 783 790 791 797 803 805 806 807
#> [271] 812 815 819 822 823 824 825 829 833 839 841 842 844 849 851 852 859 863
#> [289] 865 866 868 872 873 878 880 882 884 887 888 891 892 894 895 899 901 902
#> [307] 909 915 918 921 922

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance # table

#>       row_id rep fold
#>    1:      1   1    2
#>    2:      2   1    1
#>    3:      3   1    1
#>    4:      4   1    2
#>    5:      5   1    3
#>   ---                
#> 1840:    918   2    2
#> 1841:    919   2    1
#> 1842:    920   2    1
#> 1843:    921   2    3
#> 1844:    922   2    3