(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi: 10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters

#> [1] 6

rrcv$folds(1:6)

#> [1] 1 2 3 1 2 3

rrcv$repeats(1:6)

#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)

#>   [1]   1   3   5   6   7   8   9  10  11  12  13  14  15  16  17  19  20  21
#>  [19]  22  23  24  25  27  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  49  51  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  90  91  92  93  94  95  96  97  98
#>  [91]  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116
#> [109] 117 118 119 121 122 123 124 125 126 127 128 130 131 132 133 134 136 137
#> [127] 138 139 140 141 142 143 144 145 146 147 148 149 152 154 155 156 157 159
#> [145] 160 161 162 163 164 165 166 167 168 170 171 173 174 175 176 177 178 179
#> [163] 180 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197
#> [181] 198 199 200 201 202 203 204 205 208 209 210 211 212 213 214 215 216 217
#> [199] 218 219 220 221 222 223 224 225 226 228 229 230 231 232 233 234 235 236
#> [217] 237 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 256
#> [235] 257 258 259 260 261 262 263 264 266 267 269 270 273 274 275 276 277 278
#> [253] 279 280 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297
#> [271] 298 299 300 301 302 303 304 305 306 307 308 310 311 312 313 314 315 316
#> [289] 317 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334
#> [307] 335 336 337 338 339 340 341 342 343 344 345 347 348 349 350 351 352 353
#> [325] 354 355 356 359 360 361 362 363 364 365 366 367 368 369 371 372 374 375
#> [343] 376 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393
#> [361] 394 395 396 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411
#> [379] 413 414 415 416 417 418 419 420 421 424 425 426 427 428 430 431 432 433
#> [397] 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451
#> [415] 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469
#> [433] 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487
#> [451] 488 489 490 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505
#> [469] 506 507 508 509 510 511 512 513 514 515 516 517 519 520 521 522 523 524
#> [487] 526 527 528 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544
#> [505] 545 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 563
#> [523] 564 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581
#> [541] 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599
#> [559] 600 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617
#> [577] 618 619 620 621 622 624 625 626 627 628 629 630 631 632 633 634 635 636
#> [595] 637 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654
#> [613] 655 656 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672
#> [631] 673 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690
#> [649] 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708
#> [667] 709 710 711 712 713 714 715 716 717 718 719 720 721 722 723 724 725 726
#> [685] 728 729 730 731 732 733 734 735 736 738 739 740 741 742 743 744 745 747
#> [703] 748 749 750 751

rrcv$test_set(1)

#>  [1]   4  48  50 129 135 150 206 346 358 370 422 429 529 623 737

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance # table

#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>  [1]   4  48  50 129 135 150 206 346 358 370 422 429 529 623 737
#> 
#> [[1]]$test[[2]]
#> [1] 707
#> 
#> [[1]]$test[[3]]
#>  [1] 119 187 194 208 244 323 355 390 479 506 601 625 717
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   3   5   6   7   8   9  10  11  12  13  14  15  16  17  19  20  21
#>  [19]  22  23  24  25  27  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  49  51  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  90  91  92  93  94  95  96  97  98
#>  [91]  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116
#> [109] 117 118 119 121 122 123 124 125 126 127 128 130 131 132 133 134 136 137
#> [127] 138 139 140 141 142 143 144 145 146 147 148 149 152 154 155 156 157 159
#> [145] 160 161 162 163 164 165 166 167 168 170 171 173 174 175 176 177 178 179
#> [163] 180 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197
#> [181] 198 199 200 201 202 203 204 205 208 209 210 211 212 213 214 215 216 217
#> [199] 218 219 220 221 222 223 224 225 226 228 229 230 231 232 233 234 235 236
#> [217] 237 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 256
#> [235] 257 258 259 260 261 262 263 264 266 267 269 270 273 274 275 276 277 278
#> [253] 279 280 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297
#> [271] 298 299 300 301 302 303 304 305 306 307 308 310 311 312 313 314 315 316
#> [289] 317 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334
#> [307] 335 336 337 338 339 340 341 342 343 344 345 347 348 349 350 351 352 353
#> [325] 354 355 356 359 360 361 362 363 364 365 366 367 368 369 371 372 374 375
#> [343] 376 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393
#> [361] 394 395 396 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411
#> [379] 413 414 415 416 417 418 419 420 421 424 425 426 427 428 430 431 432 433
#> [397] 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451
#> [415] 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469
#> [433] 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487
#> [451] 488 489 490 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505
#> [469] 506 507 508 509 510 511 512 513 514 515 516 517 519 520 521 522 523 524
#> [487] 526 527 528 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544
#> [505] 545 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 563
#> [523] 564 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581
#> [541] 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599
#> [559] 600 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617
#> [577] 618 619 620 621 622 624 625 626 627 628 629 630 631 632 633 634 635 636
#> [595] 637 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654
#> [613] 655 656 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672
#> [631] 673 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690
#> [649] 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708
#> [667] 709 710 711 712 713 714 715 716 717 718 719 720 721 722 723 724 725 726
#> [685] 728 729 730 731 732 733 734 735 736 738 739 740 741 742 743 744 745 747
#> [703] 748 749 750 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  72  73
#>  [73]  74  75  76  77  78  79  80  81  82  83  84  86  87  88  89  90  91  92
#>  [91]  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110
#> [109] 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128
#> [127] 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146
#> [145] 147 148 149 150 151 152 153 154 155 156 157 159 160 161 162 163 164 165
#> [163] 166 167 168 169 171 172 173 174 175 176 177 178 179 180 181 182 183 184
#> [181] 186 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203
#> [199] 204 205 206 207 208 209 210 211 212 213 214 215 219 220 221 222 223 224
#> [217] 225 226 227 228 229 230 231 232 233 235 236 237 238 239 240 241 242 243
#> [235] 244 245 246 248 249 250 251 252 253 254 256 257 258 259 260 261 262 263
#> [253] 264 265 266 267 268 269 270 271 272 273 274 275 276 277 279 280 281 282
#> [271] 283 284 285 286 287 288 289 291 292 293 294 295 296 297 298 299 300 301
#> [289] 302 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318 319 320
#> [307] 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338
#> [325] 339 340 341 342 343 345 346 347 348 349 350 351 352 353 354 355 356 357
#> [343] 358 359 360 361 363 364 365 366 367 368 369 370 371 372 373 374 375 376
#> [361] 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394
#> [379] 395 396 397 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413
#> [397] 414 415 416 417 418 419 420 421 422 423 424 425 427 428 429 430 431 432
#> [415] 433 434 435 436 437 439 440 441 442 443 444 445 446 447 448 449 450 451
#> [433] 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469
#> [451] 471 472 473 474 475 477 478 479 480 481 482 483 484 485 486 487 488 489
#> [469] 490 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507
#> [487] 508 509 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 526
#> [505] 527 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544
#> [523] 545 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562
#> [541] 563 564 565 566 567 568 569 570 571 572 573 575 576 577 578 579 580 581
#> [559] 582 583 584 586 587 588 590 591 593 594 595 596 597 598 599 600 601 602
#> [577] 603 604 605 606 607 608 609 610 611 613 614 616 617 618 619 620 621 622
#> [595] 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640
#> [613] 641 642 643 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659
#> [631] 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677
#> [649] 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695
#> [667] 696 697 698 699 700 701 702 703 704 705 706 708 709 710 711 712 713 714
#> [685] 715 716 717 718 719 720 721 722 723 724 726 727 728 729 730 731 732 733
#> [703] 734 735 736 737 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  75  76  77  78  79  80  82  83  84  85  86  87  88  89  90  91  92
#>  [91]  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110
#> [109] 111 112 113 114 115 116 117 118 120 121 122 123 124 125 126 127 128 129
#> [127] 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147
#> [145] 148 149 150 151 152 153 154 155 156 157 158 161 162 163 164 165 166 167
#> [163] 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185
#> [181] 186 188 189 190 191 192 193 195 196 197 198 199 200 201 202 203 204 205
#> [199] 206 207 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224
#> [217] 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242
#> [235] 243 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 260 261
#> [253] 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279
#> [271] 280 281 282 283 284 285 286 287 288 289 290 291 293 294 295 296 297 298
#> [289] 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316
#> [307] 317 318 319 320 321 322 324 325 326 327 328 329 330 331 332 333 334 335
#> [325] 336 337 338 339 340 341 342 343 344 345 346 347 348 349 350 351 352 353
#> [343] 354 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372
#> [361] 373 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388 389 391
#> [379] 392 393 394 395 396 397 398 399 400 401 402 403 404 405 406 407 408 409
#> [397] 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426 427
#> [415] 428 429 430 431 432 433 434 435 437 438 439 440 441 442 443 444 445 446
#> [433] 447 448 449 450 451 452 453 455 456 457 458 459 460 461 462 463 464 465
#> [451] 466 467 468 469 470 471 472 473 474 475 476 477 478 480 481 482 483 484
#> [469] 485 486 487 488 489 490 491 492 493 494 495 496 497 498 499 500 501 502
#> [487] 503 504 505 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [505] 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 537 538 539
#> [523] 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557
#> [541] 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575
#> [559] 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591 592 594
#> [577] 595 596 598 599 600 602 603 604 605 606 607 608 609 610 611 612 613 614
#> [595] 615 616 617 618 619 620 621 622 623 624 626 627 628 629 630 631 632 633
#> [613] 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648 649 650 651
#> [631] 652 653 654 655 656 657 658 659 660 661 662 663 664 665 666 667 668 669
#> [649] 670 671 673 674 675 676 677 678 679 680 681 682 683 684 686 687 688 689
#> [667] 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706 707
#> [685] 708 709 710 711 712 713 714 715 716 718 719 720 721 722 723 724 725 726
#> [703] 727 728 729 730 731 732 733 734 735 736 737 738 739 740 741 742 743 744
#> [721] 745 746 747 748 749 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#> [1]  85 158 428 470 476 589 592 633 746
#> 
#> [[2]]$test[[2]]
#>  [1]  16 107 212 264 273 312 316 379 420 432 445 459 482 515 590 660
#> 
#> [[2]]$test[[3]]
#>  [1]  22  80  83 101 176 197 341 352 450 556 614
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  27  28  29  30  31  32  33  34  35  36  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  49  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  65  66  67  68  69  72  73  74  75  76  77  78
#>  [73]  79  80  81  82  83  84  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 100 101 103 104 105 106 107 108 109 110 111 112 113 114 115 116
#> [109] 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134
#> [127] 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 152 154 155
#> [145] 156 157 159 160 161 162 163 164 165 166 167 168 169 170 171 173 174 175
#> [163] 176 177 178 179 180 181 182 183 184 186 187 188 189 190 191 192 193 194
#> [181] 195 196 197 198 199 200 201 202 203 204 205 208 209 210 211 212 213 214
#> [199] 215 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233 235
#> [217] 236 237 239 240 241 242 243 244 245 246 248 249 250 251 252 253 254 256
#> [235] 257 258 259 260 261 262 263 264 266 267 268 269 270 271 273 274 275 276
#> [253] 277 279 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296
#> [271] 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314
#> [289] 315 316 317 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332
#> [307] 333 334 335 336 337 338 339 340 341 342 343 345 346 347 348 349 350 351
#> [325] 352 353 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370
#> [343] 371 372 373 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388
#> [361] 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403 404 405 406
#> [379] 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424
#> [397] 425 426 427 430 431 432 433 434 435 436 437 439 440 441 442 443 444 445
#> [415] 446 447 448 449 450 452 453 454 455 456 457 458 459 460 461 462 463 464
#> [433] 465 466 467 468 469 471 472 473 474 475 477 478 479 480 481 482 483 484
#> [451] 485 486 487 488 489 490 491 492 493 494 495 496 497 498 500 501 502 503
#> [469] 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [487] 522 523 524 526 527 528 529 530 531 532 533 534 535 536 537 538 539 540
#> [505] 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557 558
#> [523] 559 560 561 563 564 565 566 567 568 569 570 571 572 573 575 576 577 578
#> [541] 579 580 581 582 583 584 585 586 587 588 590 591 593 594 595 596 597 598
#> [559] 599 600 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616
#> [577] 617 618 619 620 621 622 624 625 626 627 628 629 630 631 632 634 635 636
#> [595] 637 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654
#> [613] 655 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673
#> [631] 674 675 676 677 678 680 681 682 683 684 685 686 687 688 689 690 691 692
#> [649] 693 694 695 696 697 698 699 700 701 702 703 704 705 706 708 709 710 711
#> [667] 712 713 714 715 716 717 718 719 720 721 722 723 724 725 726 728 729 730
#> [685] 731 732 733 734 735 736 737 739 740 741 742 743 744 745 747 748 749 750
#> [703] 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  15  17  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  33  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  52  53  54  55  56  57  58  59
#>  [55]  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77
#>  [73]  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 108 109 110 111 112 113 114
#> [109] 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 134 135 136 137 138 139 140 141 142 143 144 145 147 150 151 152 153
#> [145] 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171
#> [163] 172 173 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 190
#> [181] 191 192 194 195 196 197 198 199 200 201 202 203 204 205 206 207 208 209
#> [199] 210 211 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227 228
#> [217] 229 230 231 233 234 235 236 237 238 239 240 241 242 243 244 245 246 247
#> [235] 248 249 250 251 253 254 255 256 257 259 260 261 262 263 265 266 267 268
#> [253] 269 270 271 272 274 275 276 277 278 279 280 281 282 283 284 285 287 288
#> [271] 289 290 291 292 293 294 296 297 298 299 301 303 304 305 306 307 308 309
#> [289] 310 311 313 315 317 318 319 320 321 322 323 324 325 326 327 328 329 331
#> [307] 332 333 334 335 336 337 338 339 340 341 342 343 344 345 346 348 349 350
#> [325] 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368
#> [343] 370 371 372 373 374 375 376 377 378 380 381 382 383 384 385 386 388 389
#> [361] 390 391 392 393 394 395 396 397 398 400 401 402 403 404 405 406 407 408
#> [379] 409 410 411 412 413 414 415 416 417 418 419 422 423 424 425 426 427 428
#> [397] 429 430 431 433 434 435 436 437 438 440 441 442 443 444 446 448 449 450
#> [415] 451 452 453 454 455 456 457 458 460 461 462 463 464 465 466 467 468 470
#> [433] 471 472 474 475 476 477 478 479 480 481 483 484 485 486 487 488 489 490
#> [451] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 508 509 510
#> [469] 511 512 513 514 516 517 518 519 520 521 522 523 524 525 526 527 528 529
#> [487] 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547
#> [505] 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565
#> [523] 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [541] 584 585 586 587 588 589 591 592 593 594 595 596 597 598 599 600 601 602
#> [559] 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619 620
#> [577] 621 622 623 624 625 627 628 629 630 631 632 633 634 635 636 637 638 639
#> [595] 640 641 642 643 644 646 648 649 650 651 652 653 654 655 656 657 658 659
#> [613] 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 679
#> [631] 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696 697
#> [649] 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715
#> [667] 716 717 718 719 720 722 723 724 725 726 727 728 729 730 732 733 734 735
#> [685] 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  81  82  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 157 158 159 160 161 162 163 164 165 166 167
#> [163] 168 169 170 171 172 173 174 175 178 179 180 181 182 183 184 185 186 187
#> [181] 188 189 190 191 192 193 194 195 196 198 199 200 201 202 203 204 205 206
#> [199] 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224
#> [217] 225 226 227 228 229 230 231 232 233 234 235 236 237 238 240 241 242 243
#> [235] 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 260 261
#> [253] 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279
#> [271] 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 298
#> [289] 299 300 301 302 303 304 305 306 307 308 309 310 312 313 314 315 316 317
#> [307] 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335
#> [325] 336 337 338 339 340 342 343 344 345 346 347 348 349 350 351 353 354 355
#> [343] 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373
#> [361] 374 375 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392
#> [379] 393 394 395 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412
#> [397] 413 414 415 416 417 418 419 420 421 422 423 425 426 427 428 429 430 431
#> [415] 432 433 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 451
#> [433] 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469
#> [451] 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487
#> [469] 488 489 490 491 492 493 494 495 496 497 498 499 500 501 503 504 505 506
#> [487] 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524
#> [505] 525 526 527 528 529 530 531 532 533 534 535 537 538 539 540 541 542 543
#> [523] 544 545 546 547 548 549 550 551 552 553 554 555 557 558 559 560 561 562
#> [541] 563 564 565 566 567 569 570 571 572 573 574 575 576 577 578 579 580 581
#> [559] 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599
#> [577] 600 601 602 603 604 605 606 607 608 609 610 611 613 616 617 618 619 620
#> [595] 621 622 623 624 625 626 627 628 629 630 631 632 633 635 636 637 638 639
#> [613] 640 641 642 643 644 645 646 647 649 650 651 653 654 655 656 657 658 659
#> [631] 660 661 662 663 664 665 667 668 669 670 671 672 673 674 675 676 677 678
#> [649] 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696
#> [667] 697 699 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715
#> [685] 716 717 718 719 720 721 722 723 725 726 727 728 729 730 731 732 733 734
#> [703] 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#>