Repeated Environmental Block Cross Validation Resampling

Environmental Block Cross Validation. This strategy uses k-means clustering to specify blocks of similar environmental conditions. Only numeric features can be used. The features used for building blocks can be specified in the param_set. By default, all numeric features are used.

References

Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G (2018). “blockCV: an R package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models.” bioRxiv. doi: 10.1101/357798 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVEnv

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVEnv$new()

ResamplingRepeatedSpCVEnv$folds()

ResamplingRepeatedSpCVEnv$repeats()

ResamplingRepeatedSpCVEnv$instantiate()

ResamplingRepeatedSpCVEnv$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "coordinate-based" repeated resampling instance.

Usage

ResamplingRepeatedSpCVEnv$new(id = "repeated-spcv-env")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVEnv$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVEnv$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVEnv$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVEnv$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated-spcv-env", folds = 4, repeats = 2)
rrcv$instantiate(task)

# Individual sets:
rrcv$train_set(1)

#>  [1] 464 682   3 135 192 237 335 346 370 474 539 623

rrcv$test_set(1)

#>   [1]   1   2   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91
#>  [91]  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109
#> [109] 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127
#> [127] 128 129 130 131 132 133 134 136 137 138 139 140 141 142 143 144 145 146
#> [145] 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164
#> [163] 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182
#> [181] 183 184 185 186 187 188 189 190 191 193 194 195 196 197 198 199 200 201
#> [199] 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219
#> [217] 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234 235 236 238
#> [235] 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256
#> [253] 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274
#> [271] 275 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290 291 292
#> [289] 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310
#> [307] 311 312 313 314 315 316 317 318 319 320 321 322 323 324 325 326 327 328
#> [325] 329 330 331 332 333 334 336 337 338 339 340 341 342 343 344 345 347 348
#> [343] 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366
#> [361] 367 368 369 371 372 373 374 375 376 377 378 379 380 381 382 383 384 385
#> [379] 386 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403
#> [397] 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421
#> [415] 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439
#> [433] 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457
#> [451] 458 459 460 461 462 463 465 466 467 468 469 470 471 472 473 475 476 477
#> [469] 478 479 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494 495
#> [487] 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513
#> [505] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531
#> [523] 532 533 534 535 536 537 538 540 541 542 543 544 545 546 547 548 549 550
#> [541] 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [559] 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586
#> [577] 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604
#> [595] 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622
#> [613] 624 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641
#> [631] 642 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659
#> [649] 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677
#> [667] 678 679 680 681 683 684 685 686 687 688 689 690 691 692 693 694 695 696
#> [685] 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714
#> [703] 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732
#> [721] 733 734 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750
#> [739] 751

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance

#>       row_id rep fold
#>    1:      1   1    1
#>    2:      2   1    1
#>    3:      3   1    4
#>    4:      4   1    1
#>    5:      5   1    1
#>   ---                
#> 1498:    747   2    4
#> 1499:    748   2    4
#> 1500:    749   2    4
#> 1501:    750   2    4
#> 1502:    751   2    4