Repeated Spatio-Temporal Cluster Resampling

Spatio-temporal cluster partitioning via the vcluster executable of the CLUTO clustering application.

This partitioning method relies on the external CLUTO library. To use it, CLUTO's executables need to be downloaded and installed into this package.

See https://gist.github.com/pat-s/6430470cf817050e27d26c43c0e9be72 for an installation approach that should work on Windows and Linux. macOS is not supported by CLUTO.

Before using this method, please check the restrictive copyright shown below.

Details

By default, -clmethod='direct' is passed to the vcluster executable in contrast to the upstream default -clmethod='rb'. There is no evidence or research that this method is the best among the available ones ("rb", "rbr", "direct", "agglo", "graph", "bagglo"). Also, various other parameters can be set via argument cluto_parameters to achieve different clustering results.

Parameter -clusterfile is handled by skmeans and cannot be changed.

Copyright

CLUTO's copyright is as follows:

The CLUTO package is copyrighted by the Regents of the University of Minnesota. It can be freely used for educational and research purposes by non-profit institutions and US government agencies only. Other organizations are allowed to use CLUTO only for evaluation purposes, and any further uses will require prior approval. The software may not be sold or redistributed without prior approval. One may make copies of the software for their use provided that the copies, are not sold or distributed, are used under the same terms and conditions. As unestablished research software, this code is provided on an “as is” basis without warranty of any kind, either expressed or implied. The downloading, or executing any part of this software constitutes an implicit agreement to these terms. These terms and conditions are subject to change at any time without prior notice.

References

FIXME: correct? Ying Zhao and George Karypis. 11th Conference of Information and Knowledge Management (CIKM), pp. 515-524, 2002.

Super class

mlr3::Resampling -> ResamplingRepeatedSptCVCstf

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSptCVCstf$new()

ResamplingRepeatedSptCVCstf$folds()

ResamplingRepeatedSptCVCstf$repeats()

ResamplingRepeatedSptCVCstf$instantiate()

ResamplingRepeatedSptCVCstf$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "coordinate-based" repeated resampling instance.

Usage

ResamplingRepeatedSptCVCstf$new(id = "repeated-sptcv-cstf")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSptCVCstf$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSptCVCstf$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSptCVCstf$instantiate(
  task,
  space_var = NULL,
  time_var = NULL,
  class = NULL
)

Arguments

task

Task
A task to instantiate.

space_var

[character]
Column name identifying the spatial units.

time_var

[character]
Column name identifying the temporal units.

class

[character]
Column name identifying a class unit (e.g. land cover).

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSptCVCstf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(mlr3spatiotempcv)
task = tsk("cookfarm")

# Instantiate Resampling
rrcv = rsmp("repeated-sptcv-cstf", folds = 3, repeats = 5)
rrcv$instantiate(task, time_var = "Date")
# Individual sets:
rrcv$iters

#> [1] 15

rrcv$folds(1:6)

#> [1] 1 2 3 4 5 1

rrcv$repeats(1:6)

#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)

#>   [1]   1   2   4   6   7   8   9  10  11  12  15  16  20  21  22  23  25  26
#>  [19]  27  28  29  31  32  35  37  38  39  40  44  45  46  48  49  50  52  54
#>  [37]  56  58  59  61  62  64  65  66  68  69  71  73  75  76  77  78  80  81
#>  [55]  82  83  84  85  86  87  88  89  90  94  95  96  97  98  99 101 103 104
#>  [73] 105 107 109 110 111 112 115 116 118 119 120 121 123 124 125 126 127 129
#>  [91] 130 132 133 140 143 144 145 146 147 148 150 151 152 153 155 156 157 160
#> [109] 162 163 164 166 167 168 170 173 174 176 177 178 180 182 184 185 186 187
#> [127] 188 189 190 191 193 194 195 196 197 198 199 201 202 203 204 205 206 208
#> [145] 209 210 211 212 213 215 219 223 224 227 228 229 230 231 232 233 234 235
#> [163] 236 237 239 243 244 245 246 247 248 249 250 252 253 255 257 258 259 261
#> [181] 262 263 264 265 266 268 269 270 274 275 276 278 280 281 282 283 288 289
#> [199] 291 292 293 295 296 298 299 300 301 303 304 307 308 311 312 314 316 318
#> [217] 319 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 337 338
#> [235] 340 342 345 349 350 352 353 354 356 357 359 361 362 363 365 366 367 368
#> [253] 369 370 371 372 373 374 375 376 377 382 384 385 386 388 391 393 394 396
#> [271] 397 398 399 400 402 403 404 406 407 408 409 411 413 415 416 418 420 421
#> [289] 424 426 427 428 431 432 433 434 435 436 439 442 443 444 445 446 449 450
#> [307] 451 452 454 455 456 457 458 460 463 464 465 466 467 469 470 471 472 476
#> [325] 477 479 480 484 487 488 489 491 493 495 496 497 499

rrcv$test_set(1)

#>   [1]   3   5  13  14  17  18  19  24  30  33  34  36  41  42  43  47  51  53
#>  [19]  55  57  60  63  67  70  72  74  79  91  92  93 100 102 106 108 113 114
#>  [37] 117 122 128 131 134 135 136 137 138 139 141 142 149 154 158 159 161 165
#>  [55] 169 171 172 175 179 181 183 192 200 207 214 216 217 218 220 221 222 225
#>  [73] 226 238 240 241 242 251 254 256 260 267 271 272 273 277 279 284 285 286
#>  [91] 287 290 294 297 302 305 306 309 310 313 315 317 320 336 339 341 343 344
#> [109] 346 347 348 351 355 358 360 364 378 379 380 381 383 387 389 390 392 395
#> [127] 401 405 410 412 414 417 419 422 423 425 429 430 437 438 440 441 447 448
#> [145] 453 459 461 462 468 473 474 475 478 481 482 483 485 486 490 492 494 498
#> [163] 500

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance # table

#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>   [1]   3   5  13  14  17  18  19  24  30  33  34  36  41  42  43  47  51  53
#>  [19]  55  57  60  63  67  70  72  74  79  91  92  93 100 102 106 108 113 114
#>  [37] 117 122 128 131 134 135 136 137 138 139 141 142 149 154 158 159 161 165
#>  [55] 169 171 172 175 179 181 183 192 200 207 214 216 217 218 220 221 222 225
#>  [73] 226 238 240 241 242 251 254 256 260 267 271 272 273 277 279 284 285 286
#>  [91] 287 290 294 297 302 305 306 309 310 313 315 317 320 336 339 341 343 344
#> [109] 346 347 348 351 355 358 360 364 378 379 380 381 383 387 389 390 392 395
#> [127] 401 405 410 412 414 417 419 422 423 425 429 430 437 438 440 441 447 448
#> [145] 453 459 461 462 468 473 474 475 478 481 482 483 485 486 490 492 494 498
#> [163] 500
#> 
#> [[1]]$test[[2]]
#>   [1]   2   7  10  11  12  21  22  23  26  32  38  39  48  49  50  64  68  69
#>  [19]  71  73  75  76  77  80  83  85  86  88  90  95  96  97 101 103 105 107
#>  [37] 111 115 116 118 120 125 130 133 143 144 145 146 148 150 155 163 166 170
#>  [55] 173 177 180 184 187 189 193 194 195 196 198 199 201 202 204 208 209 211
#>  [73] 219 229 230 231 232 233 235 236 249 250 257 263 265 266 268 269 274 278
#>  [91] 280 282 283 289 291 292 293 295 296 298 300 303 308 312 314 316 318 319
#> [109] 321 334 335 337 342 345 349 350 354 356 359 361 366 368 369 376 377 386
#> [127] 388 391 393 398 399 404 407 415 420 426 427 428 434 436 439 442 443 444
#> [145] 446 451 452 454 456 460 464 467 469 470 471 477 479 488 495 499
#> 
#> [[1]]$test[[3]]
#>   [1]   1   4   6   8   9  15  16  20  25  27  28  29  31  35  37  40  44  45
#>  [19]  46  52  54  56  58  59  61  62  65  66  78  81  82  84  87  89  94  98
#>  [37]  99 104 109 110 112 119 121 123 124 126 127 129 132 140 147 151 152 153
#>  [55] 156 157 160 162 164 167 168 174 176 178 182 185 186 188 190 191 197 203
#>  [73] 205 206 210 212 213 215 223 224 227 228 234 237 239 243 244 245 246 247
#>  [91] 248 252 253 255 258 259 261 262 264 270 275 276 281 288 299 301 304 307
#> [109] 311 322 323 324 325 326 327 328 329 330 331 332 333 338 340 352 353 357
#> [127] 362 363 365 367 370 371 372 373 374 375 382 384 385 394 396 397 400 402
#> [145] 403 406 408 409 411 413 416 418 421 424 431 432 433 435 445 449 450 455
#> [163] 457 458 463 465 466 472 476 480 484 487 489 491 493 496 497
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   4   6   7   8   9  10  11  12  15  16  20  21  22  23  25  26
#>  [19]  27  28  29  31  32  35  37  38  39  40  44  45  46  48  49  50  52  54
#>  [37]  56  58  59  61  62  64  65  66  68  69  71  73  75  76  77  78  80  81
#>  [55]  82  83  84  85  86  87  88  89  90  94  95  96  97  98  99 101 103 104
#>  [73] 105 107 109 110 111 112 115 116 118 119 120 121 123 124 125 126 127 129
#>  [91] 130 132 133 140 143 144 145 146 147 148 150 151 152 153 155 156 157 160
#> [109] 162 163 164 166 167 168 170 173 174 176 177 178 180 182 184 185 186 187
#> [127] 188 189 190 191 193 194 195 196 197 198 199 201 202 203 204 205 206 208
#> [145] 209 210 211 212 213 215 219 223 224 227 228 229 230 231 232 233 234 235
#> [163] 236 237 239 243 244 245 246 247 248 249 250 252 253 255 257 258 259 261
#> [181] 262 263 264 265 266 268 269 270 274 275 276 278 280 281 282 283 288 289
#> [199] 291 292 293 295 296 298 299 300 301 303 304 307 308 311 312 314 316 318
#> [217] 319 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 337 338
#> [235] 340 342 345 349 350 352 353 354 356 357 359 361 362 363 365 366 367 368
#> [253] 369 370 371 372 373 374 375 376 377 382 384 385 386 388 391 393 394 396
#> [271] 397 398 399 400 402 403 404 406 407 408 409 411 413 415 416 418 420 421
#> [289] 424 426 427 428 431 432 433 434 435 436 439 442 443 444 445 446 449 450
#> [307] 451 452 454 455 456 457 458 460 463 464 465 466 467 469 470 471 472 476
#> [325] 477 479 480 484 487 488 489 491 493 495 496 497 499
#> 
#> [[1]]$train[[2]]
#>   [1]   1   3   4   5   6   8   9  13  14  15  16  17  18  19  20  24  25  27
#>  [19]  28  29  30  31  33  34  35  36  37  40  41  42  43  44  45  46  47  51
#>  [37]  52  53  54  55  56  57  58  59  60  61  62  63  65  66  67  70  72  74
#>  [55]  78  79  81  82  84  87  89  91  92  93  94  98  99 100 102 104 106 108
#>  [73] 109 110 112 113 114 117 119 121 122 123 124 126 127 128 129 131 132 134
#>  [91] 135 136 137 138 139 140 141 142 147 149 151 152 153 154 156 157 158 159
#> [109] 160 161 162 164 165 167 168 169 171 172 174 175 176 178 179 181 182 183
#> [127] 185 186 188 190 191 192 197 200 203 205 206 207 210 212 213 214 215 216
#> [145] 217 218 220 221 222 223 224 225 226 227 228 234 237 238 239 240 241 242
#> [163] 243 244 245 246 247 248 251 252 253 254 255 256 258 259 260 261 262 264
#> [181] 267 270 271 272 273 275 276 277 279 281 284 285 286 287 288 290 294 297
#> [199] 299 301 302 304 305 306 307 309 310 311 313 315 317 320 322 323 324 325
#> [217] 326 327 328 329 330 331 332 333 336 338 339 340 341 343 344 346 347 348
#> [235] 351 352 353 355 357 358 360 362 363 364 365 367 370 371 372 373 374 375
#> [253] 378 379 380 381 382 383 384 385 387 389 390 392 394 395 396 397 400 401
#> [271] 402 403 405 406 408 409 410 411 412 413 414 416 417 418 419 421 422 423
#> [289] 424 425 429 430 431 432 433 435 437 438 440 441 445 447 448 449 450 453
#> [307] 455 457 458 459 461 462 463 465 466 468 472 473 474 475 476 478 480 481
#> [325] 482 483 484 485 486 487 489 490 491 492 493 494 496 497 498 500
#> 
#> [[1]]$train[[3]]
#>   [1]   2   3   5   7  10  11  12  13  14  17  18  19  21  22  23  24  26  30
#>  [19]  32  33  34  36  38  39  41  42  43  47  48  49  50  51  53  55  57  60
#>  [37]  63  64  67  68  69  70  71  72  73  74  75  76  77  79  80  83  85  86
#>  [55]  88  90  91  92  93  95  96  97 100 101 102 103 105 106 107 108 111 113
#>  [73] 114 115 116 117 118 120 122 125 128 130 131 133 134 135 136 137 138 139
#>  [91] 141 142 143 144 145 146 148 149 150 154 155 158 159 161 163 165 166 169
#> [109] 170 171 172 173 175 177 179 180 181 183 184 187 189 192 193 194 195 196
#> [127] 198 199 200 201 202 204 207 208 209 211 214 216 217 218 219 220 221 222
#> [145] 225 226 229 230 231 232 233 235 236 238 240 241 242 249 250 251 254 256
#> [163] 257 260 263 265 266 267 268 269 271 272 273 274 277 278 279 280 282 283
#> [181] 284 285 286 287 289 290 291 292 293 294 295 296 297 298 300 302 303 305
#> [199] 306 308 309 310 312 313 314 315 316 317 318 319 320 321 334 335 336 337
#> [217] 339 341 342 343 344 345 346 347 348 349 350 351 354 355 356 358 359 360
#> [235] 361 364 366 368 369 376 377 378 379 380 381 383 386 387 388 389 390 391
#> [253] 392 393 395 398 399 401 404 405 407 410 412 414 415 417 419 420 422 423
#> [271] 425 426 427 428 429 430 434 436 437 438 439 440 441 442 443 444 446 447
#> [289] 448 451 452 453 454 456 459 460 461 462 464 467 468 469 470 471 473 474
#> [307] 475 477 478 479 481 482 483 485 486 488 490 492 494 495 498 499 500
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>   [1]   2   6   8  11  16  17  20  23  26  27  29  31  43  50  51  53  55  57
#>  [19]  58  60  63  65  68  69  70  72  78  80  83  95  96  97 102 107 111 114
#>  [37] 117 118 119 122 123 124 125 127 129 133 138 139 141 142 151 155 160 162
#>  [55] 164 165 170 173 178 184 186 187 189 194 195 198 207 210 215 218 219 220
#>  [73] 230 234 235 236 237 238 239 245 246 249 254 255 258 259 260 262 264 265
#>  [91] 271 280 288 289 290 293 295 296 299 301 305 308 310 312 315 319 320 325
#> [109] 329 334 336 340 344 345 352 353 355 359 360 361 369 370 375 376 379 381
#> [127] 382 385 390 391 392 393 395 399 403 405 406 410 411 412 413 415 417 423
#> [145] 424 426 427 430 441 443 447 450 458 459 460 465 470 471 480 483 486 490
#> [163] 499 500
#> 
#> [[2]]$test[[2]]
#>   [1]   3   5   7   9  10  12  13  14  24  25  30  34  37  39  44  45  46  48
#>  [19]  49  59  62  64  74  76  81  84  87  88  89  91  92  93 105 108 109 110
#>  [37] 112 115 116 130 134 144 145 146 148 153 154 156 157 158 161 167 168 169
#>  [55] 171 172 174 175 179 180 188 190 193 196 199 200 202 205 206 208 209 211
#>  [73] 214 217 221 222 225 226 231 233 244 250 251 252 257 261 267 268 269 278
#>  [91] 279 285 287 292 297 298 300 303 304 306 318 321 322 323 327 332 333 335
#> [109] 339 341 343 346 350 354 356 357 362 365 366 367 371 373 374 380 383 384
#> [127] 386 388 396 397 398 400 401 402 404 407 409 414 421 422 428 432 433 434
#> [145] 435 439 444 446 448 451 452 454 457 461 472 473 474 475 478 479 482 484
#> [163] 489 492 493 497
#> 
#> [[2]]$test[[3]]
#>   [1]   1   4  15  18  19  21  22  28  32  33  35  36  38  40  41  42  47  52
#>  [19]  54  56  61  66  67  71  73  75  77  79  82  85  86  90  94  98  99 100
#>  [37] 101 103 104 106 113 120 121 126 128 131 132 135 136 137 140 143 147 149
#>  [55] 150 152 159 163 166 176 177 181 182 183 185 191 192 197 201 203 204 212
#>  [73] 213 216 223 224 227 228 229 232 240 241 242 243 247 248 253 256 263 266
#>  [91] 270 272 273 274 275 276 277 281 282 283 284 286 291 294 302 307 309 311
#> [109] 313 314 316 317 324 326 328 330 331 337 338 342 347 348 349 351 358 363
#> [127] 364 368 372 377 378 387 389 394 408 416 418 419 420 425 429 431 436 437
#> [145] 438 440 442 445 449 453 455 456 462 463 464 466 467 468 469 476 477 481
#> [163] 485 487 488 491 494 495 496 498
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   3   4   5   7   9  10  12  13  14  15  18  19  21  22  24  25  28
#>  [19]  30  32  33  34  35  36  37  38  39  40  41  42  44  45  46  47  48  49
#>  [37]  52  54  56  59  61  62  64  66  67  71  73  74  75  76  77  79  81  82
#>  [55]  84  85  86  87  88  89  90  91  92  93  94  98  99 100 101 103 104 105
#>  [73] 106 108 109 110 112 113 115 116 120 121 126 128 130 131 132 134 135 136
#>  [91] 137 140 143 144 145 146 147 148 149 150 152 153 154 156 157 158 159 161
#> [109] 163 166 167 168 169 171 172 174 175 176 177 179 180 181 182 183 185 188
#> [127] 190 191 192 193 196 197 199 200 201 202 203 204 205 206 208 209 211 212
#> [145] 213 214 216 217 221 222 223 224 225 226 227 228 229 231 232 233 240 241
#> [163] 242 243 244 247 248 250 251 252 253 256 257 261 263 266 267 268 269 270
#> [181] 272 273 274 275 276 277 278 279 281 282 283 284 285 286 287 291 292 294
#> [199] 297 298 300 302 303 304 306 307 309 311 313 314 316 317 318 321 322 323
#> [217] 324 326 327 328 330 331 332 333 335 337 338 339 341 342 343 346 347 348
#> [235] 349 350 351 354 356 357 358 362 363 364 365 366 367 368 371 372 373 374
#> [253] 377 378 380 383 384 386 387 388 389 394 396 397 398 400 401 402 404 407
#> [271] 408 409 414 416 418 419 420 421 422 425 428 429 431 432 433 434 435 436
#> [289] 437 438 439 440 442 444 445 446 448 449 451 452 453 454 455 456 457 461
#> [307] 462 463 464 466 467 468 469 472 473 474 475 476 477 478 479 481 482 484
#> [325] 485 487 488 489 491 492 493 494 495 496 497 498
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   4   6   8  11  15  16  17  18  19  20  21  22  23  26  27  28
#>  [19]  29  31  32  33  35  36  38  40  41  42  43  47  50  51  52  53  54  55
#>  [37]  56  57  58  60  61  63  65  66  67  68  69  70  71  72  73  75  77  78
#>  [55]  79  80  82  83  85  86  90  94  95  96  97  98  99 100 101 102 103 104
#>  [73] 106 107 111 113 114 117 118 119 120 121 122 123 124 125 126 127 128 129
#>  [91] 131 132 133 135 136 137 138 139 140 141 142 143 147 149 150 151 152 155
#> [109] 159 160 162 163 164 165 166 170 173 176 177 178 181 182 183 184 185 186
#> [127] 187 189 191 192 194 195 197 198 201 203 204 207 210 212 213 215 216 218
#> [145] 219 220 223 224 227 228 229 230 232 234 235 236 237 238 239 240 241 242
#> [163] 243 245 246 247 248 249 253 254 255 256 258 259 260 262 263 264 265 266
#> [181] 270 271 272 273 274 275 276 277 280 281 282 283 284 286 288 289 290 291
#> [199] 293 294 295 296 299 301 302 305 307 308 309 310 311 312 313 314 315 316
#> [217] 317 319 320 324 325 326 328 329 330 331 334 336 337 338 340 342 344 345
#> [235] 347 348 349 351 352 353 355 358 359 360 361 363 364 368 369 370 372 375
#> [253] 376 377 378 379 381 382 385 387 389 390 391 392 393 394 395 399 403 405
#> [271] 406 408 410 411 412 413 415 416 417 418 419 420 423 424 425 426 427 429
#> [289] 430 431 436 437 438 440 441 442 443 445 447 449 450 453 455 456 458 459
#> [307] 460 462 463 464 465 466 467 468 469 470 471 476 477 480 481 483 485 486
#> [325] 487 488 490 491 494 495 496 498 499 500
#> 
#> [[2]]$train[[3]]
#>   [1]   2   3   5   6   7   8   9  10  11  12  13  14  16  17  20  23  24  25
#>  [19]  26  27  29  30  31  34  37  39  43  44  45  46  48  49  50  51  53  55
#>  [37]  57  58  59  60  62  63  64  65  68  69  70  72  74  76  78  80  81  83
#>  [55]  84  87  88  89  91  92  93  95  96  97 102 105 107 108 109 110 111 112
#>  [73] 114 115 116 117 118 119 122 123 124 125 127 129 130 133 134 138 139 141
#>  [91] 142 144 145 146 148 151 153 154 155 156 157 158 160 161 162 164 165 167
#> [109] 168 169 170 171 172 173 174 175 178 179 180 184 186 187 188 189 190 193
#> [127] 194 195 196 198 199 200 202 205 206 207 208 209 210 211 214 215 217 218
#> [145] 219 220 221 222 225 226 230 231 233 234 235 236 237 238 239 244 245 246
#> [163] 249 250 251 252 254 255 257 258 259 260 261 262 264 265 267 268 269 271
#> [181] 278 279 280 285 287 288 289 290 292 293 295 296 297 298 299 300 301 303
#> [199] 304 305 306 308 310 312 315 318 319 320 321 322 323 325 327 329 332 333
#> [217] 334 335 336 339 340 341 343 344 345 346 350 352 353 354 355 356 357 359
#> [235] 360 361 362 365 366 367 369 370 371 373 374 375 376 379 380 381 382 383
#> [253] 384 385 386 388 390 391 392 393 395 396 397 398 399 400 401 402 403 404
#> [271] 405 406 407 409 410 411 412 413 414 415 417 421 422 423 424 426 427 428
#> [289] 430 432 433 434 435 439 441 443 444 446 447 448 450 451 452 454 457 458
#> [307] 459 460 461 465 470 471 472 473 474 475 478 479 480 482 483 484 486 489
#> [325] 490 492 493 497 499 500
#> 
#> 
#> 
#> [[3]]
#> [[3]]$test
#> [[3]]$test[[1]]
#>   [1]   7  10  11  16  18  22  24  25  28  31  41  45  50  53  56  58  62  63
#>  [19]  66  71  72  74  78  80  83  86  88  90  91  92  99 102 108 112 113 117
#>  [37] 118 125 128 130 131 132 137 139 141 149 154 157 161 162 165 168 169 174
#>  [55] 176 180 194 197 199 200 201 203 204 208 212 213 223 226 228 229 230 232
#>  [73] 236 237 239 247 254 260 261 263 266 270 271 275 279 281 287 290 291 292
#>  [91] 295 296 304 307 308 309 310 311 312 314 317 321 328 331 338 339 342 344
#> [109] 345 346 349 352 355 361 364 365 371 374 381 382 386 387 388 394 395 396
#> [127] 398 404 409 419 421 423 425 433 434 436 437 440 441 442 443 444 445 447
#> [145] 450 452 453 456 461 465 469 473 477 478 479 484 487 494 496 497 500
#> 
#> [[3]]$test[[2]]
#>   [1]   4   5   6   8   9  13  14  17  20  23  26  27  29  32  34  35  37  38
#>  [19]  39  40  44  46  47  48  52  55  61  64  65  67  73  75  89  95 100 101
#>  [37] 103 104 105 106 111 114 116 119 122 123 124 136 138 140 147 148 151 155
#>  [55] 159 160 167 171 172 173 182 184 185 187 188 189 191 195 205 207 210 211
#>  [73] 214 215 217 219 221 224 225 227 233 235 241 242 243 244 246 249 253 255
#>  [91] 259 268 272 274 276 277 278 282 286 288 289 294 297 299 302 303 313 315
#> [109] 316 318 319 322 323 325 327 330 333 334 336 340 341 343 347 350 354 358
#> [127] 359 360 362 367 373 375 376 380 389 391 392 393 399 401 402 410 412 413
#> [145] 414 416 417 418 420 424 426 430 431 432 435 438 448 449 458 462 468 470
#> [163] 471 472 474 475 476 481 482 485 490 492 495
#> 
#> [[3]]$test[[3]]
#>   [1]   1   2   3  12  15  19  21  30  33  36  42  43  49  51  54  57  59  60
#>  [19]  68  69  70  76  77  79  81  82  84  85  87  93  94  96  97  98 107 109
#>  [37] 110 115 120 121 126 127 129 133 134 135 142 143 144 145 146 150 152 153
#>  [55] 156 158 163 164 166 170 175 177 178 179 181 183 186 190 192 193 196 198
#>  [73] 202 206 209 216 218 220 222 231 234 238 240 245 248 250 251 252 256 257
#>  [91] 258 262 264 265 267 269 273 280 283 284 285 293 298 300 301 305 306 320
#> [109] 324 326 329 332 335 337 348 351 353 356 357 363 366 368 369 370 372 377
#> [127] 378 379 383 384 385 390 397 400 403 405 406 407 408 411 415 422 427 428
#> [145] 429 439 446 451 454 455 457 459 460 463 464 466 467 480 483 486 488 489
#> [163] 491 493 498 499
#> 
#> 
#> [[3]]$train
#> [[3]]$train[[1]]
#>   [1]   1   2   3   4   5   6   8   9  12  13  14  15  17  19  20  21  23  26
#>  [19]  27  29  30  32  33  34  35  36  37  38  39  40  42  43  44  46  47  48
#>  [37]  49  51  52  54  55  57  59  60  61  64  65  67  68  69  70  73  75  76
#>  [55]  77  79  81  82  84  85  87  89  93  94  95  96  97  98 100 101 103 104
#>  [73] 105 106 107 109 110 111 114 115 116 119 120 121 122 123 124 126 127 129
#>  [91] 133 134 135 136 138 140 142 143 144 145 146 147 148 150 151 152 153 155
#> [109] 156 158 159 160 163 164 166 167 170 171 172 173 175 177 178 179 181 182
#> [127] 183 184 185 186 187 188 189 190 191 192 193 195 196 198 202 205 206 207
#> [145] 209 210 211 214 215 216 217 218 219 220 221 222 224 225 227 231 233 234
#> [163] 235 238 240 241 242 243 244 245 246 248 249 250 251 252 253 255 256 257
#> [181] 258 259 262 264 265 267 268 269 272 273 274 276 277 278 280 282 283 284
#> [199] 285 286 288 289 293 294 297 298 299 300 301 302 303 305 306 313 315 316
#> [217] 318 319 320 322 323 324 325 326 327 329 330 332 333 334 335 336 337 340
#> [235] 341 343 347 348 350 351 353 354 356 357 358 359 360 362 363 366 367 368
#> [253] 369 370 372 373 375 376 377 378 379 380 383 384 385 389 390 391 392 393
#> [271] 397 399 400 401 402 403 405 406 407 408 410 411 412 413 414 415 416 417
#> [289] 418 420 422 424 426 427 428 429 430 431 432 435 438 439 446 448 449 451
#> [307] 454 455 457 458 459 460 462 463 464 466 467 468 470 471 472 474 475 476
#> [325] 480 481 482 483 485 486 488 489 490 491 492 493 495 498 499
#> 
#> [[3]]$train[[2]]
#>   [1]   1   2   3   7  10  11  12  15  16  18  19  21  22  24  25  28  30  31
#>  [19]  33  36  41  42  43  45  49  50  51  53  54  56  57  58  59  60  62  63
#>  [37]  66  68  69  70  71  72  74  76  77  78  79  80  81  82  83  84  85  86
#>  [55]  87  88  90  91  92  93  94  96  97  98  99 102 107 108 109 110 112 113
#>  [73] 115 117 118 120 121 125 126 127 128 129 130 131 132 133 134 135 137 139
#>  [91] 141 142 143 144 145 146 149 150 152 153 154 156 157 158 161 162 163 164
#> [109] 165 166 168 169 170 174 175 176 177 178 179 180 181 183 186 190 192 193
#> [127] 194 196 197 198 199 200 201 202 203 204 206 208 209 212 213 216 218 220
#> [145] 222 223 226 228 229 230 231 232 234 236 237 238 239 240 245 247 248 250
#> [163] 251 252 254 256 257 258 260 261 262 263 264 265 266 267 269 270 271 273
#> [181] 275 279 280 281 283 284 285 287 290 291 292 293 295 296 298 300 301 304
#> [199] 305 306 307 308 309 310 311 312 314 317 320 321 324 326 328 329 331 332
#> [217] 335 337 338 339 342 344 345 346 348 349 351 352 353 355 356 357 361 363
#> [235] 364 365 366 368 369 370 371 372 374 377 378 379 381 382 383 384 385 386
#> [253] 387 388 390 394 395 396 397 398 400 403 404 405 406 407 408 409 411 415
#> [271] 419 421 422 423 425 427 428 429 433 434 436 437 439 440 441 442 443 444
#> [289] 445 446 447 450 451 452 453 454 455 456 457 459 460 461 463 464 465 466
#> [307] 467 469 473 477 478 479 480 483 484 486 487 488 489 491 493 494 496 497
#> [325] 498 499 500
#> 
#> [[3]]$train[[3]]
#>   [1]   4   5   6   7   8   9  10  11  13  14  16  17  18  20  22  23  24  25
#>  [19]  26  27  28  29  31  32  34  35  37  38  39  40  41  44  45  46  47  48
#>  [37]  50  52  53  55  56  58  61  62  63  64  65  66  67  71  72  73  74  75
#>  [55]  78  80  83  86  88  89  90  91  92  95  99 100 101 102 103 104 105 106
#>  [73] 108 111 112 113 114 116 117 118 119 122 123 124 125 128 130 131 132 136
#>  [91] 137 138 139 140 141 147 148 149 151 154 155 157 159 160 161 162 165 167
#> [109] 168 169 171 172 173 174 176 180 182 184 185 187 188 189 191 194 195 197
#> [127] 199 200 201 203 204 205 207 208 210 211 212 213 214 215 217 219 221 223
#> [145] 224 225 226 227 228 229 230 232 233 235 236 237 239 241 242 243 244 246
#> [163] 247 249 253 254 255 259 260 261 263 266 268 270 271 272 274 275 276 277
#> [181] 278 279 281 282 286 287 288 289 290 291 292 294 295 296 297 299 302 303
#> [199] 304 307 308 309 310 311 312 313 314 315 316 317 318 319 321 322 323 325
#> [217] 327 328 330 331 333 334 336 338 339 340 341 342 343 344 345 346 347 349
#> [235] 350 352 354 355 358 359 360 361 362 364 365 367 371 373 374 375 376 380
#> [253] 381 382 386 387 388 389 391 392 393 394 395 396 398 399 401 402 404 409
#> [271] 410 412 413 414 416 417 418 419 420 421 423 424 425 426 430 431 432 433
#> [289] 434 435 436 437 438 440 441 442 443 444 445 447 448 449 450 452 453 456
#> [307] 458 461 462 465 468 469 470 471 472 473 474 475 476 477 478 479 481 482
#> [325] 484 485 487 490 492 494 495 496 497 500
#> 
#> 
#> 
#> [[4]]
#> [[4]]$test
#> [[4]]$test[[1]]
#>   [1]   4   6  13  19  21  22  24  28  30  31  35  36  40  41  47  50  53  55
#>  [19]  57  59  60  64  65  71  72  75  77  80  83  87  88  93  95  96  99 102
#>  [37] 103 105 108 109 111 113 114 115 117 118 121 123 125 126 128 129 136 142
#>  [55] 143 146 151 152 153 156 165 174 176 180 183 187 189 194 195 197 206 207
#>  [73] 213 216 218 220 226 229 234 237 242 244 245 246 247 249 250 251 257 258
#>  [91] 262 263 266 267 268 269 274 275 276 277 286 290 292 298 299 300 303 305
#> [109] 306 307 309 321 328 333 337 339 340 342 349 350 357 359 361 363 364 367
#> [127] 373 382 395 397 398 399 401 415 422 426 428 430 431 432 435 436 441 442
#> [145] 443 444 449 450 458 459 461 462 463 464 465 468 474 477 479 482 488 491
#> [163] 493 499
#> 
#> [[4]]$test[[2]]
#>   [1]   2   8   9  10  12  14  20  26  29  33  38  42  43  45  48  51  52  56
#>  [19]  63  67  68  76  78  79  84  85  86  91  92  94  97  98 100 101 104 106
#>  [37] 110 112 116 120 127 131 132 138 139 141 147 148 149 158 159 166 169 170
#>  [55] 171 172 175 177 181 182 184 186 188 191 196 201 204 210 214 215 219 221
#>  [73] 223 227 228 233 235 240 243 248 252 254 255 265 270 271 272 278 280 281
#>  [91] 282 287 288 289 294 295 301 313 316 319 322 323 325 326 327 329 330 331
#> [109] 332 334 338 343 344 346 347 348 351 353 355 356 358 366 368 370 375 376
#> [127] 378 379 381 386 388 394 400 404 405 408 410 411 412 414 416 417 418 419
#> [145] 423 433 434 437 438 439 445 447 451 453 455 457 460 466 469 470 472 473
#> [163] 475 476 481 484 494 495 498
#> 
#> [[4]]$test[[3]]
#>   [1]   1   3   5   7  11  15  16  17  18  23  25  27  32  34  37  39  44  46
#>  [19]  49  54  58  61  62  66  69  70  73  74  81  82  89  90 107 119 122 124
#>  [37] 130 133 134 135 137 140 144 145 150 154 155 157 160 161 162 163 164 167
#>  [55] 168 173 178 179 185 190 192 193 198 199 200 202 203 205 208 209 211 212
#>  [73] 217 222 224 225 230 231 232 236 238 239 241 253 256 259 260 261 264 273
#>  [91] 279 283 284 285 291 293 296 297 302 304 308 310 311 312 314 315 317 318
#> [109] 320 324 335 336 341 345 352 354 360 362 365 369 371 372 374 377 380 383
#> [127] 384 385 387 389 390 391 392 393 396 402 403 406 407 409 413 420 421 424
#> [145] 425 427 429 440 446 448 452 454 456 467 471 478 480 483 485 486 487 489
#> [163] 490 492 496 497 500
#> 
#> 
#> [[4]]$train
#> [[4]]$train[[1]]
#>   [1]   1   2   3   5   7   8   9  10  11  12  14  15  16  17  18  20  23  25
#>  [19]  26  27  29  32  33  34  37  38  39  42  43  44  45  46  48  49  51  52
#>  [37]  54  56  58  61  62  63  66  67  68  69  70  73  74  76  78  79  81  82
#>  [55]  84  85  86  89  90  91  92  94  97  98 100 101 104 106 107 110 112 116
#>  [73] 119 120 122 124 127 130 131 132 133 134 135 137 138 139 140 141 144 145
#>  [91] 147 148 149 150 154 155 157 158 159 160 161 162 163 164 166 167 168 169
#> [109] 170 171 172 173 175 177 178 179 181 182 184 185 186 188 190 191 192 193
#> [127] 196 198 199 200 201 202 203 204 205 208 209 210 211 212 214 215 217 219
#> [145] 221 222 223 224 225 227 228 230 231 232 233 235 236 238 239 240 241 243
#> [163] 248 252 253 254 255 256 259 260 261 264 265 270 271 272 273 278 279 280
#> [181] 281 282 283 284 285 287 288 289 291 293 294 295 296 297 301 302 304 308
#> [199] 310 311 312 313 314 315 316 317 318 319 320 322 323 324 325 326 327 329
#> [217] 330 331 332 334 335 336 338 341 343 344 345 346 347 348 351 352 353 354
#> [235] 355 356 358 360 362 365 366 368 369 370 371 372 374 375 376 377 378 379
#> [253] 380 381 383 384 385 386 387 388 389 390 391 392 393 394 396 400 402 403
#> [271] 404 405 406 407 408 409 410 411 412 413 414 416 417 418 419 420 421 423
#> [289] 424 425 427 429 433 434 437 438 439 440 445 446 447 448 451 452 453 454
#> [307] 455 456 457 460 466 467 469 470 471 472 473 475 476 478 480 481 483 484
#> [325] 485 486 487 489 490 492 494 495 496 497 498 500
#> 
#> [[4]]$train[[2]]
#>   [1]   1   3   4   5   6   7  11  13  15  16  17  18  19  21  22  23  24  25
#>  [19]  27  28  30  31  32  34  35  36  37  39  40  41  44  46  47  49  50  53
#>  [37]  54  55  57  58  59  60  61  62  64  65  66  69  70  71  72  73  74  75
#>  [55]  77  80  81  82  83  87  88  89  90  93  95  96  99 102 103 105 107 108
#>  [73] 109 111 113 114 115 117 118 119 121 122 123 124 125 126 128 129 130 133
#>  [91] 134 135 136 137 140 142 143 144 145 146 150 151 152 153 154 155 156 157
#> [109] 160 161 162 163 164 165 167 168 173 174 176 178 179 180 183 185 187 189
#> [127] 190 192 193 194 195 197 198 199 200 202 203 205 206 207 208 209 211 212
#> [145] 213 216 217 218 220 222 224 225 226 229 230 231 232 234 236 237 238 239
#> [163] 241 242 244 245 246 247 249 250 251 253 256 257 258 259 260 261 262 263
#> [181] 264 266 267 268 269 273 274 275 276 277 279 283 284 285 286 290 291 292
#> [199] 293 296 297 298 299 300 302 303 304 305 306 307 308 309 310 311 312 314
#> [217] 315 317 318 320 321 324 328 333 335 336 337 339 340 341 342 345 349 350
#> [235] 352 354 357 359 360 361 362 363 364 365 367 369 371 372 373 374 377 380
#> [253] 382 383 384 385 387 389 390 391 392 393 395 396 397 398 399 401 402 403
#> [271] 406 407 409 413 415 420 421 422 424 425 426 427 428 429 430 431 432 435
#> [289] 436 440 441 442 443 444 446 448 449 450 452 454 456 458 459 461 462 463
#> [307] 464 465 467 468 471 474 477 478 479 480 482 483 485 486 487 488 489 490
#> [325] 491 492 493 496 497 499 500
#> 
#> [[4]]$train[[3]]
#>   [1]   2   4   6   8   9  10  12  13  14  19  20  21  22  24  26  28  29  30
#>  [19]  31  33  35  36  38  40  41  42  43  45  47  48  50  51  52  53  55  56
#>  [37]  57  59  60  63  64  65  67  68  71  72  75  76  77  78  79  80  83  84
#>  [55]  85  86  87  88  91  92  93  94  95  96  97  98  99 100 101 102 103 104
#>  [73] 105 106 108 109 110 111 112 113 114 115 116 117 118 120 121 123 125 126
#>  [91] 127 128 129 131 132 136 138 139 141 142 143 146 147 148 149 151 152 153
#> [109] 156 158 159 165 166 169 170 171 172 174 175 176 177 180 181 182 183 184
#> [127] 186 187 188 189 191 194 195 196 197 201 204 206 207 210 213 214 215 216
#> [145] 218 219 220 221 223 226 227 228 229 233 234 235 237 240 242 243 244 245
#> [163] 246 247 248 249 250 251 252 254 255 257 258 262 263 265 266 267 268 269
#> [181] 270 271 272 274 275 276 277 278 280 281 282 286 287 288 289 290 292 294
#> [199] 295 298 299 300 301 303 305 306 307 309 313 316 319 321 322 323 325 326
#> [217] 327 328 329 330 331 332 333 334 337 338 339 340 342 343 344 346 347 348
#> [235] 349 350 351 353 355 356 357 358 359 361 363 364 366 367 368 370 373 375
#> [253] 376 378 379 381 382 386 388 394 395 397 398 399 400 401 404 405 408 410
#> [271] 411 412 414 415 416 417 418 419 422 423 426 428 430 431 432 433 434 435
#> [289] 436 437 438 439 441 442 443 444 445 447 449 450 451 453 455 457 458 459
#> [307] 460 461 462 463 464 465 466 468 469 470 472 473 474 475 476 477 479 481
#> [325] 482 484 488 491 493 494 495 498 499
#> 
#> 
#> 
#> [[5]]
#> [[5]]$test
#> [[5]]$test[[1]]
#>   [1]   4   9  11  12  13  15  17  22  30  34  35  37  40  48  49  51  52  54
#>  [19]  55  56  59  62  63  64  67  73  75  77  80  82  84  85  87  88 102 103
#>  [37] 107 109 110 113 116 117 118 119 120 122 123 125 126 128 130 139 143 151
#>  [55] 153 155 156 160 166 170 171 174 175 177 180 187 189 196 199 201 202 206
#>  [73] 207 209 214 219 223 226 228 231 236 241 246 250 254 263 266 268 269 270
#>  [91] 273 276 280 284 289 294 301 303 310 314 317 320 324 326 333 336 338 339
#> [109] 349 354 355 357 358 368 370 372 379 381 386 387 388 397 400 401 408 412
#> [127] 414 415 417 422 423 426 430 432 435 436 438 439 441 442 445 454 458 461
#> [145] 463 464 466 468 473 475 477 478 484 486 487 489 490 491 495
#> 
#> [[5]]$test[[2]]
#>   [1]   1   3   5  19  20  21  23  26  29  31  32  36  43  44  46  50  53  58
#>  [19]  61  66  70  71  74  78  83  86  89  91  92  93  96  98 101 104 106 115
#>  [37] 121 127 129 131 133 138 142 146 147 149 152 154 158 159 162 164 165 172
#>  [55] 182 183 184 185 186 191 192 194 195 197 198 200 204 205 211 212 213 216
#>  [73] 217 221 224 225 227 237 238 240 244 245 247 248 251 255 256 262 264 265
#>  [91] 267 272 274 275 279 281 282 287 288 290 292 293 297 300 302 305 306 307
#> [109] 308 309 311 312 313 316 319 322 323 327 329 330 334 335 337 341 343 351
#> [127] 353 356 361 362 363 364 367 373 375 376 380 382 385 389 390 392 393 395
#> [145] 398 403 405 411 413 416 418 419 425 427 428 443 444 448 449 450 451 455
#> [163] 462 467 469 471 472 479 480 481 482 483 485 488 493 498
#> 
#> [[5]]$test[[3]]
#>   [1]   2   6   7   8  10  14  16  18  24  25  27  28  33  38  39  41  42  45
#>  [19]  47  57  60  65  68  69  72  76  79  81  90  94  95  97  99 100 105 108
#>  [37] 111 112 114 124 132 134 135 136 137 140 141 144 145 148 150 157 161 163
#>  [55] 167 168 169 173 176 178 179 181 188 190 193 203 208 210 215 218 220 222
#>  [73] 229 230 232 233 234 235 239 242 243 249 252 253 257 258 259 260 261 271
#>  [91] 277 278 283 285 286 291 295 296 298 299 304 315 318 321 325 328 331 332
#> [109] 340 342 344 345 346 347 348 350 352 359 360 365 366 369 371 374 377 378
#> [127] 383 384 391 394 396 399 402 404 406 407 409 410 420 421 424 429 431 433
#> [145] 434 437 440 446 447 452 453 456 457 459 460 465 470 474 476 492 494 496
#> [163] 497 499 500
#> 
#> 
#> [[5]]$train
#> [[5]]$train[[1]]
#>   [1]   1   2   3   5   6   7   8  10  14  16  18  19  20  21  23  24  25  26
#>  [19]  27  28  29  31  32  33  36  38  39  41  42  43  44  45  46  47  50  53
#>  [37]  57  58  60  61  65  66  68  69  70  71  72  74  76  78  79  81  83  86
#>  [55]  89  90  91  92  93  94  95  96  97  98  99 100 101 104 105 106 108 111
#>  [73] 112 114 115 121 124 127 129 131 132 133 134 135 136 137 138 140 141 142
#>  [91] 144 145 146 147 148 149 150 152 154 157 158 159 161 162 163 164 165 167
#> [109] 168 169 172 173 176 178 179 181 182 183 184 185 186 188 190 191 192 193
#> [127] 194 195 197 198 200 203 204 205 208 210 211 212 213 215 216 217 218 220
#> [145] 221 222 224 225 227 229 230 232 233 234 235 237 238 239 240 242 243 244
#> [163] 245 247 248 249 251 252 253 255 256 257 258 259 260 261 262 264 265 267
#> [181] 271 272 274 275 277 278 279 281 282 283 285 286 287 288 290 291 292 293
#> [199] 295 296 297 298 299 300 302 304 305 306 307 308 309 311 312 313 315 316
#> [217] 318 319 321 322 323 325 327 328 329 330 331 332 334 335 337 340 341 342
#> [235] 343 344 345 346 347 348 350 351 352 353 356 359 360 361 362 363 364 365
#> [253] 366 367 369 371 373 374 375 376 377 378 380 382 383 384 385 389 390 391
#> [271] 392 393 394 395 396 398 399 402 403 404 405 406 407 409 410 411 413 416
#> [289] 418 419 420 421 424 425 427 428 429 431 433 434 437 440 443 444 446 447
#> [307] 448 449 450 451 452 453 455 456 457 459 460 462 465 467 469 470 471 472
#> [325] 474 476 479 480 481 482 483 485 488 492 493 494 496 497 498 499 500
#> 
#> [[5]]$train[[2]]
#>   [1]   2   4   6   7   8   9  10  11  12  13  14  15  16  17  18  22  24  25
#>  [19]  27  28  30  33  34  35  37  38  39  40  41  42  45  47  48  49  51  52
#>  [37]  54  55  56  57  59  60  62  63  64  65  67  68  69  72  73  75  76  77
#>  [55]  79  80  81  82  84  85  87  88  90  94  95  97  99 100 102 103 105 107
#>  [73] 108 109 110 111 112 113 114 116 117 118 119 120 122 123 124 125 126 128
#>  [91] 130 132 134 135 136 137 139 140 141 143 144 145 148 150 151 153 155 156
#> [109] 157 160 161 163 166 167 168 169 170 171 173 174 175 176 177 178 179 180
#> [127] 181 187 188 189 190 193 196 199 201 202 203 206 207 208 209 210 214 215
#> [145] 218 219 220 222 223 226 228 229 230 231 232 233 234 235 236 239 241 242
#> [163] 243 246 249 250 252 253 254 257 258 259 260 261 263 266 268 269 270 271
#> [181] 273 276 277 278 280 283 284 285 286 289 291 294 295 296 298 299 301 303
#> [199] 304 310 314 315 317 318 320 321 324 325 326 328 331 332 333 336 338 339
#> [217] 340 342 344 345 346 347 348 349 350 352 354 355 357 358 359 360 365 366
#> [235] 368 369 370 371 372 374 377 378 379 381 383 384 386 387 388 391 394 396
#> [253] 397 399 400 401 402 404 406 407 408 409 410 412 414 415 417 420 421 422
#> [271] 423 424 426 429 430 431 432 433 434 435 436 437 438 439 440 441 442 445
#> [289] 446 447 452 453 454 456 457 458 459 460 461 463 464 465 466 468 470 473
#> [307] 474 475 476 477 478 484 486 487 489 490 491 492 494 495 496 497 499 500
#> 
#> [[5]]$train[[3]]
#>   [1]   1   3   4   5   9  11  12  13  15  17  19  20  21  22  23  26  29  30
#>  [19]  31  32  34  35  36  37  40  43  44  46  48  49  50  51  52  53  54  55
#>  [37]  56  58  59  61  62  63  64  66  67  70  71  73  74  75  77  78  80  82
#>  [55]  83  84  85  86  87  88  89  91  92  93  96  98 101 102 103 104 106 107
#>  [73] 109 110 113 115 116 117 118 119 120 121 122 123 125 126 127 128 129 130
#>  [91] 131 133 138 139 142 143 146 147 149 151 152 153 154 155 156 158 159 160
#> [109] 162 164 165 166 170 171 172 174 175 177 180 182 183 184 185 186 187 189
#> [127] 191 192 194 195 196 197 198 199 200 201 202 204 205 206 207 209 211 212
#> [145] 213 214 216 217 219 221 223 224 225 226 227 228 231 236 237 238 240 241
#> [163] 244 245 246 247 248 250 251 254 255 256 262 263 264 265 266 267 268 269
#> [181] 270 272 273 274 275 276 279 280 281 282 284 287 288 289 290 292 293 294
#> [199] 297 300 301 302 303 305 306 307 308 309 310 311 312 313 314 316 317 319
#> [217] 320 322 323 324 326 327 329 330 333 334 335 336 337 338 339 341 343 349
#> [235] 351 353 354 355 356 357 358 361 362 363 364 367 368 370 372 373 375 376
#> [253] 379 380 381 382 385 386 387 388 389 390 392 393 395 397 398 400 401 403
#> [271] 405 408 411 412 413 414 415 416 417 418 419 422 423 425 426 427 428 430
#> [289] 432 435 436 438 439 441 442 443 444 445 448 449 450 451 454 455 458 461
#> [307] 462 463 464 466 467 468 469 471 472 473 475 477 478 479 480 481 482 483
#> [325] 484 485 486 487 488 489 490 491 493 495 498
#> 
#> 
#>