Spatial Block Cross Validation Resampling

Spatial Block Cross validation implemented by the blockCV package.

References

Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G (2018). “blockCV: an R package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models.” bioRxiv. doi: 10.1101/357798 .

Super class

mlr3::Resampling -> ResamplingSpCVBlock

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSpCVBlock$new()

ResamplingSpCVBlock$instantiate()

ResamplingSpCVBlock$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "Environmental Block" resampling instance.

Usage

ResamplingSpCVBlock$new(id = "spcv-block")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSpCVBlock$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSpCVBlock$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv-block", range = 1000)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)

#>   [1]  27  38  41  43  53  67  72  82  98 114 116 124 141 154 186 198 199 233
#>  [19] 235 275 296 298 299 301 361 367 392 401 410 441 462 466 484 489 495 508
#>  [37] 522 526 530 551 576 579 595 600 603 694 697 709 719 721 748  10  12  14
#>  [55]  17  21  39  45  57  61  66  68  69  77  86  91  94  97 105 106 108 115
#>  [73] 140 155 171 180 190 196 201 202 219 240 242 243 263 277 282 287 289 305
#>  [91] 317 326 328 334 353 372 384 385 389 395 405 414 415 416 427 435 456 487
#> [109] 488 511 512 519 527 532 550 554 561 563 565 580 582 583 618 619 631 638
#> [127] 642 649 651 654 668 670 671 673 681 682 683 691 703 711 720 723 736 747
#> [145]  42  46  54  55  62  65  79  93 113 128 130 132 142 145 163 181 182 191
#> [163] 195 209 211 213 225 229 230 249 250 253 262 270 308 322 327 336 366 368
#> [181] 383 452 458 465 475 478 493 498 504 516 534 539 542 552 581 588 599 627
#> [199] 630 646 650 657 663 665 674 676 688 689 701 702 714 716 744 751   2   3
#> [217]   5  18  19  28  63  89 104 118 120 125 129 131 138 161 203 214 221 231
#> [235] 236 251 260 268 271 281 346 349 370 373 377 406 411 412 422 423 442 444
#> [253] 455 460 492 510 523 541 596 628 661 693 710 712 718 728 733  47  52  59
#> [271]  60  76  87  95  96  99 121 139 144 156 157 239 241 269 274 283 319 321
#> [289] 329 332 375 376 397 408 417 419 424 440 471 501 502 514 540 548 553 560
#> [307] 569 571 578 597 635 637 652 662 666 669 672 685 686 692 698 704 717 724
#> [325] 729 741 749  25  32  74 117 119 123 143 147 159 160 162 165 178 187 188
#> [343] 194 200 204 208 215 223 224 244 246 266 279 285 288 292 306 313 318 323
#> [361] 348 355 363 364 365 371 381 388 390 393 409 418 430 431 433 436 446 454
#> [379] 468 472 477 479 480 496 497 505 506 524 533 538 543 545 549 555 558 559
#> [397] 567 570 584 586 591 593 598 601 607 611 617 621 624 625 640 658 667 680
#> [415] 687 726 730 740 742 750   4   8  15  24  26  48  50  64  70  71  73  85
#> [433] 102 135 150 151 153 158 166 170 172 185 192 206 207 216 217 218 227 234
#> [451] 238 247 255 261 265 272 278 280 344 354 358 362 398 402 426 428 429 438
#> [469] 451 453 461 464 470 476 499 518 520 525 529 557 562 574 589 592 605 610
#> [487] 612 623 633 644 656 679 707 725 727 737 738 739 746   1   6  11  16  23
#> [505]  33  34  51  56  84 100 107 112 146 148 149 152 164 167 168 174 175 193
#> [523] 210 212 232 248 252 257 258 264 267 273 286 294 295 300 302 307 310 312
#> [541] 314 316 324 330 342 347 351 369 374 379 387 399 420 421 425 432 439 445
#> [559] 447 459 469 473 482 485 491 507 515 517 572 577 590 604 609 626 629 632
#> [577] 645 647 660 677 678 699 731   7   9  13  20  22  30  31  35  36  40  44
#> [595]  75  78  80  81  83  88  92 101 109 111 134 137 173 176 177 179 184 189
#> [613] 197 205 226 228 237 254 256 259 276 284 291 293 297 304 311 315 320 331
#> [631] 333 335 337 338 339 340 341 343 352 360 378 380 382 386 391 396 407 434
#> [649] 437 443 448 450 457 481 494 500 513 521 535 536 556 566 568 587 606 608
#> [667] 614 615 616 622 634 636 648 659 675 684 695 700 705 713 722 734 745

rcv$test_set(1)

#>  [1]  29  37  49  58  90 103 110 122 126 127 133 136 169 183 220 222 245 290 303
#> [20] 309 325 345 350 356 357 359 394 400 403 404 413 449 463 467 474 483 486 490
#> [39] 503 509 528 531 537 544 546 547 564 573 575 585 594 602 613 620 639 641 643
#> [58] 653 655 664 690 696 706 708 715 732 735 743

intersect(rcv$train_set(1), rcv$test_set(1))

#> integer(0)

# Internal storage:
rcv$instance

#>      row_id fold
#>   1:     29    1
#>   2:     37    1
#>   3:     49    1
#>   4:     58    1
#>   5:     90    1
#>  ---            
#> 747:    705   10
#> 748:    713   10
#> 749:    722   10
#> 750:    734   10
#> 751:    745   10