(sperrorest) Spatial "disc" resampling

partition_discpartitions the sample into training and tests set by selecting circular test areas (possibly surrounded by an exclusion buffer) and using the remaining samples as training samples (leave-one-disc-out cross-validation).partition_loo` creates training and test sets for leave-one-out cross-validation with (optional) buffer.

Note

Test area discs are centered at (random) samples, not at general random locations. Test area discs may (and likely will) overlap independently of the value of replace. replace only controls the replacement of the center point of discs when drawing center points from the samples.

mlr3spatiotempcv notes

The 'Description' and 'Note' fields are inherited from the respective upstream function.

For a list of available arguments, please see sperrorest::partition_disc.

This method is similar to ResamplingSpCVBuffer.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi: 10.1109/igarss.2012.6352393 .

See also

Super class

mlr3::Resampling -> ResamplingSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSpCVDisc$new()

ResamplingSpCVDisc$instantiate()

ResamplingSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingSpCVDisc$new(id = "spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv_disc", folds = 3L, radius = 200L, buffer = 200L)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)

#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  53  54  55
#>  [55]  56  57  58  59  61  62  63  64  65  66  67  68  69  70  71  72  73  74
#>  [73]  75  77  78  79  80  81  82  83  84  85  86  88  89  90  91  92  93  94
#>  [91]  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 157 158 159 160 161 162 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 178 179 180 181 182 183 185 186 187 188
#> [181] 189 190 191 192 193 194 195 196 198 199 200 201 202 203 204 205 206 207
#> [199] 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225
#> [217] 226 227 228 229 230 231 232 233 234 235 236 237 238 240 242 243 244 245
#> [235] 246 247 248 249 250 251 252 253 254 255 257 258 259 260 261 262 263 264
#> [253] 265 266 267 268 270 271 272 273 274 275 276 277 278 279 280 281 282 283
#> [271] 284 285 286 287 288 289 290 291 292 293 294 295 296 298 299 300 301 302
#> [289] 303 304 305 306 307 308 309 310 312 313 314 315 316 317 318 319 320 322
#> [307] 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340
#> [325] 341 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358
#> [343] 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 377
#> [361] 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395
#> [379] 398 399 400 401 402 403 404 405 406 408 409 410 411 412 413 414 415 416
#> [397] 417 418 419 420 421 422 423 425 426 427 428 429 430 431 432 433 435 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454
#> [433] 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 472 473
#> [451] 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488 489 490 491
#> [469] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509
#> [487] 510 511 512 513 515 516 517 518 519 520 521 522 523 524 525 526 527 528
#> [505] 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546
#> [523] 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564
#> [541] 565 566 567 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [559] 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601
#> [577] 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [595] 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637
#> [613] 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655
#> [631] 656 657 658 659 660 661 662 663 664 665 667 668 669 670 671 672 673 674
#> [649] 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692
#> [667] 693 694 695 696 697 699 701 702 703 704 705 706 707 708 709 710 711 712
#> [685] 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730
#> [703] 731 732 733 734 735 736 737 738 739 740 741 742 743 744 746 747 748 749
#> [721] 750 751

rcv$test_set(1)

#> [1]  52  60  76 239 376 397 471

# check that no obs are in both sets
intersect(rcv$train_set(1), rcv$test_set(1)) # good!

#> integer(0)

# Internal storage:
rcv$instance # table

#> $test
#> $test[[1]]
#> [1]  52  60  76 239 376 397 471
#> 
#> $test[[2]]
#> [1] 156 239 269 321 376 397 424 514 698
#> 
#> $test[[3]]
#> [1] 143 188 204 285 318 409 477 567 591
#> 
#> 
#> $train
#> $train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  53  54  55
#>  [55]  56  57  58  59  61  62  63  64  65  66  67  68  69  70  71  72  73  74
#>  [73]  75  77  78  79  80  81  82  83  84  85  86  88  89  90  91  92  93  94
#>  [91]  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 157 158 159 160 161 162 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 178 179 180 181 182 183 185 186 187 188
#> [181] 189 190 191 192 193 194 195 196 198 199 200 201 202 203 204 205 206 207
#> [199] 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225
#> [217] 226 227 228 229 230 231 232 233 234 235 236 237 238 240 242 243 244 245
#> [235] 246 247 248 249 250 251 252 253 254 255 257 258 259 260 261 262 263 264
#> [253] 265 266 267 268 270 271 272 273 274 275 276 277 278 279 280 281 282 283
#> [271] 284 285 286 287 288 289 290 291 292 293 294 295 296 298 299 300 301 302
#> [289] 303 304 305 306 307 308 309 310 312 313 314 315 316 317 318 319 320 322
#> [307] 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340
#> [325] 341 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358
#> [343] 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 377
#> [361] 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395
#> [379] 398 399 400 401 402 403 404 405 406 408 409 410 411 412 413 414 415 416
#> [397] 417 418 419 420 421 422 423 425 426 427 428 429 430 431 432 433 435 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454
#> [433] 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 472 473
#> [451] 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488 489 490 491
#> [469] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509
#> [487] 510 511 512 513 515 516 517 518 519 520 521 522 523 524 525 526 527 528
#> [505] 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546
#> [523] 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564
#> [541] 565 566 567 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [559] 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601
#> [577] 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [595] 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637
#> [613] 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655
#> [631] 656 657 658 659 660 661 662 663 664 665 667 668 669 670 671 672 673 674
#> [649] 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692
#> [667] 693 694 695 696 697 699 701 702 703 704 705 706 707 708 709 710 711 712
#> [685] 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730
#> [703] 731 732 733 734 735 736 737 738 739 740 741 742 743 744 746 747 748 749
#> [721] 750 751
#> 
#> $train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  53  54  55
#>  [55]  56  57  58  59  61  62  63  64  65  66  67  68  69  70  71  72  73  74
#>  [73]  75  77  78  79  80  81  82  83  84  85  86  88  89  90  91  92  93  94
#>  [91]  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 157 158 159 160 161 162 163 164 165 166 167
#> [163] 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185
#> [181] 186 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203
#> [199] 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221
#> [217] 222 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 240
#> [235] 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259
#> [253] 260 261 262 263 264 265 266 267 268 270 271 272 273 274 275 276 277 278
#> [271] 279 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296
#> [289] 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314
#> [307] 315 316 317 318 319 320 322 323 324 325 326 327 328 329 330 331 332 333
#> [325] 334 335 336 337 338 339 340 341 342 343 344 345 346 347 348 349 350 351
#> [343] 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369
#> [361] 370 371 372 373 374 375 377 378 379 380 381 382 383 384 385 386 387 388
#> [379] 389 390 391 392 393 394 395 396 398 399 400 401 402 403 404 405 406 407
#> [397] 408 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 425 426
#> [415] 427 428 429 430 431 432 433 434 435 436 437 438 439 440 441 442 443 444
#> [433] 445 446 447 448 449 450 451 452 453 454 455 456 457 458 459 460 461 462
#> [451] 463 464 465 466 467 468 469 470 472 473 474 475 476 477 478 479 480 481
#> [469] 482 483 484 485 486 487 488 489 490 491 492 493 494 495 496 497 498 499
#> [487] 500 502 503 504 505 506 507 508 509 510 511 512 513 515 516 517 518 519
#> [505] 520 521 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 537
#> [523] 538 539 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555
#> [541] 556 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573
#> [559] 574 575 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591
#> [577] 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607 608 609
#> [595] 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 625 626 627
#> [613] 628 629 630 631 632 633 634 636 637 638 639 640 641 642 643 644 645 646
#> [631] 647 648 649 650 651 653 654 655 656 657 658 659 660 661 663 664 665 667
#> [649] 668 669 670 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685
#> [667] 686 687 688 689 690 691 692 693 694 695 696 697 699 700 701 702 703 704
#> [685] 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721 722
#> [703] 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740
#> [721] 741 742 743 744 745 746 747 748 749 750 751
#> 
#> $train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 124 125 126 127
#> [127] 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 144 145 146
#> [145] 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 163 164 165
#> [163] 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183
#> [181] 184 185 186 187 189 190 191 192 193 194 195 196 197 198 199 201 202 203
#> [199] 205 206 207 208 209 210 211 212 213 214 216 217 218 219 220 221 222 224
#> [217] 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242
#> [235] 243 244 245 247 248 249 250 251 252 253 254 255 256 257 258 259 260 261
#> [253] 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 280
#> [271] 281 282 283 284 286 287 289 290 291 292 293 294 295 296 297 298 299 300
#> [289] 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317 319
#> [307] 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337
#> [325] 338 339 340 341 342 343 344 345 346 347 349 350 351 352 353 354 355 356
#> [343] 357 358 359 360 361 362 363 364 366 367 368 369 370 372 373 374 375 376
#> [361] 377 378 379 380 382 383 384 385 386 387 388 389 390 391 392 393 394 395
#> [379] 396 397 398 399 400 401 402 403 404 405 406 407 408 410 411 412 413 414
#> [397] 415 416 417 418 419 420 421 422 423 424 425 426 427 428 429 432 433 434
#> [415] 435 436 437 438 439 440 441 442 443 444 445 447 448 449 450 451 452 453
#> [433] 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471
#> [451] 472 473 474 475 476 478 479 480 481 482 483 484 485 486 487 488 489 490
#> [469] 491 492 493 494 495 497 498 499 500 501 502 503 504 506 507 508 509 510
#> [487] 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528
#> [505] 529 530 531 532 534 535 536 537 539 540 541 542 543 544 545 546 547 548
#> [523] 549 550 551 552 553 554 555 556 557 558 560 561 562 563 564 565 566 568
#> [541] 569 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586 587
#> [559] 588 589 590 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606
#> [577] 608 609 610 611 612 613 614 615 616 618 619 620 621 622 623 624 625 626
#> [595] 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642 643 644
#> [613] 645 646 647 648 649 650 651 652 653 654 655 656 657 659 660 661 662 663
#> [631] 664 665 666 668 669 670 671 672 673 674 675 676 677 678 679 681 682 683
#> [649] 684 685 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701
#> [667] 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719
#> [685] 720 721 722 723 724 725 727 728 729 730 731 732 733 734 735 736 737 738
#> [703] 739 741 742 743 744 745 746 747 748 749 751
#> 
#>