(sperrorest) Spatial "Tiles" resampling

partition_tiles divides the study area into a specified number of rectangular tiles. Optionally small partitions can be merged with adjacent tiles to achieve a minimum number or percentage of samples in each tile.

Note

Default parameter settings may change in future releases. This function, especially the rotation and shifting part of it and the algorithm for cleaning up small tiles is still a bit experimental. Use with caution. For non-zero offsets (offset!='none')), the number of tiles may actually be greater than nsplit[1]*nsplit[2] because of fractional tiles lurking into the study region. reassign=TRUE with suitable thresholds is therefore recommended for non-zero (including random) offsets.

mlr3spatiotempcv notes

The 'Description' and 'Note' fields are inherited from the respective upstream function.

For a list of available arguments, please see sperrorest::partition_tiles.

This method is similar to ResamplingSpCVBlock.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi: 10.1109/igarss.2012.6352393 .

See also

Super class

mlr3::Resampling -> ResamplingSpCVTiles

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSpCVTiles$new()

ResamplingSpCVTiles$instantiate()

ResamplingSpCVTiles$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Tiles' resampling" resampling instance.

Usage

ResamplingSpCVTiles$new(id = "spcv_Tiles")

Arguments

id

character(1)
Identifier for the resampling strategy. For a list of available arguments, please see sperrorest::partition_tiles.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSpCVTiles$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSpCVTiles$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (mlr3misc::require_namespaces("sperrorest", quietly = TRUE)) {
  library(mlr3)
  task = tsk("ecuador")

  # Instantiate Resampling
  rcv = rsmp("spcv_tiles", nsplit = c(4L, 3L), reassign = FALSE)
  rcv$instantiate(task)

  # Individual sets:
  rcv$train_set(1)
  rcv$test_set(1)
  # check that no obs are in both sets
  intersect(rcv$train_set(1), rcv$test_set(1)) # good!

  # Internal storage:
  rcv$instance # table
}

#> $train
#> $train$`1`
#>   [1]   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  34  35  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58
#>  [55]  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  75  76  77
#>  [73]  78  79  80  82  83  85  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 101 102 103 104 105 106 107 108 109 110 113 114 115 116 118 120
#> [109] 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138
#> [127] 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153 154 155 156
#> [145] 157 158 161 162 163 166 169 170 171 172 173 174 176 177 179 180 181 182
#> [163] 183 185 186 188 189 190 191 192 193 195 196 197 198 199 200 201 202 203
#> [181] 204 205 206 207 209 211 212 213 214 215 216 217 218 219 220 221 222 223
#> [199] 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241
#> [217] 242 243 245 246 247 249 250 251 252 253 254 255 258 259 260 261 262 263
#> [235] 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280 281
#> [253] 282 283 284 285 286 287 288 289 290 291 293 294 295 296 297 298 299 300
#> [271] 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318
#> [289] 319 320 321 322 325 326 327 328 329 330 331 332 333 334 335 336 337 338
#> [307] 339 340 341 342 343 344 345 346 347 348 349 350 352 353 354 356 357 358
#> [325] 359 360 361 362 365 366 367 368 369 370 371 372 373 375 376 377 378 379
#> [343] 380 381 383 384 385 386 387 388 389 392 394 395 396 397 398 399 400 401
#> [361] 402 403 404 405 406 408 409 410 411 412 413 414 415 416 417 418 419 420
#> [379] 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 437 438 439
#> [397] 440 441 442 443 444 445 446 447 448 449 450 451 452 453 455 456 457 458
#> [415] 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475 476
#> [433] 477 478 480 482 483 484 485 486 487 488 489 490 491 492 493 494 495 496
#> [451] 498 499 500 501 502 503 504 505 507 508 509 510 511 512 513 514 515 516
#> [469] 517 518 519 520 521 522 523 525 526 527 528 529 530 531 532 533 534 535
#> [487] 536 537 538 539 540 541 542 543 544 546 547 548 550 551 552 553 554 555
#> [505] 556 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573
#> [523] 574 575 576 577 578 579 580 581 582 583 585 586 587 588 589 590 591 592
#> [541] 594 595 596 598 599 600 602 603 604 605 606 607 608 609 610 611 612 613
#> [559] 614 615 617 618 619 620 621 622 623 626 627 628 629 630 631 632 633 634
#> [577] 635 636 637 638 639 641 642 643 644 645 646 647 648 649 650 651 652 653
#> [595] 654 655 656 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671
#> [613] 673 674 675 676 677 678 679 680 681 682 683 684 686 688 689 690 691 692
#> [631] 693 694 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711
#> [649] 712 713 714 715 716 718 719 720 721 722 723 724 725 726 727 728 729 730
#> [667] 731 732 733 734 735 736 737 738 739 740 741 743 744 745 746 747 748 749
#> [685] 750 751
#> 
#> $train$`2`
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  15  17  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  33  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  52  53  54  55  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78
#>  [73]  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96
#>  [91]  97  98  99 100 101 102 103 104 105 106 108 109 110 111 112 113 114 115
#> [109] 116 117 118 119 120 122 123 124 125 126 127 128 129 130 131 132 133 134
#> [127] 135 136 137 138 139 140 141 142 143 144 145 147 150 151 153 154 155 156
#> [145] 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172 173 175 176
#> [163] 177 178 179 180 181 182 183 184 185 186 187 188 189 190 191 192 194 195
#> [181] 196 197 199 200 201 202 203 204 205 206 207 208 209 210 211 213 214 215
#> [199] 216 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231 233 234
#> [217] 235 236 237 238 239 240 241 242 243 244 245 246 247 248 250 251 253 254
#> [235] 255 256 257 259 260 261 262 263 265 266 268 269 270 271 272 274 275 276
#> [253] 277 278 279 280 281 282 283 284 285 287 288 289 290 291 292 293 296 297
#> [271] 298 299 301 303 304 305 306 308 309 311 313 315 317 318 319 320 321 322
#> [289] 323 324 325 326 327 328 331 333 334 335 336 337 338 339 340 341 343 344
#> [307] 345 346 348 349 350 351 352 353 354 355 356 357 358 359 360 362 363 364
#> [325] 365 366 367 368 370 371 372 373 374 376 377 378 380 381 382 383 384 385
#> [343] 386 388 389 390 391 392 393 394 395 396 397 398 400 401 402 403 404 405
#> [361] 406 407 409 411 412 413 414 415 416 418 419 422 423 424 426 427 428 429
#> [379] 430 431 433 434 435 436 437 438 440 441 442 443 444 446 448 449 450 451
#> [397] 452 453 454 455 456 457 458 460 461 462 463 464 465 467 468 470 471 472
#> [415] 474 475 476 477 478 479 480 481 483 484 486 487 488 489 490 492 493 494
#> [433] 495 496 497 498 499 500 501 502 503 504 505 506 508 509 510 511 512 513
#> [451] 514 516 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532 533
#> [469] 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549 550 551
#> [487] 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568 569
#> [505] 570 571 573 574 575 576 578 579 580 581 582 583 584 585 586 587 589 591
#> [523] 592 593 594 595 596 597 598 599 600 601 602 603 605 606 607 608 610 611
#> [541] 612 613 614 615 616 617 618 619 620 621 622 623 624 625 627 628 630 631
#> [559] 633 634 635 636 638 639 640 641 642 643 644 646 648 649 651 652 653 654
#> [577] 655 656 657 658 659 661 662 663 664 665 666 667 668 669 670 671 672 673
#> [595] 674 675 676 679 680 681 682 683 684 685 687 688 689 690 691 692 693 694
#> [613] 695 696 697 698 700 701 702 703 704 705 706 707 708 709 710 711 712 713
#> [631] 714 715 716 717 718 719 720 722 723 724 725 726 727 728 729 730 732 733
#> [649] 734 735 736 737 738 739 740 741 742 743 744 745 746 747 749 750 751
#> 
#> $train$`3`
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  42  43  44  45  46  47  48  49  50  51  52  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  68  69  70  71  73  74  75  76  77
#>  [73]  78  79  80  81  83  84  85  86  87  88  89  90  91  92  93  94  95  96
#>  [91]  97  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113 115 117
#> [109] 118 119 120 121 122 123 125 126 127 128 129 130 131 132 133 134 135 136
#> [127] 137 138 139 140 142 143 144 145 146 147 148 149 150 151 152 153 155 156
#> [145] 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172 173 174
#> [163] 175 176 177 178 179 180 181 182 183 184 185 187 188 189 190 191 192 193
#> [181] 194 195 196 197 198 200 201 202 203 204 205 206 207 208 209 210 211 212
#> [199] 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227 228 229 230
#> [217] 231 232 234 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250
#> [235] 251 252 253 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268
#> [253] 269 270 271 272 273 274 276 277 278 279 280 281 282 283 284 285 286 287
#> [271] 288 289 290 291 292 293 294 295 297 300 302 303 304 305 306 307 308 309
#> [289] 310 311 312 313 314 315 316 317 318 319 320 321 322 323 324 325 326 327
#> [307] 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344 345
#> [325] 346 347 348 349 350 351 352 353 354 355 356 357 358 360 361 362 363 364
#> [343] 365 366 368 369 370 371 372 373 374 375 376 377 378 379 380 381 382 383
#> [361] 384 385 386 387 388 389 390 391 393 394 395 396 397 398 399 400 402 403
#> [379] 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421
#> [397] 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439
#> [415] 440 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 458
#> [433] 459 460 461 463 464 465 466 468 469 470 471 472 473 474 475 476 477 478
#> [451] 479 480 481 482 483 484 485 486 487 488 490 491 492 493 494 495 496 497
#> [469] 498 499 500 501 502 503 504 505 506 507 509 510 511 512 513 514 515 516
#> [487] 517 518 519 520 521 523 524 525 527 528 529 531 532 533 534 535 536 537
#> [505] 538 539 540 541 542 543 544 545 546 548 549 550 551 552 553 554 555 556
#> [523] 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574
#> [541] 575 577 578 580 581 582 583 584 585 586 587 588 589 590 591 592 593 594
#> [559] 595 596 597 598 599 600 601 604 605 606 607 608 609 610 611 612 613 614
#> [577] 615 616 617 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632
#> [595] 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648 649 650
#> [613] 651 652 653 654 655 656 657 658 659 660 661 662 663 664 665 666 667 668
#> [631] 669 670 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685 686
#> [649] 687 688 689 690 691 692 693 694 695 696 698 699 700 701 702 703 704 705
#> [667] 706 707 708 709 710 711 712 713 714 715 716 717 718 720 721 722 723 724
#> [685] 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740 741 742
#> [703] 743 744 745 746 747 748 749 750 751
#> 
#> $train$`4`
#>   [1]   1   2   3   4   5   6   7   8  10  11  12  13  14  15  16  17  18  19
#>  [19]  21  23  24  25  26  27  28  29  32  33  34  35  36  37  39  41  42  44
#>  [37]  45  46  47  48  49  50  51  53  54  55  56  57  58  59  61  62  63  64
#>  [55]  65  66  67  68  69  70  71  72  73  74  75  77  78  79  81  82  84  85
#>  [73]  86  87  88  89  90  91  93  94  95  96  97  98  99 100 102 103 104 105
#>  [91] 106 107 108 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123
#> [109] 124 125 126 127 128 129 130 131 132 133 135 136 137 138 139 140 141 142
#> [127] 143 144 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160
#> [145] 161 162 163 164 165 166 167 168 169 170 171 172 174 175 178 180 181 182
#> [163] 183 184 185 186 187 188 190 191 192 193 194 195 196 198 199 200 201 202
#> [181] 203 204 206 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221
#> [199] 222 223 224 225 227 229 230 231 232 233 234 235 236 238 240 241 242 243
#> [217] 244 245 246 247 248 249 250 251 252 253 255 256 257 258 259 260 261 262
#> [235] 263 264 265 266 267 268 269 270 271 272 273 274 275 277 278 279 280 281
#> [253] 282 283 285 286 287 288 289 290 292 294 295 296 298 299 300 301 302 303
#> [271] 305 306 307 308 309 310 312 313 314 316 317 318 319 320 321 322 323 324
#> [289] 325 326 327 328 329 330 331 332 334 336 337 338 342 344 345 346 347 348
#> [307] 349 350 351 353 354 355 356 357 358 359 361 362 363 364 365 366 367 368
#> [325] 369 370 371 372 373 374 375 376 377 378 379 381 382 383 384 385 387 388
#> [343] 389 390 391 392 393 394 395 398 399 400 401 402 403 404 405 406 407 408
#> [361] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426
#> [379] 427 428 429 430 431 432 433 435 436 438 439 440 441 442 444 445 446 447
#> [397] 448 449 451 452 453 454 455 456 458 459 460 461 462 463 464 465 466 467
#> [415] 468 469 470 472 473 474 475 476 477 478 479 480 481 482 483 485 486 487
#> [433] 488 489 490 491 492 493 494 496 497 498 499 501 503 504 505 506 507 508
#> [451] 509 510 511 512 514 515 516 517 518 519 520 522 523 524 525 526 527 528
#> [469] 529 530 531 532 533 534 535 537 538 539 540 541 542 543 544 545 546 547
#> [487] 548 549 550 552 553 554 555 557 558 559 560 561 562 563 564 565 566 567
#> [505] 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586
#> [523] 588 589 590 591 592 593 594 596 597 598 599 601 602 603 604 605 607 609
#> [541] 610 611 612 613 616 617 618 619 620 621 623 624 625 626 627 628 629 630
#> [559] 631 632 633 635 636 637 638 639 640 641 642 643 644 645 646 647 649 650
#> [577] 651 653 654 655 656 657 658 660 661 662 663 664 665 667 668 669 670 671
#> [595] 672 673 674 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690
#> [613] 691 692 693 695 696 697 698 699 701 702 703 704 705 706 707 708 710 711
#> [631] 712 714 715 716 717 718 719 720 721 723 725 726 727 728 729 730 731 732
#> [649] 733 734 735 736 737 738 739 740 741 742 743 744 746 747 748 749 750 751
#> 
#> $train$`5`
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  43  44  45  46  48  49  50  51  52  53  54  56  57
#>  [55]  58  60  61  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77
#>  [73]  78  79  80  81  82  83  84  85  86  88  89  90  91  92  94  97  98 100
#>  [91] 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118
#> [109] 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136
#> [127] 137 138 140 141 142 143 145 146 147 148 149 150 151 152 153 154 155 157
#> [145] 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172 173 174 175
#> [163] 176 177 178 179 180 181 182 183 184 185 186 187 188 189 190 191 192 193
#> [181] 194 195 196 197 198 199 200 201 202 203 204 205 206 207 208 209 210 211
#> [199] 212 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227 228 229
#> [217] 231 232 233 234 235 236 237 238 239 240 242 243 244 245 246 247 248 249
#> [235] 251 252 253 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268
#> [253] 270 271 272 273 275 276 277 278 279 280 281 282 284 285 286 287 288 289
#> [271] 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307
#> [289] 308 309 310 311 312 313 314 315 316 317 318 320 322 323 324 325 326 327
#> [307] 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344 345
#> [325] 346 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363
#> [343] 364 365 366 367 368 369 370 371 372 373 374 375 377 378 379 380 381 382
#> [361] 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401
#> [379] 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 420
#> [397] 421 422 423 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439
#> [415] 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 458
#> [433] 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475 476
#> [451] 477 478 479 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494
#> [469] 495 496 497 498 499 500 502 503 505 506 507 508 509 510 511 512 513 515
#> [487] 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532 533 535
#> [505] 536 537 538 541 542 543 544 545 546 547 549 550 551 554 555 556 557 558
#> [523] 559 561 562 563 564 565 566 567 568 570 572 573 574 575 576 577 579 580
#> [541] 581 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598
#> [559] 599 600 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616
#> [577] 617 618 619 620 621 622 623 624 625 626 628 629 631 632 633 634 636 637
#> [595] 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655
#> [613] 656 658 659 660 661 663 664 666 667 668 670 671 672 673 675 677 678 679
#> [631] 680 681 682 683 684 685 686 687 690 691 693 694 695 696 697 699 700 701
#> [649] 703 705 706 707 708 709 710 711 712 713 714 715 717 718 719 720 721 722
#> [667] 723 724 725 726 727 728 730 731 732 733 734 735 736 737 738 739 740 742
#> [685] 743 745 746 747 748 750
#> 
#> $train$`6`
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  13  15  16  17  18  19  20
#>  [19]  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  41  42  43  44  45  47  48  49  50  51  52  53  55  56  57  58  59
#>  [55]  60  61  62  63  64  67  68  70  71  72  73  74  75  76  77  78  80  81
#>  [73]  82  83  84  85  87  88  89  90  92  93  95  96  98  99 100 101 102 103
#>  [91] 104 106 107 109 110 111 112 114 115 116 117 118 119 120 121 122 123 124
#> [109] 125 126 127 129 131 133 134 135 136 137 138 139 141 142 143 144 146 147
#> [127] 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 164 165 166
#> [145] 167 168 169 170 172 173 174 175 176 177 178 179 180 183 184 185 186 187
#> [163] 188 189 190 192 193 194 197 198 199 200 201 203 204 205 206 207 208 210
#> [181] 212 214 215 216 217 218 219 220 221 222 223 224 226 227 228 230 231 232
#> [199] 233 234 235 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250
#> [217] 251 252 254 255 256 257 258 259 260 261 263 264 265 266 267 268 269 271
#> [235] 272 273 274 275 276 278 279 280 281 283 284 285 286 287 288 289 290 291
#> [253] 292 293 294 295 296 297 298 299 300 301 302 303 304 306 307 309 310 311
#> [271] 312 313 314 315 316 317 318 319 320 321 323 324 325 329 330 331 332 333
#> [289] 335 337 338 339 340 341 342 343 344 345 346 347 348 349 350 351 352 354
#> [307] 355 356 357 358 359 360 361 362 363 364 365 367 369 370 371 372 373 374
#> [325] 375 376 377 378 379 380 381 382 383 384 385 386 387 388 390 391 392 393
#> [343] 394 396 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412
#> [361] 413 414 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432
#> [379] 433 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450
#> [397] 451 453 454 455 457 459 460 461 462 463 464 466 467 468 469 470 471 472
#> [415] 473 474 476 477 479 480 481 482 483 484 485 486 487 488 489 490 491 492
#> [433] 494 495 496 497 499 500 501 502 503 504 505 506 507 508 509 510 512 513
#> [451] 514 515 516 517 518 520 521 522 523 524 525 526 528 529 530 531 532 533
#> [469] 534 535 536 537 538 539 540 541 543 544 545 546 547 548 549 550 551 552
#> [487] 553 554 555 556 557 558 559 560 561 562 564 566 567 568 569 570 571 572
#> [505] 573 574 575 576 577 578 579 580 582 584 585 586 587 588 589 590 591 592
#> [523] 593 594 595 596 597 598 600 601 602 603 604 605 606 607 608 609 610 611
#> [541] 612 613 614 615 616 617 620 621 622 623 624 625 626 627 628 629 630 632
#> [559] 633 634 635 636 637 638 639 640 641 642 643 644 645 647 648 649 650 652
#> [577] 653 654 655 656 657 658 659 660 661 662 664 665 666 667 669 672 673 674
#> [595] 675 676 677 678 679 680 682 683 684 685 686 687 688 689 690 692 693 694
#> [613] 695 696 697 698 699 700 702 704 705 706 707 708 709 710 711 712 713 715
#> [631] 716 717 718 719 721 722 723 724 725 726 727 728 729 730 731 732 733 734
#> [649] 735 736 737 738 739 740 741 742 743 744 745 746 748 749 750 751
#> 
#> $train$`7`
#>   [1]   1   2   3   5   6   7   9  10  11  12  13  14  16  17  18  19  20  21
#>  [19]  22  23  24  25  27  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  49  50  51  52  53  54  55  56  57  59  60
#>  [55]  61  62  63  65  66  67  68  69  72  74  75  76  77  78  79  80  81  82
#>  [73]  83  84  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 101
#>  [91] 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119 120
#> [109] 121 122 123 124 125 127 128 129 130 131 132 134 135 136 137 138 139 140
#> [127] 141 142 143 144 145 146 147 148 149 152 153 154 155 156 157 159 160 161
#> [145] 162 163 164 165 167 168 171 172 173 174 175 176 177 178 179 181 182 183
#> [163] 184 186 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202
#> [181] 203 204 205 208 209 210 211 212 213 214 215 219 221 222 223 224 225 226
#> [199] 227 228 229 230 231 232 233 235 236 237 239 240 241 242 243 244 246 248
#> [217] 249 250 251 252 253 254 256 257 258 259 260 262 263 264 266 267 268 269
#> [235] 270 271 273 274 275 276 277 279 281 282 283 284 285 286 288 289 291 292
#> [253] 293 294 295 296 297 298 299 300 301 302 304 305 306 307 308 310 311 312
#> [271] 313 314 315 316 317 318 319 320 321 322 323 324 325 326 327 328 329 330
#> [289] 331 332 333 334 335 336 337 338 339 340 341 342 343 345 346 347 348 349
#> [307] 350 351 352 353 355 358 359 360 361 363 364 365 366 367 368 369 370 371
#> [325] 372 373 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388 389
#> [343] 390 391 392 393 395 396 397 399 400 401 403 404 405 406 407 408 409 410
#> [361] 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 427 430 431
#> [379] 432 433 434 436 437 439 440 441 442 443 444 445 446 447 448 449 450 452
#> [397] 453 454 455 456 457 458 459 460 462 463 464 465 466 467 468 469 471 472
#> [415] 473 474 475 477 478 479 480 481 482 483 484 485 486 487 489 490 491 492
#> [433] 493 494 495 496 497 498 500 501 502 504 505 506 507 508 510 511 512 513
#> [451] 514 515 516 517 518 519 521 522 523 524 526 527 528 530 531 532 533 534
#> [469] 535 536 537 538 539 540 541 542 543 545 546 547 548 549 550 551 552 553
#> [487] 554 555 556 557 558 559 560 563 564 565 566 567 568 569 570 571 572 573
#> [505] 575 576 577 578 579 580 581 583 584 586 587 588 590 591 593 594 595 596
#> [523] 597 598 599 600 601 602 603 604 606 607 608 609 611 613 614 615 616 617
#> [541] 618 619 620 621 622 624 625 626 627 628 629 630 631 632 634 635 636 637
#> [559] 638 639 640 641 642 643 645 646 647 648 649 650 651 652 653 654 655 657
#> [577] 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675
#> [595] 676 677 678 680 681 682 684 685 686 687 688 689 690 691 692 693 694 695
#> [613] 696 697 698 699 700 701 702 703 704 705 709 710 711 712 713 714 715 716
#> [631] 717 718 719 720 721 722 723 724 726 728 729 730 731 732 733 734 735 736
#> [649] 737 739 740 741 742 743 744 745 747 748 749 750 751
#> 
#> $train$`8`
#>   [1]   1   2   3   4   5   6   7   8   9  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38  40
#>  [37]  41  42  43  44  45  46  47  48  50  51  52  53  54  55  56  58  59  60
#>  [55]  62  63  64  65  66  67  69  70  71  72  73  74  75  76  78  79  80  81
#>  [73]  82  83  84  85  86  87  88  89  91  92  93  94  95  96  97  98  99 100
#>  [91] 101 102 104 105 107 108 109 111 112 113 114 116 117 118 119 120 121 123
#> [109] 124 125 126 127 128 129 130 131 132 133 134 135 137 138 139 140 141 142
#> [127] 143 144 145 146 147 148 149 150 151 152 153 154 156 157 158 159 160 161
#> [145] 162 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179
#> [163] 180 181 182 184 185 186 187 188 189 191 192 193 194 195 196 197 198 199
#> [181] 200 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218
#> [199] 220 221 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238
#> [217] 239 240 241 242 244 245 246 247 248 249 250 251 252 253 254 255 256 257
#> [235] 258 259 260 261 262 264 265 266 267 268 269 270 271 272 273 274 275 276
#> [253] 277 278 279 280 281 282 283 284 285 286 287 288 290 291 292 293 294 295
#> [271] 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313
#> [289] 314 315 316 318 319 320 321 322 323 324 326 327 328 329 330 331 332 333
#> [307] 334 335 336 337 338 339 340 341 342 343 344 346 347 348 349 351 352 353
#> [325] 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371
#> [343] 373 374 375 376 377 378 379 380 381 382 383 384 386 387 388 389 390 391
#> [361] 392 393 394 395 396 397 398 399 401 402 406 407 408 409 410 411 412 415
#> [379] 416 417 418 419 420 421 422 423 424 425 426 428 429 430 431 432 433 434
#> [397] 435 436 437 438 439 440 441 442 443 444 445 446 447 448 450 451 452 453
#> [415] 454 455 456 457 458 459 460 461 462 464 465 466 467 468 469 470 471 472
#> [433] 473 475 476 477 478 479 480 481 482 484 485 488 489 491 492 493 494 495
#> [451] 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513
#> [469] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 529 530 531 532
#> [487] 533 534 535 536 538 539 540 541 542 543 544 545 547 548 549 551 552 553
#> [505] 554 555 556 557 558 559 560 561 562 563 565 566 567 568 569 570 571 572
#> [523] 574 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591 592
#> [541] 593 595 596 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611
#> [559] 612 614 615 616 617 618 619 621 622 623 624 625 626 627 628 629 630 631
#> [577] 632 633 634 635 636 637 638 640 644 645 646 647 648 650 651 652 653 655
#> [595] 656 657 658 659 660 661 662 663 665 666 667 668 669 670 671 672 674 675
#> [613] 676 677 678 679 680 681 683 684 685 686 687 688 689 691 692 693 694 695
#> [631] 697 698 699 700 701 702 703 704 705 706 707 708 709 710 712 713 714 716
#> [649] 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734
#> [667] 735 737 738 739 740 741 742 744 745 746 747 748 749 750 751
#> 
#> $train$`9`
#>   [1]   1   2   3   4   5   6   8   9  10  11  12  14  15  16  18  19  20  21
#>  [19]  22  23  24  25  26  27  28  29  30  31  32  33  34  36  37  38  39  40
#>  [37]  41  42  43  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59
#>  [55]  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  76  77  79
#>  [73]  80  81  82  83  84  85  86  87  89  90  91  92  93  94  95  96  97  98
#>  [91]  99 100 101 102 103 104 105 106 107 108 110 111 112 113 114 115 116 117
#> [109] 118 119 120 121 122 124 125 126 127 128 129 130 131 132 133 134 135 136
#> [127] 138 139 140 141 144 145 146 148 149 150 151 152 153 154 155 156 157 158
#> [145] 159 160 161 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177
#> [163] 178 179 180 181 182 183 184 185 186 187 189 190 191 192 193 194 195 196
#> [181] 197 198 199 201 202 203 205 206 207 208 209 210 211 212 213 214 216 217
#> [199] 218 219 220 221 222 224 225 226 227 228 229 230 231 232 233 234 235 236
#> [217] 237 238 239 241 243 244 245 247 248 249 250 251 252 253 254 255 256 257
#> [235] 258 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276
#> [253] 277 278 280 281 282 283 284 286 287 289 290 291 292 293 294 295 296 297
#> [271] 298 299 300 301 302 303 304 305 307 308 309 310 311 312 314 315 316 317
#> [289] 319 321 322 323 324 325 326 327 328 329 330 332 333 334 335 336 339 340
#> [307] 341 342 343 344 345 346 347 349 350 351 352 353 354 355 356 357 358 359
#> [325] 360 361 362 363 364 366 367 368 369 370 372 373 374 375 376 377 379 380
#> [343] 382 383 385 386 387 388 389 390 391 392 393 394 395 396 397 398 399 400
#> [361] 401 402 403 404 405 406 407 408 410 411 412 413 414 415 416 417 419 420
#> [379] 421 422 423 424 425 426 427 428 429 432 433 434 435 436 437 438 439 440
#> [397] 441 442 443 444 445 447 449 450 451 452 453 454 455 456 457 458 459 460
#> [415] 461 462 463 464 465 466 467 468 469 470 471 473 474 475 476 478 479 480
#> [433] 481 482 483 484 485 486 487 488 489 490 491 492 493 495 497 498 499 500
#> [451] 501 502 503 504 506 507 508 509 510 511 513 514 515 516 517 518 519 520
#> [469] 521 522 523 524 525 526 527 528 529 530 531 534 536 537 539 540 541 542
#> [487] 543 544 545 546 547 548 549 550 551 552 553 555 556 557 558 560 561 562
#> [505] 563 564 565 568 569 571 572 573 574 575 576 577 578 579 581 582 583 584
#> [523] 585 587 588 589 590 592 593 594 595 596 597 598 599 600 601 602 603 604
#> [541] 605 606 608 609 610 612 613 614 615 616 618 619 620 622 623 624 625 626
#> [559] 627 628 629 630 631 632 633 634 635 637 639 640 641 642 643 644 645 646
#> [577] 647 648 649 650 651 652 653 654 655 656 657 659 660 661 662 663 664 665
#> [595] 666 668 669 670 671 672 673 674 675 676 677 678 679 681 682 683 685 686
#> [613] 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704
#> [631] 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721 722 724
#> [649] 725 727 728 729 730 731 732 733 735 736 737 738 739 741 742 743 744 745
#> [667] 746 747 748 749 751
#> 
#> $train$`10`
#>   [1]   1   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  19  20
#>  [19]  21  22  23  26  27  29  30  31  33  34  35  36  37  38  39  40  41  42
#>  [37]  43  44  45  46  47  48  49  51  52  53  54  55  56  57  58  59  60  61
#>  [55]  62  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80
#>  [73]  81  82  83  84  85  86  87  88  90  91  92  93  94  95  96  97  98  99
#>  [91] 100 101 102 103 105 106 107 108 109 110 111 112 113 114 115 116 117 119
#> [109] 121 122 123 124 125 126 128 130 132 133 134 136 137 139 140 141 142 143
#> [127] 144 145 146 147 148 149 150 151 152 154 155 156 157 158 159 160 162 163
#> [145] 164 165 166 167 168 169 170 171 173 174 175 176 177 178 179 180 181 182
#> [163] 183 184 185 186 187 188 189 190 191 193 194 195 196 197 198 199 200 201
#> [181] 202 203 204 205 206 207 208 209 210 211 212 213 215 216 217 218 219 220
#> [199] 222 223 225 226 228 229 230 231 232 233 234 235 236 237 238 239 240 241
#> [217] 242 243 244 245 246 247 248 249 250 252 253 254 255 256 257 258 259 260
#> [235] 261 262 263 264 265 267 269 270 272 273 274 275 276 277 278 279 280 282
#> [253] 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299 300
#> [271] 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318
#> [289] 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336
#> [307] 337 338 339 340 341 342 343 344 345 347 348 349 350 351 352 353 354 355
#> [325] 356 357 359 360 361 362 363 364 365 366 367 368 369 371 372 374 375 376
#> [343] 378 379 380 381 382 383 384 385 386 387 389 390 391 392 393 394 395 396
#> [361] 397 398 399 400 401 402 403 404 405 407 408 409 410 413 414 415 416 417
#> [379] 418 419 420 421 424 425 426 427 428 429 430 431 432 434 435 436 437 438
#> [397] 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454 456 457
#> [415] 458 459 461 462 463 464 465 466 467 469 470 471 472 473 474 475 476 477
#> [433] 478 479 481 482 483 484 485 486 487 488 489 490 491 492 493 494 495 496
#> [451] 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514
#> [469] 515 516 517 519 520 521 522 524 525 526 527 528 529 530 532 533 534 535
#> [487] 536 537 538 539 540 542 544 545 546 547 548 549 550 551 552 553 554 556
#> [505] 557 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575
#> [523] 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591 592 593
#> [541] 594 595 596 597 599 600 601 602 603 604 605 606 607 608 609 610 611 612
#> [559] 613 614 615 616 617 618 619 620 621 622 623 624 625 626 627 628 629 630
#> [577] 631 632 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648
#> [595] 649 650 651 652 653 654 655 656 657 658 659 660 662 663 664 665 666 667
#> [613] 668 669 670 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685
#> [631] 686 687 688 689 690 691 692 694 695 696 697 698 699 700 701 702 703 704
#> [649] 705 706 707 708 709 710 711 713 714 715 716 717 718 719 720 721 722 723
#> [667] 724 725 726 727 728 729 731 732 733 734 735 736 738 739 740 741 742 743
#> [685] 744 745 746 747 748 749 750 751
#> 
#> $train$`11`
#>   [1]   1   2   4   6   7   8   9  10  11  12  13  14  15  16  17  18  20  21
#>  [19]  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  41  42  43  44  46  47  48  49  50  51  52  53  54  55  56  57  58
#>  [55]  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76
#>  [73]  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94
#>  [91]  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167
#> [163] 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185
#> [181] 186 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 204
#> [199] 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222
#> [217] 223 224 225 226 227 228 229 230 232 233 234 235 237 238 239 240 241 242
#> [235] 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 261
#> [253] 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279
#> [271] 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297
#> [289] 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315
#> [307] 316 317 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333
#> [325] 334 335 336 337 338 339 340 341 342 343 344 345 346 347 348 350 351 352
#> [343] 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370
#> [361] 371 372 373 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388
#> [379] 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403 404 405 406
#> [397] 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424
#> [415] 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439 440 441 443
#> [433] 445 446 447 448 449 450 451 452 454 455 456 457 458 459 460 461 462 463
#> [451] 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481 482
#> [469] 483 484 485 486 487 488 489 490 491 493 494 495 496 497 498 499 500 501
#> [487] 502 503 504 505 506 507 508 509 511 512 513 514 515 516 517 518 519 520
#> [505] 521 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 537 538
#> [523] 539 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556
#> [541] 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575
#> [559] 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591 592 593
#> [577] 594 595 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 612
#> [595] 613 614 615 616 617 618 619 620 621 622 623 624 625 626 627 629 630 631
#> [613] 632 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648 649
#> [631] 650 651 652 654 656 657 658 659 660 661 662 663 664 665 666 667 668 669
#> [649] 670 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685 686 687
#> [667] 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705
#> [685] 706 707 708 709 711 712 713 714 715 716 717 719 720 721 722 723 724 725
#> [703] 726 727 729 730 731 734 736 737 738 740 741 742 743 744 745 746 747 748
#> [721] 749 750 751
#> 
#> 
#> $test
#> $test$`1`
#>  [1]   1   6  33  36  74  81  84 100 111 112 117 119 159 160 164 165 167 168 175
#> [20] 178 184 187 194 208 210 244 248 256 257 292 323 324 351 355 363 364 374 382
#> [39] 390 391 393 407 436 454 479 481 497 506 524 545 549 584 593 597 601 616 624
#> [58] 625 640 672 685 687 695 717 742
#> 
#> $test$`2`
#>  [1]  11  16  23  34  51  56 107 121 146 148 149 152 157 174 193 198 212 232 249
#> [20] 252 258 264 267 273 286 294 295 300 302 307 310 312 314 316 329 330 332 342
#> [39] 347 361 369 375 379 387 399 408 410 417 420 421 425 432 439 445 447 459 466
#> [58] 469 473 482 485 491 507 515 517 572 577 588 590 604 609 626 629 632 637 645
#> [77] 647 650 660 677 678 686 699 721 731 748
#> 
#> $test$`3`
#>  [1]  27  41  53  67  72  82  98 114 116 124 141 154 186 199 233 235 275 296 298
#> [20] 299 301 359 367 392 401 441 462 467 489 508 522 526 530 547 576 579 602 603
#> [39] 697 719
#> 
#> $test$`4`
#>  [1]   9  20  22  30  31  38  40  43  52  60  76  80  83  92 101 134 173 176 177
#> [20] 179 189 197 205 226 228 237 239 254 276 284 291 293 297 304 311 315 333 335
#> [39] 339 340 341 343 352 360 380 386 396 397 434 437 443 450 457 471 484 495 500
#> [58] 502 513 521 536 551 556 568 587 595 600 606 608 614 615 622 634 648 652 659
#> [77] 666 675 694 700 709 713 722 724 745
#> 
#> $test$`5`
#>  [1]  42  47  55  59  62  87  93  95  96  99 139 144 156 230 241 250 269 274 283
#> [20] 319 321 376 383 419 424 440 501 504 514 516 534 539 540 548 552 553 560 569
#> [39] 571 578 627 630 635 657 662 665 669 674 676 688 689 692 698 702 704 716 729
#> [58] 741 744 749 751
#> 
#> $test$`6`
#>  [1]  12  14  21  46  54  65  66  69  79  86  91  94  97 105 108 113 128 130 132
#> [20] 140 145 163 171 181 182 191 195 196 202 209 211 213 225 229 253 262 270 277
#> [39] 282 305 308 322 326 327 328 334 336 353 366 368 389 395 415 416 452 456 458
#> [58] 465 475 478 493 498 511 519 527 542 563 565 581 583 599 618 619 631 646 651
#> [77] 663 668 670 671 681 691 701 703 714 720 747
#> 
#> $test$`7`
#>  [1]   4   8  15  26  48  58  64  70  71  73  85 102 126 133 150 151 158 166 169
#> [20] 170 180 185 206 207 216 217 218 220 234 238 245 247 255 261 265 272 278 280
#> [39] 287 290 303 309 344 354 356 357 362 394 398 402 426 428 429 435 438 451 461
#> [58] 470 476 488 499 503 509 520 525 529 544 561 562 574 582 585 589 592 605 610
#> [77] 612 623 633 644 656 679 683 706 707 708 725 727 738 746
#> 
#> $test$`8`
#>  [1]  10  29  37  39  49  57  61  68  77  90 103 106 110 115 122 136 155 183 190
#> [20] 201 219 222 243 263 289 317 325 345 350 372 385 400 403 404 405 413 414 427
#> [39] 449 463 474 483 486 487 490 528 537 546 550 564 573 575 594 613 620 639 641
#> [58] 642 643 649 654 664 673 682 690 696 711 715 736 743
#> 
#> $test$`9`
#>  [1]   7  13  17  35  44  75  78  88 109 123 137 142 143 147 162 188 200 204 215
#> [20] 223 240 242 246 259 279 285 288 306 313 318 320 331 337 338 348 365 371 378
#> [39] 381 384 409 418 430 431 446 448 472 477 494 496 505 512 532 533 535 538 554
#> [58] 559 566 567 570 580 586 591 607 611 617 621 636 638 658 667 680 684 705 723
#> [77] 726 734 740 750
#> 
#> $test$`10`
#>  [1]   2  18  24  25  28  32  50  63  89 104 118 120 127 129 131 135 138 153 161
#> [20] 172 192 214 221 224 227 251 266 268 271 281 346 358 370 373 377 388 406 411
#> [39] 412 422 423 433 455 460 468 480 518 523 531 541 543 555 558 598 661 693 712
#> [58] 730 737
#> 
#> $test$`11`
#>  [1]   3   5  19  45 125 203 231 236 260 349 442 444 453 464 492 510 557 596 628
#> [20] 653 655 710 718 728 732 733 735 739
#> 
#>