Create spatio-temporal Folds Using Predefined Groups

Implementation of CAST::CreateSpaceTimeFolds().

Details

Using "class" is helpful in the case that data are clustered in space and are categorical. E.g This is the case for land cover classifications when training data come as training polygons. In this case the data should be split in a way that entire polygons are held back (spacevar="polygonID") but at the same time the distribution of classes should be similar in each fold (class="LUC").

References

Meyer H, Reudenbach C, Hengl T, Katurji M, Nauss T (2018). “Improving performance of spatio-temporal machine learning models using forward feature selection and target-oriented validation.” Environmental Modelling \& Software, 101, 1--9. doi: 10.1016/j.envsoft.2017.12.001 , https://doi.org/10.1016%2Fj.envsoft.2017.12.001.

Super class

mlr3::Resampling -> ResamplingSptCVCstf

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSptCVCstf$new()

ResamplingSptCVCstf$instantiate()

ResamplingSptCVCstf$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spacetime Folds" resampling instance.

Usage

ResamplingSptCVCstf$new(id = "spcv-cstf")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSptCVCstf$instantiate(
  task,
  space_var = NULL,
  time_var = NULL,
  class = NULL
)

Arguments

task

Task
A task to instantiate.

space_var

[character]
Column name identifying the spatial units.

time_var

[character]
Column name identifying the temporal units.

class

[character]
Column name identifying a class unit (e.g. land cover).

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSptCVCstf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("cookfarm")

# Instantiate Resampling
rcv = rsmp("spcv-cstf", folds = 5)
rcv$instantiate(task, time_var = "Date", space_var = "SOURCEID")

# Individual sets:
rcv$train_set(1)

#>  [1]  17  23  35  40  41  50  62  67  78  91  92  96 101 108 113 116 120 122 123
#> [20] 126 128 131 139 142 151 154 157 170 178 181 193 204 205 212 213 218 230 233
#> [39] 243 254 259 260 262 263 266 273 282 287 290 296 311 317 331 332 339 345 355
#> [58] 358 359 370 371 374 375 378 393 396 406 412 413 417 419 423 441 446 458 461
#> [77] 465 475 483 488 490 500

rcv$test_set(1)

#>  [1]   4  14  22  24  38  42  86  89 165 180 220 245 280 285 307 342 349 362 380
#> [20] 422 429 459 473

# check that no obs are in both sets
intersect(rcv$train_set(1), rcv$test_set(1)) # good!

#> integer(0)

# Internal storage:
rcv$instance # table

#> $test
#> $test[[1]]
#>  [1]   4  14  22  24  38  42  86  89 165 180 220 245 280 285 307 342 349 362 380
#> [20] 422 429 459 473
#> 
#> $test[[2]]
#>  [1]   8  12  20  32  43  84 112 146 209 302 346 356 381 410 439 454 472 484
#> 
#> $test[[3]]
#>  [1]  44  80 118 186 197 225 240 281 286 291 316 318 329 338 403 424 460 476 486
#> [20] 495 496
#> 
#> $test[[4]]
#>  [1]   3  11  15  29  72  87 107 109 140 174 183 228 235 253 255 284 369 383 385
#> [20] 444 453 491
#> 
#> $test[[5]]
#>  [1]  21  25  36  53  85  90 103 105 150 182 221 238 257 298 328 330 352 353 357
#> [20] 384 411 415 418 430 487
#> 
#> 
#> $train
#> $train[[1]]
#>  [1]  17  23  35  40  41  50  62  67  78  91  92  96 101 108 113 116 120 122 123
#> [20] 126 128 131 139 142 151 154 157 170 178 181 193 204 205 212 213 218 230 233
#> [39] 243 254 259 260 262 263 266 273 282 287 290 296 311 317 331 332 339 345 355
#> [58] 358 359 370 371 374 375 378 393 396 406 412 413 417 419 423 441 446 458 461
#> [77] 465 475 483 488 490 500
#> 
#> $train[[2]]
#>  [1]   2   7  10  30  31  34  51  52  58  64  79  88  97  98 102 104 110 115 119
#> [20] 124 153 158 161 162 169 175 191 192 194 195 196 200 210 211 215 226 229 231
#> [39] 237 244 256 261 288 294 295 300 301 321 322 325 327 333 336 348 354 365 367
#> [58] 373 382 386 392 416 420 425 426 438 450 451 455 463 464 468 482 489 497 499
#> 
#> $train[[3]]
#>  [1]   5   9  16  19  26  27  37  46  47  48  49  68  71  74  77  82  93  99 111
#> [20] 114 121 127 132 135 143 147 148 149 152 155 164 167 168 176 190 202 216 217
#> [39] 239 241 246 252 264 267 270 274 276 278 279 297 306 310 313 320 323 335 341
#> [58] 344 351 376 379 387 394 398 405 408 409 421 437 442 447 448 456 467 478 479
#> [77] 492 494
#> 
#> $train[[4]]
#>  [1]   1  13  18  39  45  54  56  60  63  66  69  70  73  76 106 125 133 134 138
#> [20] 141 145 160 173 179 184 188 198 201 208 214 222 227 234 242 247 248 251 258
#> [39] 271 275 292 293 305 308 309 319 324 334 337 340 361 364 372 377 391 397 400
#> [58] 401 414 427 431 432 434 435 440 452 462 466 469 470 480 481 493 498
#> 
#> $train[[5]]
#>  [1]   6  28  33  55  57  59  61  65  75  81  83  94  95 100 117 129 130 136 137
#> [20] 144 156 159 163 166 171 172 177 185 187 189 199 203 206 207 219 223 224 232
#> [39] 236 249 250 265 268 269 272 277 283 289 299 303 304 312 314 315 326 343 347
#> [58] 350 360 363 366 368 388 389 390 395 399 402 404 407 428 433 436 443 445 449
#> [77] 457 471 474 477 485
#> 
#>