SpatioTemporal Classification Task

This task specializes Task and TaskSupervised for spatio-temporal classification problems. The target column is assumed to be a factor. The task_type is set to "classif" and "spatiotemporal".

A spatial example task is available via mlr_tasks$get("ecuador"). During initialization, coordinates need to be passed. By default, coordinates are not used as features. This can be changed by setting coords_as_features = TRUE.

See also

Super classes

mlr3::Task -> mlr3::TaskSupervised -> mlr3::TaskClassif -> TaskClassifST

Public fields

coordinate_names

character
The variables names of the coordinates in the data.

crs

character
The proj4string of the coordinate system.

Methods

Public methods

TaskClassifST$new()

TaskClassifST$coordinates()

TaskClassifST$print()

TaskClassifST$clone()

Inherited methods

mlr3::Task$cbind()

mlr3::Task$filter()

mlr3::Task$format()

mlr3::Task$formula()

mlr3::Task$head()

mlr3::Task$help()

mlr3::Task$levels()

mlr3::Task$missings()

mlr3::Task$rbind()

mlr3::Task$rename()

mlr3::Task$select()

mlr3::Task$set_col_role()

mlr3::Task$set_row_role()

mlr3::TaskClassif$data()

mlr3::TaskClassif$droplevels()

mlr3::TaskClassif$truth()

Method new()

Create a new spatiotemporal resampling Task

Usage

TaskClassifST$new(
  id,
  backend,
  target,
  positive = NULL,
  coords_as_features = FALSE,
  crs = NA,
  coordinate_names = NA
)

Arguments

id

character(1)
Identifier for the task.

backend

DataBackend
Either a DataBackend, or any object which is convertible to a DataBackend with as_data_backend(). E.g., a data.frame() will be converted to a DataBackendDataTable.

target

character(1)
Name of the target column.

positive

character(1)
Only for binary classification: Name of the positive class. The levels of the target columns are reordered accordingly, so that the first element of $class_names is the positive class, and the second element is the negative class.

coords_as_features

logical(1)
Whether the coordinates should also be used as features. Default is FALSE.

crs

character(1)
Coordinates reference system

coordinate_names

character(2)
The variables names of the coordinates in the data.

Method coordinates()

Return the coordinates of the task

Usage

TaskClassifST$coordinates(rows = NULL)

Arguments

rows

Row IDs. Can be used to subset the returned coordinates.

Method print()

Print the task.

Usage

TaskClassifST$print(...)

Arguments

...

Arguments passed to the $print() method of the superclass.

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskClassifST$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

data = data.table::as.data.table(readRDS(system.file("extdata",
  "ecuador.rda",
  package = "mlr3spatiotempcv")))
task = TaskClassifST$new("ecuador",
  backend = data, target = "slides",
  positive = "TRUE", coordinate_names = c("x", "y"))

task$task_type

#> [1] "classif"

task$formula()

#> slides ~ .
#> NULL

task$class_names

#> [1] "TRUE"  "FALSE"

task$positive

#> [1] "TRUE"

task$negative

#> [1] "FALSE"

task$coordinates()

#>             x       y
#>   1: 712882.5 9560002
#>   2: 715232.5 9559582
#>   3: 715392.5 9560172
#>   4: 715042.5 9559312
#>   5: 715382.5 9560142
#>  ---                 
#> 747: 714472.5 9558482
#> 748: 713142.5 9560992
#> 749: 713322.5 9560562
#> 750: 715392.5 9557932
#> 751: 713802.5 9560862

task$coordinate_names

#> [1] "x" "y"