(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

mlr3::Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  91  92  93  94
#>  [91]  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 111 112 113
#> [109] 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167
#> [163] 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 184 185 186
#> [181] 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204
#> [199] 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222
#> [217] 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240
#> [235] 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258
#> [253] 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276
#> [271] 277 278 279 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294
#> [289] 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312
#> [307] 313 314 315 316 317 318 319 320 321 322 323 324 326 327 328 329 330 331
#> [325] 332 333 334 335 336 337 338 339 340 341 342 343 344 346 347 348 349 351
#> [343] 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369
#> [361] 370 371 372 373 374 375 376 377 378 379 380 381 382 383 384 385 386 387
#> [379] 388 389 390 391 392 393 394 395 396 397 398 399 401 402 405 406 407 408
#> [397] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426
#> [415] 427 428 429 430 431 432 433 434 435 436 437 438 439 440 441 442 443 444
#> [433] 445 446 447 448 450 451 452 453 454 455 456 457 458 459 460 461 462 464
#> [451] 465 466 467 468 469 470 471 472 473 475 476 477 478 479 480 481 482 484
#> [469] 485 486 487 488 489 491 492 493 494 495 496 497 498 499 500 501 502 503
#> [487] 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [505] 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 538 539 540
#> [523] 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557 558
#> [541] 559 560 561 562 563 565 566 567 568 569 570 571 572 573 574 576 577 578
#> [559] 579 580 581 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596
#> [577] 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 612 614 615
#> [595] 616 617 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632 633
#> [613] 634 635 636 637 638 639 640 641 644 645 646 647 648 650 651 652 653 654
#> [631] 655 656 657 658 659 660 661 662 663 665 666 667 668 669 670 671 672 673
#> [649] 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691
#> [667] 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709
#> [685] 710 711 712 713 714 716 717 718 719 720 721 722 723 724 725 726 727 728
#> [703] 729 730 731 732 733 734 735 736 737 738 739 740 741 742 744 745 746 747
#> [721] 748 749 750 751
rrcv$test_set(1)
#>  [1]  29  37  49 183 325 350 400 403 404 449 575 643 664 715
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>  [1]  29  37  49 183 325 350 400 403 404 449 575 643 664 715
#> 
#> [[1]]$test[[2]]
#>  [1]  10  39  57  61  68  77 106 115 190 201 243 263 289 317 372 405 414 427 487
#> [20] 736
#> 
#> [[1]]$test[[3]]
#>  [1]  25  63 161 221 224 251 266 377 406 411 468
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  91  92  93  94
#>  [91]  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 111 112 113
#> [109] 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167
#> [163] 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 184 185 186
#> [181] 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204
#> [199] 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222
#> [217] 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240
#> [235] 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258
#> [253] 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276
#> [271] 277 278 279 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294
#> [289] 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312
#> [307] 313 314 315 316 317 318 319 320 321 322 323 324 326 327 328 329 330 331
#> [325] 332 333 334 335 336 337 338 339 340 341 342 343 344 346 347 348 349 351
#> [343] 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369
#> [361] 370 371 372 373 374 375 376 377 378 379 380 381 382 383 384 385 386 387
#> [379] 388 389 390 391 392 393 394 395 396 397 398 399 401 402 405 406 407 408
#> [397] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426
#> [415] 427 428 429 430 431 432 433 434 435 436 437 438 439 440 441 442 443 444
#> [433] 445 446 447 448 450 451 452 453 454 455 456 457 458 459 460 461 462 464
#> [451] 465 466 467 468 469 470 471 472 473 475 476 477 478 479 480 481 482 484
#> [469] 485 486 487 488 489 491 492 493 494 495 496 497 498 499 500 501 502 503
#> [487] 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [505] 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 538 539 540
#> [523] 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557 558
#> [541] 559 560 561 562 563 565 566 567 568 569 570 571 572 573 574 576 577 578
#> [559] 579 580 581 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596
#> [577] 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 612 614 615
#> [595] 616 617 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632 633
#> [613] 634 635 636 637 638 639 640 641 644 645 646 647 648 650 651 652 653 654
#> [631] 655 656 657 658 659 660 661 662 663 665 666 667 668 669 670 671 672 673
#> [649] 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691
#> [667] 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709
#> [685] 710 711 712 713 714 716 717 718 719 720 721 722 723 724 725 726 727 728
#> [703] 729 730 731 732 733 734 735 736 737 738 739 740 741 742 744 745 746 747
#> [721] 748 749 750 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  40  41  42  43  44  46  47  48  49  50  51  52  53  54  55  56  58
#>  [55]  59  60  62  63  64  65  66  67  69  70  71  72  73  74  75  76  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 100 101 102 103 104 105 107 108 109 110 111 112 113 114 116 117
#> [109] 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135
#> [127] 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153
#> [145] 154 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172
#> [163] 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 191
#> [181] 192 193 194 195 196 197 198 199 200 202 203 204 205 206 207 208 209 210
#> [199] 211 212 213 214 215 216 217 218 220 221 222 223 224 225 226 227 228 229
#> [217] 230 231 232 233 234 235 236 237 238 239 240 241 242 244 245 246 247 248
#> [235] 249 250 251 252 253 254 255 256 257 258 259 260 261 262 264 265 266 267
#> [253] 268 269 270 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285
#> [271] 286 287 288 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304
#> [289] 305 306 307 308 309 310 311 312 313 314 315 316 318 319 320 321 322 323
#> [307] 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341
#> [325] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [343] 360 361 362 363 364 365 366 367 368 369 370 371 373 374 375 376 377 378
#> [361] 379 380 381 382 383 384 386 387 388 389 390 391 392 393 394 395 396 397
#> [379] 398 399 400 401 402 403 404 406 407 408 409 410 411 412 413 415 416 417
#> [397] 418 419 420 421 422 423 424 425 426 428 429 430 431 432 433 434 435 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454
#> [433] 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472
#> [451] 473 474 475 476 477 478 479 480 481 482 483 484 485 486 488 489 490 491
#> [469] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509
#> [487] 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527
#> [505] 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545
#> [523] 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563
#> [541] 564 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581
#> [559] 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599
#> [577] 600 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617
#> [595] 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635
#> [613] 636 637 638 639 640 641 643 644 645 646 647 648 650 651 652 653 654 655
#> [631] 656 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 674
#> [649] 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692
#> [667] 693 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710
#> [685] 711 712 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728
#> [703] 729 730 731 732 733 734 735 737 738 739 740 741 742 743 744 745 746 747
#> [721] 748 749 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  26  27  28  29  30  31  33  34  35  36  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  59  60  61  62  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 156 157 158 159 160 162 163 164 165 166 167
#> [163] 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185
#> [181] 186 187 188 189 190 191 192 193 194 195 196 197 198 199 201 202 203 204
#> [199] 205 206 207 208 209 210 211 212 213 214 216 217 218 219 220 222 223 225
#> [217] 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243
#> [235] 244 245 247 248 249 250 252 253 254 255 256 257 258 259 260 261 262 263
#> [253] 264 265 267 268 269 270 271 272 273 274 275 276 277 278 280 281 282 283
#> [271] 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299 300 301
#> [289] 302 303 304 305 307 308 309 310 311 312 314 315 316 317 318 319 320 321
#> [307] 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339
#> [325] 340 341 342 343 344 345 346 347 349 350 351 352 353 354 355 356 357 358
#> [343] 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376
#> [361] 378 379 380 381 382 383 384 385 386 387 389 390 391 392 393 394 395 396
#> [379] 397 398 399 400 401 402 403 404 405 407 408 409 410 412 413 414 415 416
#> [397] 417 418 419 420 421 422 423 424 425 426 427 428 429 431 432 434 435 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454
#> [433] 455 456 457 458 459 461 462 463 464 465 466 467 469 470 471 472 473 474
#> [451] 475 476 477 478 479 481 482 483 484 485 486 487 488 489 490 491 492 493
#> [469] 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511
#> [487] 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529
#> [505] 530 531 532 533 534 535 536 537 538 539 540 542 544 545 546 547 548 549
#> [523] 550 551 552 553 554 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [541] 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 587
#> [559] 588 589 590 591 592 593 594 595 596 597 599 600 601 602 603 604 605 606
#> [577] 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624
#> [595] 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642
#> [613] 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660
#> [631] 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678
#> [649] 679 680 681 682 683 684 685 686 687 688 689 690 691 692 694 695 696 697
#> [667] 698 699 700 701 702 703 704 705 706 707 708 709 710 711 713 714 715 716
#> [685] 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734
#> [703] 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>  [1]  16 107 212 258 264 273 312 316 369 379 420 432 445 459 507 515 645 647 660
#> 
#> [[2]]$test[[2]]
#> [1]  10  39  68 289 404 427 642 649
#> 
#> [[2]]$test[[3]]
#>  [1]  10  39  57  61  68  77 106 115 155 190 201 219 243 263 289 317 372 385 405
#> [20] 414 427 487 642 673
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  15  17  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  33  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  52  53  54  55  56  57  58  59
#>  [55]  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77
#>  [73]  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 108 109 110 111 112 113 114
#> [109] 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 134 135 136 137 138 139 140 141 142 143 144 145 147 150 151 153 154
#> [145] 155 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172
#> [163] 173 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 190 191
#> [181] 192 194 195 196 197 198 199 200 201 202 203 204 205 206 207 208 209 210
#> [199] 211 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227 228 229
#> [217] 230 231 233 234 235 236 237 238 239 240 241 242 243 244 245 246 247 248
#> [235] 249 250 251 253 254 255 256 257 259 260 261 262 263 265 266 268 269 270
#> [253] 271 272 274 275 276 277 278 279 280 281 282 283 284 285 287 288 289 290
#> [271] 291 292 293 294 296 297 298 299 301 303 304 305 306 308 309 310 311 313
#> [289] 315 317 318 319 320 321 322 323 324 325 326 327 328 329 331 332 333 334
#> [307] 335 336 337 338 339 340 341 342 343 344 345 346 348 349 350 351 352 353
#> [325] 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 370 371 372
#> [343] 373 374 375 376 377 378 380 381 382 383 384 385 386 388 389 390 391 392
#> [361] 393 394 395 396 397 398 400 401 402 403 404 405 406 407 408 409 410 411
#> [379] 412 413 414 415 416 417 418 419 422 423 424 425 426 427 428 429 430 431
#> [397] 433 434 435 436 437 438 440 441 442 443 444 446 447 448 449 450 451 452
#> [415] 453 454 455 456 457 458 460 461 462 463 464 465 466 467 468 470 471 472
#> [433] 474 475 476 477 478 479 480 481 483 484 486 487 488 489 490 492 493 494
#> [451] 495 496 497 498 499 500 501 502 503 504 505 506 508 509 510 511 512 513
#> [469] 514 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532
#> [487] 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549 550
#> [505] 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [523] 569 570 571 572 573 574 575 576 578 579 580 581 582 583 584 585 586 587
#> [541] 588 589 591 592 593 594 595 596 597 598 599 600 601 602 603 605 606 607
#> [559] 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 625
#> [577] 627 628 629 630 631 633 634 635 636 638 639 640 641 642 643 644 646 648
#> [595] 649 650 651 652 653 654 655 656 657 658 659 661 662 663 664 665 666 667
#> [613] 668 669 670 671 672 673 674 675 676 677 679 680 681 682 683 684 685 686
#> [631] 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704
#> [649] 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721 722
#> [667] 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740
#> [685] 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  58  59
#>  [55]  60  62  63  64  65  66  67  69  70  71  72  73  74  75  76  78  79  80
#>  [73]  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98
#>  [91]  99 100 101 102 103 104 105 107 108 109 110 111 112 113 114 116 117 118
#> [109] 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136
#> [127] 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153 154
#> [145] 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172 173
#> [163] 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 191 192
#> [181] 193 194 195 196 197 198 199 200 202 203 204 205 206 207 208 209 210 211
#> [199] 212 213 214 215 216 217 218 220 221 222 223 224 225 226 227 228 229 230
#> [217] 231 232 233 234 235 236 237 238 239 240 241 242 244 245 246 247 248 249
#> [235] 250 251 252 253 254 255 256 257 258 259 260 261 262 264 265 266 267 268
#> [253] 269 270 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285 286
#> [271] 287 288 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305
#> [289] 306 307 308 309 310 311 312 313 314 315 316 318 319 320 321 322 323 324
#> [307] 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343
#> [325] 344 345 346 347 348 349 351 352 353 354 355 356 357 358 359 360 361 362
#> [343] 363 364 365 366 367 368 369 370 371 373 374 375 376 377 378 379 380 381
#> [361] 382 383 384 386 387 388 389 390 391 392 393 394 395 396 397 398 399 401
#> [379] 402 403 406 407 408 409 410 411 412 413 415 416 417 418 419 420 421 422
#> [397] 423 424 425 426 428 429 430 431 432 433 434 435 436 437 438 439 440 441
#> [415] 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 458 459
#> [433] 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475 476 477
#> [451] 478 479 480 481 482 483 484 485 486 488 489 490 491 492 493 494 495 496
#> [469] 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514
#> [487] 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532
#> [505] 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549 551
#> [523] 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568 569
#> [541] 570 571 572 573 574 576 577 578 579 580 581 582 583 584 585 586 587 588
#> [559] 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606
#> [577] 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624
#> [595] 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 644
#> [613] 645 646 647 648 650 651 652 653 654 655 656 657 658 659 660 661 662 663
#> [631] 665 666 667 668 669 670 671 672 674 675 676 677 678 679 680 681 682 683
#> [649] 684 685 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701
#> [667] 702 703 704 705 706 707 708 709 710 711 712 713 714 716 717 718 719 720
#> [685] 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738
#> [703] 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  40  41  42  43  44  46  47  48  49  50  51  52  53  54  55  56  58
#>  [55]  59  60  62  63  64  65  66  67  69  70  71  72  73  74  75  76  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 100 101 102 103 104 105 107 108 109 110 111 112 113 114 116 117
#> [109] 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135
#> [127] 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153
#> [145] 154 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172
#> [163] 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 191
#> [181] 192 193 194 195 196 197 198 199 200 202 203 204 205 206 207 208 209 210
#> [199] 211 212 213 214 215 216 217 218 220 221 222 223 224 225 226 227 228 229
#> [217] 230 231 232 233 234 235 236 237 238 239 240 241 242 244 245 246 247 248
#> [235] 249 250 251 252 253 254 255 256 257 258 259 260 261 262 264 265 266 267
#> [253] 268 269 270 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285
#> [271] 286 287 288 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304
#> [289] 305 306 307 308 309 310 311 312 313 314 315 316 318 319 320 321 322 323
#> [307] 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341
#> [325] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [343] 360 361 362 363 364 365 366 367 368 369 370 371 373 374 375 376 377 378
#> [361] 379 380 381 382 383 384 386 387 388 389 390 391 392 393 394 395 396 397
#> [379] 398 399 400 401 402 403 406 407 408 409 410 411 412 413 415 416 417 418
#> [397] 419 420 421 422 423 424 425 426 428 429 430 431 432 433 434 435 436 437
#> [415] 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455
#> [433] 456 457 458 459 460 461 462 463 465 466 467 468 469 470 471 472 473 474
#> [451] 475 476 477 478 479 480 481 482 483 484 485 486 488 489 490 491 492 493
#> [469] 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511
#> [487] 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529
#> [505] 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547
#> [523] 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565
#> [541] 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [559] 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601
#> [577] 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [595] 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637
#> [613] 638 639 640 641 643 644 645 646 647 648 650 651 652 653 654 655 656 657
#> [631] 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 674 675 676
#> [649] 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694
#> [667] 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712
#> [685] 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730
#> [703] 731 732 733 734 735 737 738 739 740 741 742 743 744 745 746 747 748 749
#> [721] 750 751
#> 
#> 
#>