(sperrorest) Coordinate-based k-means clustering

Splits data by clustering in the coordinate space. See the upstream implementation at sperrorest::partition_kmeans() and Brenning (2012) for further information.

Details

Universal partitioning method that splits the data in the coordinate space. Useful for spatially homogeneous datasets that cannot be split well with rectangular approaches like ResamplingSpCVBlock.

Parameters

folds (integer(1))
Number of folds.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingSpCVCoords

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSpCVCoords$new()

ResamplingSpCVCoords$instantiate()

ResamplingSpCVCoords$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "coordinate-based" repeated resampling instance.

For a list of available arguments, please see sperrorest::partition_cv.

Usage

ResamplingSpCVCoords$new(id = "spcv_coords")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSpCVCoords$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSpCVCoords$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv_coords", folds = 5)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)
#>   [1]   1   6   9  22  23  30  33  34  36  51  52  56  60  74  76  80  81  83
#>  [19]  84  92 100 101 111 112 117 119 134 146 148 149 159 160 164 165 167 168
#>  [37] 173 174 175 176 177 178 179 184 187 189 193 194 197 205 208 210 228 232
#>  [55] 237 239 244 248 252 254 256 257 276 286 291 292 293 294 295 297 302 304
#>  [73] 307 310 311 314 315 323 324 330 333 335 339 340 341 342 347 351 352 355
#>  [91] 363 364 374 376 380 382 386 390 391 393 396 397 407 425 434 436 437 439
#> [109] 443 447 450 454 457 469 471 479 481 485 491 497 506 517 524 536 545 549
#> [127] 551 556 568 572 577 584 593 597 601 606 609 614 616 622 624 625 626 629
#> [145] 632 640 645 648 666 672 677 678 685 687 695 699 700 717 722 742 745   2
#> [163]   3   4   5  10  18  19  26  28  29  37  39  45  48  49  50  57  58  61
#> [181]  64  68  77  89  90 103 106 110 115 120 122 125 126 127 129 131 133 135
#> [199] 136 138 150 151 153 155 169 170 172 183 190 201 203 206 216 218 219 220
#> [217] 222 231 234 236 238 243 245 255 260 263 265 268 271 272 281 289 290 303
#> [235] 309 317 325 344 345 346 349 350 356 357 358 362 370 372 373 385 394 400
#> [253] 403 404 405 412 413 414 422 423 426 427 429 442 444 449 453 455 463 464
#> [271] 474 483 486 487 490 492 502 503 509 510 518 523 525 528 529 531 537 544
#> [289] 546 557 564 573 575 585 594 596 612 613 620 623 628 634 639 641 642 643
#> [307] 644 649 653 655 661 664 673 690 693 706 707 708 710 712 715 718 724 727
#> [325] 728 732 733 735 736 737 739 743   7  12  13  14  17  21  25  32  35  44
#> [343]  46  54  63  65  66  75  78  79  86  88  91  94  97 104 105 109 113 118
#> [361] 123 128 130 132 137 140 142 143 145 147 161 162 163 171 181 182 188 191
#> [379] 192 196 200 202 204 209 211 213 214 215 221 223 224 225 229 240 242 246
#> [397] 251 253 259 262 266 270 277 279 285 288 305 306 308 313 318 320 322 327
#> [415] 331 336 337 338 348 365 366 368 371 377 378 381 384 388 406 409 411 416
#> [433] 418 430 431 433 446 448 452 458 460 465 468 472 475 477 478 480 493 494
#> [451] 496 498 505 512 527 532 533 535 538 541 542 543 554 555 558 559 561 566
#> [469] 567 570 580 581 586 591 598 599 607 611 617 619 621 636 638 646 658 663
#> [487] 667 680 684 701 703 705 714 720 723 726 730 734 740 750   8  15  20  24
#> [505]  27  31  38  40  41  43  53  67  69  70  71  72  73  82  85  98 102 108
#> [523] 114 116 124 141 154 158 166 180 185 186 195 199 207 217 226 227 233 235
#> [541] 247 261 275 278 280 282 284 287 296 298 299 301 326 328 334 343 353 354
#> [559] 359 360 367 389 392 395 398 401 402 415 428 435 438 441 451 456 461 462
#> [577] 467 470 476 484 488 489 495 499 500 508 511 513 519 520 521 522 526 530
#> [595] 547 562 563 565 574 576 579 582 583 587 589 592 595 600 602 603 605 608
#> [613] 610 615 618 631 633 651 656 659 668 670 671 675 679 681 683 691 694 697
#> [631] 709 713 719 725 738 746 747
rcv$test_set(1)
#>   [1]  11  16  42  47  55  59  62  87  93  95  96  99 107 121 139 144 152 156
#>  [19] 157 198 212 230 241 249 250 258 264 267 269 273 274 283 300 312 316 319
#>  [37] 321 329 332 361 369 375 379 383 387 399 408 410 417 419 420 421 424 432
#>  [55] 440 445 459 466 473 482 501 504 507 514 515 516 534 539 540 548 550 552
#>  [73] 553 560 569 571 578 588 590 604 627 630 635 637 647 650 652 654 657 660
#>  [91] 662 665 669 674 676 682 686 688 689 692 696 698 702 704 711 716 721 729
#> [109] 731 741 744 748 749 751
# check that no obs are in both sets
intersect(rcv$train_set(1), rcv$test_set(1)) # good!
#> integer(0)

# Internal storage:
rcv$instance # table
#>      row_id fold
#>   1:     11    1
#>   2:     16    1
#>   3:     42    1
#>   4:     47    1
#>   5:     55    1
#>  ---            
#> 747:    719    5
#> 748:    725    5
#> 749:    738    5
#> 750:    746    5
#> 751:    747    5