Repeated Spatial Cross Validation Resampling

Spatial Block Cross validation implemented by the blockCV package.

Details

By default blockCV::spatialBlock() does not allow the creation of multiple repetitions. mlr3spatiotempcv adds support for this when using the range argument for fold creation. When supplying a vector of length(repeats) for argument range, these different settings will be used to create folds which differ among the repetitions.

Multiple repetitions are not possible when using the "row & cols" approach because the created folds will always be the same.

References

Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G (2018). “blockCV: an R package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models.” bioRxiv. doi: 10.1101/357798 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVBlock

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVBlock$new()

ResamplingRepeatedSpCVBlock$folds()

ResamplingRepeatedSpCVBlock$repeats()

ResamplingRepeatedSpCVBlock$instantiate()

ResamplingRepeatedSpCVBlock$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "coordinate-based" repeated resampling instance.

Usage

ResamplingRepeatedSpCVBlock$new(id = "repeated_spcv_block")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVBlock$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVBlock$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVBlock$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVBlock$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("diplodia")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_block",
  folds = 3, repeats = 2,
  range = c(5000, 10000))
rrcv$instantiate(task)

# Individual sets:
rrcv$iters

#> [1] 6

rrcv$folds(1:6)

#> [1] 1 2 1 2 1 2

rrcv$repeats(1:6)

#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)

#>   [1]   6   7  10  13  15  16  18  25  26  28  30  31  32  36  37  38  39  40
#>  [19]  48  53  63  65  72  75  76  80  90  93 103 104 105 108 115 124 127 128
#>  [37] 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146
#>  [55] 147 150 156 157 165 166 171 172 173 174 176 177 178 179 190 191 213 214
#>  [73] 215 221 224 232 233 234 236 238 242 247 248 249 252 255 256 257 258 260
#>  [91] 261 266 272 273 276 277 278 285 291 292 300 301 302 303 305 306 307 308
#> [109] 309 310 313 315 316 317 318 322 324 325 330 331 332 333 334 335 336 341
#> [127] 342 344 346 349 350 351 352 358 370 414 415 416 420 421 424 431 433 434
#> [145] 436 438 444 447 450 453 455 457 458 462 464 465 467 468 470 473 474 475
#> [163] 476 481 488 489 490 498 499 500 501 502 503 507 524 546 547 548 549 550
#> [181] 564 565 566 582 583 584 586 587 588 601 602 604 605 616 617 618 619 621
#> [199] 623 624 637 638 650 651 653 658 659 661 662 663 664 672 676 677 679 688
#> [217] 691 692 693 697 700 701 703 704 705 706 707 708 709 710 711 712 713 714
#> [235] 715 716 717 718 726 747 748 749 750 756 757 760 761 766 767 768 769 770
#> [253] 775 778 784 789 796 799 802 804 808 812 817 818 820 821 823 826 828 832
#> [271] 834 836 839 841 842 844 851 852 853 855 859 862 865 867 871 873 874 875
#> [289] 878 880 887 888 899 901 902 906 907 909 913 916 917 920 921 922   1   2
#> [307]   3   4   5  11  12  14  17  19  20  23  24  29  42  43  44  45  47  49
#> [325]  50  51  57  60  67  68  69  70  78  81  82  83  88  91  92  94  95  96
#> [343]  98  99 101 110 111 113 114 116 117 118 119 120 121 122 125 126 151 152
#> [361] 153 154 155 158 159 162 168 169 170 175 184 185 186 187 188 192 193 194
#> [379] 195 197 198 202 206 209 211 212 219 220 222 223 225 235 237 239 240 241
#> [397] 246 265 267 269 271 286 288 289 290 294 295 296 297 299 304 314 321 326
#> [415] 337 338 340 345 347 348 354 355 357 408 409 410 411 412 413 419 422 423
#> [433] 427 428 429 430 435 440 446 449 451 452 456 459 460 461 463 469 471 477
#> [451] 478 479 480 483 484 491 494 495 504 510 513 514 515 517 518 519 521 522
#> [469] 523 539 540 541 542 543 544 545 551 552 553 554 557 558 559 560 561 562
#> [487] 563 575 576 577 578 579 580 581 589 598 599 606 607 612 614 615 620 625
#> [505] 626 627 628 629 630 631 632 633 634 635 636 640 641 642 643 644 645 647
#> [523] 648 649 652 654 655 656 660 665 666 669 671 673 678 682 683 686 689 696
#> [541] 699 702 719 722 724 725 727 728 729 730 731 732 737 738 741 744 745 746
#> [559] 751 752 762 771 774 776 777 781 782 783 793 795 797 798 800 801 803 805
#> [577] 809 813 814 816 822 824 825 827 833 835 837 838 840 843 847 848 856 857
#> [595] 858 863 876 879 881 885 886 889 890 891 893 894 895 897 898 900 903 904
#> [613] 918 919

rrcv$test_set(1)

#>   [1]   8   9  21  22  27  33  34  35  41  46  52  54  55  56  58  59  61  62
#>  [19]  64  66  71  73  74  77  79  84  85  86  87  89  97 100 102 106 107 109
#>  [37] 112 123 148 149 160 161 163 164 167 180 181 182 183 189 196 199 200 201
#>  [55] 203 204 205 207 208 210 216 217 218 226 227 228 229 230 231 243 244 245
#>  [73] 250 251 253 254 259 262 263 264 268 270 274 275 279 280 281 282 283 284
#>  [91] 287 293 298 311 312 319 320 323 327 328 329 339 343 353 356 359 360 361
#> [109] 362 363 364 365 366 367 368 369 371 372 373 374 375 376 377 378 379 380
#> [127] 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398
#> [145] 399 400 401 402 403 404 405 406 407 417 418 425 426 432 437 439 441 442
#> [163] 443 445 448 454 466 472 482 485 486 487 492 493 496 497 505 506 508 509
#> [181] 511 512 516 520 525 526 527 528 529 530 531 532 533 534 535 536 537 538
#> [199] 555 556 567 568 569 570 571 572 573 574 585 590 591 592 593 594 595 596
#> [217] 597 600 603 608 609 610 611 613 622 639 646 657 667 668 670 674 675 680
#> [235] 681 684 685 687 690 694 695 698 720 721 723 733 734 735 736 739 740 742
#> [253] 743 753 754 755 758 759 763 764 765 772 773 779 780 785 786 787 788 790
#> [271] 791 792 794 806 807 810 811 815 819 829 830 831 845 846 849 850 854 860
#> [289] 861 864 866 868 869 870 872 877 882 883 884 892 896 905 908 910 911 912
#> [307] 914 915

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance # table

#>       row_id rep fold
#>    1:      1   1    3
#>    2:      2   1    3
#>    3:      3   1    3
#>    4:      4   1    3
#>    5:      5   1    3
#>   ---                
#> 1840:    918   2    1
#> 1841:    919   2    2
#> 1842:    920   2    1
#> 1843:    921   2    1
#> 1844:    922   2    1