Repeated "Leave-location-and-time-out" resampling.

Implementation of CAST::CreateSpaceTimeFolds().

Details

Using "class" is helpful in the case that data are clustered in space and are categorical. E.g This is the case for land cover classifications when training data come as training polygons. In this case the data should be split in a way that entire polygons are held back (spacevar="polygonID") but at the same time the distribution of classes should be similar in each fold (class="LUC").

References

Zhao Y, Karypis G (2002). “Evaluation of Hierarchical Clustering Algorithms for Document Datasets.” 11th Conference of Information and Knowledge Management (CIKM), 51-524. http://glaros.dtc.umn.edu/gkhome/node/167.

Super class

mlr3::Resampling -> ResamplingRepeatedSptCVCstf

Public fields

space_var

character(1)
Column name identifying the spatial units.

time_var

character(1)
Column name identifying the temporal units.

class

character(1)
Column name identifying a class unit (e.g. land cover).

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSptCVCstf$new()

ResamplingRepeatedSptCVCstf$folds()

ResamplingRepeatedSptCVCstf$repeats()

ResamplingRepeatedSptCVCstf$instantiate()

ResamplingRepeatedSptCVCstf$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spacetime Folds" resampling instance.

Usage

ResamplingRepeatedSptCVCstf$new(
  id = "repeated_sptcv_cstf",
  space_var = NULL,
  time_var = NULL,
  class = NULL
)

Arguments

id

character(1)
Identifier for the resampling strategy.

space_var

character(1)
Column name identifying the spatial units.

time_var

character(1)
Column name identifying the temporal units.

class

character(1)
Column name identifying a class unit (e.g. land cover).

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSptCVCstf$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSptCVCstf$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSptCVCstf$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSptCVCstf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(mlr3spatiotempcv)
task = tsk("cookfarm")

# Instantiate Resampling
rrcv = rsmp("repeated_sptcv_cstf", folds = 3, repeats = 5, time_var = "Date")
rrcv$instantiate(task)
# Individual sets:
rrcv$iters

#> [1] 15

rrcv$folds(1:6)

#> [1] 1 2 3 4 5 1

rrcv$repeats(1:6)

#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)

#>   [1]   2   4   5   6   7   8  10  12  13  14  15  16  17  18  20  22  23  24
#>  [19]  27  28  29  30  31  32  33  37  38  40  42  43  44  45  46  47  48  49
#>  [37]  50  52  53  54  56  57  58  60  61  63  64  65  66  68  69  70  74  78
#>  [55]  81  82  83  84  86  87  88  89  90  93  94  97  99 100 104 105 106 107
#>  [73] 108 109 112 113 115 116 117 118 119 120 121 123 126 128 129 130 131 132
#>  [91] 133 136 137 139 141 142 143 146 149 150 152 153 154 155 157 158 161 163
#> [109] 165 166 168 169 170 171 172 173 174 176 177 178 179 180 182 183 184 188
#> [127] 191 192 193 194 195 197 199 200 201 202 203 204 205 206 208 210 214 215
#> [145] 216 217 219 222 223 224 225 226 227 228 229 230 231 232 233 234 235 237
#> [163] 239 240 242 243 244 245 247 248 249 250 252 253 256 258 259 260 262 265
#> [181] 267 268 269 270 271 273 274 275 276 277 278 280 281 282 287 288 290 293
#> [199] 295 296 297 298 299 301 304 306 308 313 315 316 317 318 320 321 322 323
#> [217] 325 326 327 329 330 331 332 335 337 338 339 340 342 343 344 347 348 349
#> [235] 351 352 353 354 356 358 359 360 361 362 363 365 367 368 369 370 371 372
#> [253] 375 377 378 380 383 384 385 386 387 388 389 390 391 393 394 396 398 399
#> [271] 402 403 404 406 409 410 411 414 415 416 417 419 420 421 422 424 425 426
#> [289] 427 429 430 431 432 433 434 435 436 437 438 439 440 442 443 444 445 446
#> [307] 447 448 449 450 453 454 455 456 457 459 460 462 464 465 466 467 470 471
#> [325] 474 475 476 477 478 479 481 484 487 488 490 491 492 493 496 498 499

rrcv$test_set(1)

#>   [1]   1   3   9  11  19  21  25  26  34  35  36  39  41  51  55  59  62  67
#>  [19]  71  72  73  75  76  77  79  80  85  91  92  95  96  98 101 102 103 110
#>  [37] 111 114 122 124 125 127 134 135 138 140 144 145 147 148 151 156 159 160
#>  [55] 162 164 167 175 181 185 186 187 189 190 196 198 207 209 211 212 213 218
#>  [73] 220 221 236 238 241 246 251 254 255 257 261 263 264 266 272 279 283 284
#>  [91] 285 286 289 291 292 294 300 302 303 305 307 309 310 311 312 314 319 324
#> [109] 328 333 334 336 341 345 346 350 355 357 364 366 373 374 376 379 381 382
#> [127] 392 395 397 400 401 405 407 408 412 413 418 423 428 441 451 452 458 461
#> [145] 463 468 469 472 473 480 482 483 485 486 489 494 495 497 500

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance # table

#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>   [1]   1   3   9  11  19  21  25  26  34  35  36  39  41  51  55  59  62  67
#>  [19]  71  72  73  75  76  77  79  80  85  91  92  95  96  98 101 102 103 110
#>  [37] 111 114 122 124 125 127 134 135 138 140 144 145 147 148 151 156 159 160
#>  [55] 162 164 167 175 181 185 186 187 189 190 196 198 207 209 211 212 213 218
#>  [73] 220 221 236 238 241 246 251 254 255 257 261 263 264 266 272 279 283 284
#>  [91] 285 286 289 291 292 294 300 302 303 305 307 309 310 311 312 314 319 324
#> [109] 328 333 334 336 341 345 346 350 355 357 364 366 373 374 376 379 381 382
#> [127] 392 395 397 400 401 405 407 408 412 413 418 423 428 441 451 452 458 461
#> [145] 463 468 469 472 473 480 482 483 485 486 489 494 495 497 500
#> 
#> [[1]]$test[[2]]
#>   [1]   5  10  14  15  16  18  22  24  29  31  37  38  40  43  44  46  48  49
#>  [19]  50  54  58  60  61  63  66  78  81  82  84  97  99 100 104 106 107 108
#>  [37] 116 117 119 120 123 126 128 131 142 150 154 155 157 163 166 168 169 170
#>  [55] 176 177 179 184 192 193 197 201 202 203 204 205 210 217 219 223 225 226
#>  [73] 228 230 231 233 234 235 240 242 244 245 247 248 249 250 253 259 260 273
#>  [91] 276 277 278 288 290 293 295 296 298 306 308 313 315 317 327 329 332 337
#> [109] 338 339 340 342 343 344 348 349 352 356 358 361 362 367 368 370 371 380
#> [127] 394 396 402 404 406 409 414 416 419 420 424 426 427 430 432 433 434 435
#> [145] 438 440 442 445 448 449 456 457 459 460 462 466 470 477 479 484 488 490
#> [163] 493 496 498 499
#> 
#> [[1]]$test[[3]]
#>   [1]   2   4   6   7   8  12  13  17  20  23  27  28  30  32  33  42  45  47
#>  [19]  52  53  56  57  64  65  68  69  70  74  83  86  87  88  89  90  93  94
#>  [37] 105 109 112 113 115 118 121 129 130 132 133 136 137 139 141 143 146 149
#>  [55] 152 153 158 161 165 171 172 173 174 178 180 182 183 188 191 194 195 199
#>  [73] 200 206 208 214 215 216 222 224 227 229 232 237 239 243 252 256 258 262
#>  [91] 265 267 268 269 270 271 274 275 280 281 282 287 297 299 301 304 316 318
#> [109] 320 321 322 323 325 326 330 331 335 347 351 353 354 359 360 363 365 369
#> [127] 372 375 377 378 383 384 385 386 387 388 389 390 391 393 398 399 403 410
#> [145] 411 415 417 421 422 425 429 431 436 437 439 443 444 446 447 450 453 454
#> [163] 455 464 465 467 471 474 475 476 478 481 487 491 492
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   2   4   5   6   7   8  10  12  13  14  15  16  17  18  20  22  23  24
#>  [19]  27  28  29  30  31  32  33  37  38  40  42  43  44  45  46  47  48  49
#>  [37]  50  52  53  54  56  57  58  60  61  63  64  65  66  68  69  70  74  78
#>  [55]  81  82  83  84  86  87  88  89  90  93  94  97  99 100 104 105 106 107
#>  [73] 108 109 112 113 115 116 117 118 119 120 121 123 126 128 129 130 131 132
#>  [91] 133 136 137 139 141 142 143 146 149 150 152 153 154 155 157 158 161 163
#> [109] 165 166 168 169 170 171 172 173 174 176 177 178 179 180 182 183 184 188
#> [127] 191 192 193 194 195 197 199 200 201 202 203 204 205 206 208 210 214 215
#> [145] 216 217 219 222 223 224 225 226 227 228 229 230 231 232 233 234 235 237
#> [163] 239 240 242 243 244 245 247 248 249 250 252 253 256 258 259 260 262 265
#> [181] 267 268 269 270 271 273 274 275 276 277 278 280 281 282 287 288 290 293
#> [199] 295 296 297 298 299 301 304 306 308 313 315 316 317 318 320 321 322 323
#> [217] 325 326 327 329 330 331 332 335 337 338 339 340 342 343 344 347 348 349
#> [235] 351 352 353 354 356 358 359 360 361 362 363 365 367 368 369 370 371 372
#> [253] 375 377 378 380 383 384 385 386 387 388 389 390 391 393 394 396 398 399
#> [271] 402 403 404 406 409 410 411 414 415 416 417 419 420 421 422 424 425 426
#> [289] 427 429 430 431 432 433 434 435 436 437 438 439 440 442 443 444 445 446
#> [307] 447 448 449 450 453 454 455 456 457 459 460 462 464 465 466 467 470 471
#> [325] 474 475 476 477 478 479 481 484 487 488 490 491 492 493 496 498 499
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   6   7   8   9  11  12  13  17  19  20  21  23  25  26
#>  [19]  27  28  30  32  33  34  35  36  39  41  42  45  47  51  52  53  55  56
#>  [37]  57  59  62  64  65  67  68  69  70  71  72  73  74  75  76  77  79  80
#>  [55]  83  85  86  87  88  89  90  91  92  93  94  95  96  98 101 102 103 105
#>  [73] 109 110 111 112 113 114 115 118 121 122 124 125 127 129 130 132 133 134
#>  [91] 135 136 137 138 139 140 141 143 144 145 146 147 148 149 151 152 153 156
#> [109] 158 159 160 161 162 164 165 167 171 172 173 174 175 178 180 181 182 183
#> [127] 185 186 187 188 189 190 191 194 195 196 198 199 200 206 207 208 209 211
#> [145] 212 213 214 215 216 218 220 221 222 224 227 229 232 236 237 238 239 241
#> [163] 243 246 251 252 254 255 256 257 258 261 262 263 264 265 266 267 268 269
#> [181] 270 271 272 274 275 279 280 281 282 283 284 285 286 287 289 291 292 294
#> [199] 297 299 300 301 302 303 304 305 307 309 310 311 312 314 316 318 319 320
#> [217] 321 322 323 324 325 326 328 330 331 333 334 335 336 341 345 346 347 350
#> [235] 351 353 354 355 357 359 360 363 364 365 366 369 372 373 374 375 376 377
#> [253] 378 379 381 382 383 384 385 386 387 388 389 390 391 392 393 395 397 398
#> [271] 399 400 401 403 405 407 408 410 411 412 413 415 417 418 421 422 423 425
#> [289] 428 429 431 436 437 439 441 443 444 446 447 450 451 452 453 454 455 458
#> [307] 461 463 464 465 467 468 469 471 472 473 474 475 476 478 480 481 482 483
#> [325] 485 486 487 489 491 492 494 495 497 500
#> 
#> [[1]]$train[[3]]
#>   [1]   1   3   5   9  10  11  14  15  16  18  19  21  22  24  25  26  29  31
#>  [19]  34  35  36  37  38  39  40  41  43  44  46  48  49  50  51  54  55  58
#>  [37]  59  60  61  62  63  66  67  71  72  73  75  76  77  78  79  80  81  82
#>  [55]  84  85  91  92  95  96  97  98  99 100 101 102 103 104 106 107 108 110
#>  [73] 111 114 116 117 119 120 122 123 124 125 126 127 128 131 134 135 138 140
#>  [91] 142 144 145 147 148 150 151 154 155 156 157 159 160 162 163 164 166 167
#> [109] 168 169 170 175 176 177 179 181 184 185 186 187 189 190 192 193 196 197
#> [127] 198 201 202 203 204 205 207 209 210 211 212 213 217 218 219 220 221 223
#> [145] 225 226 228 230 231 233 234 235 236 238 240 241 242 244 245 246 247 248
#> [163] 249 250 251 253 254 255 257 259 260 261 263 264 266 272 273 276 277 278
#> [181] 279 283 284 285 286 288 289 290 291 292 293 294 295 296 298 300 302 303
#> [199] 305 306 307 308 309 310 311 312 313 314 315 317 319 324 327 328 329 332
#> [217] 333 334 336 337 338 339 340 341 342 343 344 345 346 348 349 350 352 355
#> [235] 356 357 358 361 362 364 366 367 368 370 371 373 374 376 379 380 381 382
#> [253] 392 394 395 396 397 400 401 402 404 405 406 407 408 409 412 413 414 416
#> [271] 418 419 420 423 424 426 427 428 430 432 433 434 435 438 440 441 442 445
#> [289] 448 449 451 452 456 457 458 459 460 461 462 463 466 468 469 470 472 473
#> [307] 477 479 480 482 483 484 485 486 488 489 490 493 494 495 496 497 498 499
#> [325] 500
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>   [1]   1   3  17  19  21  22  24  25  27  28  41  42  46  47  48  49  50  53
#>  [19]  54  56  58  61  62  66  70  74  76  77  79  84  86  88  93  99 101 102
#>  [37] 107 108 110 112 113 116 118 121 129 130 136 137 140 141 142 145 149 150
#>  [55] 153 161 167 168 169 171 174 175 178 179 185 188 189 190 193 199 204 207
#>  [73] 210 211 212 215 216 218 224 225 226 227 235 236 239 243 249 250 263 264
#>  [91] 265 266 267 269 274 275 277 283 285 289 292 295 298 300 301 305 306 315
#> [109] 320 321 322 325 327 328 329 330 337 339 340 348 349 357 361 362 363 367
#> [127] 368 369 370 373 374 376 377 381 385 392 394 397 402 403 405 406 410 411
#> [145] 413 414 415 423 424 430 431 432 435 436 437 439 440 441 442 445 451 454
#> [163] 459 460 461 462 470 474 477 479 483 489 499
#> 
#> [[2]]$test[[2]]
#>   [1]   4   5   8  10  11  13  14  15  18  20  23  26  31  33  34  36  37  40
#>  [19]  44  52  57  59  63  64  67  69  73  81  83  90  91  92  95  96  97  98
#>  [37] 103 104 105 114 119 120 123 124 126 127 131 133 134 138 139 146 147 148
#>  [55] 151 154 155 160 163 164 180 183 186 187 194 195 196 197 205 206 209 214
#>  [73] 220 221 228 232 233 237 240 241 246 251 255 258 259 261 262 272 278 279
#>  [91] 280 282 287 288 293 296 297 302 304 307 309 311 312 313 314 316 317 319
#> [109] 324 332 336 341 342 345 351 352 371 375 379 380 382 383 384 386 390 391
#> [127] 393 395 398 401 404 407 408 412 417 418 419 421 422 426 427 429 434 438
#> [145] 444 446 453 455 457 458 463 465 466 469 480 482 485 488 491 494 495 496
#> [163] 498
#> 
#> [[2]]$test[[3]]
#>   [1]   2   6   7   9  12  16  29  30  32  35  38  39  43  45  51  55  60  65
#>  [19]  68  71  72  75  78  80  82  85  87  89  94 100 106 109 111 115 117 122
#>  [37] 125 128 132 135 143 144 152 156 157 158 159 162 165 166 170 172 173 176
#>  [55] 177 181 182 184 191 192 198 200 201 202 203 208 213 217 219 222 223 229
#>  [73] 230 231 234 238 242 244 245 247 248 252 253 254 256 257 260 268 270 271
#>  [91] 273 276 281 284 286 290 291 294 299 303 308 310 318 323 326 331 333 334
#> [109] 335 338 343 344 346 347 350 353 354 355 356 358 359 360 364 365 366 372
#> [127] 378 387 388 389 396 399 400 409 416 420 425 428 433 443 447 448 449 450
#> [145] 452 456 464 467 468 471 472 473 475 476 478 481 484 486 487 490 492 493
#> [163] 497 500
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   2   4   5   6   7   8   9  10  11  12  13  14  15  16  18  20  23  26
#>  [19]  29  30  31  32  33  34  35  36  37  38  39  40  43  44  45  51  52  55
#>  [37]  57  59  60  63  64  65  67  68  69  71  72  73  75  78  80  81  82  83
#>  [55]  85  87  89  90  91  92  94  95  96  97  98 100 103 104 105 106 109 111
#>  [73] 114 115 117 119 120 122 123 124 125 126 127 128 131 132 133 134 135 138
#>  [91] 139 143 144 146 147 148 151 152 154 155 156 157 158 159 160 162 163 164
#> [109] 165 166 170 172 173 176 177 180 181 182 183 184 186 187 191 192 194 195
#> [127] 196 197 198 200 201 202 203 205 206 208 209 213 214 217 219 220 221 222
#> [145] 223 228 229 230 231 232 233 234 237 238 240 241 242 244 245 246 247 248
#> [163] 251 252 253 254 255 256 257 258 259 260 261 262 268 270 271 272 273 276
#> [181] 278 279 280 281 282 284 286 287 288 290 291 293 294 296 297 299 302 303
#> [199] 304 307 308 309 310 311 312 313 314 316 317 318 319 323 324 326 331 332
#> [217] 333 334 335 336 338 341 342 343 344 345 346 347 350 351 352 353 354 355
#> [235] 356 358 359 360 364 365 366 371 372 375 378 379 380 382 383 384 386 387
#> [253] 388 389 390 391 393 395 396 398 399 400 401 404 407 408 409 412 416 417
#> [271] 418 419 420 421 422 425 426 427 428 429 433 434 438 443 444 446 447 448
#> [289] 449 450 452 453 455 456 457 458 463 464 465 466 467 468 469 471 472 473
#> [307] 475 476 478 480 481 482 484 485 486 487 488 490 491 492 493 494 495 496
#> [325] 497 498 500
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   6   7   9  12  16  17  19  21  22  24  25  27  28  29  30
#>  [19]  32  35  38  39  41  42  43  45  46  47  48  49  50  51  53  54  55  56
#>  [37]  58  60  61  62  65  66  68  70  71  72  74  75  76  77  78  79  80  82
#>  [55]  84  85  86  87  88  89  93  94  99 100 101 102 106 107 108 109 110 111
#>  [73] 112 113 115 116 117 118 121 122 125 128 129 130 132 135 136 137 140 141
#>  [91] 142 143 144 145 149 150 152 153 156 157 158 159 161 162 165 166 167 168
#> [109] 169 170 171 172 173 174 175 176 177 178 179 181 182 184 185 188 189 190
#> [127] 191 192 193 198 199 200 201 202 203 204 207 208 210 211 212 213 215 216
#> [145] 217 218 219 222 223 224 225 226 227 229 230 231 234 235 236 238 239 242
#> [163] 243 244 245 247 248 249 250 252 253 254 256 257 260 263 264 265 266 267
#> [181] 268 269 270 271 273 274 275 276 277 281 283 284 285 286 289 290 291 292
#> [199] 294 295 298 299 300 301 303 305 306 308 310 315 318 320 321 322 323 325
#> [217] 326 327 328 329 330 331 333 334 335 337 338 339 340 343 344 346 347 348
#> [235] 349 350 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368
#> [253] 369 370 372 373 374 376 377 378 381 385 387 388 389 392 394 396 397 399
#> [271] 400 402 403 405 406 409 410 411 413 414 415 416 420 423 424 425 428 430
#> [289] 431 432 433 435 436 437 439 440 441 442 443 445 447 448 449 450 451 452
#> [307] 454 456 459 460 461 462 464 467 468 470 471 472 473 474 475 476 477 478
#> [325] 479 481 483 484 486 487 489 490 492 493 497 499 500
#> 
#> [[2]]$train[[3]]
#>   [1]   1   3   4   5   8  10  11  13  14  15  17  18  19  20  21  22  23  24
#>  [19]  25  26  27  28  31  33  34  36  37  40  41  42  44  46  47  48  49  50
#>  [37]  52  53  54  56  57  58  59  61  62  63  64  66  67  69  70  73  74  76
#>  [55]  77  79  81  83  84  86  88  90  91  92  93  95  96  97  98  99 101 102
#>  [73] 103 104 105 107 108 110 112 113 114 116 118 119 120 121 123 124 126 127
#>  [91] 129 130 131 133 134 136 137 138 139 140 141 142 145 146 147 148 149 150
#> [109] 151 153 154 155 160 161 163 164 167 168 169 171 174 175 178 179 180 183
#> [127] 185 186 187 188 189 190 193 194 195 196 197 199 204 205 206 207 209 210
#> [145] 211 212 214 215 216 218 220 221 224 225 226 227 228 232 233 235 236 237
#> [163] 239 240 241 243 246 249 250 251 255 258 259 261 262 263 264 265 266 267
#> [181] 269 272 274 275 277 278 279 280 282 283 285 287 288 289 292 293 295 296
#> [199] 297 298 300 301 302 304 305 306 307 309 311 312 313 314 315 316 317 319
#> [217] 320 321 322 324 325 327 328 329 330 332 336 337 339 340 341 342 345 348
#> [235] 349 351 352 357 361 362 363 367 368 369 370 371 373 374 375 376 377 379
#> [253] 380 381 382 383 384 385 386 390 391 392 393 394 395 397 398 401 402 403
#> [271] 404 405 406 407 408 410 411 412 413 414 415 417 418 419 421 422 423 424
#> [289] 426 427 429 430 431 432 434 435 436 437 438 439 440 441 442 444 445 446
#> [307] 451 453 454 455 457 458 459 460 461 462 463 465 466 469 470 474 477 479
#> [325] 480 482 483 485 488 489 491 494 495 496 498 499
#> 
#> 
#> 
#> [[3]]
#> [[3]]$test
#> [[3]]$test[[1]]
#>   [1]   1   3   5   7  12  13  14  17  19  20  21  24  25  33  41  44  49  50
#>  [19]  54  63  68  76  77  78  84  85  86  91  92  93  97 104 108 111 113 117
#>  [37] 121 122 123 124 132 133 144 148 151 152 154 161 162 167 168 170 171 174
#>  [55] 176 178 184 186 188 189 190 196 198 200 201 203 204 205 208 209 210 211
#>  [73] 212 214 215 216 218 225 228 230 232 234 235 238 239 240 243 244 246 252
#>  [91] 253 257 258 259 261 262 263 264 269 277 280 286 294 296 297 308 309 310
#> [109] 315 321 323 324 329 331 336 338 345 346 351 352 355 358 362 364 366 373
#> [127] 375 377 379 382 384 385 386 390 393 399 400 402 404 406 407 411 415 416
#> [145] 418 419 427 429 432 435 437 438 440 449 450 455 456 460 463 464 465 466
#> [163] 467 468 469 471 479 480 481 483 489 490
#> 
#> [[3]]$test[[2]]
#>   [1]   2   4   8  11  22  23  28  36  38  42  43  45  46  47  48  51  52  56
#>  [19]  57  59  60  61  64  65  67  69  72  73  74  75  82  83  88  94  96 101
#>  [37] 103 107 110 112 115 116 118 120 127 128 130 134 136 138 139 140 143 149
#>  [55] 150 158 160 163 165 166 173 175 177 180 182 187 191 193 197 199 206 213
#>  [73] 220 221 222 223 229 231 233 236 249 251 254 256 265 266 267 268 270 271
#>  [91] 273 275 276 278 279 281 284 285 287 288 289 290 300 305 306 307 311 312
#> [109] 316 318 319 322 325 330 333 334 337 340 344 347 353 356 359 361 365 370
#> [127] 372 374 381 383 387 388 389 394 395 396 405 410 412 417 421 422 424 428
#> [145] 433 434 441 442 443 446 448 451 452 454 457 459 461 462 470 472 473 474
#> [163] 478 482 484 487 494 496 499
#> 
#> [[3]]$test[[3]]
#>   [1]   6   9  10  15  16  18  26  27  29  30  31  32  34  35  37  39  40  53
#>  [19]  55  58  62  66  70  71  79  80  81  87  89  90  95  98  99 100 102 105
#>  [37] 106 109 114 119 125 126 129 131 135 137 141 142 145 146 147 153 155 156
#>  [55] 157 159 164 169 172 179 181 183 185 192 194 195 202 207 217 219 224 226
#>  [73] 227 237 241 242 245 247 248 250 255 260 272 274 282 283 291 292 293 295
#>  [91] 298 299 301 302 303 304 313 314 317 320 326 327 328 332 335 339 341 342
#> [109] 343 348 349 350 354 357 360 363 367 368 369 371 376 378 380 391 392 397
#> [127] 398 401 403 408 409 413 414 420 423 425 426 430 431 436 439 444 445 447
#> [145] 453 458 475 476 477 485 486 488 491 492 493 495 497 498 500
#> 
#> 
#> [[3]]$train
#> [[3]]$train[[1]]
#>   [1]   2   4   6   8   9  10  11  15  16  18  22  23  26  27  28  29  30  31
#>  [19]  32  34  35  36  37  38  39  40  42  43  45  46  47  48  51  52  53  55
#>  [37]  56  57  58  59  60  61  62  64  65  66  67  69  70  71  72  73  74  75
#>  [55]  79  80  81  82  83  87  88  89  90  94  95  96  98  99 100 101 102 103
#>  [73] 105 106 107 109 110 112 114 115 116 118 119 120 125 126 127 128 129 130
#>  [91] 131 134 135 136 137 138 139 140 141 142 143 145 146 147 149 150 153 155
#> [109] 156 157 158 159 160 163 164 165 166 169 172 173 175 177 179 180 181 182
#> [127] 183 185 187 191 192 193 194 195 197 199 202 206 207 213 217 219 220 221
#> [145] 222 223 224 226 227 229 231 233 236 237 241 242 245 247 248 249 250 251
#> [163] 254 255 256 260 265 266 267 268 270 271 272 273 274 275 276 278 279 281
#> [181] 282 283 284 285 287 288 289 290 291 292 293 295 298 299 300 301 302 303
#> [199] 304 305 306 307 311 312 313 314 316 317 318 319 320 322 325 326 327 328
#> [217] 330 332 333 334 335 337 339 340 341 342 343 344 347 348 349 350 353 354
#> [235] 356 357 359 360 361 363 365 367 368 369 370 371 372 374 376 378 380 381
#> [253] 383 387 388 389 391 392 394 395 396 397 398 401 403 405 408 409 410 412
#> [271] 413 414 417 420 421 422 423 424 425 426 428 430 431 433 434 436 439 441
#> [289] 442 443 444 445 446 447 448 451 452 453 454 457 458 459 461 462 470 472
#> [307] 473 474 475 476 477 478 482 484 485 486 487 488 491 492 493 494 495 496
#> [325] 497 498 499 500
#> 
#> [[3]]$train[[2]]
#>   [1]   1   3   5   6   7   9  10  12  13  14  15  16  17  18  19  20  21  24
#>  [19]  25  26  27  29  30  31  32  33  34  35  37  39  40  41  44  49  50  53
#>  [37]  54  55  58  62  63  66  68  70  71  76  77  78  79  80  81  84  85  86
#>  [55]  87  89  90  91  92  93  95  97  98  99 100 102 104 105 106 108 109 111
#>  [73] 113 114 117 119 121 122 123 124 125 126 129 131 132 133 135 137 141 142
#>  [91] 144 145 146 147 148 151 152 153 154 155 156 157 159 161 162 164 167 168
#> [109] 169 170 171 172 174 176 178 179 181 183 184 185 186 188 189 190 192 194
#> [127] 195 196 198 200 201 202 203 204 205 207 208 209 210 211 212 214 215 216
#> [145] 217 218 219 224 225 226 227 228 230 232 234 235 237 238 239 240 241 242
#> [163] 243 244 245 246 247 248 250 252 253 255 257 258 259 260 261 262 263 264
#> [181] 269 272 274 277 280 282 283 286 291 292 293 294 295 296 297 298 299 301
#> [199] 302 303 304 308 309 310 313 314 315 317 320 321 323 324 326 327 328 329
#> [217] 331 332 335 336 338 339 341 342 343 345 346 348 349 350 351 352 354 355
#> [235] 357 358 360 362 363 364 366 367 368 369 371 373 375 376 377 378 379 380
#> [253] 382 384 385 386 390 391 392 393 397 398 399 400 401 402 403 404 406 407
#> [271] 408 409 411 413 414 415 416 418 419 420 423 425 426 427 429 430 431 432
#> [289] 435 436 437 438 439 440 444 445 447 449 450 453 455 456 458 460 463 464
#> [307] 465 466 467 468 469 471 475 476 477 479 480 481 483 485 486 488 489 490
#> [325] 491 492 493 495 497 498 500
#> 
#> [[3]]$train[[3]]
#>   [1]   1   2   3   4   5   7   8  11  12  13  14  17  19  20  21  22  23  24
#>  [19]  25  28  33  36  38  41  42  43  44  45  46  47  48  49  50  51  52  54
#>  [37]  56  57  59  60  61  63  64  65  67  68  69  72  73  74  75  76  77  78
#>  [55]  82  83  84  85  86  88  91  92  93  94  96  97 101 103 104 107 108 110
#>  [73] 111 112 113 115 116 117 118 120 121 122 123 124 127 128 130 132 133 134
#>  [91] 136 138 139 140 143 144 148 149 150 151 152 154 158 160 161 162 163 165
#> [109] 166 167 168 170 171 173 174 175 176 177 178 180 182 184 186 187 188 189
#> [127] 190 191 193 196 197 198 199 200 201 203 204 205 206 208 209 210 211 212
#> [145] 213 214 215 216 218 220 221 222 223 225 228 229 230 231 232 233 234 235
#> [163] 236 238 239 240 243 244 246 249 251 252 253 254 256 257 258 259 261 262
#> [181] 263 264 265 266 267 268 269 270 271 273 275 276 277 278 279 280 281 284
#> [199] 285 286 287 288 289 290 294 296 297 300 305 306 307 308 309 310 311 312
#> [217] 315 316 318 319 321 322 323 324 325 329 330 331 333 334 336 337 338 340
#> [235] 344 345 346 347 351 352 353 355 356 358 359 361 362 364 365 366 370 372
#> [253] 373 374 375 377 379 381 382 383 384 385 386 387 388 389 390 393 394 395
#> [271] 396 399 400 402 404 405 406 407 410 411 412 415 416 417 418 419 421 422
#> [289] 424 427 428 429 432 433 434 435 437 438 440 441 442 443 446 448 449 450
#> [307] 451 452 454 455 456 457 459 460 461 462 463 464 465 466 467 468 469 470
#> [325] 471 472 473 474 478 479 480 481 482 483 484 487 489 490 494 496 499
#> 
#> 
#> 
#> [[4]]
#> [[4]]$test
#> [[4]]$test[[1]]
#>   [1]   4   6   7  10  15  18  19  23  26  27  30  31  34  35  37  38  39  40
#>  [19]  49  52  55  57  64  66  67  70  75  77  82  83  86  88  90 105 106 110
#>  [37] 115 116 119 126 129 131 134 137 139 141 143 144 150 153 158 159 160 165
#>  [55] 169 170 172 179 184 188 190 191 192 193 196 200 213 215 217 219 221 223
#>  [73] 229 234 236 237 239 245 249 254 256 260 261 269 270 271 272 281 282 283
#>  [91] 286 287 288 290 292 293 294 305 307 310 312 314 318 321 323 324 328 330
#> [109] 335 336 337 344 354 358 359 364 365 369 371 372 373 374 386 387 390 392
#> [127] 396 397 403 406 408 409 413 414 421 425 426 431 432 433 436 438 439 445
#> [145] 446 447 448 451 457 462 468 472 476 477 483 486 491 492 494 498 500
#> 
#> [[4]]$test[[2]]
#>   [1]   2   3   5   9  11  12  13  14  17  25  28  36  42  43  46  47  50  51
#>  [19]  53  58  59  62  65  71  72  74  78  80  84  87  91  92  96 102 103 104
#>  [37] 107 108 109 112 118 121 123 124 125 128 130 133 135 140 146 147 148 149
#>  [55] 152 154 156 157 162 168 173 175 176 178 183 186 187 189 194 199 205 206
#>  [73] 209 210 211 218 220 225 230 231 233 235 238 241 242 244 247 248 252 253
#>  [91] 258 264 273 278 279 280 289 295 296 297 299 300 303 306 308 311 317 319
#> [109] 320 322 325 326 327 329 331 339 340 343 346 347 348 351 352 355 356 357
#> [127] 360 363 367 370 375 376 378 379 381 382 384 389 391 394 395 398 402 405
#> [145] 407 419 423 427 429 430 435 440 441 444 449 452 453 454 455 459 460 464
#> [163] 465 469 470 471 475 478 479 480 487 490 495
#> 
#> [[4]]$test[[3]]
#>   [1]   1   8  16  20  21  22  24  29  32  33  41  44  45  48  54  56  60  61
#>  [19]  63  68  69  73  76  79  81  85  89  93  94  95  97  98  99 100 101 111
#>  [37] 113 114 117 120 122 127 132 136 138 142 145 151 155 161 163 164 166 167
#>  [55] 171 174 177 180 181 182 185 195 197 198 201 202 203 204 207 208 212 214
#>  [73] 216 222 224 226 227 228 232 240 243 246 250 251 255 257 259 262 263 265
#>  [91] 266 267 268 274 275 276 277 284 285 291 298 301 302 304 309 313 315 316
#> [109] 332 333 334 338 341 342 345 349 350 353 361 362 366 368 377 380 383 385
#> [127] 388 393 399 400 401 404 410 411 412 415 416 417 418 420 422 424 428 434
#> [145] 437 442 443 450 456 458 461 463 466 467 473 474 481 482 484 485 488 489
#> [163] 493 496 497 499
#> 
#> 
#> [[4]]$train
#> [[4]]$train[[1]]
#>   [1]   1   2   3   5   8   9  11  12  13  14  16  17  20  21  22  24  25  28
#>  [19]  29  32  33  36  41  42  43  44  45  46  47  48  50  51  53  54  56  58
#>  [37]  59  60  61  62  63  65  68  69  71  72  73  74  76  78  79  80  81  84
#>  [55]  85  87  89  91  92  93  94  95  96  97  98  99 100 101 102 103 104 107
#>  [73] 108 109 111 112 113 114 117 118 120 121 122 123 124 125 127 128 130 132
#>  [91] 133 135 136 138 140 142 145 146 147 148 149 151 152 154 155 156 157 161
#> [109] 162 163 164 166 167 168 171 173 174 175 176 177 178 180 181 182 183 185
#> [127] 186 187 189 194 195 197 198 199 201 202 203 204 205 206 207 208 209 210
#> [145] 211 212 214 216 218 220 222 224 225 226 227 228 230 231 232 233 235 238
#> [163] 240 241 242 243 244 246 247 248 250 251 252 253 255 257 258 259 262 263
#> [181] 264 265 266 267 268 273 274 275 276 277 278 279 280 284 285 289 291 295
#> [199] 296 297 298 299 300 301 302 303 304 306 308 309 311 313 315 316 317 319
#> [217] 320 322 325 326 327 329 331 332 333 334 338 339 340 341 342 343 345 346
#> [235] 347 348 349 350 351 352 353 355 356 357 360 361 362 363 366 367 368 370
#> [253] 375 376 377 378 379 380 381 382 383 384 385 388 389 391 393 394 395 398
#> [271] 399 400 401 402 404 405 407 410 411 412 415 416 417 418 419 420 422 423
#> [289] 424 427 428 429 430 434 435 437 440 441 442 443 444 449 450 452 453 454
#> [307] 455 456 458 459 460 461 463 464 465 466 467 469 470 471 473 474 475 478
#> [325] 479 480 481 482 484 485 487 488 489 490 493 495 496 497 499
#> 
#> [[4]]$train[[2]]
#>   [1]   1   4   6   7   8  10  15  16  18  19  20  21  22  23  24  26  27  29
#>  [19]  30  31  32  33  34  35  37  38  39  40  41  44  45  48  49  52  54  55
#>  [37]  56  57  60  61  63  64  66  67  68  69  70  73  75  76  77  79  81  82
#>  [55]  83  85  86  88  89  90  93  94  95  97  98  99 100 101 105 106 110 111
#>  [73] 113 114 115 116 117 119 120 122 126 127 129 131 132 134 136 137 138 139
#>  [91] 141 142 143 144 145 150 151 153 155 158 159 160 161 163 164 165 166 167
#> [109] 169 170 171 172 174 177 179 180 181 182 184 185 188 190 191 192 193 195
#> [127] 196 197 198 200 201 202 203 204 207 208 212 213 214 215 216 217 219 221
#> [145] 222 223 224 226 227 228 229 232 234 236 237 239 240 243 245 246 249 250
#> [163] 251 254 255 256 257 259 260 261 262 263 265 266 267 268 269 270 271 272
#> [181] 274 275 276 277 281 282 283 284 285 286 287 288 290 291 292 293 294 298
#> [199] 301 302 304 305 307 309 310 312 313 314 315 316 318 321 323 324 328 330
#> [217] 332 333 334 335 336 337 338 341 342 344 345 349 350 353 354 358 359 361
#> [235] 362 364 365 366 368 369 371 372 373 374 377 380 383 385 386 387 388 390
#> [253] 392 393 396 397 399 400 401 403 404 406 408 409 410 411 412 413 414 415
#> [271] 416 417 418 420 421 422 424 425 426 428 431 432 433 434 436 437 438 439
#> [289] 442 443 445 446 447 448 450 451 456 457 458 461 462 463 466 467 468 472
#> [307] 473 474 476 477 481 482 483 484 485 486 488 489 491 492 493 494 496 497
#> [325] 498 499 500
#> 
#> [[4]]$train[[3]]
#>   [1]   2   3   4   5   6   7   9  10  11  12  13  14  15  17  18  19  23  25
#>  [19]  26  27  28  30  31  34  35  36  37  38  39  40  42  43  46  47  49  50
#>  [37]  51  52  53  55  57  58  59  62  64  65  66  67  70  71  72  74  75  77
#>  [55]  78  80  82  83  84  86  87  88  90  91  92  96 102 103 104 105 106 107
#>  [73] 108 109 110 112 115 116 118 119 121 123 124 125 126 128 129 130 131 133
#>  [91] 134 135 137 139 140 141 143 144 146 147 148 149 150 152 153 154 156 157
#> [109] 158 159 160 162 165 168 169 170 172 173 175 176 178 179 183 184 186 187
#> [127] 188 189 190 191 192 193 194 196 199 200 205 206 209 210 211 213 215 217
#> [145] 218 219 220 221 223 225 229 230 231 233 234 235 236 237 238 239 241 242
#> [163] 244 245 247 248 249 252 253 254 256 258 260 261 264 269 270 271 272 273
#> [181] 278 279 280 281 282 283 286 287 288 289 290 292 293 294 295 296 297 299
#> [199] 300 303 305 306 307 308 310 311 312 314 317 318 319 320 321 322 323 324
#> [217] 325 326 327 328 329 330 331 335 336 337 339 340 343 344 346 347 348 351
#> [235] 352 354 355 356 357 358 359 360 363 364 365 367 369 370 371 372 373 374
#> [253] 375 376 378 379 381 382 384 386 387 389 390 391 392 394 395 396 397 398
#> [271] 402 403 405 406 407 408 409 413 414 419 421 423 425 426 427 429 430 431
#> [289] 432 433 435 436 438 439 440 441 444 445 446 447 448 449 451 452 453 454
#> [307] 455 457 459 460 462 464 465 468 469 470 471 472 475 476 477 478 479 480
#> [325] 483 486 487 490 491 492 494 495 498 500
#> 
#> 
#> 
#> [[5]]
#> [[5]]$test
#> [[5]]$test[[1]]
#>   [1]   4   6   9  10  13  18  25  27  32  33  34  37  40  42  43  44  47  49
#>  [19]  51  52  53  60  62  63  70  74  77  81  84  86  89  94  95 102 106 112
#>  [37] 113 114 115 117 118 119 121 123 125 126 130 134 139 144 146 149 152 153
#>  [55] 159 160 164 165 170 171 172 173 174 176 177 181 190 194 201 203 204 208
#>  [73] 216 218 226 227 228 237 238 245 248 255 256 257 260 269 270 272 273 274
#>  [91] 277 280 281 283 285 293 298 301 302 305 308 310 312 313 314 318 320 321
#> [109] 325 327 330 331 332 338 348 350 351 352 356 362 365 369 370 371 373 374
#> [127] 377 384 386 387 390 392 396 398 399 401 404 405 406 416 421 423 425 427
#> [145] 429 432 444 445 446 448 452 454 459 464 465 472 476 479 482 483 490 491
#> [163] 498 500
#> 
#> [[5]]$test[[2]]
#>   [1]   1   2   7  11  12  21  22  23  24  26  30  35  41  46  50  54  55  57
#>  [19]  58  64  67  68  73  76  78  79  80  82  83  88  90  93  96 104 105 107
#>  [37] 109 111 120 127 128 129 131 132 133 135 137 140 142 145 150 151 156 157
#>  [55] 158 161 162 175 182 184 185 188 191 193 195 197 202 206 215 217 221 225
#>  [73] 231 232 233 234 235 239 240 241 243 244 246 247 249 252 253 254 258 263
#>  [91] 265 268 278 284 287 288 292 297 299 300 303 304 309 311 317 326 328 329
#> [109] 333 334 335 337 339 340 343 344 345 347 354 355 358 359 360 361 363 364
#> [127] 366 367 375 389 391 393 395 400 407 408 409 411 412 418 420 434 437 440
#> [145] 447 449 450 451 455 457 461 462 466 467 468 469 471 474 475 478 481 489
#> [163] 493 495 496 499
#> 
#> [[5]]$test[[3]]
#>   [1]   3   5   8  14  15  16  17  19  20  28  29  31  36  38  39  45  48  56
#>  [19]  59  61  65  66  69  71  72  75  85  87  91  92  97  98  99 100 101 103
#>  [37] 108 110 116 122 124 136 138 141 143 147 148 154 155 163 166 167 168 169
#>  [55] 178 179 180 183 186 187 189 192 196 198 199 200 205 207 209 210 211 212
#>  [73] 213 214 219 220 222 223 224 229 230 236 242 250 251 259 261 262 264 266
#>  [91] 267 271 275 276 279 282 286 289 290 291 294 295 296 306 307 315 316 319
#> [109] 322 323 324 336 341 342 346 349 353 357 368 372 376 378 379 380 381 382
#> [127] 383 385 388 394 397 402 403 410 413 414 415 417 419 422 424 426 428 430
#> [145] 431 433 435 436 438 439 441 442 443 453 456 458 460 463 470 473 477 480
#> [163] 484 485 486 487 488 492 494 497
#> 
#> 
#> [[5]]$train
#> [[5]]$train[[1]]
#>   [1]   1   2   3   5   7   8  11  12  14  15  16  17  19  20  21  22  23  24
#>  [19]  26  28  29  30  31  35  36  38  39  41  45  46  48  50  54  55  56  57
#>  [37]  58  59  61  64  65  66  67  68  69  71  72  73  75  76  78  79  80  82
#>  [55]  83  85  87  88  90  91  92  93  96  97  98  99 100 101 103 104 105 107
#>  [73] 108 109 110 111 116 120 122 124 127 128 129 131 132 133 135 136 137 138
#>  [91] 140 141 142 143 145 147 148 150 151 154 155 156 157 158 161 162 163 166
#> [109] 167 168 169 175 178 179 180 182 183 184 185 186 187 188 189 191 192 193
#> [127] 195 196 197 198 199 200 202 205 206 207 209 210 211 212 213 214 215 217
#> [145] 219 220 221 222 223 224 225 229 230 231 232 233 234 235 236 239 240 241
#> [163] 242 243 244 246 247 249 250 251 252 253 254 258 259 261 262 263 264 265
#> [181] 266 267 268 271 275 276 278 279 282 284 286 287 288 289 290 291 292 294
#> [199] 295 296 297 299 300 303 304 306 307 309 311 315 316 317 319 322 323 324
#> [217] 326 328 329 333 334 335 336 337 339 340 341 342 343 344 345 346 347 349
#> [235] 353 354 355 357 358 359 360 361 363 364 366 367 368 372 375 376 378 379
#> [253] 380 381 382 383 385 388 389 391 393 394 395 397 400 402 403 407 408 409
#> [271] 410 411 412 413 414 415 417 418 419 420 422 424 426 428 430 431 433 434
#> [289] 435 436 437 438 439 440 441 442 443 447 449 450 451 453 455 456 457 458
#> [307] 460 461 462 463 466 467 468 469 470 471 473 474 475 477 478 480 481 484
#> [325] 485 486 487 488 489 492 493 494 495 496 497 499
#> 
#> [[5]]$train[[2]]
#>   [1]   3   4   5   6   8   9  10  13  14  15  16  17  18  19  20  25  27  28
#>  [19]  29  31  32  33  34  36  37  38  39  40  42  43  44  45  47  48  49  51
#>  [37]  52  53  56  59  60  61  62  63  65  66  69  70  71  72  74  75  77  81
#>  [55]  84  85  86  87  89  91  92  94  95  97  98  99 100 101 102 103 106 108
#>  [73] 110 112 113 114 115 116 117 118 119 121 122 123 124 125 126 130 134 136
#>  [91] 138 139 141 143 144 146 147 148 149 152 153 154 155 159 160 163 164 165
#> [109] 166 167 168 169 170 171 172 173 174 176 177 178 179 180 181 183 186 187
#> [127] 189 190 192 194 196 198 199 200 201 203 204 205 207 208 209 210 211 212
#> [145] 213 214 216 218 219 220 222 223 224 226 227 228 229 230 236 237 238 242
#> [163] 245 248 250 251 255 256 257 259 260 261 262 264 266 267 269 270 271 272
#> [181] 273 274 275 276 277 279 280 281 282 283 285 286 289 290 291 293 294 295
#> [199] 296 298 301 302 305 306 307 308 310 312 313 314 315 316 318 319 320 321
#> [217] 322 323 324 325 327 330 331 332 336 338 341 342 346 348 349 350 351 352
#> [235] 353 356 357 362 365 368 369 370 371 372 373 374 376 377 378 379 380 381
#> [253] 382 383 384 385 386 387 388 390 392 394 396 397 398 399 401 402 403 404
#> [271] 405 406 410 413 414 415 416 417 419 421 422 423 424 425 426 427 428 429
#> [289] 430 431 432 433 435 436 438 439 441 442 443 444 445 446 448 452 453 454
#> [307] 456 458 459 460 463 464 465 470 472 473 476 477 479 480 482 483 484 485
#> [325] 486 487 488 490 491 492 494 497 498 500
#> 
#> [[5]]$train[[3]]
#>   [1]   1   2   4   6   7   9  10  11  12  13  18  21  22  23  24  25  26  27
#>  [19]  30  32  33  34  35  37  40  41  42  43  44  46  47  49  50  51  52  53
#>  [37]  54  55  57  58  60  62  63  64  67  68  70  73  74  76  77  78  79  80
#>  [55]  81  82  83  84  86  88  89  90  93  94  95  96 102 104 105 106 107 109
#>  [73] 111 112 113 114 115 117 118 119 120 121 123 125 126 127 128 129 130 131
#>  [91] 132 133 134 135 137 139 140 142 144 145 146 149 150 151 152 153 156 157
#> [109] 158 159 160 161 162 164 165 170 171 172 173 174 175 176 177 181 182 184
#> [127] 185 188 190 191 193 194 195 197 201 202 203 204 206 208 215 216 217 218
#> [145] 221 225 226 227 228 231 232 233 234 235 237 238 239 240 241 243 244 245
#> [163] 246 247 248 249 252 253 254 255 256 257 258 260 263 265 268 269 270 272
#> [181] 273 274 277 278 280 281 283 284 285 287 288 292 293 297 298 299 300 301
#> [199] 302 303 304 305 308 309 310 311 312 313 314 317 318 320 321 325 326 327
#> [217] 328 329 330 331 332 333 334 335 337 338 339 340 343 344 345 347 348 350
#> [235] 351 352 354 355 356 358 359 360 361 362 363 364 365 366 367 369 370 371
#> [253] 373 374 375 377 384 386 387 389 390 391 392 393 395 396 398 399 400 401
#> [271] 404 405 406 407 408 409 411 412 416 418 420 421 423 425 427 429 432 434
#> [289] 437 440 444 445 446 447 448 449 450 451 452 454 455 457 459 461 462 464
#> [307] 465 466 467 468 469 471 472 474 475 476 478 479 481 482 483 489 490 491
#> [325] 493 495 496 498 499 500
#> 
#> 
#>