Repeated Spatio-Temporal Cluster Resampling

Spatio-temporal cluster partitioning via the vcluster executable of the CLUTO clustering application.

This partitioning method relies on the external CLUTO library. To use it, CLUTO's executables need to be downloaded and installed into this package.

See https://gist.github.com/pat-s/6430470cf817050e27d26c43c0e9be72 for an installation approach that should work on Windows and Linux. macOS is not supported by CLUTO.

Before using this method, please check the restrictive copyright shown below.

Details

By default, -clmethod='direct' is passed to the vcluster executable in contrast to the upstream default -clmethod='rb'. There is no evidence or research that this method is the best among the available ones ("rb", "rbr", "direct", "agglo", "graph", "bagglo"). Also, various other parameters can be set via argument cluto_parameters to achieve different clustering results.

Parameter -clusterfile is handled by skmeans and cannot be changed.

Copyright

CLUTO's copyright is as follows:

The CLUTO package is copyrighted by the Regents of the University of Minnesota. It can be freely used for educational and research purposes by non-profit institutions and US government agencies only. Other organizations are allowed to use CLUTO only for evaluation purposes, and any further uses will require prior approval. The software may not be sold or redistributed without prior approval. One may make copies of the software for their use provided that the copies, are not sold or distributed, are used under the same terms and conditions. As unestablished research software, this code is provided on an “as is” basis without warranty of any kind, either expressed or implied. The downloading, or executing any part of this software constitutes an implicit agreement to these terms. These terms and conditions are subject to change at any time without prior notice.

References

FIXME: correct? Ying Zhao and George Karypis. 11th Conference of Information and Knowledge Management (CIKM), pp. 515-524, 2002.

Super class

mlr3::Resampling -> ResamplingRepeatedSptCVCstf

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSptCVCstf$new()

ResamplingRepeatedSptCVCstf$folds()

ResamplingRepeatedSptCVCstf$repeats()

ResamplingRepeatedSptCVCstf$instantiate()

ResamplingRepeatedSptCVCstf$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "coordinate-based" repeated resampling instance.

Usage

ResamplingRepeatedSptCVCstf$new(id = "repeated-sptcv-cstf")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSptCVCstf$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSptCVCstf$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSptCVCstf$instantiate(
  task,
  space_var = NULL,
  time_var = NULL,
  class = NULL
)

Arguments

task

Task
A task to instantiate.

space_var

[character]
Column name identifying the spatial units.

time_var

[character]
Column name identifying the temporal units.

class

[character]
Column name identifying a class unit (e.g. land cover).

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSptCVCstf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(mlr3spatiotempcv)
task = tsk("cookfarm")

# Instantiate Resampling
rrcv = rsmp("repeated-sptcv-cstf", folds = 3, repeats = 5)
rrcv$instantiate(task, time_var = "Date")
# Individual sets:
rrcv$iters

#> [1] 15

rrcv$folds(1:6)

#> [1] 1 2 3 4 5 1

rrcv$repeats(1:6)

#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)

#>   [1]   2   3   4   5   6   7   8   9  10  12  14  17  19  20  23  24  25  27
#>  [19]  28  29  30  32  35  37  38  40  42  44  45  46  49  50  51  55  56  58
#>  [37]  59  60  61  63  65  67  68  69  70  72  73  74  75  76  77  78  83  84
#>  [55]  85  86  87  88  89  90  91  93  94  95  96  97  98  99 100 101 103 104
#>  [73] 107 108 109 111 112 113 114 115 117 118 120 121 122 124 125 127 129 130
#>  [91] 131 132 134 135 136 137 143 144 145 146 149 150 152 153 154 155 156 157
#> [109] 158 159 160 161 162 163 164 166 167 170 172 173 174 175 176 177 178 179
#> [127] 181 182 183 184 185 186 187 189 191 192 193 194 196 198 202 204 205 206
#> [145] 208 209 210 211 213 215 216 217 219 220 222 223 224 225 227 228 229 230
#> [163] 231 232 233 234 235 236 238 240 242 243 244 245 246 248 249 251 254 255
#> [181] 256 258 259 260 262 264 267 272 273 274 276 277 279 280 282 283 284 288
#> [199] 289 291 292 293 294 295 296 297 298 299 300 301 302 304 305 306 307 308
#> [217] 310 312 313 314 315 318 319 320 321 324 325 328 329 330 333 334 335 336
#> [235] 339 341 342 343 344 345 347 351 354 356 357 358 359 360 361 363 367 368
#> [253] 370 371 372 373 374 376 377 378 381 382 386 387 390 391 392 393 394 396
#> [271] 397 398 400 402 403 404 406 407 408 410 413 416 417 418 419 422 423 424
#> [289] 425 426 427 428 429 430 432 433 434 435 437 438 439 441 442 443 444 445
#> [307] 446 447 448 449 450 451 452 453 457 460 461 463 464 466 467 468 469 470
#> [325] 471 474 475 476 478 482 486 489 490 491 492 493 494 496 498 500

rrcv$test_set(1)

#>   [1]   1  11  13  15  16  18  21  22  26  31  33  34  36  39  41  43  47  48
#>  [19]  52  53  54  57  62  64  66  71  79  80  81  82  92 102 105 106 110 116
#>  [37] 119 123 126 128 133 138 139 140 141 142 147 148 151 165 168 169 171 180
#>  [55] 188 190 195 197 199 200 201 203 207 212 214 218 221 226 237 239 241 247
#>  [73] 250 252 253 257 261 263 265 266 268 269 270 271 275 278 281 285 286 287
#>  [91] 290 303 309 311 316 317 322 323 326 327 331 332 337 338 340 346 348 349
#> [109] 350 352 353 355 362 364 365 366 369 375 379 380 383 384 385 388 389 395
#> [127] 399 401 405 409 411 412 414 415 420 421 431 436 440 454 455 456 458 459
#> [145] 462 465 472 473 477 479 480 481 483 484 485 487 488 495 497 499

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance # table

#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>   [1]   1  11  13  15  16  18  21  22  26  31  33  34  36  39  41  43  47  48
#>  [19]  52  53  54  57  62  64  66  71  79  80  81  82  92 102 105 106 110 116
#>  [37] 119 123 126 128 133 138 139 140 141 142 147 148 151 165 168 169 171 180
#>  [55] 188 190 195 197 199 200 201 203 207 212 214 218 221 226 237 239 241 247
#>  [73] 250 252 253 257 261 263 265 266 268 269 270 271 275 278 281 285 286 287
#>  [91] 290 303 309 311 316 317 322 323 326 327 331 332 337 338 340 346 348 349
#> [109] 350 352 353 355 362 364 365 366 369 375 379 380 383 384 385 388 389 395
#> [127] 399 401 405 409 411 412 414 415 420 421 431 436 440 454 455 456 458 459
#> [145] 462 465 472 473 477 479 480 481 483 484 485 487 488 495 497 499
#> 
#> [[1]]$test[[2]]
#>   [1]   4   6   7   9  12  14  17  23  24  25  28  29  32  35  37  38  40  42
#>  [19]  45  49  59  61  65  67  72  74  77  78  85  86  87  88  89  94  99 101
#>  [37] 103 108 112 113 117 122 127 129 135 136 137 143 144 145 149 153 154 155
#>  [55] 156 157 160 167 170 172 173 174 176 177 181 183 185 186 187 189 193 194
#>  [73] 196 202 204 205 208 209 210 220 223 224 229 231 233 235 236 240 242 244
#>  [91] 246 248 254 255 260 276 277 279 280 282 295 298 299 304 306 308 314 315
#> [109] 320 321 324 328 329 339 342 343 351 356 357 358 360 361 363 367 368 370
#> [127] 371 372 373 374 376 378 381 386 390 391 393 404 407 413 417 422 423 425
#> [145] 426 429 432 435 437 439 444 445 446 447 450 451 452 461 463 464 466 468
#> [163] 469 470 471 482 486 489 490 494 500
#> 
#> [[1]]$test[[3]]
#>   [1]   2   3   5   8  10  19  20  27  30  44  46  50  51  55  56  58  60  63
#>  [19]  68  69  70  73  75  76  83  84  90  91  93  95  96  97  98 100 104 107
#>  [37] 109 111 114 115 118 120 121 124 125 130 131 132 134 146 150 152 158 159
#>  [55] 161 162 163 164 166 175 178 179 182 184 191 192 198 206 211 213 215 216
#>  [73] 217 219 222 225 227 228 230 232 234 238 243 245 249 251 256 258 259 262
#>  [91] 264 267 272 273 274 283 284 288 289 291 292 293 294 296 297 300 301 302
#> [109] 305 307 310 312 313 318 319 325 330 333 334 335 336 341 344 345 347 354
#> [127] 359 377 382 387 392 394 396 397 398 400 402 403 406 408 410 416 418 419
#> [145] 424 427 428 430 433 434 438 441 442 443 448 449 453 457 460 467 474 475
#> [163] 476 478 491 492 493 496 498
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   2   3   4   5   6   7   8   9  10  12  14  17  19  20  23  24  25  27
#>  [19]  28  29  30  32  35  37  38  40  42  44  45  46  49  50  51  55  56  58
#>  [37]  59  60  61  63  65  67  68  69  70  72  73  74  75  76  77  78  83  84
#>  [55]  85  86  87  88  89  90  91  93  94  95  96  97  98  99 100 101 103 104
#>  [73] 107 108 109 111 112 113 114 115 117 118 120 121 122 124 125 127 129 130
#>  [91] 131 132 134 135 136 137 143 144 145 146 149 150 152 153 154 155 156 157
#> [109] 158 159 160 161 162 163 164 166 167 170 172 173 174 175 176 177 178 179
#> [127] 181 182 183 184 185 186 187 189 191 192 193 194 196 198 202 204 205 206
#> [145] 208 209 210 211 213 215 216 217 219 220 222 223 224 225 227 228 229 230
#> [163] 231 232 233 234 235 236 238 240 242 243 244 245 246 248 249 251 254 255
#> [181] 256 258 259 260 262 264 267 272 273 274 276 277 279 280 282 283 284 288
#> [199] 289 291 292 293 294 295 296 297 298 299 300 301 302 304 305 306 307 308
#> [217] 310 312 313 314 315 318 319 320 321 324 325 328 329 330 333 334 335 336
#> [235] 339 341 342 343 344 345 347 351 354 356 357 358 359 360 361 363 367 368
#> [253] 370 371 372 373 374 376 377 378 381 382 386 387 390 391 392 393 394 396
#> [271] 397 398 400 402 403 404 406 407 408 410 413 416 417 418 419 422 423 424
#> [289] 425 426 427 428 429 430 432 433 434 435 437 438 439 441 442 443 444 445
#> [307] 446 447 448 449 450 451 452 453 457 460 461 463 464 466 467 468 469 470
#> [325] 471 474 475 476 478 482 486 489 490 491 492 493 494 496 498 500
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   5   8  10  11  13  15  16  18  19  20  21  22  26  27  30
#>  [19]  31  33  34  36  39  41  43  44  46  47  48  50  51  52  53  54  55  56
#>  [37]  57  58  60  62  63  64  66  68  69  70  71  73  75  76  79  80  81  82
#>  [55]  83  84  90  91  92  93  95  96  97  98 100 102 104 105 106 107 109 110
#>  [73] 111 114 115 116 118 119 120 121 123 124 125 126 128 130 131 132 133 134
#>  [91] 138 139 140 141 142 146 147 148 150 151 152 158 159 161 162 163 164 165
#> [109] 166 168 169 171 175 178 179 180 182 184 188 190 191 192 195 197 198 199
#> [127] 200 201 203 206 207 211 212 213 214 215 216 217 218 219 221 222 225 226
#> [145] 227 228 230 232 234 237 238 239 241 243 245 247 249 250 251 252 253 256
#> [163] 257 258 259 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275
#> [181] 278 281 283 284 285 286 287 288 289 290 291 292 293 294 296 297 300 301
#> [199] 302 303 305 307 309 310 311 312 313 316 317 318 319 322 323 325 326 327
#> [217] 330 331 332 333 334 335 336 337 338 340 341 344 345 346 347 348 349 350
#> [235] 352 353 354 355 359 362 364 365 366 369 375 377 379 380 382 383 384 385
#> [253] 387 388 389 392 394 395 396 397 398 399 400 401 402 403 405 406 408 409
#> [271] 410 411 412 414 415 416 418 419 420 421 424 427 428 430 431 433 434 436
#> [289] 438 440 441 442 443 448 449 453 454 455 456 457 458 459 460 462 465 467
#> [307] 472 473 474 475 476 477 478 479 480 481 483 484 485 487 488 491 492 493
#> [325] 495 496 497 498 499
#> 
#> [[1]]$train[[3]]
#>   [1]   1   4   6   7   9  11  12  13  14  15  16  17  18  21  22  23  24  25
#>  [19]  26  28  29  31  32  33  34  35  36  37  38  39  40  41  42  43  45  47
#>  [37]  48  49  52  53  54  57  59  61  62  64  65  66  67  71  72  74  77  78
#>  [55]  79  80  81  82  85  86  87  88  89  92  94  99 101 102 103 105 106 108
#>  [73] 110 112 113 116 117 119 122 123 126 127 128 129 133 135 136 137 138 139
#>  [91] 140 141 142 143 144 145 147 148 149 151 153 154 155 156 157 160 165 167
#> [109] 168 169 170 171 172 173 174 176 177 180 181 183 185 186 187 188 189 190
#> [127] 193 194 195 196 197 199 200 201 202 203 204 205 207 208 209 210 212 214
#> [145] 218 220 221 223 224 226 229 231 233 235 236 237 239 240 241 242 244 246
#> [163] 247 248 250 252 253 254 255 257 260 261 263 265 266 268 269 270 271 275
#> [181] 276 277 278 279 280 281 282 285 286 287 290 295 298 299 303 304 306 308
#> [199] 309 311 314 315 316 317 320 321 322 323 324 326 327 328 329 331 332 337
#> [217] 338 339 340 342 343 346 348 349 350 351 352 353 355 356 357 358 360 361
#> [235] 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 378 379 380
#> [253] 381 383 384 385 386 388 389 390 391 393 395 399 401 404 405 407 409 411
#> [271] 412 413 414 415 417 420 421 422 423 425 426 429 431 432 435 436 437 439
#> [289] 440 444 445 446 447 450 451 452 454 455 456 458 459 461 462 463 464 465
#> [307] 466 468 469 470 471 472 473 477 479 480 481 482 483 484 485 486 487 488
#> [325] 489 490 494 495 497 499 500
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>   [1]   5  13  16  18  23  26  36  37  38  39  55  56  58  59  62  63  67  70
#>  [19]  72  75  77  78  79  80  82  83  85  88  90  96  98 109 110 113 116 119
#>  [37] 120 125 126 127 129 132 133 135 141 143 144 153 154 156 157 161 162 167
#>  [55] 170 172 173 177 178 180 183 184 186 188 201 209 210 212 217 219 220 225
#>  [73] 227 228 229 231 232 234 235 239 240 242 244 246 248 251 252 254 258 273
#>  [91] 280 281 285 289 297 298 311 312 316 319 320 322 327 329 334 335 336 339
#> [109] 341 343 348 351 353 357 358 361 367 368 369 370 371 373 376 381 383 386
#> [127] 393 394 397 398 401 404 405 406 407 411 415 417 420 423 428 430 431 437
#> [145] 440 441 443 444 448 453 456 461 465 467 470 471 472 474 483 486 488 491
#> [163] 494 496 498 499
#> 
#> [[2]]$test[[2]]
#>   [1]   1   3   6   8  12  17  19  20  21  28  30  31  32  34  41  42  43  47
#>  [19]  48  49  50  53  57  60  61  65  66  74  84  87  91  93  99 100 101 102
#>  [37] 103 104 105 106 124 128 131 137 138 140 146 148 149 158 159 160 165 171
#>  [55] 174 175 181 182 185 187 189 191 192 193 195 196 198 202 206 207 208 213
#>  [73] 215 216 218 223 224 237 238 241 243 245 247 253 256 257 259 261 262 265
#>  [91] 267 268 270 271 272 274 275 276 278 279 282 286 288 290 293 299 300 302
#> [109] 303 305 307 308 313 321 325 328 330 331 332 338 340 344 347 350 354 359
#> [127] 362 363 365 375 378 379 380 387 389 395 400 408 413 416 418 419 426 427
#> [145] 429 433 434 436 439 442 445 446 447 450 452 454 457 459 464 466 473 475
#> [163] 476 477 478 481 482 484 485 492 497
#> 
#> [[2]]$test[[3]]
#>   [1]   2   4   7   9  10  11  14  15  22  24  25  27  29  33  35  40  44  45
#>  [19]  46  51  52  54  64  68  69  71  73  76  81  86  89  92  94  95  97 107
#>  [37] 108 111 112 114 115 117 118 121 122 123 130 134 136 139 142 145 147 150
#>  [55] 151 152 155 163 164 166 168 169 176 179 190 194 197 199 200 203 204 205
#>  [73] 211 214 221 222 226 230 233 236 249 250 255 260 263 264 266 269 277 283
#>  [91] 284 287 291 292 294 295 296 301 304 306 309 310 314 315 317 318 323 324
#> [109] 326 333 337 342 345 346 349 352 355 356 360 364 366 372 374 377 382 384
#> [127] 385 388 390 391 392 396 399 402 403 409 410 412 414 421 422 424 425 432
#> [145] 435 438 449 451 455 458 460 462 463 468 469 479 480 487 489 490 493 495
#> [163] 500
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   6   7   8   9  10  11  12  14  15  17  19  20  21  22
#>  [19]  24  25  27  28  29  30  31  32  33  34  35  40  41  42  43  44  45  46
#>  [37]  47  48  49  50  51  52  53  54  57  60  61  64  65  66  68  69  71  73
#>  [55]  74  76  81  84  86  87  89  91  92  93  94  95  97  99 100 101 102 103
#>  [73] 104 105 106 107 108 111 112 114 115 117 118 121 122 123 124 128 130 131
#>  [91] 134 136 137 138 139 140 142 145 146 147 148 149 150 151 152 155 158 159
#> [109] 160 163 164 165 166 168 169 171 174 175 176 179 181 182 185 187 189 190
#> [127] 191 192 193 194 195 196 197 198 199 200 202 203 204 205 206 207 208 211
#> [145] 213 214 215 216 218 221 222 223 224 226 230 233 236 237 238 241 243 245
#> [163] 247 249 250 253 255 256 257 259 260 261 262 263 264 265 266 267 268 269
#> [181] 270 271 272 274 275 276 277 278 279 282 283 284 286 287 288 290 291 292
#> [199] 293 294 295 296 299 300 301 302 303 304 305 306 307 308 309 310 313 314
#> [217] 315 317 318 321 323 324 325 326 328 330 331 332 333 337 338 340 342 344
#> [235] 345 346 347 349 350 352 354 355 356 359 360 362 363 364 365 366 372 374
#> [253] 375 377 378 379 380 382 384 385 387 388 389 390 391 392 395 396 399 400
#> [271] 402 403 408 409 410 412 413 414 416 418 419 421 422 424 425 426 427 429
#> [289] 432 433 434 435 436 438 439 442 445 446 447 449 450 451 452 454 455 457
#> [307] 458 459 460 462 463 464 466 468 469 473 475 476 477 478 479 480 481 482
#> [325] 484 485 487 489 490 492 493 495 497 500
#> 
#> [[2]]$train[[2]]
#>   [1]   2   4   5   7   9  10  11  13  14  15  16  18  22  23  24  25  26  27
#>  [19]  29  33  35  36  37  38  39  40  44  45  46  51  52  54  55  56  58  59
#>  [37]  62  63  64  67  68  69  70  71  72  73  75  76  77  78  79  80  81  82
#>  [55]  83  85  86  88  89  90  92  94  95  96  97  98 107 108 109 110 111 112
#>  [73] 113 114 115 116 117 118 119 120 121 122 123 125 126 127 129 130 132 133
#>  [91] 134 135 136 139 141 142 143 144 145 147 150 151 152 153 154 155 156 157
#> [109] 161 162 163 164 166 167 168 169 170 172 173 176 177 178 179 180 183 184
#> [127] 186 188 190 194 197 199 200 201 203 204 205 209 210 211 212 214 217 219
#> [145] 220 221 222 225 226 227 228 229 230 231 232 233 234 235 236 239 240 242
#> [163] 244 246 248 249 250 251 252 254 255 258 260 263 264 266 269 273 277 280
#> [181] 281 283 284 285 287 289 291 292 294 295 296 297 298 301 304 306 309 310
#> [199] 311 312 314 315 316 317 318 319 320 322 323 324 326 327 329 333 334 335
#> [217] 336 337 339 341 342 343 345 346 348 349 351 352 353 355 356 357 358 360
#> [235] 361 364 366 367 368 369 370 371 372 373 374 376 377 381 382 383 384 385
#> [253] 386 388 390 391 392 393 394 396 397 398 399 401 402 403 404 405 406 407
#> [271] 409 410 411 412 414 415 417 420 421 422 423 424 425 428 430 431 432 435
#> [289] 437 438 440 441 443 444 448 449 451 453 455 456 458 460 461 462 463 465
#> [307] 467 468 469 470 471 472 474 479 480 483 486 487 488 489 490 491 493 494
#> [325] 495 496 498 499 500
#> 
#> [[2]]$train[[3]]
#>   [1]   1   3   5   6   8  12  13  16  17  18  19  20  21  23  26  28  30  31
#>  [19]  32  34  36  37  38  39  41  42  43  47  48  49  50  53  55  56  57  58
#>  [37]  59  60  61  62  63  65  66  67  70  72  74  75  77  78  79  80  82  83
#>  [55]  84  85  87  88  90  91  93  96  98  99 100 101 102 103 104 105 106 109
#>  [73] 110 113 116 119 120 124 125 126 127 128 129 131 132 133 135 137 138 140
#>  [91] 141 143 144 146 148 149 153 154 156 157 158 159 160 161 162 165 167 170
#> [109] 171 172 173 174 175 177 178 180 181 182 183 184 185 186 187 188 189 191
#> [127] 192 193 195 196 198 201 202 206 207 208 209 210 212 213 215 216 217 218
#> [145] 219 220 223 224 225 227 228 229 231 232 234 235 237 238 239 240 241 242
#> [163] 243 244 245 246 247 248 251 252 253 254 256 257 258 259 261 262 265 267
#> [181] 268 270 271 272 273 274 275 276 278 279 280 281 282 285 286 288 289 290
#> [199] 293 297 298 299 300 302 303 305 307 308 311 312 313 316 319 320 321 322
#> [217] 325 327 328 329 330 331 332 334 335 336 338 339 340 341 343 344 347 348
#> [235] 350 351 353 354 357 358 359 361 362 363 365 367 368 369 370 371 373 375
#> [253] 376 378 379 380 381 383 386 387 389 393 394 395 397 398 400 401 404 405
#> [271] 406 407 408 411 413 415 416 417 418 419 420 423 426 427 428 429 430 431
#> [289] 433 434 436 437 439 440 441 442 443 444 445 446 447 448 450 452 453 454
#> [307] 456 457 459 461 464 465 466 467 470 471 472 473 474 475 476 477 478 481
#> [325] 482 483 484 485 486 488 491 492 494 496 497 498 499
#> 
#> 
#> 
#> [[3]]
#> [[3]]$test
#> [[3]]$test[[1]]
#>   [1]   2   4  15  16  19  20  35  36  40  47  48  49  50  51  54  55  57  60
#>  [19]  72  73  75  77  78  79  80  82  83  84  86  87  91  94  96  97  99 106
#>  [37] 109 110 111 112 126 128 129 134 137 140 141 143 144 145 147 148 153 154
#>  [55] 158 159 165 167 174 176 178 179 194 195 197 201 202 204 207 208 210 212
#>  [73] 213 214 216 221 222 227 230 235 236 238 239 242 253 257 261 268 272 278
#>  [91] 284 286 288 290 292 296 301 303 307 308 311 312 314 315 316 317 332 336
#> [109] 337 345 348 353 354 362 365 366 370 374 381 386 389 390 392 397 405 406
#> [127] 407 411 414 426 428 430 432 433 435 441 442 443 447 449 452 454 455 463
#> [145] 467 469 470 474 479 481 484 485 487 489 491 496 497
#> 
#> [[3]]$test[[2]]
#>   [1]   5   7   8  12  13  21  22  23  25  27  28  30  32  33  37  39  42  44
#>  [19]  46  56  58  61  63  64  67  69  74  81  85  89  92  93  98 100 101 103
#>  [37] 104 115 118 119 120 122 123 124 136 139 151 161 162 170 173 175 177 180
#>  [55] 181 183 184 185 187 188 189 190 191 192 193 196 198 199 200 203 205 209
#>  [73] 215 217 218 223 224 225 228 229 234 241 243 244 245 246 248 250 251 254
#>  [91] 256 258 262 263 265 266 267 269 274 276 277 279 280 282 285 287 293 294
#> [109] 297 298 304 306 309 313 318 319 320 321 322 325 327 328 334 341 342 344
#> [127] 347 349 352 355 358 359 360 364 367 368 369 373 376 378 379 380 382 383
#> [145] 384 391 393 398 400 401 402 404 408 416 417 420 423 424 427 438 439 444
#> [163] 445 446 448 450 458 459 465 466 468 472 473 475 476 482 486 490 492 493
#> [181] 495 499
#> 
#> [[3]]$test[[3]]
#>   [1]   1   3   6   9  10  11  14  17  18  24  26  29  31  34  38  41  43  45
#>  [19]  52  53  59  62  65  66  68  70  71  76  88  90  95 102 105 107 108 113
#>  [37] 114 116 117 121 125 127 130 131 132 133 135 138 142 146 149 150 152 155
#>  [55] 156 157 160 163 164 166 168 169 171 172 182 186 206 211 219 220 226 231
#>  [73] 232 233 237 240 247 249 252 255 259 260 264 270 271 273 275 281 283 289
#>  [91] 291 295 299 300 302 305 310 323 324 326 329 330 331 333 335 338 339 340
#> [109] 343 346 350 351 356 357 361 363 371 372 375 377 385 387 388 394 395 396
#> [127] 399 403 409 410 412 413 415 418 419 421 422 425 429 431 434 436 437 440
#> [145] 451 453 456 457 460 461 462 464 471 477 478 480 483 488 494 498 500
#> 
#> 
#> [[3]]$train
#> [[3]]$train[[1]]
#>   [1]   1   3   5   6   7   8   9  10  11  12  13  14  17  18  21  22  23  24
#>  [19]  25  26  27  28  29  30  31  32  33  34  37  38  39  41  42  43  44  45
#>  [37]  46  52  53  56  58  59  61  62  63  64  65  66  67  68  69  70  71  74
#>  [55]  76  81  85  88  89  90  92  93  95  98 100 101 102 103 104 105 107 108
#>  [73] 113 114 115 116 117 118 119 120 121 122 123 124 125 127 130 131 132 133
#>  [91] 135 136 138 139 142 146 149 150 151 152 155 156 157 160 161 162 163 164
#> [109] 166 168 169 170 171 172 173 175 177 180 181 182 183 184 185 186 187 188
#> [127] 189 190 191 192 193 196 198 199 200 203 205 206 209 211 215 217 218 219
#> [145] 220 223 224 225 226 228 229 231 232 233 234 237 240 241 243 244 245 246
#> [163] 247 248 249 250 251 252 254 255 256 258 259 260 262 263 264 265 266 267
#> [181] 269 270 271 273 274 275 276 277 279 280 281 282 283 285 287 289 291 293
#> [199] 294 295 297 298 299 300 302 304 305 306 309 310 313 318 319 320 321 322
#> [217] 323 324 325 326 327 328 329 330 331 333 334 335 338 339 340 341 342 343
#> [235] 344 346 347 349 350 351 352 355 356 357 358 359 360 361 363 364 367 368
#> [253] 369 371 372 373 375 376 377 378 379 380 382 383 384 385 387 388 391 393
#> [271] 394 395 396 398 399 400 401 402 403 404 408 409 410 412 413 415 416 417
#> [289] 418 419 420 421 422 423 424 425 427 429 431 434 436 437 438 439 440 444
#> [307] 445 446 448 450 451 453 456 457 458 459 460 461 462 464 465 466 468 471
#> [325] 472 473 475 476 477 478 480 482 483 486 488 490 492 493 494 495 498 499
#> [343] 500
#> 
#> [[3]]$train[[2]]
#>   [1]   1   2   3   4   6   9  10  11  14  15  16  17  18  19  20  24  26  29
#>  [19]  31  34  35  36  38  40  41  43  45  47  48  49  50  51  52  53  54  55
#>  [37]  57  59  60  62  65  66  68  70  71  72  73  75  76  77  78  79  80  82
#>  [55]  83  84  86  87  88  90  91  94  95  96  97  99 102 105 106 107 108 109
#>  [73] 110 111 112 113 114 116 117 121 125 126 127 128 129 130 131 132 133 134
#>  [91] 135 137 138 140 141 142 143 144 145 146 147 148 149 150 152 153 154 155
#> [109] 156 157 158 159 160 163 164 165 166 167 168 169 171 172 174 176 178 179
#> [127] 182 186 194 195 197 201 202 204 206 207 208 210 211 212 213 214 216 219
#> [145] 220 221 222 226 227 230 231 232 233 235 236 237 238 239 240 242 247 249
#> [163] 252 253 255 257 259 260 261 264 268 270 271 272 273 275 278 281 283 284
#> [181] 286 288 289 290 291 292 295 296 299 300 301 302 303 305 307 308 310 311
#> [199] 312 314 315 316 317 323 324 326 329 330 331 332 333 335 336 337 338 339
#> [217] 340 343 345 346 348 350 351 353 354 356 357 361 362 363 365 366 370 371
#> [235] 372 374 375 377 381 385 386 387 388 389 390 392 394 395 396 397 399 403
#> [253] 405 406 407 409 410 411 412 413 414 415 418 419 421 422 425 426 428 429
#> [271] 430 431 432 433 434 435 436 437 440 441 442 443 447 449 451 452 453 454
#> [289] 455 456 457 460 461 462 463 464 467 469 470 471 474 477 478 479 480 481
#> [307] 483 484 485 487 488 489 491 494 496 497 498 500
#> 
#> [[3]]$train[[3]]
#>   [1]   2   4   5   7   8  12  13  15  16  19  20  21  22  23  25  27  28  30
#>  [19]  32  33  35  36  37  39  40  42  44  46  47  48  49  50  51  54  55  56
#>  [37]  57  58  60  61  63  64  67  69  72  73  74  75  77  78  79  80  81  82
#>  [55]  83  84  85  86  87  89  91  92  93  94  96  97  98  99 100 101 103 104
#>  [73] 106 109 110 111 112 115 118 119 120 122 123 124 126 128 129 134 136 137
#>  [91] 139 140 141 143 144 145 147 148 151 153 154 158 159 161 162 165 167 170
#> [109] 173 174 175 176 177 178 179 180 181 183 184 185 187 188 189 190 191 192
#> [127] 193 194 195 196 197 198 199 200 201 202 203 204 205 207 208 209 210 212
#> [145] 213 214 215 216 217 218 221 222 223 224 225 227 228 229 230 234 235 236
#> [163] 238 239 241 242 243 244 245 246 248 250 251 253 254 256 257 258 261 262
#> [181] 263 265 266 267 268 269 272 274 276 277 278 279 280 282 284 285 286 287
#> [199] 288 290 292 293 294 296 297 298 301 303 304 306 307 308 309 311 312 313
#> [217] 314 315 316 317 318 319 320 321 322 325 327 328 332 334 336 337 341 342
#> [235] 344 345 347 348 349 352 353 354 355 358 359 360 362 364 365 366 367 368
#> [253] 369 370 373 374 376 378 379 380 381 382 383 384 386 389 390 391 392 393
#> [271] 397 398 400 401 402 404 405 406 407 408 411 414 416 417 420 423 424 426
#> [289] 427 428 430 432 433 435 438 439 441 442 443 444 445 446 447 448 449 450
#> [307] 452 454 455 458 459 463 465 466 467 468 469 470 472 473 474 475 476 479
#> [325] 481 482 484 485 486 487 489 490 491 492 493 495 496 497 499
#> 
#> 
#> 
#> [[4]]
#> [[4]]$test
#> [[4]]$test[[1]]
#>   [1]   9  12  14  16  24  25  27  30  31  34  36  39  41  42  49  56  60  61
#>  [19]  64  68  69  70  71  74  85  88  89  90  94  95  99 102 104 106 108 109
#>  [37] 112 114 117 119 126 130 133 139 141 145 155 158 159 165 166 171 176 177
#>  [55] 180 181 182 183 185 187 189 190 191 192 196 198 200 202 205 206 207 208
#>  [73] 220 221 224 227 228 230 232 237 239 241 243 244 245 247 249 250 256 262
#>  [91] 269 270 272 273 274 275 279 290 292 293 295 296 298 304 306 310 313 314
#> [109] 316 321 330 335 336 337 338 339 342 345 355 367 368 369 373 374 377 378
#> [127] 381 382 383 385 387 388 389 390 391 397 400 412 413 416 417 418 422 424
#> [145] 425 426 427 429 436 439 442 450 452 460 461 471 473 475 477 478 480 481
#> [163] 484 485 499
#> 
#> [[4]]$test[[2]]
#>   [1]   5   6   7   8  13  15  20  23  32  33  37  43  44  45  46  47  50  51
#>  [19]  54  57  58  62  65  67  73  75  76  78  79  80  83  86  91  93 100 103
#>  [37] 118 120 121 123 127 131 132 135 136 137 138 140 142 144 147 150 151 152
#>  [55] 154 160 161 162 163 167 168 170 184 186 188 193 197 199 203 204 209 210
#>  [73] 213 214 215 217 218 223 225 226 234 236 240 242 248 252 253 257 258 260
#>  [91] 261 265 267 277 280 281 283 284 285 286 288 297 299 302 307 308 312 319
#> [109] 320 323 324 325 326 329 332 334 341 347 348 351 354 356 358 360 361 365
#> [127] 370 372 376 379 393 394 398 401 404 405 407 409 414 419 421 428 432 433
#> [145] 435 437 441 443 445 446 447 448 449 451 453 455 456 458 459 464 465 466
#> [163] 476 482 483 486 487 488 491 493 494 496 497 498 500
#> 
#> [[4]]$test[[3]]
#>   [1]   1   2   3   4  10  11  17  18  19  21  22  26  28  29  35  38  40  48
#>  [19]  52  53  55  59  63  66  72  77  81  82  84  87  92  96  97  98 101 105
#>  [37] 107 110 111 113 115 116 122 124 125 128 129 134 143 146 148 149 153 156
#>  [55] 157 164 169 172 173 174 175 178 179 194 195 201 211 212 216 219 222 229
#>  [73] 231 233 235 238 246 251 254 255 259 263 264 266 268 271 276 278 282 287
#>  [91] 289 291 294 300 301 303 305 309 311 315 317 318 322 327 328 331 333 340
#> [109] 343 344 346 349 350 352 353 357 359 362 363 364 366 371 375 380 384 386
#> [127] 392 395 396 399 402 403 406 408 410 411 415 420 423 430 431 434 438 440
#> [145] 444 454 457 462 463 467 468 469 470 472 474 479 489 490 492 495
#> 
#> 
#> [[4]]$train
#> [[4]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8  10  11  13  15  17  18  19  20  21  22
#>  [19]  23  26  28  29  32  33  35  37  38  40  43  44  45  46  47  48  50  51
#>  [37]  52  53  54  55  57  58  59  62  63  65  66  67  72  73  75  76  77  78
#>  [55]  79  80  81  82  83  84  86  87  91  92  93  96  97  98 100 101 103 105
#>  [73] 107 110 111 113 115 116 118 120 121 122 123 124 125 127 128 129 131 132
#>  [91] 134 135 136 137 138 140 142 143 144 146 147 148 149 150 151 152 153 154
#> [109] 156 157 160 161 162 163 164 167 168 169 170 172 173 174 175 178 179 184
#> [127] 186 188 193 194 195 197 199 201 203 204 209 210 211 212 213 214 215 216
#> [145] 217 218 219 222 223 225 226 229 231 233 234 235 236 238 240 242 246 248
#> [163] 251 252 253 254 255 257 258 259 260 261 263 264 265 266 267 268 271 276
#> [181] 277 278 280 281 282 283 284 285 286 287 288 289 291 294 297 299 300 301
#> [199] 302 303 305 307 308 309 311 312 315 317 318 319 320 322 323 324 325 326
#> [217] 327 328 329 331 332 333 334 340 341 343 344 346 347 348 349 350 351 352
#> [235] 353 354 356 357 358 359 360 361 362 363 364 365 366 370 371 372 375 376
#> [253] 379 380 384 386 392 393 394 395 396 398 399 401 402 403 404 405 406 407
#> [271] 408 409 410 411 414 415 419 420 421 423 428 430 431 432 433 434 435 437
#> [289] 438 440 441 443 444 445 446 447 448 449 451 453 454 455 456 457 458 459
#> [307] 462 463 464 465 466 467 468 469 470 472 474 476 479 482 483 486 487 488
#> [325] 489 490 491 492 493 494 495 496 497 498 500
#> 
#> [[4]]$train[[2]]
#>   [1]   1   2   3   4   9  10  11  12  14  16  17  18  19  21  22  24  25  26
#>  [19]  27  28  29  30  31  34  35  36  38  39  40  41  42  48  49  52  53  55
#>  [37]  56  59  60  61  63  64  66  68  69  70  71  72  74  77  81  82  84  85
#>  [55]  87  88  89  90  92  94  95  96  97  98  99 101 102 104 105 106 107 108
#>  [73] 109 110 111 112 113 114 115 116 117 119 122 124 125 126 128 129 130 133
#>  [91] 134 139 141 143 145 146 148 149 153 155 156 157 158 159 164 165 166 169
#> [109] 171 172 173 174 175 176 177 178 179 180 181 182 183 185 187 189 190 191
#> [127] 192 194 195 196 198 200 201 202 205 206 207 208 211 212 216 219 220 221
#> [145] 222 224 227 228 229 230 231 232 233 235 237 238 239 241 243 244 245 246
#> [163] 247 249 250 251 254 255 256 259 262 263 264 266 268 269 270 271 272 273
#> [181] 274 275 276 278 279 282 287 289 290 291 292 293 294 295 296 298 300 301
#> [199] 303 304 305 306 309 310 311 313 314 315 316 317 318 321 322 327 328 330
#> [217] 331 333 335 336 337 338 339 340 342 343 344 345 346 349 350 352 353 355
#> [235] 357 359 362 363 364 366 367 368 369 371 373 374 375 377 378 380 381 382
#> [253] 383 384 385 386 387 388 389 390 391 392 395 396 397 399 400 402 403 406
#> [271] 408 410 411 412 413 415 416 417 418 420 422 423 424 425 426 427 429 430
#> [289] 431 434 436 438 439 440 442 444 450 452 454 457 460 461 462 463 467 468
#> [307] 469 470 471 472 473 474 475 477 478 479 480 481 484 485 489 490 492 495
#> [325] 499
#> 
#> [[4]]$train[[3]]
#>   [1]   5   6   7   8   9  12  13  14  15  16  20  23  24  25  27  30  31  32
#>  [19]  33  34  36  37  39  41  42  43  44  45  46  47  49  50  51  54  56  57
#>  [37]  58  60  61  62  64  65  67  68  69  70  71  73  74  75  76  78  79  80
#>  [55]  83  85  86  88  89  90  91  93  94  95  99 100 102 103 104 106 108 109
#>  [73] 112 114 117 118 119 120 121 123 126 127 130 131 132 133 135 136 137 138
#>  [91] 139 140 141 142 144 145 147 150 151 152 154 155 158 159 160 161 162 163
#> [109] 165 166 167 168 170 171 176 177 180 181 182 183 184 185 186 187 188 189
#> [127] 190 191 192 193 196 197 198 199 200 202 203 204 205 206 207 208 209 210
#> [145] 213 214 215 217 218 220 221 223 224 225 226 227 228 230 232 234 236 237
#> [163] 239 240 241 242 243 244 245 247 248 249 250 252 253 256 257 258 260 261
#> [181] 262 265 267 269 270 272 273 274 275 277 279 280 281 283 284 285 286 288
#> [199] 290 292 293 295 296 297 298 299 302 304 306 307 308 310 312 313 314 316
#> [217] 319 320 321 323 324 325 326 329 330 332 334 335 336 337 338 339 341 342
#> [235] 345 347 348 351 354 355 356 358 360 361 365 367 368 369 370 372 373 374
#> [253] 376 377 378 379 381 382 383 385 387 388 389 390 391 393 394 397 398 400
#> [271] 401 404 405 407 409 412 413 414 416 417 418 419 421 422 424 425 426 427
#> [289] 428 429 432 433 435 436 437 439 441 442 443 445 446 447 448 449 450 451
#> [307] 452 453 455 456 458 459 460 461 464 465 466 471 473 475 476 477 478 480
#> [325] 481 482 483 484 485 486 487 488 491 493 494 496 497 498 499 500
#> 
#> 
#> 
#> [[5]]
#> [[5]]$test
#> [[5]]$test[[1]]
#>   [1]   4   5   6  15  18  25  26  29  30  32  35  38  40  43  44  46  51  52
#>  [19]  54  55  57  60  62  65  67  69  71  72  73  76  81  89  91  95  96  98
#>  [37] 104 112 113 115 120 122 123 127 131 134 135 136 137 138 144 145 147 149
#>  [55] 151 160 161 166 167 170 178 179 186 187 189 191 194 197 198 201 205 206
#>  [73] 208 209 210 214 217 218 220 222 225 227 240 241 248 249 251 252 254 255
#>  [91] 261 265 268 270 273 277 280 281 291 292 293 294 295 297 299 301 303 304
#> [109] 308 310 313 324 329 333 338 341 351 358 359 362 365 369 370 372 376 379
#> [127] 380 384 385 387 388 390 392 406 407 414 416 419 420 423 425 427 430 431
#> [145] 434 438 440 448 452 454 455 456 457 458 459 465 468 475 479 480 484 487
#> [163] 488 490 491 493 495 497
#> 
#> [[5]]$test[[2]]
#>   [1]   2   7   9  10  12  14  16  17  23  31  36  37  45  48  50  56  58  59
#>  [19]  61  63  66  68  75  77  79  80  82  85  90  93  94  97  99 101 103 107
#>  [37] 110 116 118 121 124 126 132 139 141 143 146 148 152 154 155 156 157 159
#>  [55] 162 168 171 174 175 176 177 182 183 185 190 193 196 199 200 207 213 219
#>  [73] 223 224 226 228 229 231 232 233 234 236 237 238 239 242 243 245 250 258
#>  [91] 260 262 267 269 271 272 276 278 282 284 289 300 307 309 316 321 322 327
#> [109] 330 331 332 336 337 339 344 348 353 355 357 361 363 364 366 368 371 386
#> [127] 389 393 394 396 398 399 400 405 409 411 412 415 417 418 421 426 428 429
#> [145] 437 439 441 442 445 446 460 461 462 463 464 466 471 472 474 482 483 485
#> [163] 492 494 496 499 500
#> 
#> [[5]]$test[[3]]
#>   [1]   1   3   8  11  13  19  20  21  22  24  27  28  33  34  39  41  42  47
#>  [19]  49  53  64  70  74  78  83  84  86  87  88  92 100 102 105 106 108 109
#>  [37] 111 114 117 119 125 128 129 130 133 140 142 150 153 158 163 164 165 169
#>  [55] 172 173 180 181 184 188 192 195 202 203 204 211 212 215 216 221 230 235
#>  [73] 244 246 247 253 256 257 259 263 264 266 274 275 279 283 285 286 287 288
#>  [91] 290 296 298 302 305 306 311 312 314 315 317 318 319 320 323 325 326 328
#> [109] 334 335 340 342 343 345 346 347 349 350 352 354 356 360 367 373 374 375
#> [127] 377 378 381 382 383 391 395 397 401 402 403 404 408 410 413 422 424 432
#> [145] 433 435 436 443 444 447 449 450 451 453 467 469 470 473 476 477 478 481
#> [163] 486 489 498
#> 
#> 
#> [[5]]$train
#> [[5]]$train[[1]]
#>   [1]   1   2   3   7   8   9  10  11  12  13  14  16  17  19  20  21  22  23
#>  [19]  24  27  28  31  33  34  36  37  39  41  42  45  47  48  49  50  53  56
#>  [37]  58  59  61  63  64  66  68  70  74  75  77  78  79  80  82  83  84  85
#>  [55]  86  87  88  90  92  93  94  97  99 100 101 102 103 105 106 107 108 109
#>  [73] 110 111 114 116 117 118 119 121 124 125 126 128 129 130 132 133 139 140
#>  [91] 141 142 143 146 148 150 152 153 154 155 156 157 158 159 162 163 164 165
#> [109] 168 169 171 172 173 174 175 176 177 180 181 182 183 184 185 188 190 192
#> [127] 193 195 196 199 200 202 203 204 207 211 212 213 215 216 219 221 223 224
#> [145] 226 228 229 230 231 232 233 234 235 236 237 238 239 242 243 244 245 246
#> [163] 247 250 253 256 257 258 259 260 262 263 264 266 267 269 271 272 274 275
#> [181] 276 278 279 282 283 284 285 286 287 288 289 290 296 298 300 302 305 306
#> [199] 307 309 311 312 314 315 316 317 318 319 320 321 322 323 325 326 327 328
#> [217] 330 331 332 334 335 336 337 339 340 342 343 344 345 346 347 348 349 350
#> [235] 352 353 354 355 356 357 360 361 363 364 366 367 368 371 373 374 375 377
#> [253] 378 381 382 383 386 389 391 393 394 395 396 397 398 399 400 401 402 403
#> [271] 404 405 408 409 410 411 412 413 415 417 418 421 422 424 426 428 429 432
#> [289] 433 435 436 437 439 441 442 443 444 445 446 447 449 450 451 453 460 461
#> [307] 462 463 464 466 467 469 470 471 472 473 474 476 477 478 481 482 483 485
#> [325] 486 489 492 494 496 498 499 500
#> 
#> [[5]]$train[[2]]
#>   [1]   1   3   4   5   6   8  11  13  15  18  19  20  21  22  24  25  26  27
#>  [19]  28  29  30  32  33  34  35  38  39  40  41  42  43  44  46  47  49  51
#>  [37]  52  53  54  55  57  60  62  64  65  67  69  70  71  72  73  74  76  78
#>  [55]  81  83  84  86  87  88  89  91  92  95  96  98 100 102 104 105 106 108
#>  [73] 109 111 112 113 114 115 117 119 120 122 123 125 127 128 129 130 131 133
#>  [91] 134 135 136 137 138 140 142 144 145 147 149 150 151 153 158 160 161 163
#> [109] 164 165 166 167 169 170 172 173 178 179 180 181 184 186 187 188 189 191
#> [127] 192 194 195 197 198 201 202 203 204 205 206 208 209 210 211 212 214 215
#> [145] 216 217 218 220 221 222 225 227 230 235 240 241 244 246 247 248 249 251
#> [163] 252 253 254 255 256 257 259 261 263 264 265 266 268 270 273 274 275 277
#> [181] 279 280 281 283 285 286 287 288 290 291 292 293 294 295 296 297 298 299
#> [199] 301 302 303 304 305 306 308 310 311 312 313 314 315 317 318 319 320 323
#> [217] 324 325 326 328 329 333 334 335 338 340 341 342 343 345 346 347 349 350
#> [235] 351 352 354 356 358 359 360 362 365 367 369 370 372 373 374 375 376 377
#> [253] 378 379 380 381 382 383 384 385 387 388 390 391 392 395 397 401 402 403
#> [271] 404 406 407 408 410 413 414 416 419 420 422 423 424 425 427 430 431 432
#> [289] 433 434 435 436 438 440 443 444 447 448 449 450 451 452 453 454 455 456
#> [307] 457 458 459 465 467 468 469 470 473 475 476 477 478 479 480 481 484 486
#> [325] 487 488 489 490 491 493 495 497 498
#> 
#> [[5]]$train[[3]]
#>   [1]   2   4   5   6   7   9  10  12  14  15  16  17  18  23  25  26  29  30
#>  [19]  31  32  35  36  37  38  40  43  44  45  46  48  50  51  52  54  55  56
#>  [37]  57  58  59  60  61  62  63  65  66  67  68  69  71  72  73  75  76  77
#>  [55]  79  80  81  82  85  89  90  91  93  94  95  96  97  98  99 101 103 104
#>  [73] 107 110 112 113 115 116 118 120 121 122 123 124 126 127 131 132 134 135
#>  [91] 136 137 138 139 141 143 144 145 146 147 148 149 151 152 154 155 156 157
#> [109] 159 160 161 162 166 167 168 170 171 174 175 176 177 178 179 182 183 185
#> [127] 186 187 189 190 191 193 194 196 197 198 199 200 201 205 206 207 208 209
#> [145] 210 213 214 217 218 219 220 222 223 224 225 226 227 228 229 231 232 233
#> [163] 234 236 237 238 239 240 241 242 243 245 248 249 250 251 252 254 255 258
#> [181] 260 261 262 265 267 268 269 270 271 272 273 276 277 278 280 281 282 284
#> [199] 289 291 292 293 294 295 297 299 300 301 303 304 307 308 309 310 313 316
#> [217] 321 322 324 327 329 330 331 332 333 336 337 338 339 341 344 348 351 353
#> [235] 355 357 358 359 361 362 363 364 365 366 368 369 370 371 372 376 379 380
#> [253] 384 385 386 387 388 389 390 392 393 394 396 398 399 400 405 406 407 409
#> [271] 411 412 414 415 416 417 418 419 420 421 423 425 426 427 428 429 430 431
#> [289] 434 437 438 439 440 441 442 445 446 448 452 454 455 456 457 458 459 460
#> [307] 461 462 463 464 465 466 468 471 472 474 475 479 480 482 483 484 485 487
#> [325] 488 490 491 492 493 494 495 496 497 499 500
#> 
#> 
#>