(CAST) Repeated "leave-location-and-time-out" resampling

Create spatial, temporal or spatio-temporal Folds for cross validation

Details

Using "class" is helpful in the case that data are clustered in space and are categorical. E.g This is the case for land cover classifications when training data come as training polygons. In this case the data should be split in a way that entire polygons are held back (spacevar="polygonID") but at the same time the distribution of classes should be similar in each fold (class="LUC").

Note

Standard k-fold cross-validation can lead to considerable misinterpretation in spatial-temporal modelling tasks. This function can be used to prepare a Leave-Location-Out, Leave-Time-Out or Leave-Location-and-Time-Out cross-validation as target-oriented validation strategies for spatial-temporal prediction tasks. See Meyer et al. (2018) for further information.

mlr3spatiotempcv notes

The 'Description', 'Details' and 'Note' fields are inherited from the respective upstream function.

For a list of available arguments, please see CAST::CreateSpacetimeFolds.

References

Zhao Y, Karypis G (2002). “Evaluation of Hierarchical Clustering Algorithms for Document Datasets.” 11th Conference of Information and Knowledge Management (CIKM), 51-524. http://glaros.dtc.umn.edu/gkhome/node/167.

Super class

mlr3::Resampling -> ResamplingRepeatedSptCVCstf

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSptCVCstf$new()

ResamplingRepeatedSptCVCstf$folds()

ResamplingRepeatedSptCVCstf$repeats()

ResamplingRepeatedSptCVCstf$instantiate()

ResamplingRepeatedSptCVCstf$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spacetime Folds" resampling instance.

For a list of available arguments, please see CAST::CreateSpacetimeFolds.

Usage

ResamplingRepeatedSptCVCstf$new(id = "repeated_sptcv_cstf")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSptCVCstf$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSptCVCstf$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSptCVCstf$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSptCVCstf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(mlr3spatiotempcv)
task = tsk("cookfarm")

# Instantiate Resampling
rrcv = rsmp("repeated_sptcv_cstf", folds = 3, repeats = 5, time_var = "Date")
rrcv$instantiate(task)
# Individual sets:
rrcv$iters

#> [1] 15

rrcv$folds(1:6)

#> [1] 1 2 3 4 5 1

rrcv$repeats(1:6)

#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)

#>   [1]   1   2   4   5   6   7   8   9  10  14  15  18  19  20  21  22  23  24
#>  [19]  25  29  30  32  33  36  38  39  41  42  44  45  46  47  49  51  54  55
#>  [37]  57  58  61  62  64  66  67  68  69  73  74  76  77  78  79  80  81  83
#>  [55]  84  85  86  87  88  89  90  91  92  93  95  98  99 100 101 102 103 105
#>  [73] 106 108 110 111 112 114 115 116 117 118 120 121 123 124 125 129 130 133
#>  [91] 136 137 138 141 142 143 144 145 146 147 149 150 151 153 154 157 158 159
#> [109] 160 161 162 163 164 166 167 171 172 173 174 176 180 181 183 185 186 187
#> [127] 188 189 190 191 195 196 197 198 204 205 206 207 209 212 213 214 217 218
#> [145] 220 221 222 224 226 227 228 229 232 236 237 238 239 240 241 242 245 246
#> [163] 247 250 251 255 256 257 258 259 260 262 263 264 266 268 269 270 271 272
#> [181] 276 277 278 279 280 281 284 285 286 287 288 289 290 291 292 293 294 295
#> [199] 296 297 298 301 302 306 307 308 309 311 312 314 315 316 318 319 321 322
#> [217] 323 324 326 327 328 330 331 332 333 334 335 336 338 339 341 343 345 346
#> [235] 347 348 349 353 354 357 358 359 360 361 362 365 366 367 368 369 371 372
#> [253] 373 374 375 376 377 378 379 380 381 382 384 385 386 388 389 390 393 394
#> [271] 396 397 399 400 403 405 406 407 409 411 413 414 415 418 419 420 423 424
#> [289] 425 426 427 428 429 430 431 432 434 435 436 437 439 441 442 443 445 446
#> [307] 447 448 450 452 454 455 456 457 458 461 464 465 467 472 473 474 475 476
#> [325] 478 479 480 481 482 483 484 485 488 489 490 491 494 495 497 500

rrcv$test_set(1)

#>   [1]   3  11  12  13  16  17  26  27  28  31  34  35  37  40  43  48  50  52
#>  [19]  53  56  59  60  63  65  70  71  72  75  82  94  96  97 104 107 109 113
#>  [37] 119 122 126 127 128 131 132 134 135 139 140 148 152 155 156 165 168 169
#>  [55] 170 175 177 178 179 182 184 192 193 194 199 200 201 202 203 208 210 211
#>  [73] 215 216 219 223 225 230 231 233 234 235 243 244 248 249 252 253 254 261
#>  [91] 265 267 273 274 275 282 283 299 300 303 304 305 310 313 317 320 325 329
#> [109] 337 340 342 344 350 351 352 355 356 363 364 370 383 387 391 392 395 398
#> [127] 401 402 404 408 410 412 416 417 421 422 433 438 440 444 449 451 453 459
#> [145] 460 462 463 466 468 469 470 471 477 486 487 492 493 496 498 499

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance # table

#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>   [1]   3  11  12  13  16  17  26  27  28  31  34  35  37  40  43  48  50  52
#>  [19]  53  56  59  60  63  65  70  71  72  75  82  94  96  97 104 107 109 113
#>  [37] 119 122 126 127 128 131 132 134 135 139 140 148 152 155 156 165 168 169
#>  [55] 170 175 177 178 179 182 184 192 193 194 199 200 201 202 203 208 210 211
#>  [73] 215 216 219 223 225 230 231 233 234 235 243 244 248 249 252 253 254 261
#>  [91] 265 267 273 274 275 282 283 299 300 303 304 305 310 313 317 320 325 329
#> [109] 337 340 342 344 350 351 352 355 356 363 364 370 383 387 391 392 395 398
#> [127] 401 402 404 408 410 412 416 417 421 422 433 438 440 444 449 451 453 459
#> [145] 460 462 463 466 468 469 470 471 477 486 487 492 493 496 498 499
#> 
#> [[1]]$test[[2]]
#>   [1]   4   6   8   9  10  14  23  24  25  32  33  39  42  45  46  47  49  54
#>  [19]  57  58  62  64  68  73  74  76  78  79  81  83  84  87  89  92  93  99
#>  [37] 100 102 103 106 111 115 118 120 121 125 129 133 136 137 143 144 145 146
#>  [55] 151 153 162 163 166 171 172 174 176 180 181 183 187 188 189 191 195 197
#>  [73] 204 207 209 213 222 229 239 240 241 242 245 246 250 251 257 259 260 263
#>  [91] 264 269 270 271 272 276 277 278 281 285 286 287 288 290 292 293 294 296
#> [109] 297 307 308 312 316 318 321 326 327 328 331 332 334 335 339 343 345 353
#> [127] 359 360 362 365 366 367 368 369 371 372 375 376 377 381 385 386 388 390
#> [145] 393 396 399 400 411 415 418 419 420 425 426 428 431 434 435 436 437 441
#> [163] 443 446 454 455 456 457 458 464 467 473 475 476 478 479 484 490 494 495
#> 
#> [[1]]$test[[3]]
#>   [1]   1   2   5   7  15  18  19  20  21  22  29  30  36  38  41  44  51  55
#>  [19]  61  66  67  69  77  80  85  86  88  90  91  95  98 101 105 108 110 112
#>  [37] 114 116 117 123 124 130 138 141 142 147 149 150 154 157 158 159 160 161
#>  [55] 164 167 173 185 186 190 196 198 205 206 212 214 217 218 220 221 224 226
#>  [73] 227 228 232 236 237 238 247 255 256 258 262 266 268 279 280 284 289 291
#>  [91] 295 298 301 302 306 309 311 314 315 319 322 323 324 330 333 336 338 341
#> [109] 346 347 348 349 354 357 358 361 373 374 378 379 380 382 384 389 394 397
#> [127] 403 405 406 407 409 413 414 423 424 427 429 430 432 439 442 445 447 448
#> [145] 450 452 461 465 472 474 480 481 482 483 485 488 489 491 497 500
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   4   5   6   7   8   9  10  14  15  18  19  20  21  22  23  24
#>  [19]  25  29  30  32  33  36  38  39  41  42  44  45  46  47  49  51  54  55
#>  [37]  57  58  61  62  64  66  67  68  69  73  74  76  77  78  79  80  81  83
#>  [55]  84  85  86  87  88  89  90  91  92  93  95  98  99 100 101 102 103 105
#>  [73] 106 108 110 111 112 114 115 116 117 118 120 121 123 124 125 129 130 133
#>  [91] 136 137 138 141 142 143 144 145 146 147 149 150 151 153 154 157 158 159
#> [109] 160 161 162 163 164 166 167 171 172 173 174 176 180 181 183 185 186 187
#> [127] 188 189 190 191 195 196 197 198 204 205 206 207 209 212 213 214 217 218
#> [145] 220 221 222 224 226 227 228 229 232 236 237 238 239 240 241 242 245 246
#> [163] 247 250 251 255 256 257 258 259 260 262 263 264 266 268 269 270 271 272
#> [181] 276 277 278 279 280 281 284 285 286 287 288 289 290 291 292 293 294 295
#> [199] 296 297 298 301 302 306 307 308 309 311 312 314 315 316 318 319 321 322
#> [217] 323 324 326 327 328 330 331 332 333 334 335 336 338 339 341 343 345 346
#> [235] 347 348 349 353 354 357 358 359 360 361 362 365 366 367 368 369 371 372
#> [253] 373 374 375 376 377 378 379 380 381 382 384 385 386 388 389 390 393 394
#> [271] 396 397 399 400 403 405 406 407 409 411 413 414 415 418 419 420 423 424
#> [289] 425 426 427 428 429 430 431 432 434 435 436 437 439 441 442 443 445 446
#> [307] 447 448 450 452 454 455 456 457 458 461 464 465 467 472 473 474 475 476
#> [325] 478 479 480 481 482 483 484 485 488 489 490 491 494 495 497 500
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   5   7  11  12  13  15  16  17  18  19  20  21  22  26  27
#>  [19]  28  29  30  31  34  35  36  37  38  40  41  43  44  48  50  51  52  53
#>  [37]  55  56  59  60  61  63  65  66  67  69  70  71  72  75  77  80  82  85
#>  [55]  86  88  90  91  94  95  96  97  98 101 104 105 107 108 109 110 112 113
#>  [73] 114 116 117 119 122 123 124 126 127 128 130 131 132 134 135 138 139 140
#>  [91] 141 142 147 148 149 150 152 154 155 156 157 158 159 160 161 164 165 167
#> [109] 168 169 170 173 175 177 178 179 182 184 185 186 190 192 193 194 196 198
#> [127] 199 200 201 202 203 205 206 208 210 211 212 214 215 216 217 218 219 220
#> [145] 221 223 224 225 226 227 228 230 231 232 233 234 235 236 237 238 243 244
#> [163] 247 248 249 252 253 254 255 256 258 261 262 265 266 267 268 273 274 275
#> [181] 279 280 282 283 284 289 291 295 298 299 300 301 302 303 304 305 306 309
#> [199] 310 311 313 314 315 317 319 320 322 323 324 325 329 330 333 336 337 338
#> [217] 340 341 342 344 346 347 348 349 350 351 352 354 355 356 357 358 361 363
#> [235] 364 370 373 374 378 379 380 382 383 384 387 389 391 392 394 395 397 398
#> [253] 401 402 403 404 405 406 407 408 409 410 412 413 414 416 417 421 422 423
#> [271] 424 427 429 430 432 433 438 439 440 442 444 445 447 448 449 450 451 452
#> [289] 453 459 460 461 462 463 465 466 468 469 470 471 472 474 477 480 481 482
#> [307] 483 485 486 487 488 489 491 492 493 496 497 498 499 500
#> 
#> [[1]]$train[[3]]
#>   [1]   3   4   6   8   9  10  11  12  13  14  16  17  23  24  25  26  27  28
#>  [19]  31  32  33  34  35  37  39  40  42  43  45  46  47  48  49  50  52  53
#>  [37]  54  56  57  58  59  60  62  63  64  65  68  70  71  72  73  74  75  76
#>  [55]  78  79  81  82  83  84  87  89  92  93  94  96  97  99 100 102 103 104
#>  [73] 106 107 109 111 113 115 118 119 120 121 122 125 126 127 128 129 131 132
#>  [91] 133 134 135 136 137 139 140 143 144 145 146 148 151 152 153 155 156 162
#> [109] 163 165 166 168 169 170 171 172 174 175 176 177 178 179 180 181 182 183
#> [127] 184 187 188 189 191 192 193 194 195 197 199 200 201 202 203 204 207 208
#> [145] 209 210 211 213 215 216 219 222 223 225 229 230 231 233 234 235 239 240
#> [163] 241 242 243 244 245 246 248 249 250 251 252 253 254 257 259 260 261 263
#> [181] 264 265 267 269 270 271 272 273 274 275 276 277 278 281 282 283 285 286
#> [199] 287 288 290 292 293 294 296 297 299 300 303 304 305 307 308 310 312 313
#> [217] 316 317 318 320 321 325 326 327 328 329 331 332 334 335 337 339 340 342
#> [235] 343 344 345 350 351 352 353 355 356 359 360 362 363 364 365 366 367 368
#> [253] 369 370 371 372 375 376 377 381 383 385 386 387 388 390 391 392 393 395
#> [271] 396 398 399 400 401 402 404 408 410 411 412 415 416 417 418 419 420 421
#> [289] 422 425 426 428 431 433 434 435 436 437 438 440 441 443 444 446 449 451
#> [307] 453 454 455 456 457 458 459 460 462 463 464 466 467 468 469 470 471 473
#> [325] 475 476 477 478 479 484 486 487 490 492 493 494 495 496 498 499
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>   [1]   5   6   8  10  12  14  27  30  34  36  37  40  45  46  53  55  58  61
#>  [19]  62  69  70  72  77  78  80  82  85  86  88  90  91  95  99 102 106 107
#>  [37] 108 112 116 119 120 121 123 127 135 136 141 154 156 163 167 169 170 172
#>  [55] 177 179 180 181 186 188 192 193 194 195 196 197 198 200 205 206 208 209
#>  [73] 212 215 220 223 224 225 229 232 239 240 242 244 245 248 249 250 252 253
#>  [91] 254 256 260 268 279 280 283 287 290 303 306 309 310 312 316 317 319 322
#> [109] 323 324 329 331 335 336 337 343 344 346 348 349 353 357 367 368 374 381
#> [127] 382 384 388 390 392 393 394 396 398 401 403 405 407 412 423 424 425 427
#> [145] 431 434 436 438 439 440 441 442 443 449 450 456 461 462 464 469 471 474
#> [163] 476 477 481 482 486
#> 
#> [[2]]$test[[2]]
#>   [1]   2   3   4  13  17  21  23  24  28  29  31  33  35  39  43  44  51  54
#>  [19]  57  60  63  64  66  67  68  75  76  79  81  83  84  87  96 101 109 113
#>  [37] 115 117 122 124 130 134 142 144 145 147 148 149 150 153 155 157 158 160
#>  [55] 162 165 168 178 182 183 185 189 190 191 199 204 210 211 213 216 217 219
#>  [73] 227 228 230 238 247 251 258 259 261 263 264 265 266 269 270 272 273 277
#>  [91] 282 285 286 288 293 294 296 299 301 302 307 313 314 320 321 326 332 338
#> [109] 340 342 345 347 350 351 352 354 356 358 359 360 362 363 366 369 371 372
#> [127] 375 379 380 383 386 389 395 399 402 404 409 411 414 416 417 418 419 420
#> [145] 422 428 433 437 445 446 447 448 453 460 467 470 472 473 475 478 479 480
#> [163] 484 488 494 499
#> 
#> [[2]]$test[[3]]
#>   [1]   1   7   9  11  15  16  18  19  20  22  25  26  32  38  41  42  47  48
#>  [19]  49  50  52  56  59  65  71  73  74  89  92  93  94  97  98 100 103 104
#>  [37] 105 110 111 114 118 125 126 128 129 131 132 133 137 138 139 140 143 146
#>  [55] 151 152 159 161 164 166 171 173 174 175 176 184 187 201 202 203 207 214
#>  [73] 218 221 222 226 231 233 234 235 236 237 241 243 246 255 257 262 267 271
#>  [91] 274 275 276 278 281 284 289 291 292 295 297 298 300 304 305 308 311 315
#> [109] 318 325 327 328 330 333 334 339 341 355 361 364 365 370 373 376 377 378
#> [127] 385 387 391 397 400 406 408 410 413 415 421 426 429 430 432 435 444 451
#> [145] 452 454 455 457 458 459 463 465 466 468 483 485 487 489 490 491 492 493
#> [163] 495 496 497 498 500
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   7   9  11  13  15  16  17  18  19  20  21  22  23  24
#>  [19]  25  26  28  29  31  32  33  35  38  39  41  42  43  44  47  48  49  50
#>  [37]  51  52  54  56  57  59  60  63  64  65  66  67  68  71  73  74  75  76
#>  [55]  79  81  83  84  87  89  92  93  94  96  97  98 100 101 103 104 105 109
#>  [73] 110 111 113 114 115 117 118 122 124 125 126 128 129 130 131 132 133 134
#>  [91] 137 138 139 140 142 143 144 145 146 147 148 149 150 151 152 153 155 157
#> [109] 158 159 160 161 162 164 165 166 168 171 173 174 175 176 178 182 183 184
#> [127] 185 187 189 190 191 199 201 202 203 204 207 210 211 213 214 216 217 218
#> [145] 219 221 222 226 227 228 230 231 233 234 235 236 237 238 241 243 246 247
#> [163] 251 255 257 258 259 261 262 263 264 265 266 267 269 270 271 272 273 274
#> [181] 275 276 277 278 281 282 284 285 286 288 289 291 292 293 294 295 296 297
#> [199] 298 299 300 301 302 304 305 307 308 311 313 314 315 318 320 321 325 326
#> [217] 327 328 330 332 333 334 338 339 340 341 342 345 347 350 351 352 354 355
#> [235] 356 358 359 360 361 362 363 364 365 366 369 370 371 372 373 375 376 377
#> [253] 378 379 380 383 385 386 387 389 391 395 397 399 400 402 404 406 408 409
#> [271] 410 411 413 414 415 416 417 418 419 420 421 422 426 428 429 430 432 433
#> [289] 435 437 444 445 446 447 448 451 452 453 454 455 457 458 459 460 463 465
#> [307] 466 467 468 470 472 473 475 478 479 480 483 484 485 487 488 489 490 491
#> [325] 492 493 494 495 496 497 498 499 500
#> 
#> [[2]]$train[[2]]
#>   [1]   1   5   6   7   8   9  10  11  12  14  15  16  18  19  20  22  25  26
#>  [19]  27  30  32  34  36  37  38  40  41  42  45  46  47  48  49  50  52  53
#>  [37]  55  56  58  59  61  62  65  69  70  71  72  73  74  77  78  80  82  85
#>  [55]  86  88  89  90  91  92  93  94  95  97  98  99 100 102 103 104 105 106
#>  [73] 107 108 110 111 112 114 116 118 119 120 121 123 125 126 127 128 129 131
#>  [91] 132 133 135 136 137 138 139 140 141 143 146 151 152 154 156 159 161 163
#> [109] 164 166 167 169 170 171 172 173 174 175 176 177 179 180 181 184 186 187
#> [127] 188 192 193 194 195 196 197 198 200 201 202 203 205 206 207 208 209 212
#> [145] 214 215 218 220 221 222 223 224 225 226 229 231 232 233 234 235 236 237
#> [163] 239 240 241 242 243 244 245 246 248 249 250 252 253 254 255 256 257 260
#> [181] 262 267 268 271 274 275 276 278 279 280 281 283 284 287 289 290 291 292
#> [199] 295 297 298 300 303 304 305 306 308 309 310 311 312 315 316 317 318 319
#> [217] 322 323 324 325 327 328 329 330 331 333 334 335 336 337 339 341 343 344
#> [235] 346 348 349 353 355 357 361 364 365 367 368 370 373 374 376 377 378 381
#> [253] 382 384 385 387 388 390 391 392 393 394 396 397 398 400 401 403 405 406
#> [271] 407 408 410 412 413 415 421 423 424 425 426 427 429 430 431 432 434 435
#> [289] 436 438 439 440 441 442 443 444 449 450 451 452 454 455 456 457 458 459
#> [307] 461 462 463 464 465 466 468 469 471 474 476 477 481 482 483 485 486 487
#> [325] 489 490 491 492 493 495 496 497 498 500
#> 
#> [[2]]$train[[3]]
#>   [1]   2   3   4   5   6   8  10  12  13  14  17  21  23  24  27  28  29  30
#>  [19]  31  33  34  35  36  37  39  40  43  44  45  46  51  53  54  55  57  58
#>  [37]  60  61  62  63  64  66  67  68  69  70  72  75  76  77  78  79  80  81
#>  [55]  82  83  84  85  86  87  88  90  91  95  96  99 101 102 106 107 108 109
#>  [73] 112 113 115 116 117 119 120 121 122 123 124 127 130 134 135 136 141 142
#>  [91] 144 145 147 148 149 150 153 154 155 156 157 158 160 162 163 165 167 168
#> [109] 169 170 172 177 178 179 180 181 182 183 185 186 188 189 190 191 192 193
#> [127] 194 195 196 197 198 199 200 204 205 206 208 209 210 211 212 213 215 216
#> [145] 217 219 220 223 224 225 227 228 229 230 232 238 239 240 242 244 245 247
#> [163] 248 249 250 251 252 253 254 256 258 259 260 261 263 264 265 266 268 269
#> [181] 270 272 273 277 279 280 282 283 285 286 287 288 290 293 294 296 299 301
#> [199] 302 303 306 307 309 310 312 313 314 316 317 319 320 321 322 323 324 326
#> [217] 329 331 332 335 336 337 338 340 342 343 344 345 346 347 348 349 350 351
#> [235] 352 353 354 356 357 358 359 360 362 363 366 367 368 369 371 372 374 375
#> [253] 379 380 381 382 383 384 386 388 389 390 392 393 394 395 396 398 399 401
#> [271] 402 403 404 405 407 409 411 412 414 416 417 418 419 420 422 423 424 425
#> [289] 427 428 431 433 434 436 437 438 439 440 441 442 443 445 446 447 448 449
#> [307] 450 453 456 460 461 462 464 467 469 470 471 472 473 474 475 476 477 478
#> [325] 479 480 481 482 484 486 488 494 499
#> 
#> 
#> 
#> [[3]]
#> [[3]]$test
#> [[3]]$test[[1]]
#>   [1]   1   3   4   7   9  10  13  15  18  21  27  32  33  39  45  46  53  56
#>  [19]  58  65  66  67  68  70  71  75  76  80  84  87  89  96  98  99 100 101
#>  [37] 104 105 106 108 109 112 113 114 116 118 124 125 127 132 133 138 144 151
#>  [55] 152 158 164 171 172 173 174 184 191 200 206 208 209 211 215 216 223 227
#>  [73] 234 237 238 243 244 245 246 247 250 251 253 255 258 260 263 265 267 272
#>  [91] 274 275 280 284 285 290 297 298 299 301 302 306 307 309 315 325 333 334
#> [109] 339 343 348 349 351 353 359 367 368 370 375 377 379 381 383 384 388 390
#> [127] 394 395 398 406 407 410 412 414 416 417 421 422 425 428 429 431 432 435
#> [145] 436 438 439 440 441 443 445 446 449 450 455 458 459 465 467 472 473 474
#> [163] 479 480 481 482 487 494
#> 
#> [[3]]$test[[2]]
#>   [1]   2  14  16  17  19  22  23  24  25  26  29  30  37  38  42  44  47  49
#>  [19]  50  57  60  61  62  64  72  74  77  81  83  85  88  90  92  93  95  97
#>  [37] 102 103 107 110 111 115 117 120 129 130 134 137 140 141 143 149 150 160
#>  [55] 161 168 169 175 178 181 182 187 188 189 190 192 193 194 195 197 198 201
#>  [73] 202 203 207 210 213 218 219 221 222 225 228 233 236 239 248 249 259 261
#>  [91] 266 270 271 282 286 287 288 292 293 294 296 300 304 312 313 314 317 320
#> [109] 323 327 328 329 331 332 337 338 341 346 347 354 355 361 363 364 371 372
#> [127] 386 389 391 393 396 397 399 400 401 402 403 404 405 408 411 413 419 424
#> [145] 426 434 437 442 447 453 454 456 457 460 462 464 468 469 470 484 485 486
#> [163] 491 493 499
#> 
#> [[3]]$test[[3]]
#>   [1]   5   6   8  11  12  20  28  31  34  35  36  40  41  43  48  51  52  54
#>  [19]  55  59  63  69  73  78  79  82  86  91  94 119 121 122 123 126 128 131
#>  [37] 135 136 139 142 145 146 147 148 153 154 155 156 157 159 162 163 165 166
#>  [55] 167 170 176 177 179 180 183 185 186 196 199 204 205 212 214 217 220 224
#>  [73] 226 229 230 231 232 235 240 241 242 252 254 256 257 262 264 268 269 273
#>  [91] 276 277 278 279 281 283 289 291 295 303 305 308 310 311 316 318 319 321
#> [109] 322 324 326 330 335 336 340 342 344 345 350 352 356 357 358 360 362 365
#> [127] 366 369 373 374 376 378 380 382 385 387 392 409 415 418 420 423 427 430
#> [145] 433 444 448 451 452 461 463 466 471 475 476 477 478 483 488 489 490 492
#> [163] 495 496 497 498 500
#> 
#> 
#> [[3]]$train
#> [[3]]$train[[1]]
#>   [1]   2   5   6   8  11  12  14  16  17  19  20  22  23  24  25  26  28  29
#>  [19]  30  31  34  35  36  37  38  40  41  42  43  44  47  48  49  50  51  52
#>  [37]  54  55  57  59  60  61  62  63  64  69  72  73  74  77  78  79  81  82
#>  [55]  83  85  86  88  90  91  92  93  94  95  97 102 103 107 110 111 115 117
#>  [73] 119 120 121 122 123 126 128 129 130 131 134 135 136 137 139 140 141 142
#>  [91] 143 145 146 147 148 149 150 153 154 155 156 157 159 160 161 162 163 165
#> [109] 166 167 168 169 170 175 176 177 178 179 180 181 182 183 185 186 187 188
#> [127] 189 190 192 193 194 195 196 197 198 199 201 202 203 204 205 207 210 212
#> [145] 213 214 217 218 219 220 221 222 224 225 226 228 229 230 231 232 233 235
#> [163] 236 239 240 241 242 248 249 252 254 256 257 259 261 262 264 266 268 269
#> [181] 270 271 273 276 277 278 279 281 282 283 286 287 288 289 291 292 293 294
#> [199] 295 296 300 303 304 305 308 310 311 312 313 314 316 317 318 319 320 321
#> [217] 322 323 324 326 327 328 329 330 331 332 335 336 337 338 340 341 342 344
#> [235] 345 346 347 350 352 354 355 356 357 358 360 361 362 363 364 365 366 369
#> [253] 371 372 373 374 376 378 380 382 385 386 387 389 391 392 393 396 397 399
#> [271] 400 401 402 403 404 405 408 409 411 413 415 418 419 420 423 424 426 427
#> [289] 430 433 434 437 442 444 447 448 451 452 453 454 456 457 460 461 462 463
#> [307] 464 466 468 469 470 471 475 476 477 478 483 484 485 486 488 489 490 491
#> [325] 492 493 495 496 497 498 499 500
#> 
#> [[3]]$train[[2]]
#>   [1]   1   3   4   5   6   7   8   9  10  11  12  13  15  18  20  21  27  28
#>  [19]  31  32  33  34  35  36  39  40  41  43  45  46  48  51  52  53  54  55
#>  [37]  56  58  59  63  65  66  67  68  69  70  71  73  75  76  78  79  80  82
#>  [55]  84  86  87  89  91  94  96  98  99 100 101 104 105 106 108 109 112 113
#>  [73] 114 116 118 119 121 122 123 124 125 126 127 128 131 132 133 135 136 138
#>  [91] 139 142 144 145 146 147 148 151 152 153 154 155 156 157 158 159 162 163
#> [109] 164 165 166 167 170 171 172 173 174 176 177 179 180 183 184 185 186 191
#> [127] 196 199 200 204 205 206 208 209 211 212 214 215 216 217 220 223 224 226
#> [145] 227 229 230 231 232 234 235 237 238 240 241 242 243 244 245 246 247 250
#> [163] 251 252 253 254 255 256 257 258 260 262 263 264 265 267 268 269 272 273
#> [181] 274 275 276 277 278 279 280 281 283 284 285 289 290 291 295 297 298 299
#> [199] 301 302 303 305 306 307 308 309 310 311 315 316 318 319 321 322 324 325
#> [217] 326 330 333 334 335 336 339 340 342 343 344 345 348 349 350 351 352 353
#> [235] 356 357 358 359 360 362 365 366 367 368 369 370 373 374 375 376 377 378
#> [253] 379 380 381 382 383 384 385 387 388 390 392 394 395 398 406 407 409 410
#> [271] 412 414 415 416 417 418 420 421 422 423 425 427 428 429 430 431 432 433
#> [289] 435 436 438 439 440 441 443 444 445 446 448 449 450 451 452 455 458 459
#> [307] 461 463 465 466 467 471 472 473 474 475 476 477 478 479 480 481 482 483
#> [325] 487 488 489 490 492 494 495 496 497 498 500
#> 
#> [[3]]$train[[3]]
#>   [1]   1   2   3   4   7   9  10  13  14  15  16  17  18  19  21  22  23  24
#>  [19]  25  26  27  29  30  32  33  37  38  39  42  44  45  46  47  49  50  53
#>  [37]  56  57  58  60  61  62  64  65  66  67  68  70  71  72  74  75  76  77
#>  [55]  80  81  83  84  85  87  88  89  90  92  93  95  96  97  98  99 100 101
#>  [73] 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 120
#>  [91] 124 125 127 129 130 132 133 134 137 138 140 141 143 144 149 150 151 152
#> [109] 158 160 161 164 168 169 171 172 173 174 175 178 181 182 184 187 188 189
#> [127] 190 191 192 193 194 195 197 198 200 201 202 203 206 207 208 209 210 211
#> [145] 213 215 216 218 219 221 222 223 225 227 228 233 234 236 237 238 239 243
#> [163] 244 245 246 247 248 249 250 251 253 255 258 259 260 261 263 265 266 267
#> [181] 270 271 272 274 275 280 282 284 285 286 287 288 290 292 293 294 296 297
#> [199] 298 299 300 301 302 304 306 307 309 312 313 314 315 317 320 323 325 327
#> [217] 328 329 331 332 333 334 337 338 339 341 343 346 347 348 349 351 353 354
#> [235] 355 359 361 363 364 367 368 370 371 372 375 377 379 381 383 384 386 388
#> [253] 389 390 391 393 394 395 396 397 398 399 400 401 402 403 404 405 406 407
#> [271] 408 410 411 412 413 414 416 417 419 421 422 424 425 426 428 429 431 432
#> [289] 434 435 436 437 438 439 440 441 442 443 445 446 447 449 450 453 454 455
#> [307] 456 457 458 459 460 462 464 465 467 468 469 470 472 473 474 479 480 481
#> [325] 482 484 485 486 487 491 493 494 499
#> 
#> 
#> 
#> [[4]]
#> [[4]]$test
#> [[4]]$test[[1]]
#>   [1]   1   5   8  12  16  18  20  22  26  28  29  30  32  36  37  43  45  49
#>  [19]  53  62  65  72  76  77  78  81  82  86  94 100 102 106 107 110 112 114
#>  [37] 115 121 126 127 128 131 135 143 146 147 151 154 156 157 158 159 160 163
#>  [55] 164 168 169 171 172 174 180 181 182 195 199 204 205 207 208 210 214 217
#>  [73] 218 220 221 222 223 229 231 232 233 236 237 241 242 244 245 246 252 253
#>  [91] 254 255 256 260 262 263 269 270 271 273 277 279 281 282 284 293 298 303
#> [109] 311 316 318 319 321 322 325 326 327 330 332 333 334 335 340 343 344 348
#> [127] 350 353 356 360 361 362 364 365 371 372 374 375 377 380 388 391 393 398
#> [145] 402 408 409 411 412 420 425 428 437 440 443 445 448 449 458 460 461 466
#> [163] 468 469 470 471 473 474 475 484 487 488 495 496 498 500
#> 
#> [[4]]$test[[2]]
#>   [1]   2   3   7  10  13  14  15  21  27  33  34  35  38  42  44  47  48  51
#>  [19]  55  56  58  59  60  63  69  70  71  74  79  90  91  93  95  96  99 101
#>  [37] 103 104 108 111 113 116 123 129 132 134 137 142 145 148 152 155 161 162
#>  [55] 165 166 173 183 184 186 187 189 193 194 200 201 202 203 206 211 216 219
#>  [73] 225 226 227 228 230 234 235 247 249 250 257 261 265 267 274 275 276 280
#>  [91] 283 286 288 294 295 301 304 305 306 307 310 313 315 320 329 331 337 339
#> [109] 342 346 347 351 357 358 367 368 370 373 376 378 381 382 384 386 387 389
#> [127] 392 394 395 397 400 401 405 406 407 410 413 416 421 423 427 430 431 432
#> [145] 436 438 439 441 444 450 453 454 456 459 462 464 481 483 486 491 493
#> 
#> [[4]]$test[[3]]
#>   [1]   4   6   9  11  17  19  23  24  25  31  39  40  41  46  50  52  54  57
#>  [19]  61  64  66  67  68  73  75  80  83  84  85  87  88  89  92  97  98 105
#>  [37] 109 117 118 119 120 122 124 125 130 133 136 138 139 140 141 144 149 150
#>  [55] 153 167 170 175 176 177 178 179 185 188 190 191 192 196 197 198 209 212
#>  [73] 213 215 224 238 239 240 243 248 251 258 259 264 266 268 272 278 285 287
#>  [91] 289 290 291 292 296 297 299 300 302 308 309 312 314 317 323 324 328 336
#> [109] 338 341 345 349 352 354 355 359 363 366 369 379 383 385 390 396 399 403
#> [127] 404 414 415 417 418 419 422 424 426 429 433 434 435 442 446 447 451 452
#> [145] 455 457 463 465 467 472 476 477 478 479 480 482 485 489 490 492 494 497
#> [163] 499
#> 
#> 
#> [[4]]$train
#> [[4]]$train[[1]]
#>   [1]   2   3   4   6   7   9  10  11  13  14  15  17  19  21  23  24  25  27
#>  [19]  31  33  34  35  38  39  40  41  42  44  46  47  48  50  51  52  54  55
#>  [37]  56  57  58  59  60  61  63  64  66  67  68  69  70  71  73  74  75  79
#>  [55]  80  83  84  85  87  88  89  90  91  92  93  95  96  97  98  99 101 103
#>  [73] 104 105 108 109 111 113 116 117 118 119 120 122 123 124 125 129 130 132
#>  [91] 133 134 136 137 138 139 140 141 142 144 145 148 149 150 152 153 155 161
#> [109] 162 165 166 167 170 173 175 176 177 178 179 183 184 185 186 187 188 189
#> [127] 190 191 192 193 194 196 197 198 200 201 202 203 206 209 211 212 213 215
#> [145] 216 219 224 225 226 227 228 230 234 235 238 239 240 243 247 248 249 250
#> [163] 251 257 258 259 261 264 265 266 267 268 272 274 275 276 278 280 283 285
#> [181] 286 287 288 289 290 291 292 294 295 296 297 299 300 301 302 304 305 306
#> [199] 307 308 309 310 312 313 314 315 317 320 323 324 328 329 331 336 337 338
#> [217] 339 341 342 345 346 347 349 351 352 354 355 357 358 359 363 366 367 368
#> [235] 369 370 373 376 378 379 381 382 383 384 385 386 387 389 390 392 394 395
#> [253] 396 397 399 400 401 403 404 405 406 407 410 413 414 415 416 417 418 419
#> [271] 421 422 423 424 426 427 429 430 431 432 433 434 435 436 438 439 441 442
#> [289] 444 446 447 450 451 452 453 454 455 456 457 459 462 463 464 465 467 472
#> [307] 476 477 478 479 480 481 482 483 485 486 489 490 491 492 493 494 497 499
#> 
#> [[4]]$train[[2]]
#>   [1]   1   4   5   6   8   9  11  12  16  17  18  19  20  22  23  24  25  26
#>  [19]  28  29  30  31  32  36  37  39  40  41  43  45  46  49  50  52  53  54
#>  [37]  57  61  62  64  65  66  67  68  72  73  75  76  77  78  80  81  82  83
#>  [55]  84  85  86  87  88  89  92  94  97  98 100 102 105 106 107 109 110 112
#>  [73] 114 115 117 118 119 120 121 122 124 125 126 127 128 130 131 133 135 136
#>  [91] 138 139 140 141 143 144 146 147 149 150 151 153 154 156 157 158 159 160
#> [109] 163 164 167 168 169 170 171 172 174 175 176 177 178 179 180 181 182 185
#> [127] 188 190 191 192 195 196 197 198 199 204 205 207 208 209 210 212 213 214
#> [145] 215 217 218 220 221 222 223 224 229 231 232 233 236 237 238 239 240 241
#> [163] 242 243 244 245 246 248 251 252 253 254 255 256 258 259 260 262 263 264
#> [181] 266 268 269 270 271 272 273 277 278 279 281 282 284 285 287 289 290 291
#> [199] 292 293 296 297 298 299 300 302 303 308 309 311 312 314 316 317 318 319
#> [217] 321 322 323 324 325 326 327 328 330 332 333 334 335 336 338 340 341 343
#> [235] 344 345 348 349 350 352 353 354 355 356 359 360 361 362 363 364 365 366
#> [253] 369 371 372 374 375 377 379 380 383 385 388 390 391 393 396 398 399 402
#> [271] 403 404 408 409 411 412 414 415 417 418 419 420 422 424 425 426 428 429
#> [289] 433 434 435 437 440 442 443 445 446 447 448 449 451 452 455 457 458 460
#> [307] 461 463 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480
#> [325] 482 484 485 487 488 489 490 492 494 495 496 497 498 499 500
#> 
#> [[4]]$train[[3]]
#>   [1]   1   2   3   5   7   8  10  12  13  14  15  16  18  20  21  22  26  27
#>  [19]  28  29  30  32  33  34  35  36  37  38  42  43  44  45  47  48  49  51
#>  [37]  53  55  56  58  59  60  62  63  65  69  70  71  72  74  76  77  78  79
#>  [55]  81  82  86  90  91  93  94  95  96  99 100 101 102 103 104 106 107 108
#>  [73] 110 111 112 113 114 115 116 121 123 126 127 128 129 131 132 134 135 137
#>  [91] 142 143 145 146 147 148 151 152 154 155 156 157 158 159 160 161 162 163
#> [109] 164 165 166 168 169 171 172 173 174 180 181 182 183 184 186 187 189 193
#> [127] 194 195 199 200 201 202 203 204 205 206 207 208 210 211 214 216 217 218
#> [145] 219 220 221 222 223 225 226 227 228 229 230 231 232 233 234 235 236 237
#> [163] 241 242 244 245 246 247 249 250 252 253 254 255 256 257 260 261 262 263
#> [181] 265 267 269 270 271 273 274 275 276 277 279 280 281 282 283 284 286 288
#> [199] 293 294 295 298 301 303 304 305 306 307 310 311 313 315 316 318 319 320
#> [217] 321 322 325 326 327 329 330 331 332 333 334 335 337 339 340 342 343 344
#> [235] 346 347 348 350 351 353 356 357 358 360 361 362 364 365 367 368 370 371
#> [253] 372 373 374 375 376 377 378 380 381 382 384 386 387 388 389 391 392 393
#> [271] 394 395 397 398 400 401 402 405 406 407 408 409 410 411 412 413 416 420
#> [289] 421 423 425 427 428 430 431 432 436 437 438 439 440 441 443 444 445 448
#> [307] 449 450 453 454 456 458 459 460 461 462 464 466 468 469 470 471 473 474
#> [325] 475 481 483 484 486 487 488 491 493 495 496 498 500
#> 
#> 
#> 
#> [[5]]
#> [[5]]$test
#> [[5]]$test[[1]]
#>   [1]   1   2   4   7  11  16  20  23  24  26  27  30  33  36  39  40  44  48
#>  [19]  50  52  55  57  61  63  64  68  70  73  77  83  85  91  92  95  96  98
#>  [37] 100 103 116 117 119 123 124 132 133 134 136 139 144 146 147 148 150 152
#>  [55] 155 159 160 161 171 173 174 177 182 187 188 193 198 200 206 214 215 217
#>  [73] 219 220 228 229 238 239 242 247 248 249 251 256 257 258 259 262 265 266
#>  [91] 274 275 278 279 281 286 293 294 296 297 311 312 314 318 319 322 326 335
#> [109] 337 338 340 342 343 346 347 350 352 354 360 363 365 371 377 379 380 385
#> [127] 387 389 391 393 395 406 407 408 409 410 415 419 423 427 428 429 432 435
#> [145] 442 445 446 447 450 452 455 456 457 462 465 467 473 474 475 480 481 484
#> [163] 486 488 496 499 500
#> 
#> [[5]]$test[[2]]
#>   [1]   8   9  12  15  21  25  28  29  31  32  41  43  45  46  51  56  60  62
#>  [19]  65  66  71  75  76  79  81  86  89  90  94 101 102 105 107 108 111 114
#>  [37] 115 121 125 126 127 128 129 130 131 137 138 143 145 149 151 157 158 162
#>  [55] 163 164 165 166 167 168 176 180 181 186 192 199 209 210 212 218 221 224
#>  [73] 225 227 231 233 237 240 241 243 246 254 255 261 263 264 267 268 269 270
#>  [91] 271 273 276 280 282 283 285 288 291 292 301 304 306 307 308 309 310 316
#> [109] 321 325 328 329 331 332 334 339 341 349 353 356 370 372 375 376 378 381
#> [127] 382 383 384 388 392 394 396 397 398 399 400 402 411 412 413 414 417 420
#> [145] 421 422 424 431 433 436 437 440 441 443 458 459 460 463 464 469 470 478
#> [163] 485 487 489 490 491 495 497 498
#> 
#> [[5]]$test[[3]]
#>   [1]   3   5   6  10  13  14  17  18  19  22  34  35  37  38  42  47  49  53
#>  [19]  54  58  59  67  69  72  74  78  80  82  84  87  88  93  97  99 104 106
#>  [37] 109 110 112 113 118 120 122 135 140 141 142 153 154 156 169 170 172 175
#>  [55] 178 179 183 184 185 189 190 191 194 195 196 197 201 202 203 204 205 207
#>  [73] 208 211 213 216 222 223 226 230 232 234 235 236 244 245 250 252 253 260
#>  [91] 272 277 284 287 289 290 295 298 299 300 302 303 305 313 315 317 320 323
#> [109] 324 327 330 333 336 344 345 348 351 355 357 358 359 361 362 364 366 367
#> [127] 368 369 373 374 386 390 401 403 404 405 416 418 425 426 430 434 438 439
#> [145] 444 448 449 451 453 454 461 466 468 471 472 476 477 479 482 483 492 493
#> [163] 494
#> 
#> 
#> [[5]]$train
#> [[5]]$train[[1]]
#>   [1]   3   5   6   8   9  10  12  13  14  15  17  18  19  21  22  25  28  29
#>  [19]  31  32  34  35  37  38  41  42  43  45  46  47  49  51  53  54  56  58
#>  [37]  59  60  62  65  66  67  69  71  72  74  75  76  78  79  80  81  82  84
#>  [55]  86  87  88  89  90  93  94  97  99 101 102 104 105 106 107 108 109 110
#>  [73] 111 112 113 114 115 118 120 121 122 125 126 127 128 129 130 131 135 137
#>  [91] 138 140 141 142 143 145 149 151 153 154 156 157 158 162 163 164 165 166
#> [109] 167 168 169 170 172 175 176 178 179 180 181 183 184 185 186 189 190 191
#> [127] 192 194 195 196 197 199 201 202 203 204 205 207 208 209 210 211 212 213
#> [145] 216 218 221 222 223 224 225 226 227 230 231 232 233 234 235 236 237 240
#> [163] 241 243 244 245 246 250 252 253 254 255 260 261 263 264 267 268 269 270
#> [181] 271 272 273 276 277 280 282 283 284 285 287 288 289 290 291 292 295 298
#> [199] 299 300 301 302 303 304 305 306 307 308 309 310 313 315 316 317 320 321
#> [217] 323 324 325 327 328 329 330 331 332 333 334 336 339 341 344 345 348 349
#> [235] 351 353 355 356 357 358 359 361 362 364 366 367 368 369 370 372 373 374
#> [253] 375 376 378 381 382 383 384 386 388 390 392 394 396 397 398 399 400 401
#> [271] 402 403 404 405 411 412 413 414 416 417 418 420 421 422 424 425 426 430
#> [289] 431 433 434 436 437 438 439 440 441 443 444 448 449 451 453 454 458 459
#> [307] 460 461 463 464 466 468 469 470 471 472 476 477 478 479 482 483 485 487
#> [325] 489 490 491 492 493 494 495 497 498
#> 
#> [[5]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7  10  11  13  14  16  17  18  19  20  22  23
#>  [19]  24  26  27  30  33  34  35  36  37  38  39  40  42  44  47  48  49  50
#>  [37]  52  53  54  55  57  58  59  61  63  64  67  68  69  70  72  73  74  77
#>  [55]  78  80  82  83  84  85  87  88  91  92  93  95  96  97  98  99 100 103
#>  [73] 104 106 109 110 112 113 116 117 118 119 120 122 123 124 132 133 134 135
#>  [91] 136 139 140 141 142 144 146 147 148 150 152 153 154 155 156 159 160 161
#> [109] 169 170 171 172 173 174 175 177 178 179 182 183 184 185 187 188 189 190
#> [127] 191 193 194 195 196 197 198 200 201 202 203 204 205 206 207 208 211 213
#> [145] 214 215 216 217 219 220 222 223 226 228 229 230 232 234 235 236 238 239
#> [163] 242 244 245 247 248 249 250 251 252 253 256 257 258 259 260 262 265 266
#> [181] 272 274 275 277 278 279 281 284 286 287 289 290 293 294 295 296 297 298
#> [199] 299 300 302 303 305 311 312 313 314 315 317 318 319 320 322 323 324 326
#> [217] 327 330 333 335 336 337 338 340 342 343 344 345 346 347 348 350 351 352
#> [235] 354 355 357 358 359 360 361 362 363 364 365 366 367 368 369 371 373 374
#> [253] 377 379 380 385 386 387 389 390 391 393 395 401 403 404 405 406 407 408
#> [271] 409 410 415 416 418 419 423 425 426 427 428 429 430 432 434 435 438 439
#> [289] 442 444 445 446 447 448 449 450 451 452 453 454 455 456 457 461 462 465
#> [307] 466 467 468 471 472 473 474 475 476 477 479 480 481 482 483 484 486 488
#> [325] 492 493 494 496 499 500
#> 
#> [[5]]$train[[3]]
#>   [1]   1   2   4   7   8   9  11  12  15  16  20  21  23  24  25  26  27  28
#>  [19]  29  30  31  32  33  36  39  40  41  43  44  45  46  48  50  51  52  55
#>  [37]  56  57  60  61  62  63  64  65  66  68  70  71  73  75  76  77  79  81
#>  [55]  83  85  86  89  90  91  92  94  95  96  98 100 101 102 103 105 107 108
#>  [73] 111 114 115 116 117 119 121 123 124 125 126 127 128 129 130 131 132 133
#>  [91] 134 136 137 138 139 143 144 145 146 147 148 149 150 151 152 155 157 158
#> [109] 159 160 161 162 163 164 165 166 167 168 171 173 174 176 177 180 181 182
#> [127] 186 187 188 192 193 198 199 200 206 209 210 212 214 215 217 218 219 220
#> [145] 221 224 225 227 228 229 231 233 237 238 239 240 241 242 243 246 247 248
#> [163] 249 251 254 255 256 257 258 259 261 262 263 264 265 266 267 268 269 270
#> [181] 271 273 274 275 276 278 279 280 281 282 283 285 286 288 291 292 293 294
#> [199] 296 297 301 304 306 307 308 309 310 311 312 314 316 318 319 321 322 325
#> [217] 326 328 329 331 332 334 335 337 338 339 340 341 342 343 346 347 349 350
#> [235] 352 353 354 356 360 363 365 370 371 372 375 376 377 378 379 380 381 382
#> [253] 383 384 385 387 388 389 391 392 393 394 395 396 397 398 399 400 402 406
#> [271] 407 408 409 410 411 412 413 414 415 417 419 420 421 422 423 424 427 428
#> [289] 429 431 432 433 435 436 437 440 441 442 443 445 446 447 450 452 455 456
#> [307] 457 458 459 460 462 463 464 465 467 469 470 473 474 475 478 480 481 484
#> [325] 485 486 487 488 489 490 491 495 496 497 498 499 500
#> 
#> 
#>