Spatial Block Cross Validation Resampling

Spatial Block Cross validation implemented by the blockCV package.

References

Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G (2018). “blockCV: an R package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models.” bioRxiv. doi: 10.1101/357798 .

Super class

mlr3::Resampling -> ResamplingSpCVBlock

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSpCVBlock$new()

ResamplingSpCVBlock$instantiate()

ResamplingSpCVBlock$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "Environmental Block" resampling instance.

Usage

ResamplingSpCVBlock$new(id = "spcv-block")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSpCVBlock$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSpCVBlock$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv-block")

#> Registered S3 method overwritten by 'xts':
#>   method     from
#>   as.zoo.xts zoo 

rcv$instantiate(task)

#> Warning: Hyperparameters 'rows' and 'cols' not set.
#>                  Using the default value of '4' set by 'mlr3spatiotempcv' for
#>                  both which results in a grid of 16.
#>                  You might want to set these values yourself during resampling
#>                  construction.

# Individual sets:
rcv$train_set(1)

#>   [1]   7  13  17  35  44  45  75  78  88 109 137 142 143 147 188 204 240 242
#>  [19] 259 270 285 318 320 331 337 338 349 365 371 378 381 384 409 418 431 444
#>  [37] 446 448 453 464 477 494 496 505 510 528 533 535 538 554 557 566 567 570
#>  [55] 580 591 596 621 628 636 638 655 658 667 684 705 710 723 728 733 734 735
#>  [73] 739 740 750   9  20  27  30  31  38  40  41  43  53  67  72  82  92  98
#>  [91] 114 116 124 141 154 173 179 186 189 199 205 226 228 233 235 237 254 275
#> [109] 276 284 291 293 296 298 299 301 304 315 333 335 339 340 343 360 367 380
#> [127] 386 392 401 437 441 443 457 462 484 495 500 508 513 521 522 526 551 576
#> [145] 587 595 600 608 622 659 675 694 709 713 722  11  16 107 121 152 157 198
#> [163] 212 249 258 264 267 273 300 312 316 329 332 361 369 375 379 399 408 410
#> [181] 417 420 432 445 459 466 473 482 515 588 590 604 637 650 660 686 721 731
#> [199] 748  22  24  25  32  52  60  63  76  80  83  89 101 104 118 120 123 134
#> [217] 151 153 156 161 162 172 176 177 192 197 200 214 215 221 223 224 227 239
#> [235] 246 251 266 269 271 279 288 297 306 311 313 321 341 348 352 376 377 388
#> [253] 396 397 406 411 424 429 430 433 434 450 460 468 471 472 480 502 512 514
#> [271] 532 536 541 543 555 556 558 559 568 586 598 606 607 611 614 615 617 634
#> [289] 648 652 662 666 680 693 698 700 712 724 726 730 745  42  47  55  59  62
#> [307]  87  93  95  96  99 139 144 230 241 250 274 283 319 383 419 440 501 504
#> [325] 516 534 539 540 548 552 553 560 569 571 578 627 630 635 657 665 669 674
#> [343] 676 688 689 692 702 704 716 729 741 744 749 751   4   8  15  26  48  64
#> [361]  69  70  71  73  74  81  85  91  97 102 108 119 140 150 158 159 160 166
#> [379] 180 185 187 194 206 207 208 217 238 244 247 261 265 272 278 280 287 292
#> [397] 305 323 326 328 334 353 354 355 389 390 398 402 415 428 435 436 438 451
#> [415] 454 456 461 470 476 479 488 499 506 520 524 549 561 562 563 574 582 584
#> [433] 589 593 597 601 605 610 616 625 633 651 656 672 679 681 683 685 687 717
#> [451] 720 738 746 747   1   6  23  33  34  36  51  56  84 100 111 112 117 146
#> [469] 148 149 164 165 167 168 174 175 178 184 193 210 232 248 252 256 257 286
#> [487] 294 295 302 307 310 314 324 330 342 347 351 363 364 374 382 387 391 393
#> [505] 407 421 425 439 447 469 481 485 491 497 507 517 545 572 577 609 624 626
#> [523] 629 632 640 645 647 677 678 695 699 742   2   3   5  12  14  18  19  21
#> [541]  28  46  50  54  65  66  79  86  94 105 113 125 127 128 129 130 131 132
#> [559] 135 138 145 163 171 181 182 191 195 196 202 203 209 211 213 225 229 231
#> [577] 236 253 260 262 268 277 281 282 308 322 327 336 346 358 366 368 370 373
#> [595] 395 412 416 422 423 442 452 455 458 465 475 478 492 493 498 511 518 519
#> [613] 523 527 531 542 565 581 583 599 618 619 631 646 653 661 663 668 670 671
#> [631] 691 701 703 714 718 732 737  10  29  37  39  49  57  61  68  77 106 110
#> [649] 115 155 183 190 201 219 243 263 289 317 325 345 350 372 385 400 403 404
#> [667] 405 414 427 449 463 474 483 487 490 537 546 550 564 575 620 639 642 643
#> [685] 649 654 664 673 682 696 711 715 736 743

rcv$test_set(1)

#>  [1]  58  90 103 122 126 133 136 169 170 216 218 220 222 234 245 255 290 303 309
#> [20] 344 356 357 359 362 394 413 426 467 486 489 503 509 525 529 530 544 547 573
#> [39] 579 585 592 594 602 603 612 613 623 641 644 690 697 706 707 708 719 725 727

intersect(rcv$train_set(1), rcv$test_set(1))

#> integer(0)

# Internal storage:
rcv$instance

#>      row_id fold
#>   1:     58    1
#>   2:     90    1
#>   3:    103    1
#>   4:    122    1
#>   5:    126    1
#>  ---            
#> 747:    696   10
#> 748:    711   10
#> 749:    715   10
#> 750:    736   10
#> 751:    743   10