Spatial Block Cross Validation Resampling

Spatial Block Cross validation implemented by the blockCV package.

Details

By default blockCV::spatialBlock() does not allow the creation of multiple repetitions. mlr3spatiotempcv adds support for this when using the range argument for fold creation. When supplying a vector of length(repeats) for argument range, these different settings will be used to create folds which differ among the repetitions.

Multiple repetitions are not possible when using the "row & cols" approach because the created folds will always be the same.

References

Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G (2018). “blockCV: an R package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models.” bioRxiv. doi: 10.1101/357798 .

Super class

mlr3::Resampling -> ResamplingSpCVBlock

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSpCVBlock$new()

ResamplingSpCVBlock$instantiate()

ResamplingSpCVBlock$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "Environmental Block" resampling instance.

Usage

ResamplingSpCVBlock$new(id = "spcv-block")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSpCVBlock$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSpCVBlock$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv-block", range = 1000)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)

#>   [1]  29  37  49  58  90 103 110 122 126 127 133 136 169 183 220 222 245 290
#>  [19] 303 309 325 345 350 356 357 394 400 403 404 413 449 463 474 483 486 490
#>  [37] 503 509 528 531 537 544 546 564 573 575 585 594 613 620 639 641 643 653
#>  [55] 655 664 690 696 706 708 710 715 728 732 735 743  10  27  38  39  41  43
#>  [73]  45  53  57  61  67  68  72  77  82  98 106 114 115 116 124 141 154 155
#>  [91] 186 190 199 201 219 233 235 243 263 275 289 296 298 299 301 317 367 372
#> [109] 385 392 401 405 414 427 441 462 484 487 489 495 508 522 526 530 550 551
#> [127] 576 579 595 600 603 642 649 654 673 682 694 697 709 711 719 736   9  20
#> [145]  22  30  31  36  40  80  81  83  92 101 111 134 173 176 177 179 184 189
#> [163] 197 205 226 228 237 254 256 276 284 291 293 297 304 311 315 333 335 339
#> [181] 340 341 343 352 359 360 380 382 386 391 396 407 434 437 443 450 457 467
#> [199] 481 500 513 521 536 547 556 568 587 602 606 608 614 615 616 622 634 648
#> [217] 659 675 695 700 713 722 745  46  54  65  74  79 113 117 119 128 130 132
#> [235] 142 145 159 160 163 165 178 181 182 187 191 194 195 208 209 211 213 225
#> [253] 229 244 253 262 270 292 308 322 323 327 336 355 363 364 366 368 390 393
#> [271] 436 452 454 458 465 475 478 479 493 497 498 506 524 542 545 549 581 584
#> [289] 593 599 601 624 625 640 646 663 687 701 714 742   1   6  11  16  23  33
#> [307]  34  51  56  84 100 107 112 146 148 149 152 164 167 168 174 175 193 210
#> [325] 212 232 248 252 257 258 264 267 273 286 294 295 300 302 307 310 312 314
#> [343] 316 324 330 342 347 351 369 374 379 387 399 420 421 425 432 439 445 447
#> [361] 459 469 473 482 485 491 507 515 517 572 577 590 597 604 609 626 629 632
#> [379] 645 647 660 672 677 678 685 699 717 731   3   5  19  25  32 123 125 143
#> [397] 147 162 188 200 203 204 215 223 224 231 236 246 260 266 279 285 288 306
#> [415] 313 318 348 349 365 371 381 388 409 418 430 431 433 442 444 446 468 472
#> [433] 477 480 492 496 505 510 523 533 538 543 554 555 558 559 567 570 586 591
#> [451] 596 598 607 611 617 621 628 658 661 667 680 718 726 730 733 740 750  12
#> [469]  14  17  21  66  69  86  91  94  97 105 108 140 171 180 196 202 240 242
#> [487] 277 282 287 305 326 328 334 353 384 389 395 415 416 435 453 456 464 488
#> [505] 511 512 519 527 532 557 561 563 565 580 582 583 618 619 631 638 651 668
#> [523] 670 671 681 683 691 703 720 723 739 747   7  13  35  42  44  55  62  75
#> [541]  78  88  93 109 137 230 249 250 259 320 331 337 338 378 383 448 494 504
#> [559] 516 534 535 539 552 566 588 627 630 636 650 657 665 674 676 684 688 689
#> [577] 702 705 716 734 744 751   4   8  15  24  26  47  48  50  52  59  60  64
#> [595]  70  71  73  76  85  87  95  96  99 102 121 135 139 144 150 151 153 156
#> [613] 157 158 166 170 172 185 192 206 207 216 217 218 227 234 238 239 241 247
#> [631] 255 261 265 269 272 274 278 280 283 319 321 329 332 344 354 358 362 375
#> [649] 376 397 398 402 408 417 419 424 426 428 429 438 440 451 461 470 471 476
#> [667] 499 501 502 514 518 520 525 529 540 548 553 560 562 569 571 574 578 589
#> [685] 592 605 610 612 623 633 635 637 644 652 656 662 666 669 679 686 692 698
#> [703] 704 707 724 725 727 729 737 738 741 746 749

rcv$test_set(1)

#>  [1]   2  18  28  63  89 104 118 120 129 131 138 161 198 214 221 251 268 271 281
#> [20] 346 361 370 373 377 406 410 411 412 422 423 455 460 466 541 693 712 721 748

intersect(rcv$train_set(1), rcv$test_set(1))

#> integer(0)

# Internal storage:
rcv$instance

#>      row_id fold
#>   1:      2    1
#>   2:     18    1
#>   3:     28    1
#>   4:     63    1
#>   5:     89    1
#>  ---            
#> 747:    737   10
#> 748:    738   10
#> 749:    741   10
#> 750:    746   10
#> 751:    749   10