Spatial Cross Validation Resampling

Spatial Cross validation following the "k-means" approach after Brenning 2012.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi: 10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingSpCVCoords

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSpCVCoords$new()

ResamplingSpCVCoords$instantiate()

ResamplingSpCVCoords$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create an "Environmental Block" resampling instance.

Usage

ResamplingSpCVCoords$new(id = "spcv-coords")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSpCVCoords$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSpCVCoords$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv-coords")
rcv$param_set$values = list(folds = 3)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)

#>   [1]   2   3   5  10  18  19  29  37  39  42  45  49  55  57  58  61  62  68
#>  [19]  77  90  93 103 106 110 115 122 125 126 127 129 131 133 135 136 155 169
#>  [37] 170 183 190 201 203 216 218 219 220 222 230 231 234 236 243 245 250 255
#>  [55] 260 263 268 281 289 290 303 309 317 325 344 345 346 349 350 356 357 362
#>  [73] 370 372 373 383 385 394 400 403 404 405 412 413 414 422 423 426 427 442
#>  [91] 444 449 453 463 464 474 483 486 487 490 492 502 503 504 509 510 518 523
#> [109] 525 528 529 531 537 539 544 546 548 550 552 557 564 569 573 575 578 585
#> [127] 594 596 612 613 615 620 627 628 634 639 641 642 643 644 649 653 654 655
#> [145] 657 661 662 664 669 673 682 688 689 690 692 696 706 707 708 710 711 715
#> [163] 718 724 725 727 728 729 732 733 735 736 737 739 741 743 744 751   1   6
#> [181]   9  11  16  22  23  30  33  34  36  47  51  52  56  59  60  74  76  80
#> [199]  81  83  84  87  92  95  96  99 100 101 107 111 112 117 119 121 134 139
#> [217] 144 146 148 149 152 156 157 159 160 164 165 167 168 173 174 175 176 177
#> [235] 178 179 184 186 187 189 193 194 197 198 205 208 210 212 228 232 237 239
#> [253] 241 244 248 249 252 254 256 257 258 264 267 269 273 274 276 283 286 291
#> [271] 292 293 294 295 297 300 302 304 307 310 311 312 314 315 316 319 321 323
#> [289] 324 329 330 332 333 335 339 340 341 342 347 351 352 355 361 363 364 369
#> [307] 374 375 376 379 380 382 386 387 390 391 393 396 397 399 407 408 410 417
#> [325] 419 420 421 424 425 432 434 436 437 439 440 443 445 447 450 454 457 459
#> [343] 466 469 471 473 479 481 482 485 491 497 501 506 507 514 515 516 517 521
#> [361] 522 524 534 536 540 545 549 551 553 556 560 568 571 572 577 584 587 588
#> [379] 590 593 597 601 604 606 609 614 616 622 624 625 626 629 630 632 635 637
#> [397] 640 645 647 648 650 652 659 660 665 666 672 674 675 676 677 678 685 686
#> [415] 687 695 698 699 700 702 704 713 716 717 721 722 731 742 745 748 749

rcv$test_set(1)

#>   [1]   4   7   8  12  13  14  15  17  20  21  24  25  26  27  28  31  32  35
#>  [19]  38  40  41  43  44  46  48  50  53  54  63  64  65  66  67  69  70  71
#>  [37]  72  73  75  78  79  82  85  86  88  89  91  94  97  98 102 104 105 108
#>  [55] 109 113 114 116 118 120 123 124 128 130 132 137 138 140 141 142 143 145
#>  [73] 147 150 151 153 154 158 161 162 163 166 171 172 180 181 182 185 188 191
#>  [91] 192 195 196 199 200 202 204 206 207 209 211 213 214 215 217 221 223 224
#> [109] 225 226 227 229 233 235 238 240 242 246 247 251 253 259 261 262 265 266
#> [127] 270 271 272 275 277 278 279 280 282 284 285 287 288 296 298 299 301 305
#> [145] 306 308 313 318 320 322 326 327 328 331 334 336 337 338 343 348 353 354
#> [163] 358 359 360 365 366 367 368 371 377 378 381 384 388 389 392 395 398 401
#> [181] 402 406 409 411 415 416 418 428 429 430 431 433 435 438 441 446 448 451
#> [199] 452 455 456 458 460 461 462 465 467 468 470 472 475 476 477 478 480 484
#> [217] 488 489 493 494 495 496 498 499 500 505 508 511 512 513 519 520 526 527
#> [235] 530 532 533 535 538 541 542 543 547 554 555 558 559 561 562 563 565 566
#> [253] 567 570 574 576 579 580 581 582 583 586 589 591 592 595 598 599 600 602
#> [271] 603 605 607 608 610 611 617 618 619 621 623 631 633 636 638 646 651 656
#> [289] 658 663 667 668 670 671 679 680 681 683 684 691 693 694 697 701 703 705
#> [307] 709 712 714 719 720 723 726 730 734 738 740 746 747 750

# check that no obs are in both sets
intersect(rcv$train_set(1), rcv$test_set(1)) # good!

#> integer(0)

# Internal storage:
rcv$instance # table

#>      row_id fold
#>   1:      4    1
#>   2:      7    1
#>   3:      8    1
#>   4:     12    1
#>   5:     13    1
#>  ---            
#> 747:    731    3
#> 748:    742    3
#> 749:    745    3
#> 750:    748    3
#> 751:    749    3