Create spatio-temporal Folds Using Predefined Groups

Implementation of CAST::CreateSpaceTimeFolds().

Details

Using "class" is helpful in the case that data are clustered in space and are categorical. E.g This is the case for land cover classifications when training data come as training polygons. In this case the data should be split in a way that entire polygons are held back (spacevar="polygonID") but at the same time the distribution of classes should be similar in each fold (class="LUC").

References

Meyer H, Reudenbach C, Hengl T, Katurji M, Nauss T (2018). “Improving performance of spatio-temporal machine learning models using forward feature selection and target-oriented validation.” Environmental Modelling \& Software, 101, 1--9. doi: 10.1016/j.envsoft.2017.12.001 , https://doi.org/10.1016%2Fj.envsoft.2017.12.001.

Super class

mlr3::Resampling -> ResamplingSptCVCstf

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSptCVCstf$new()

ResamplingSptCVCstf$folds()

ResamplingSptCVCstf$repeats()

ResamplingSptCVCstf$instantiate()

ResamplingSptCVCstf$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spacetime Folds" resampling instance.

Usage

ResamplingSptCVCstf$new(id = "sptcv-cstf")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingSptCVCstf$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingSptCVCstf$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSptCVCstf$instantiate(
  task,
  space_var = NULL,
  time_var = NULL,
  class = NULL
)

Arguments

task

Task
A task to instantiate.

space_var

[character]
Column name identifying the spatial units.

time_var

[character]
Column name identifying the temporal units.

class

[character]
Column name identifying a class unit (e.g. land cover).

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSptCVCstf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("cookfarm")

# Instantiate Resampling
rcv = rsmp("sptcv-cstf", folds = 5)
rcv$instantiate(task, time_var = "Date", space_var = "SOURCEID")

# Individual sets:
rcv$train_set(1)

#>  [1]   7  12  23  27  42  47  50  67  71  74  86  93  98 110 125 135 136 141 155
#> [20] 157 160 162 169 178 181 200 203 204 213 220 223 228 234 240 243 245 247 261
#> [39] 262 265 267 275 277 279 285 286 317 328 329 332 335 340 346 351 353 360 365
#> [58] 371 374 375 381 385 389 392 393 406 439 442 445 447 452 453 454 462 467 471
#> [77] 480 482 483 493 497

rcv$test_set(1)

#>  [1]   1  28  37  56  58  70 103 105 119 127 142 186 196 208 258 295 307 350 358
#> [20] 378 411 473 485

# check that no obs are in both sets
intersect(rcv$train_set(1), rcv$test_set(1)) # good!

#> integer(0)

# Internal storage:
rcv$instance # table

#> $test
#> $test[[1]]
#>  [1]   1  28  37  56  58  70 103 105 119 127 142 186 196 208 258 295 307 350 358
#> [20] 378 411 473 485
#> 
#> $test[[2]]
#>  [1]  24  30  36  73 129 153 164 165 173 180 216 231 233 264 268 297 344 348 349
#> [20] 362 398 408 412 448
#> 
#> $test[[3]]
#>  [1]  26  43  63  82  94 113 116 167 185 241 288 369 390 402 414 421 423 428 470
#> [20] 476
#> 
#> $test[[4]]
#>  [1]   3  16  33 109 111 115 146 151 156 158 159 175 252 299 321 382 446 456
#> 
#> $test[[5]]
#>  [1]   2  13  18  41  97 143 174 229 248 249 278 296 343 379 384 441
#> 
#> 
#> $train
#> $train[[1]]
#>  [1]   7  12  23  27  42  47  50  67  71  74  86  93  98 110 125 135 136 141 155
#> [20] 157 160 162 169 178 181 200 203 204 213 220 223 228 234 240 243 245 247 261
#> [39] 262 265 267 275 277 279 285 286 317 328 329 332 335 340 346 351 353 360 365
#> [58] 371 374 375 381 385 389 392 393 406 439 442 445 447 452 453 454 462 467 471
#> [77] 480 482 483 493 497
#> 
#> $train[[2]]
#>  [1]   5  19  22  31  38  49  57  62  68  75  79  80  85  87  88  99 101 108 114
#> [20] 118 121 124 130 144 152 154 177 182 190 198 202 212 217 221 225 236 237 242
#> [39] 257 259 263 266 270 281 282 284 290 293 302 303 311 318 323 330 336 338 341
#> [58] 342 359 368 370 391 396 399 400 403 404 407 417 418 427 436 457 460 465 466
#> [77] 477 488 498
#> 
#> $train[[3]]
#>  [1]   9  20  21  32  39  44  45  46  53  55  60  61  66  92 100 104 126 132 133
#> [20] 140 147 149 150 161 163 168 170 187 189 191 195 201 214 218 219 224 227 235
#> [39] 238 244 246 251 253 254 274 280 283 287 289 292 298 310 313 322 327 347 352
#> [58] 367 373 376 377 386 387 394 395 409 416 430 433 444 449 450 459 472 474 475
#> [77] 479 491 495
#> 
#> $train[[4]]
#>  [1]   4   6  10  11  15  29  34  35  40  54  59  64  65  69  72  78  84  90  96
#> [20] 123 128 131 137 145 166 183 192 193 197 199 210 211 232 239 250 255 256 260
#> [39] 269 271 272 276 294 300 301 305 306 309 312 316 320 333 337 357 361 363 364
#> [58] 366 383 397 405 410 413 419 420 424 426 429 432 435 438 458 463 464 486 487
#> [77] 499 500
#> 
#> $train[[5]]
#>  [1]   8  14  17  25  48  51  52  76  77  81  83  89  91  95 102 106 107 112 117
#> [20] 120 122 134 138 139 148 171 172 176 179 184 188 194 205 206 207 209 215 222
#> [39] 226 230 273 291 304 308 314 315 319 324 325 326 331 334 339 345 354 355 356
#> [58] 372 380 388 401 415 422 425 431 434 437 440 443 451 455 461 468 469 478 481
#> [77] 484 489 490 492 494 496
#> 
#>