(CAST) "Leave-location-and-time-out" resampling

Create spatial, temporal or spatio-temporal Folds for cross validation

Details

Using "class" is helpful in the case that data are clustered in space and are categorical. E.g This is the case for land cover classifications when training data come as training polygons. In this case the data should be split in a way that entire polygons are held back (spacevar="polygonID") but at the same time the distribution of classes should be similar in each fold (class="LUC").

Note

Standard k-fold cross-validation can lead to considerable misinterpretation in spatial-temporal modelling tasks. This function can be used to prepare a Leave-Location-Out, Leave-Time-Out or Leave-Location-and-Time-Out cross-validation as target-oriented validation strategies for spatial-temporal prediction tasks. See Meyer et al. (2018) for further information.

mlr3spatiotempcv notes

The 'Description', 'Details' and 'Note' fields are inherited from the respective upstream function.

For a list of available arguments, please see CAST::CreateSpacetimeFolds.

References

Meyer H, Reudenbach C, Hengl T, Katurji M, Nauss T (2018). “Improving performance of spatio-temporal machine learning models using forward feature selection and target-oriented validation.” Environmental Modelling & Software, 101, 1--9. doi: 10.1016/j.envsoft.2017.12.001 .

Super class

mlr3::Resampling -> ResamplingSptCVCstf

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingSptCVCstf$new()

ResamplingSptCVCstf$instantiate()

ResamplingSptCVCstf$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spacetime Folds" resampling instance.

For a list of available arguments, please see CAST::CreateSpacetimeFolds.

Usage

ResamplingSptCVCstf$new(id = "sptcv_cstf")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingSptCVCstf$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingSptCVCstf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("cookfarm")

# Instantiate Resampling
rcv = rsmp("sptcv_cstf",
  folds = 5,
  time_var = "Date", space_var = "SOURCEID")
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)

#>   [1]   1   2   3   4   5   6   7   8   9  10  13  14  17  19  22  23  24  25
#>  [19]  26  27  30  33  34  35  36  39  41  42  43  46  47  49  50  51  54  57
#>  [37]  58  59  61  62  63  64  65  68  70  71  73  74  75  76  77  78  79  82
#>  [55]  85  86  87  88  90  91  92  93  95  96  97  98  99 101 102 103 105 106
#>  [73] 108 109 111 112 113 115 120 121 122 123 125 126 127 129 130 132 134 137
#>  [91] 141 142 143 144 145 150 151 155 156 157 160 166 168 169 170 172 175 177
#> [109] 178 179 180 181 183 184 186 187 188 190 193 195 196 197 198 200 201 202
#> [127] 204 207 208 209 216 217 218 221 222 224 225 226 227 229 232 233 235 237
#> [145] 238 240 241 242 243 245 246 248 249 250 251 254 255 256 259 260 261 262
#> [163] 263 264 268 270 271 273 275 276 277 278 279 280 281 282 283 284 288 289
#> [181] 291 293 295 296 297 299 300 301 302 303 305 306 308 309 310 311 312 314
#> [199] 315 316 317 318 319 320 322 323 325 326 333 334 335 336 337 338 342 343
#> [217] 344 345 346 348 349 351 352 353 354 359 360 361 362 365 366 367 368 369
#> [235] 370 372 373 375 376 380 382 383 384 386 387 388 389 390 391 393 394 395
#> [253] 397 398 399 400 401 402 403 404 406 407 408 409 412 413 416 417 419 420
#> [271] 422 423 424 426 427 428 429 430 432 434 437 438 439 441 442 443 444 446
#> [289] 450 453 454 455 456 458 460 461 463 465 466 467 468 472 473 475 476 477
#> [307] 478 479 482 483 484 486 487 488 489 490 492 493 494 495 496 497 499

rcv$test_set(1)

#>  [1]  89 118 139 173 182 228 230 253 266 267 269 298 340 341 377 385 396 445 457
#> [20] 469 500

# check that no obs are in both sets
intersect(rcv$train_set(1), rcv$test_set(1)) # good!

#> integer(0)

# Internal storage:
rcv$instance # table

#> $test
#> $test[[1]]
#>  [1]  89 118 139 173 182 228 230 253 266 267 269 298 340 341 377 385 396 445 457
#> [20] 469 500
#> 
#> $test[[2]]
#>  [1]  34  54  63 105 126 127 137 169 175 180 188 237 276 279 295 308 394 407 430
#> [20] 486 496
#> 
#> $test[[3]]
#>  [1]   2  13  98 106 155 221 245 259 260 277 362 402 427 432 450 458 460 463 487
#> 
#> $test[[4]]
#>  [1]  49  79  82 145 200 224 233 250 256 263 337 344 346 373 376 422 437 466
#> 
#> $test[[5]]
#>  [1]   1  33  51  68  87 108 160 225 322 323 369 386 391 395 438 443
#> 
#> 
#> $train
#> $train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  13  14  17  19  22  23  24  25
#>  [19]  26  27  30  33  34  35  36  39  41  42  43  46  47  49  50  51  54  57
#>  [37]  58  59  61  62  63  64  65  68  70  71  73  74  75  76  77  78  79  82
#>  [55]  85  86  87  88  90  91  92  93  95  96  97  98  99 101 102 103 105 106
#>  [73] 108 109 111 112 113 115 120 121 122 123 125 126 127 129 130 132 134 137
#>  [91] 141 142 143 144 145 150 151 155 156 157 160 166 168 169 170 172 175 177
#> [109] 178 179 180 181 183 184 186 187 188 190 193 195 196 197 198 200 201 202
#> [127] 204 207 208 209 216 217 218 221 222 224 225 226 227 229 232 233 235 237
#> [145] 238 240 241 242 243 245 246 248 249 250 251 254 255 256 259 260 261 262
#> [163] 263 264 268 270 271 273 275 276 277 278 279 280 281 282 283 284 288 289
#> [181] 291 293 295 296 297 299 300 301 302 303 305 306 308 309 310 311 312 314
#> [199] 315 316 317 318 319 320 322 323 325 326 333 334 335 336 337 338 342 343
#> [217] 344 345 346 348 349 351 352 353 354 359 360 361 362 365 366 367 368 369
#> [235] 370 372 373 375 376 380 382 383 384 386 387 388 389 390 391 393 394 395
#> [253] 397 398 399 400 401 402 403 404 406 407 408 409 412 413 416 417 419 420
#> [271] 422 423 424 426 427 428 429 430 432 434 437 438 439 441 442 443 444 446
#> [289] 450 453 454 455 456 458 460 461 463 465 466 467 468 472 473 475 476 477
#> [307] 478 479 482 483 484 486 487 488 489 490 492 493 494 495 496 497 499
#> 
#> $train[[2]]
#>   [1]   1   2   3   4   5   7  10  11  12  13  14  15  16  17  18  21  22  23
#>  [19]  24  28  29  30  31  32  33  37  38  40  42  45  47  48  49  50  51  52
#>  [37]  53  55  59  61  65  66  67  68  69  74  76  78  79  80  81  82  83  87
#>  [55]  88  89  90  96  97  98  99 100 103 104 106 107 108 109 110 111 113 114
#>  [73] 116 117 118 119 121 123 124 128 132 133 134 135 136 138 139 142 143 144
#>  [91] 145 146 147 149 150 151 152 153 154 155 156 157 159 160 161 162 163 164
#> [109] 165 172 173 174 177 178 179 182 185 189 191 192 193 194 196 197 198 199
#> [127] 200 202 204 205 206 207 210 211 215 216 218 219 221 224 225 226 228 229
#> [145] 230 233 234 235 236 238 240 241 242 243 244 245 246 247 249 250 253 254
#> [163] 255 256 257 259 260 261 263 264 265 266 267 269 271 272 275 277 278 283
#> [181] 284 287 289 290 294 296 297 298 302 304 305 307 309 311 312 313 314 317
#> [199] 318 320 321 322 323 324 325 327 328 329 331 332 335 336 337 338 340 341
#> [217] 343 344 346 347 348 349 351 352 355 357 359 360 361 362 363 364 367 368
#> [235] 369 370 372 373 374 376 377 378 379 380 383 384 385 386 388 389 391 392
#> [253] 393 395 396 399 400 402 404 406 409 410 411 415 417 419 421 422 424 425
#> [271] 426 427 428 429 431 432 433 435 437 438 442 443 445 446 447 448 449 450
#> [289] 452 453 454 457 458 459 460 461 462 463 464 466 467 469 470 471 472 473
#> [307] 474 475 476 477 480 481 485 487 491 492 493 494 500
#> 
#> $train[[3]]
#>   [1]   1   3   6   8   9  11  12  15  17  19  20  25  26  27  29  31  32  33
#>  [19]  34  35  36  41  42  43  44  45  47  48  49  51  52  54  56  58  60  61
#>  [37]  62  63  66  68  69  70  72  73  74  75  77  78  79  81  82  84  85  87
#>  [55]  89  90  91  92  93  94  95  96  97  99 100 102 103 104 105 107 108 110
#>  [73] 112 113 114 115 116 117 118 119 122 125 126 127 129 130 131 135 137 138
#>  [91] 139 140 141 145 147 148 149 152 153 154 158 159 160 161 162 163 164 165
#> [109] 166 167 169 171 172 173 174 175 176 177 179 180 181 182 184 185 186 187
#> [127] 188 190 191 200 201 203 205 206 208 209 211 212 213 214 215 219 220 222
#> [145] 223 224 225 226 227 228 230 231 232 233 234 235 237 238 239 244 246 250
#> [163] 251 252 253 254 255 256 257 258 262 263 265 266 267 268 269 270 272 273
#> [181] 274 275 276 279 280 281 284 285 286 287 289 291 292 293 294 295 296 298
#> [199] 299 300 301 303 304 306 307 308 310 312 313 314 315 316 318 319 320 321
#> [217] 322 323 327 328 329 330 331 332 333 334 337 339 340 341 344 345 346 350
#> [235] 353 354 355 356 358 359 361 363 364 365 366 368 369 371 373 376 377 378
#> [253] 379 381 382 384 385 386 387 388 390 391 394 395 396 397 401 403 405 407
#> [271] 410 411 412 414 415 418 422 425 426 430 436 437 438 439 440 441 443 444
#> [289] 445 447 448 449 451 453 456 457 459 461 462 464 465 466 468 469 470 471
#> [307] 477 478 479 480 482 483 484 486 489 490 491 492 496 498 499 500
#> 
#> $train[[4]]
#>   [1]   1   2   4   5   7  10  11  12  13  16  18  20  21  23  27  28  30  33
#>  [19]  34  36  37  38  39  40  44  45  46  48  51  53  54  55  56  57  58  60
#>  [37]  63  64  67  68  71  72  73  77  80  81  83  84  85  86  87  89  92  94
#>  [55]  95  98 100 101 104 105 106 107 108 111 112 114 118 119 120 123 124 126
#>  [73] 127 128 131 133 135 136 137 139 140 142 144 146 147 148 152 153 154 155
#>  [91] 156 157 158 159 160 162 164 165 167 168 169 170 171 173 174 175 176 178
#> [109] 180 181 182 183 187 188 189 191 192 194 195 198 199 203 205 207 208 210
#> [127] 211 212 213 214 215 217 219 220 221 223 225 228 230 231 232 236 237 239
#> [145] 240 241 243 244 245 247 248 251 252 253 257 258 259 260 261 262 264 266
#> [163] 267 269 270 272 274 276 277 279 280 282 283 285 286 288 290 292 295 297
#> [181] 298 300 301 302 303 304 307 308 310 316 317 319 321 322 323 324 326 328
#> [199] 329 330 333 334 339 340 341 342 343 345 347 348 349 350 353 354 355 356
#> [217] 357 358 362 364 366 367 369 371 374 375 377 379 381 382 385 386 391 392
#> [235] 393 394 395 396 398 399 400 402 403 405 407 408 410 413 414 416 418 420
#> [253] 421 423 425 427 428 429 430 431 432 433 434 435 436 438 439 440 441 442
#> [271] 443 445 447 449 450 451 452 455 457 458 460 463 464 465 467 468 469 470
#> [289] 473 474 478 480 481 482 484 485 486 487 488 490 491 493 495 496 497 498
#> [307] 499 500
#> 
#> $train[[5]]
#>   [1]   2   6   8   9  13  14  15  16  18  19  20  21  22  24  25  26  28  29
#>  [19]  31  32  34  35  37  38  39  40  41  43  44  46  49  50  52  53  54  55
#>  [37]  56  57  59  60  62  63  64  65  66  67  69  70  71  72  75  76  79  80
#>  [55]  82  83  84  86  88  89  91  93  94  98 101 102 105 106 109 110 115 116
#>  [73] 117 118 120 121 122 124 125 126 127 128 129 130 131 132 133 134 136 137
#>  [91] 138 139 140 141 143 145 146 148 149 150 151 155 158 161 163 166 167 168
#> [109] 169 170 171 173 175 176 180 182 183 184 185 186 188 189 190 192 193 194
#> [127] 195 196 197 199 200 201 202 203 204 206 209 210 212 213 214 216 217 218
#> [145] 220 221 222 223 224 227 228 229 230 231 233 234 236 237 239 242 245 247
#> [163] 248 249 250 252 253 256 258 259 260 263 265 266 267 268 269 271 273 274
#> [181] 276 277 278 279 281 282 285 286 287 288 290 291 292 293 294 295 298 299
#> [199] 305 306 308 309 311 313 315 324 325 326 327 330 331 332 335 336 337 338
#> [217] 339 340 341 342 344 346 347 350 351 352 356 357 358 360 362 363 365 370
#> [235] 371 372 373 374 375 376 377 378 380 381 383 385 387 389 390 392 394 396
#> [253] 397 398 401 402 404 405 406 407 408 409 411 412 413 414 415 416 417 418
#> [271] 419 420 421 422 423 424 427 430 431 432 433 434 435 436 437 440 444 445
#> [289] 446 448 450 451 452 454 455 456 457 458 459 460 462 463 466 469 471 472
#> [307] 474 475 476 479 481 483 485 486 487 488 489 494 495 496 497 498 500
#> 
#>