SpatioTemporal Regression Task

This task specializes Task and TaskSupervised for spatiotemporal classification problems.

A spatial example task is available via tsk("ecuador"), a spatiotemporal one via tsk("cookfarm").

The coordinate reference system passed during initialization must match the one which was used during data creation, otherwise offsets of multiple meters may occur. By default, coordinates are not used as features. This can be changed by setting extra_args$coords_as_features = TRUE.

See also

Super classes

mlr3::Task -> mlr3::TaskSupervised -> mlr3::TaskRegr -> TaskRegrST

Public fields

extra_args

(named list())
Additional task arguments set during construction. Required for convert_task().

Methods

Public methods

TaskRegrST$new()

TaskRegrST$coordinates()

TaskRegrST$print()

TaskRegrST$clone()

Inherited methods

mlr3::Task$cbind()

mlr3::Task$data()

mlr3::Task$droplevels()

mlr3::Task$filter()

mlr3::Task$format()

mlr3::Task$formula()

mlr3::Task$head()

mlr3::Task$help()

mlr3::Task$levels()

mlr3::Task$missings()

mlr3::Task$rbind()

mlr3::Task$rename()

mlr3::Task$select()

mlr3::Task$set_col_roles()

mlr3::Task$set_row_roles()

mlr3::TaskRegr$truth()

Method new()

Create a new spatiotemporal resampling Task

Usage

TaskRegrST$new(
  id,
  backend,
  target,
  extra_args = list(coords_as_features = FALSE, crs = NA, coordinate_names = NA)
)

Arguments

id

[character(1)]
Identifier for the task.

backend

DataBackend
Either a DataBackend, or any object which is convertible to a DataBackend with as_data_backend(). E.g., a data.frame() will be converted to a DataBackendDataTable.

target

[character(1)]
Name of the target column.

extra_args

[named list]
Additional task arguments set during construction. Required for convert_task().

crs [character(1)]
Coordinate reference system. Either a PROJ string or an EPSG code.

coords_as_features [logical(1)]
Whether the coordinates should also be used as features.

coordinate_names [character(2)]
The variables names of the coordinates in the data.

Method coordinates()

Return the coordinates of the task

Usage

TaskRegrST$coordinates(rows = NULL)

Arguments

rows

Row IDs. Can be used to subset the returned coordinates.

Method print()

Print the task.

Usage

TaskRegrST$print(...)

Arguments

...

Arguments passed to the $print() method of the superclass.

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskRegrST$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.