Visualization Functions for SptCV Cstf Methods.

Generic S3 plot() and autoplot() (ggplot2) methods.

# S3 method for ResamplingSptCVCstf
autoplot(
  object,
  task,
  fold_id = NULL,
  plot_as_grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  tickformat_date = "%Y-%m",
  crs = NULL,
  nticks_x = 3,
  nticks_y = 3,
  point_size = 3,
  axis_label_fontsize = 11,
  ...
)

# S3 method for ResamplingRepeatedSptCVCstf
autoplot(
  object,
  task,
  fold_id = NULL,
  repeats_id = NULL,
  plot_as_grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  tickformat_date = "%Y-%m",
  crs = NULL,
  nticks_x = 3,
  nticks_y = 3,
  point_size = 3,
  axis_label_fontsize = 11,
  ...
)

# S3 method for ResamplingSptCVCstf
plot(x, ...)

# S3 method for ResamplingRepeatedSptCVCstf
plot(x, ...)

Arguments

object
[Resampling]
mlr3 spatial resampling object of class ResamplingSptCVCstf or ResamplingRepeatedSptCVCstf.

task
[TaskClassifST]/[TaskRegrST]
mlr3 task object.

fold_id
[numeric]
Fold IDs to plot.

plot_as_grid
[logical(1)]
Should a gridded plot using via patchwork be created? If FALSE only a list with all ggplot2 resamplings is returned. Only applies if a numeric vector is passed to argument fold_id.

train_color
[character(1)]
The color to use for the training set observations.

test_color
[character(1)]
The color to use for the test set observations.

tickformat_date
[character]
Date format for z-axis.

crs
[character]
EPSG code of the CRS for x and y axes. If not set, EPSG 4326 (WGS84) is used.

nticks_x
[integer]
Number of x axis breaks.

nticks_y
[integer]
Number of y axis breaks.

point_size
[numeric]
Point size of markers.

axis_label_fontsize
[integer]
Font size of axis labels.

...
Not used.

repeats_id
[numeric]
Repetition ID to plot.

x
[Resampling]
mlr3 spatial resampling object of class ResamplingSptCVCstf or ResamplingRepeatedSptCVCstf.

Examples

# \donttest{
library(mlr3)
library(mlr3spatiotempcv)
task_st = tsk("cookfarm")
resampling = rsmp("sptcv_cstf", folds = 5, time_var = "Date")
resampling$instantiate(task_st)

# plot
autoplot(resampling, task_st)

#> Warning: `arrange_()` is deprecated as of dplyr 0.7.0.
#> Please use `arrange()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

autoplot(resampling, task_st, 1)
autoplot(resampling, task_st, c(1, 2))

#> Unfortunately plotly does not support a dynamic arrangement of multiple
#> subplots. See article 'Visualization of spatiotemporal clusters'
#> (https://mlr3spatiotempcv.mlr-org.com/articles/spatiotemp-viz) for a
#> manual workaround. Use the objects in the returned list to arrange a
#> custom grid.

# }