mlr3spatiotempcv: Spatiotemporal Resampling Methods for 'mlr3'

Extends the mlr3 machine learning framework with spatio-temporal resampling methods to account for the presence of spatiotemporal autocorrelation (STAC) in predictor variables. STAC may cause highly biased performance estimates in cross-validation if ignored. A JSS article is available at doi:10.18637/jss.v111.i07 .

Main resources

Book on mlr3: https://mlr3book.mlr-org.com

mlr3book section about spatiotemporal data: https://mlr3book.mlr-org.com/chapters/chapter13/beyond_regression_and_classification.html#spatiotemp-cv

package vignettes: https://mlr3spatiotempcv.mlr-org.com/dev/articles/

Miscellaneous mlr3 content

Use cases and examples: https://mlr3gallery.mlr-org.com

More classification and regression tasks: mlr3data

Connector to OpenML: mlr3oml

More classification and regression learners: mlr3learners

Even more learners: https://github.com/mlr-org/mlr3extralearners

Preprocessing and machine learning pipelines: mlr3pipelines

Tuning of hyperparameters: mlr3tuning

Visualizations for many mlr3 objects: mlr3viz

Survival analysis and probabilistic regression: mlr3proba

Cluster analysis: mlr3cluster

Feature selection filters: mlr3filters

Feature selection wrappers: mlr3fselect

Interface to real (out-of-memory) data bases: mlr3db

Performance measures as plain functions: mlr3measures

Parallelization framework: future

Progress bars: progressr

References

Schratz P, Muenchow J, Iturritxa E, Richter J, Brenning A (2019). “Hyperparameter tuning and performance assessment of statistical and machine-learning algorithms using spatial data.” Ecological Modelling, 406, 109–120. doi:10.1016/j.ecolmodel.2019.06.002 .

Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G (2018). “blockCV: an R package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models.” bioRxiv. doi:10.1101/357798 .

Meyer H, Reudenbach C, Hengl T, Katurji M, Nauss T (2018). “Improving performance of spatio-temporal machine learning models using forward feature selection and target-oriented validation.” Environmental Modelling & Software, 101, 1–9. doi:10.1016/j.envsoft.2017.12.001 .

Zhao Y, Karypis G (2002). “Evaluation of Hierarchical Clustering Algorithms for Document Datasets.” 11th Conference of Information and Knowledge Management (CIKM), 51-524. doi:10.1145/584792.584877 .

See also

Author

Maintainer: Patrick Schratz patrick.schratz@gmail.com (ORCID)

Authors:

Marc Becker marcbecker@posteo.de (ORCID)

Other contributors:

Jannes Muenchow jannes.muenchow@uni-jena.de (ORCID) [contributor]

Michel Lang michellang@gmail.com (ORCID) [contributor]