(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   4   6   7   8   9  10  11  12  13  14  15  16  17  18  20  21
#>  [19]  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 118 119 120 121 122 123 124 126 128 129 130 131
#> [127] 132 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150
#> [145] 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186
#> [181] 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 204 205
#> [199] 206 207 208 209 210 211 212 213 214 215 216 217 218 219 221 222 223 224
#> [217] 225 226 227 228 229 230 232 233 234 235 237 238 239 240 241 242 243 244
#> [235] 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 261 262 263
#> [253] 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280 281
#> [271] 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299
#> [289] 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317
#> [307] 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335
#> [325] 336 337 338 339 340 341 342 343 344 345 346 347 348 349 350 351 352 353
#> [343] 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371
#> [361] 372 373 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388 389
#> [379] 390 391 392 393 395 396 397 398 399 400 401 402 403 404 405 406 407 408
#> [397] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426
#> [415] 427 428 429 430 431 432 433 434 435 436 437 438 439 440 441 443 444 445
#> [433] 446 447 448 449 450 451 452 453 454 455 456 457 458 459 460 461 462 463
#> [451] 464 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481
#> [469] 482 483 484 485 486 487 488 489 490 491 493 494 495 496 497 498 499 500
#> [487] 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518
#> [505] 519 520 521 522 524 525 526 527 529 530 532 533 534 535 536 537 538 539
#> [523] 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557
#> [541] 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575
#> [559] 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591 592 593
#> [577] 594 595 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 612
#> [595] 613 614 615 616 617 618 619 620 621 622 623 624 625 626 627 629 630 631
#> [613] 632 633 634 635 636 637 638 639 640 642 643 644 645 646 647 648 649 650
#> [631] 651 652 654 655 656 657 658 659 660 662 663 664 665 666 667 668 669 670
#> [649] 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688
#> [667] 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706
#> [685] 707 708 709 710 711 712 713 714 715 716 717 719 720 721 722 723 724 725
#> [703] 726 727 729 730 731 733 734 736 737 738 739 740 741 742 743 744 745 746
#> [721] 747 748 749 750 751
rrcv$test_set(1)
#>  [1]   3   5  19 125 203 231 236 260 442 492 523 661 718
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>  [1]   3   5  19 125 203 231 236 260 442 492 523 661 718
#> 
#> [[1]]$test[[2]]
#>  [1]  96 121 139 152 157 267 274 300 329 332 375 408 417 473 553 571 604 686 704
#> [20] 749
#> 
#> [[1]]$test[[3]]
#>  [1]  10  39  57  68 106 155 201 219 263 289 317 372 385 414 427 642 649
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   4   6   7   8   9  10  11  12  13  14  15  16  17  18  20  21
#>  [19]  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 118 119 120 121 122 123 124 126 128 129 130 131
#> [127] 132 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150
#> [145] 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186
#> [181] 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 204 205
#> [199] 206 207 208 209 210 211 212 213 214 215 216 217 218 219 221 222 223 224
#> [217] 225 226 227 228 229 230 232 233 234 235 237 238 239 240 241 242 243 244
#> [235] 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 261 262 263
#> [253] 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280 281
#> [271] 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299
#> [289] 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317
#> [307] 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335
#> [325] 336 337 338 339 340 341 342 343 344 345 346 347 348 349 350 351 352 353
#> [343] 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371
#> [361] 372 373 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388 389
#> [379] 390 391 392 393 395 396 397 398 399 400 401 402 403 404 405 406 407 408
#> [397] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426
#> [415] 427 428 429 430 431 432 433 434 435 436 437 438 439 440 441 443 444 445
#> [433] 446 447 448 449 450 451 452 453 454 455 456 457 458 459 460 461 462 463
#> [451] 464 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481
#> [469] 482 483 484 485 486 487 488 489 490 491 493 494 495 496 497 498 499 500
#> [487] 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518
#> [505] 519 520 521 522 524 525 526 527 529 530 532 533 534 535 536 537 538 539
#> [523] 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557
#> [541] 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575
#> [559] 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591 592 593
#> [577] 594 595 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 612
#> [595] 613 614 615 616 617 618 619 620 621 622 623 624 625 626 627 629 630 631
#> [613] 632 633 634 635 636 637 638 639 640 642 643 644 645 646 647 648 649 650
#> [631] 651 652 654 655 656 657 658 659 660 662 663 664 665 666 667 668 669 670
#> [649] 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688
#> [667] 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706
#> [685] 707 708 709 710 711 712 713 714 715 716 717 719 720 721 722 723 724 725
#> [703] 726 727 729 730 731 733 734 736 737 738 739 740 741 742 743 744 745 746
#> [721] 747 748 749 750 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 140 141 142 143 144 145 146 147 148 149 150
#> [145] 151 153 154 155 156 158 159 160 161 162 163 164 165 166 167 168 169 170
#> [163] 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188
#> [181] 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206
#> [199] 207 208 209 210 211 213 214 215 216 217 218 219 220 221 222 223 224 225
#> [217] 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243
#> [235] 244 245 246 247 248 249 250 251 252 253 254 255 256 257 259 260 261 262
#> [253] 263 264 265 266 268 269 270 271 272 273 275 276 277 278 279 280 281 282
#> [271] 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299 301
#> [289] 302 303 304 305 306 307 308 309 310 311 313 314 315 316 317 318 320 321
#> [307] 322 323 324 325 326 327 328 330 331 333 334 335 336 337 338 339 340 341
#> [325] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [343] 360 361 362 363 364 365 366 367 368 370 371 372 373 374 376 377 378 379
#> [361] 380 381 382 383 384 385 386 388 389 390 391 392 393 394 395 396 397 398
#> [379] 400 401 402 403 404 405 406 407 409 410 411 412 413 414 415 416 418 419
#> [397] 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437
#> [415] 438 439 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456
#> [433] 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472 474 475
#> [451] 476 477 478 479 480 481 482 483 484 485 486 487 488 489 490 491 492 493
#> [469] 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511
#> [487] 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529
#> [505] 530 531 532 533 535 536 537 538 539 540 541 542 543 544 545 546 547 548
#> [523] 549 550 551 552 554 555 556 557 558 559 561 562 563 564 565 566 567 568
#> [541] 569 570 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586 587
#> [559] 588 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 605 606
#> [577] 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624
#> [595] 625 626 627 628 629 630 631 632 633 634 635 636 638 639 640 641 642 643
#> [613] 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661
#> [631] 662 663 664 665 666 667 668 669 670 671 672 673 674 675 677 678 679 680
#> [649] 681 682 683 684 685 687 688 689 690 691 692 693 694 695 696 697 698 699
#> [667] 700 701 702 703 705 706 707 708 709 710 711 712 713 714 715 717 718 719
#> [685] 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737
#> [703] 738 739 740 741 742 743 744 745 746 747 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38  40
#>  [37]  41  42  43  44  46  47  48  49  50  51  52  53  54  55  56  58  59  60
#>  [55]  62  63  64  65  66  67  69  70  71  72  73  74  75  76  78  79  80  81
#>  [73]  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 107 108 109 110 111 112 113 114 116 117 118 119
#> [109] 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137
#> [127] 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153 154 156
#> [145] 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172 173 174
#> [163] 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 191 192 193
#> [181] 194 195 196 197 198 199 200 202 203 204 205 206 207 208 209 210 211 212
#> [199] 213 214 215 216 217 218 220 221 222 223 224 225 226 227 228 229 230 231
#> [217] 232 233 234 235 236 237 238 239 240 241 242 244 245 246 247 248 249 250
#> [235] 251 252 253 254 255 256 257 258 259 260 261 262 264 265 266 267 268 269
#> [253] 270 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285 286 287
#> [271] 288 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306
#> [289] 307 308 309 310 311 312 313 314 315 316 318 319 320 321 322 323 324 326
#> [307] 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344
#> [325] 345 346 347 348 349 351 352 353 354 355 356 357 358 359 360 361 362 363
#> [343] 364 365 366 367 368 369 370 371 373 374 375 376 377 378 379 380 381 382
#> [361] 383 384 386 387 388 389 390 391 392 393 394 395 396 397 398 399 401 402
#> [379] 403 406 407 408 409 410 411 412 413 415 416 417 418 419 420 421 422 423
#> [397] 424 425 426 428 429 430 431 432 433 434 435 436 437 438 439 440 441 442
#> [415] 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 458 459 460
#> [433] 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475 476 477 478
#> [451] 479 480 481 482 483 484 485 486 488 489 490 491 492 493 494 495 496 497
#> [469] 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515
#> [487] 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532 533
#> [505] 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549 551 552
#> [523] 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568 569 570
#> [541] 571 572 573 574 576 577 578 579 580 581 582 583 584 585 586 587 588 589
#> [559] 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607
#> [577] 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 625
#> [595] 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 644 645
#> [613] 646 647 648 650 651 652 653 654 655 656 657 658 659 660 661 662 663 665
#> [631] 666 667 668 669 670 671 672 674 675 676 677 678 679 680 681 682 683 684
#> [649] 685 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702
#> [667] 703 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720
#> [685] 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738
#> [703] 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>  [1]  27  53  67  69  72  82  98 108 114 141 154 233 235 275 296 328 353 389 392
#> [20] 401 415 441 462 576 600
#> 
#> [[2]]$test[[2]]
#> [1]  73 217 520 605 610 679
#> 
#> [[2]]$test[[3]]
#>  [1]   7  13  35  44  75  78  79  88 109 259 320 331 378 448 494 705
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  21  22  23  24  25  26  28  29  30  32  33  34  35  36  37  38  39
#>  [37]  42  43  44  45  46  47  48  49  50  51  52  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  68  70  71  73  74  75  76  77  78  79  80  81
#>  [73]  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  99 100 101
#>  [91] 102 103 104 105 106 107 109 110 111 112 113 115 116 117 118 119 120 121
#> [109] 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139
#> [127] 140 142 143 144 145 146 147 148 149 150 151 152 153 155 156 157 158 159
#> [145] 160 161 162 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177
#> [163] 178 179 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196
#> [181] 197 198 199 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214
#> [199] 215 216 217 218 219 220 221 222 223 224 225 227 228 229 230 231 232 234
#> [217] 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253
#> [235] 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271
#> [253] 272 273 274 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290
#> [271] 291 292 293 294 295 297 300 302 303 304 305 306 307 308 309 310 311 312
#> [289] 313 314 315 316 317 318 319 320 321 322 323 324 325 327 329 330 331 332
#> [307] 333 335 336 337 338 339 340 341 342 344 345 346 347 348 349 350 351 352
#> [325] 354 355 356 357 358 359 361 362 363 364 365 366 368 369 370 371 372 373
#> [343] 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388 390 391 393
#> [361] 394 395 396 397 398 399 400 402 403 404 405 406 407 408 409 410 411 412
#> [379] 413 414 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431
#> [397] 432 433 434 435 436 437 438 439 440 442 443 444 445 446 447 448 449 450
#> [415] 451 452 453 454 455 457 458 459 460 461 463 464 465 466 467 468 469 470
#> [433] 471 472 473 474 475 476 477 478 479 480 481 482 483 485 486 487 488 489
#> [451] 490 491 492 493 494 495 496 497 498 499 501 502 503 504 505 506 507 509
#> [469] 510 511 512 513 514 515 516 517 518 520 521 522 523 524 525 526 527 528
#> [487] 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546
#> [505] 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564
#> [523] 566 567 568 569 570 571 572 573 574 575 577 578 579 580 581 582 584 585
#> [541] 586 587 588 589 590 591 592 593 594 596 597 598 599 601 602 604 605 606
#> [559] 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624
#> [577] 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642
#> [595] 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660
#> [613] 661 662 663 664 665 666 667 669 671 672 673 674 675 676 677 678 679 680
#> [631] 681 682 683 684 685 686 687 688 689 690 691 692 693 695 696 697 698 699
#> [649] 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 717
#> [667] 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735
#> [685] 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  21  22  23  24  25  26  27  28  29  30  32  33  34  35  36  37  38
#>  [37]  39  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  72  74  75  76  77  78
#>  [73]  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96
#>  [91]  97  98  99 100 101 103 104 105 106 107 109 110 111 112 113 114 115 116
#> [109] 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134
#> [127] 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152
#> [145] 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170
#> [163] 171 172 173 174 175 176 177 178 179 181 182 183 184 186 187 188 189 190
#> [181] 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206 207 208
#> [199] 209 210 211 212 213 214 215 216 218 219 220 221 222 223 224 225 227 228
#> [217] 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 244 245 246
#> [235] 248 249 250 251 252 253 254 255 256 257 258 259 260 261 262 263 264 265
#> [253] 266 267 268 269 270 271 272 273 274 275 276 277 279 280 281 282 283 284
#> [271] 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299 300 301 302
#> [289] 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318 319 320
#> [307] 321 322 323 324 325 326 327 329 330 331 332 333 334 335 336 337 338 339
#> [325] 340 341 342 344 345 346 347 348 349 350 351 352 354 355 356 357 358 359
#> [343] 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378
#> [361] 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396
#> [379] 397 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415
#> [397] 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433
#> [415] 434 435 436 437 439 440 441 442 443 444 445 446 447 448 449 450 451 452
#> [433] 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470
#> [451] 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488
#> [469] 489 490 491 492 493 494 495 496 497 498 499 501 502 503 504 505 506 507
#> [487] 508 509 510 511 512 514 515 516 517 518 519 521 522 523 524 525 526 527
#> [505] 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545
#> [523] 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563
#> [541] 564 565 566 567 568 569 570 571 572 573 575 576 577 578 579 580 581 583
#> [559] 584 585 586 587 588 590 591 592 593 594 595 596 597 598 599 601 602 603
#> [577] 604 606 607 609 611 612 613 614 615 616 617 618 619 620 621 622 623 624
#> [595] 625 626 627 628 629 630 631 632 634 635 636 637 638 639 640 641 642 643
#> [613] 644 645 646 647 648 649 650 651 652 653 654 655 657 658 659 660 661 662
#> [631] 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678 680 681
#> [649] 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696 697 698 699
#> [667] 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 717
#> [685] 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735
#> [703] 736 737 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   6   8   9  10  11  12  14  15  16  17  18  19  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  33  34  36  37  38  39
#>  [37]  40  41  42  43  45  46  47  48  49  50  51  52  53  55  56  57  58  59
#>  [55]  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  76  77  80
#>  [73]  81  82  83  84  85  86  87  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 106 107 108 110 111 112 114 115 116 117 118 119
#> [109] 120 121 122 123 124 125 126 127 129 131 133 134 135 136 138 139 140 141
#> [127] 143 144 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161
#> [145] 162 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180
#> [163] 183 184 185 186 187 188 189 190 192 193 194 195 196 197 198 199 200 201
#> [181] 202 203 204 205 206 207 208 210 211 212 214 215 216 217 218 219 220 221
#> [199] 222 223 224 226 227 228 230 231 232 233 234 235 236 237 238 239 240 241
#> [217] 242 243 244 245 246 247 248 249 250 251 252 254 255 256 257 258 260 261
#> [235] 263 264 265 266 267 268 269 271 272 273 274 275 276 277 278 279 280 281
#> [253] 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299
#> [271] 300 301 302 303 304 305 306 307 309 310 311 312 313 314 315 316 317 318
#> [289] 319 321 323 324 325 326 328 329 330 332 333 334 335 339 340 341 342 343
#> [307] 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361
#> [325] 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 379 380
#> [343] 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398
#> [361] 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416
#> [379] 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434
#> [397] 435 436 437 438 439 440 441 442 443 444 445 446 447 449 450 451 452 453
#> [415] 454 455 456 457 459 460 461 462 463 464 466 467 468 469 470 471 472 473
#> [433] 474 476 477 479 480 481 482 483 484 485 486 487 488 489 490 491 492 495
#> [451] 496 497 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514
#> [469] 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532
#> [487] 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549 550
#> [505] 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [523] 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586
#> [541] 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604
#> [559] 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622
#> [577] 623 624 625 626 627 628 629 630 631 632 633 634 635 637 638 639 640 641
#> [595] 642 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659
#> [613] 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677
#> [631] 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695
#> [649] 696 697 698 699 700 702 703 704 706 707 708 709 710 711 712 713 715 716
#> [667] 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734
#> [685] 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#>