(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

mlr3::Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  43  44  45  46  47  48  49  50  51  52  53  54  56
#>  [55]  57  58  59  60  61  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129
#> [127] 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147
#> [145] 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165
#> [163] 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183
#> [181] 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201
#> [199] 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219
#> [217] 220 221 222 223 224 225 226 227 228 229 231 232 233 234 235 236 237 238
#> [235] 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256
#> [253] 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274
#> [271] 275 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290 291 292
#> [289] 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310
#> [307] 311 312 313 314 315 316 317 318 319 320 321 322 323 324 325 326 327 328
#> [325] 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344 345 346
#> [343] 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364
#> [361] 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379 380 381 382
#> [379] 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401
#> [397] 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419
#> [415] 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437
#> [433] 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455
#> [451] 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473
#> [469] 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488 489 490 491
#> [487] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509
#> [505] 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527
#> [523] 528 529 530 531 532 533 534 535 536 537 538 540 541 542 543 544 545 546
#> [541] 547 548 549 551 553 554 555 556 557 558 559 560 561 562 563 564 565 566
#> [559] 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584
#> [577] 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601 602
#> [595] 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619 620
#> [613] 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637 638
#> [631] 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655 656
#> [649] 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674
#> [667] 675 676 677 678 679 680 681 683 684 685 686 687 689 690 691 692 693 694
#> [685] 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 712 713
#> [703] 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731
#> [721] 732 733 734 735 736 737 738 739 740 741 742 743 745 746 747 748 749 750
#> [739] 751
rrcv$test_set(1)
#> [1] 682 711
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#> [1] 682 711
#> 
#> [[1]]$test[[2]]
#>  [1]  11  96 121 139 152 157 267 274 300 329 332 375 399 408 417 473 553 571 604
#> [20] 637 686 704 749
#> 
#> [[1]]$test[[3]]
#>  [1]  16 107 212 264 273 312 316 379 420 432 445 459 482 515 590 660
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  43  44  45  46  47  48  49  50  51  52  53  54  56
#>  [55]  57  58  59  60  61  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129
#> [127] 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147
#> [145] 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165
#> [163] 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183
#> [181] 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201
#> [199] 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219
#> [217] 220 221 222 223 224 225 226 227 228 229 231 232 233 234 235 236 237 238
#> [235] 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256
#> [253] 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274
#> [271] 275 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290 291 292
#> [289] 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310
#> [307] 311 312 313 314 315 316 317 318 319 320 321 322 323 324 325 326 327 328
#> [325] 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344 345 346
#> [343] 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364
#> [361] 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379 380 381 382
#> [379] 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401
#> [397] 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419
#> [415] 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437
#> [433] 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455
#> [451] 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473
#> [469] 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488 489 490 491
#> [487] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509
#> [505] 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527
#> [523] 528 529 530 531 532 533 534 535 536 537 538 540 541 542 543 544 545 546
#> [541] 547 548 549 551 553 554 555 556 557 558 559 560 561 562 563 564 565 566
#> [559] 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584
#> [577] 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601 602
#> [595] 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619 620
#> [613] 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637 638
#> [631] 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655 656
#> [649] 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674
#> [667] 675 676 677 678 679 680 681 683 684 685 686 687 689 690 691 692 693 694
#> [685] 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 712 713
#> [703] 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731
#> [721] 732 733 734 735 736 737 738 739 740 741 742 743 745 746 747 748 749 750
#> [739] 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 140 141 142 143 144 145 146 147 148 149 150
#> [145] 151 153 154 155 156 158 159 160 161 162 163 164 165 166 167 168 169 170
#> [163] 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188
#> [181] 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206
#> [199] 207 208 209 210 211 213 214 215 216 217 218 219 220 221 222 223 224 225
#> [217] 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243
#> [235] 244 245 246 247 248 249 250 251 252 253 254 255 256 257 259 260 261 262
#> [253] 263 264 265 266 268 269 270 271 272 273 275 276 277 278 279 280 281 282
#> [271] 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299 301
#> [289] 302 303 304 305 306 307 308 309 310 311 313 314 315 316 317 318 320 321
#> [307] 322 323 324 325 326 327 328 330 331 333 334 335 336 337 338 339 340 341
#> [325] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [343] 360 361 362 363 364 365 366 367 368 370 371 372 373 374 376 377 378 379
#> [361] 380 381 382 383 384 385 386 388 389 390 391 392 393 394 395 396 397 398
#> [379] 400 401 402 403 404 405 406 407 409 410 411 412 413 414 415 416 418 419
#> [397] 420 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438
#> [415] 439 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457
#> [433] 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472 474 475 476
#> [451] 477 478 479 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494
#> [469] 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512
#> [487] 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530
#> [505] 531 532 533 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549
#> [523] 550 551 552 554 555 556 557 558 559 561 562 563 564 565 566 567 568 569
#> [541] 570 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586 587 588
#> [559] 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 605 606 607
#> [577] 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 625
#> [595] 626 627 628 629 630 631 632 633 634 636 638 639 640 641 642 643 644 645
#> [613] 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662 663
#> [631] 664 665 666 667 668 669 670 671 672 673 674 675 677 678 679 680 681 682
#> [649] 683 684 685 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701
#> [667] 702 703 705 706 707 708 709 710 711 712 713 714 715 717 718 719 720 721
#> [685] 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739
#> [703] 740 741 742 743 744 745 746 747 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  15  17  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  33  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  52  53  54  55  56  57  58  59
#>  [55]  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77
#>  [73]  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 108 109 110 111 112 113 114
#> [109] 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 134 135 136 137 138 139 140 141 142 143 144 145 147 150 151 152 153
#> [145] 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171
#> [163] 172 173 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 190
#> [181] 191 192 194 195 196 197 198 199 200 201 202 203 204 205 206 207 208 209
#> [199] 210 211 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227 228
#> [217] 229 230 231 233 234 235 236 237 238 239 240 241 242 243 244 245 246 247
#> [235] 248 249 250 251 253 254 255 256 257 259 260 261 262 263 265 266 267 268
#> [253] 269 270 271 272 274 275 276 277 278 279 280 281 282 283 284 285 287 288
#> [271] 289 290 291 292 293 294 296 297 298 299 301 303 304 305 306 307 308 309
#> [289] 310 311 313 315 317 318 319 320 321 322 323 324 325 326 327 328 329 331
#> [307] 332 333 334 335 336 337 338 339 340 341 342 343 344 345 346 348 349 350
#> [325] 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368
#> [343] 370 371 372 373 374 375 376 377 378 380 381 382 383 384 385 386 388 389
#> [361] 390 391 392 393 394 395 396 397 398 400 401 402 403 404 405 406 407 408
#> [379] 409 410 411 412 413 414 415 416 417 418 419 422 423 424 425 426 427 428
#> [397] 429 430 431 433 434 435 436 437 438 440 441 442 443 444 446 448 449 450
#> [415] 451 452 453 454 455 456 457 458 460 461 462 463 464 465 466 467 468 470
#> [433] 471 472 474 475 476 477 478 479 480 481 483 484 485 486 487 488 489 490
#> [451] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 508 509 510
#> [469] 511 512 513 514 516 517 518 519 520 521 522 523 524 525 526 527 528 529
#> [487] 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547
#> [505] 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565
#> [523] 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [541] 584 585 586 587 588 589 591 592 593 594 595 596 597 598 599 600 601 602
#> [559] 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619 620
#> [577] 621 622 623 624 625 627 628 629 630 631 632 633 634 635 636 637 638 639
#> [595] 640 641 642 643 644 646 648 649 650 651 652 653 654 655 656 657 658 659
#> [613] 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 679
#> [631] 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696 697
#> [649] 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715
#> [667] 716 717 718 719 720 722 723 724 725 726 727 728 729 730 732 733 734 735
#> [685] 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>  [1]  22  80  83 101 176 197 341 352 450 556 614 634 648
#> 
#> [[2]]$test[[2]]
#>  [1] 258 295 369 387 399 421 507 515 637 647
#> 
#> [[2]]$test[[3]]
#> [1]  94 147 337 418 554 638 723
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  81  82  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166
#> [163] 167 168 169 170 171 172 173 174 175 177 178 179 180 181 182 183 184 185
#> [181] 186 187 188 189 190 191 192 193 194 195 196 198 199 200 201 202 203 204
#> [199] 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222
#> [217] 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 240 241
#> [235] 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259
#> [253] 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277
#> [271] 278 279 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295
#> [289] 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310 312 313 314
#> [307] 315 316 317 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332
#> [325] 333 334 335 336 337 338 339 340 342 343 344 345 346 347 348 349 350 351
#> [343] 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370
#> [361] 371 372 373 374 375 377 378 379 380 381 382 383 384 385 386 387 388 389
#> [379] 390 391 392 393 394 395 398 399 400 401 402 403 404 405 406 407 408 409
#> [397] 410 411 412 413 414 415 416 417 418 419 420 421 422 423 425 426 427 428
#> [415] 429 430 431 432 433 435 436 437 438 439 440 441 442 443 444 445 446 447
#> [433] 448 449 451 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466
#> [451] 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484
#> [469] 485 486 487 488 489 490 491 492 493 494 495 496 497 498 499 500 501 503
#> [487] 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [505] 522 523 524 525 526 527 528 529 530 531 532 533 534 535 537 538 539 540
#> [523] 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 557 558 559
#> [541] 560 561 562 563 564 565 566 567 569 570 571 572 573 574 575 576 577 578
#> [559] 579 580 581 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596
#> [577] 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 613 616 617
#> [595] 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632 633 635 636
#> [613] 637 638 639 640 641 642 643 644 645 646 647 649 650 651 653 654 655 656
#> [631] 657 658 659 660 661 662 663 664 665 667 668 669 670 671 672 673 674 675
#> [649] 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693
#> [667] 694 695 696 697 699 700 701 702 703 704 705 706 707 708 709 710 711 712
#> [685] 713 714 715 716 717 718 719 720 721 722 723 726 727 728 729 730 731 732
#> [703] 733 734 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750
#> [721] 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  15  17  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  33  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58
#>  [55]  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76
#>  [73]  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94
#>  [91]  95  96  97  98  99 100 101 102 103 104 105 106 108 109 110 111 112 113
#> [109] 114 115 116 117 118 119 120 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 134 135 136 137 138 139 140 141 142 143 144 145 147 149 150 151 153
#> [145] 154 155 156 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172
#> [163] 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 190
#> [181] 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206 207 208
#> [199] 209 210 211 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227
#> [217] 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 244 245
#> [235] 246 247 248 249 250 251 253 254 255 256 257 259 260 261 262 263 265 266
#> [253] 268 269 270 271 272 274 275 276 277 278 279 280 281 282 283 284 285 286
#> [271] 287 288 289 290 291 292 293 294 296 297 298 299 301 302 303 304 305 306
#> [289] 308 309 310 311 313 314 315 317 318 319 320 321 322 323 324 325 326 327
#> [307] 328 331 333 334 335 336 337 338 339 340 341 342 343 344 345 346 347 348
#> [325] 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366
#> [343] 367 368 370 371 372 373 374 375 376 377 378 380 381 382 383 384 385 386
#> [361] 388 389 390 391 392 393 394 395 396 397 398 400 401 402 403 404 405 406
#> [379] 407 409 410 411 412 413 414 415 416 417 418 419 422 423 424 425 426 427
#> [397] 428 429 430 431 433 434 435 436 437 438 441 442 443 444 446 447 448 449
#> [415] 450 451 452 453 454 455 456 457 458 460 461 462 463 464 465 466 467 468
#> [433] 469 470 471 472 474 475 476 477 478 479 480 481 482 483 484 485 486 487
#> [451] 488 489 490 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505
#> [469] 506 508 509 510 511 512 513 514 516 517 518 519 520 521 522 523 524 525
#> [487] 526 527 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543
#> [505] 544 545 546 547 548 549 550 551 552 553 554 555 556 557 558 559 561 562
#> [523] 563 564 565 566 567 568 569 570 572 573 574 575 576 578 579 580 581 582
#> [541] 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600
#> [559] 601 602 603 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [577] 620 621 622 623 624 625 627 628 630 631 633 634 635 636 638 639 640 641
#> [595] 642 643 644 646 648 649 650 651 652 653 654 655 656 657 658 659 661 662
#> [613] 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 679 680 681
#> [631] 682 683 684 685 687 688 689 690 691 692 693 694 695 696 697 698 699 700
#> [649] 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718
#> [667] 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736
#> [685] 737 738 739 740 741 742 743 744 745 746 747 748 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   6   8   9  10  11  13  15  16  18  19  20  22  23
#>  [19]  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40  41
#>  [37]  42  43  44  45  46  47  48  49  50  51  52  53  55  56  57  58  59  60
#>  [55]  61  62  63  64  67  68  69  70  71  72  73  74  75  76  77  79  80  81
#>  [73]  82  83  84  85  86  87  88  89  90  91  92  93  95  96  97  98  99 100
#>  [91] 101 102 103 104 106 107 108 110 111 112 114 115 116 117 118 119 120 121
#> [109] 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 138 139 140
#> [127] 141 143 144 145 146 148 149 150 151 152 153 154 155 156 157 158 159 160
#> [145] 161 162 164 165 166 167 168 169 170 172 173 174 175 176 177 178 179 180
#> [163] 183 184 185 186 187 188 189 190 192 193 194 195 197 198 199 200 201 203
#> [181] 204 205 206 207 208 209 210 211 212 214 215 216 217 218 219 220 221 222
#> [199] 224 225 226 227 228 230 231 232 233 234 235 236 237 238 239 241 243 244
#> [217] 245 246 247 248 249 250 251 252 254 255 256 257 258 259 260 261 262 263
#> [235] 264 265 266 267 268 269 271 272 273 274 275 276 277 278 279 280 281 282
#> [253] 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299 300
#> [271] 301 302 303 304 305 306 307 309 310 311 312 313 314 315 316 317 318 319
#> [289] 320 321 322 323 324 325 326 328 329 330 332 333 334 335 336 339 340 341
#> [307] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [325] 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377
#> [343] 379 380 381 382 383 385 386 387 388 389 390 391 392 393 394 395 396 397
#> [361] 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415
#> [379] 416 417 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434
#> [397] 435 436 437 438 439 440 441 442 443 444 445 447 448 449 450 451 452 453
#> [415] 454 455 456 457 458 459 460 461 462 463 464 466 467 468 469 470 471 472
#> [433] 473 474 475 476 477 479 480 481 482 483 484 485 486 487 488 489 490 491
#> [451] 492 494 495 496 497 499 500 501 502 503 504 505 506 507 508 509 510 511
#> [469] 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 528 529 530
#> [487] 531 532 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549
#> [505] 550 551 552 553 555 556 557 558 559 560 561 562 563 564 565 566 568 569
#> [523] 571 572 573 574 575 576 577 578 579 581 582 583 584 585 586 587 588 589
#> [541] 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607
#> [559] 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 625
#> [577] 626 627 628 629 630 631 632 633 634 635 636 637 639 640 641 642 643 644
#> [595] 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662
#> [613] 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679 680
#> [631] 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696 697 698
#> [649] 699 700 702 703 704 706 707 708 709 710 711 712 713 715 716 717 718 719
#> [667] 720 721 722 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738
#> [685] 739 740 741 742 743 744 745 746 747 748 749 751
#> 
#> 
#>