(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  53  54  55  56
#>  [55]  57  58  59  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  77  78  79  80  81  82  83  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 112 113 114
#> [109] 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151
#> [145] 152 153 154 155 156 157 158 159 160 161 162 163 164 166 167 168 169 170
#> [163] 171 172 173 174 175 176 178 179 180 181 182 183 185 186 187 188 189 190
#> [181] 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206 207 208
#> [199] 209 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227
#> [217] 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 244 245
#> [235] 246 247 248 249 250 251 252 253 254 255 258 259 260 261 262 263 264 265
#> [253] 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280 281 282 283
#> [271] 284 285 286 287 288 289 290 291 292 293 294 295 296 298 299 300 301 302
#> [289] 303 304 305 306 307 308 309 310 312 313 314 315 316 317 318 319 320 321
#> [307] 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339
#> [325] 340 341 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357
#> [343] 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375
#> [361] 376 377 378 379 380 381 383 384 385 386 387 388 389 390 392 393 394 395
#> [379] 397 398 399 400 401 402 403 404 405 406 408 409 410 411 412 413 414 415
#> [397] 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433
#> [415] 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452
#> [433] 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470
#> [451] 472 473 474 475 476 477 478 479 480 482 483 484 485 486 487 488 489 490
#> [469] 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508
#> [487] 509 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526
#> [505] 527 528 529 530 531 532 533 534 535 537 538 539 540 541 542 543 544 545
#> [523] 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563
#> [541] 564 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581
#> [559] 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599
#> [577] 600 601 602 603 604 605 607 608 609 610 611 612 613 614 615 616 617 618
#> [595] 619 620 621 622 623 625 626 627 628 629 630 631 632 633 634 635 636 637
#> [613] 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655
#> [631] 656 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673
#> [649] 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691
#> [667] 692 693 694 696 697 698 699 701 702 703 704 705 706 707 708 709 710 711
#> [685] 712 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729
#> [703] 730 731 732 733 734 735 736 737 738 739 740 741 743 744 746 747 748 749
#> [721] 750 751
rrcv$test_set(1)
#>  [1]  36 111 165 184 256 382 391 407 481 695 700 745
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>  [1]  36 111 165 184 256 382 391 407 481 695 700 745
#> 
#> [[1]]$test[[2]]
#>  [1]  32 200 215 224 246 266 279 306 313 348 388 430 468 559 607 617 726
#> 
#> [[1]]$test[[3]]
#> [1]  59  87 241 269 440 635
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  53  54  55  56
#>  [55]  57  58  59  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  77  78  79  80  81  82  83  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 112 113 114
#> [109] 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151
#> [145] 152 153 154 155 156 157 158 159 160 161 162 163 164 166 167 168 169 170
#> [163] 171 172 173 174 175 176 178 179 180 181 182 183 185 186 187 188 189 190
#> [181] 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206 207 208
#> [199] 209 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227
#> [217] 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 244 245
#> [235] 246 247 248 249 250 251 252 253 254 255 258 259 260 261 262 263 264 265
#> [253] 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280 281 282 283
#> [271] 284 285 286 287 288 289 290 291 292 293 294 295 296 298 299 300 301 302
#> [289] 303 304 305 306 307 308 309 310 312 313 314 315 316 317 318 319 320 321
#> [307] 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339
#> [325] 340 341 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357
#> [343] 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375
#> [361] 376 377 378 379 380 381 383 384 385 386 387 388 389 390 392 393 394 395
#> [379] 397 398 399 400 401 402 403 404 405 406 408 409 410 411 412 413 414 415
#> [397] 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433
#> [415] 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452
#> [433] 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470
#> [451] 472 473 474 475 476 477 478 479 480 482 483 484 485 486 487 488 489 490
#> [469] 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508
#> [487] 509 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526
#> [505] 527 528 529 530 531 532 533 534 535 537 538 539 540 541 542 543 544 545
#> [523] 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563
#> [541] 564 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581
#> [559] 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599
#> [577] 600 601 602 603 604 605 607 608 609 610 611 612 613 614 615 616 617 618
#> [595] 619 620 621 622 623 625 626 627 628 629 630 631 632 633 634 635 636 637
#> [613] 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655
#> [631] 656 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673
#> [649] 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691
#> [667] 692 693 694 696 697 698 699 701 702 703 704 705 706 707 708 709 710 711
#> [685] 712 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729
#> [703] 730 731 732 733 734 735 736 737 738 739 740 741 743 744 746 747 748 749
#> [721] 750 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  26  27  28  29  30  31  33  34  35  36  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74
#>  [73]  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92
#>  [91]  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110
#> [109] 111 112 113 114 115 116 117 119 120 121 122 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 159 160 163 164 165 166 167 168 169
#> [163] 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187
#> [181] 188 189 190 191 192 193 194 195 196 197 198 199 201 202 203 204 205 206
#> [199] 207 208 209 210 211 212 213 214 216 217 218 219 220 222 223 225 226 227
#> [217] 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 244 245
#> [235] 247 248 249 250 252 253 254 255 256 257 258 259 260 261 262 263 264 265
#> [253] 267 268 269 270 271 272 273 274 275 276 277 278 280 281 282 283 284 285
#> [271] 286 287 289 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304
#> [289] 305 307 308 309 310 311 312 314 315 316 317 318 319 320 321 322 323 324
#> [307] 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342
#> [325] 343 344 345 346 347 349 350 351 352 353 354 355 356 357 358 359 360 361
#> [343] 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 378 379 380
#> [361] 381 382 383 384 385 386 387 389 390 391 392 393 394 395 396 397 398 399
#> [379] 400 401 402 403 404 405 407 408 410 412 413 414 415 416 417 418 419 420
#> [397] 421 422 423 424 425 426 427 428 429 431 432 434 435 436 437 438 439 440
#> [415] 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 458
#> [433] 459 460 461 462 463 464 465 466 467 469 470 471 473 474 475 476 477 478
#> [451] 479 481 482 483 484 485 486 487 488 489 490 491 492 493 494 495 496 497
#> [469] 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515
#> [487] 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532 533
#> [505] 534 535 536 537 538 539 540 542 544 545 546 547 548 549 550 551 552 553
#> [523] 554 556 557 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574
#> [541] 575 576 577 578 579 580 581 582 583 584 585 587 588 589 590 591 592 593
#> [559] 594 595 596 597 599 600 601 602 603 604 605 606 608 609 610 612 613 614
#> [577] 615 616 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632 633
#> [595] 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648 649 650 651
#> [613] 652 653 654 655 656 657 658 659 660 661 662 663 664 665 666 667 668 669
#> [631] 670 671 672 673 674 675 676 677 678 679 681 682 683 684 685 686 687 688
#> [649] 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706
#> [667] 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721 722 723 724
#> [685] 725 727 728 729 730 731 732 733 734 735 736 737 738 739 740 741 742 743
#> [703] 744 745 746 747 748 749 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  48  49  50  51  53  54  55  56
#>  [55]  57  58  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  77  78  79  80  81  82  83  84  85  86  88  89  90  91  92  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113
#> [109] 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 140 141 142 143 145 146 147 148 149 150 151
#> [145] 152 153 154 155 158 159 160 161 162 163 164 165 166 167 168 169 170 171
#> [163] 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189
#> [181] 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206 207
#> [199] 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225
#> [217] 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 242 243 244
#> [235] 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 260 261 262
#> [253] 263 264 265 266 267 268 270 271 272 273 274 275 276 277 278 279 280 281
#> [271] 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299
#> [289] 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317
#> [307] 318 320 322 323 324 325 326 327 328 329 330 331 333 334 335 336 337 338
#> [325] 339 340 341 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356
#> [343] 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374
#> [361] 375 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393
#> [379] 394 395 396 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411
#> [397] 412 413 414 415 416 417 418 420 421 422 423 425 426 427 428 429 430 431
#> [415] 432 433 434 435 436 437 438 439 441 442 443 444 445 446 447 448 449 450
#> [433] 451 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468
#> [451] 469 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486
#> [469] 487 488 489 490 491 492 493 494 495 496 497 498 499 500 502 503 504 505
#> [487] 506 507 508 509 510 511 512 513 515 516 517 518 519 520 521 522 523 524
#> [505] 525 526 527 528 529 530 531 532 533 534 535 536 537 538 539 541 542 543
#> [523] 544 545 546 547 548 549 550 551 552 554 555 556 557 558 559 561 562 563
#> [541] 564 565 566 567 568 569 570 572 573 574 575 576 577 578 579 580 581 582
#> [559] 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600
#> [577] 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618
#> [595] 619 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 636 638
#> [613] 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655 656
#> [631] 657 658 659 660 661 662 663 664 665 666 667 668 670 671 672 673 674 675
#> [649] 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693
#> [667] 694 695 696 697 698 699 700 701 702 703 705 706 707 708 709 710 711 712
#> [685] 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730
#> [703] 731 732 733 734 735 736 737 738 739 740 741 742 743 744 745 746 747 748
#> [721] 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>  [1]   1   6  33  84 100 112 164 167 168 175 210 248 257 294 307 310 314 324 330
#> [20] 342 351 374 425 485 491 572 577 632 699
#> 
#> [[2]]$test[[2]]
#>  [1]   8  15  24 166 207 261 280 287 354 428 435 451 461 499 562
#> 
#> [[2]]$test[[3]]
#>  [1]  58 170 245 290 303 356 362 413 544 594 706
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  35  36  37  38  39  40  41
#>  [37]  42  43  44  45  46  47  48  49  50  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78
#>  [73]  79  80  81  82  83  85  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 101 102 103 104 105 106 107 108 109 110 111 113 114 115 116 118
#> [109] 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136
#> [127] 137 138 139 140 141 142 143 144 145 147 149 150 151 152 153 154 155 156
#> [145] 157 158 159 160 161 162 163 166 169 170 171 172 173 174 176 177 179 180
#> [163] 181 182 183 184 185 186 187 188 189 190 191 192 194 195 196 197 198 199
#> [181] 200 201 202 203 204 205 206 207 208 209 211 212 213 214 215 216 217 218
#> [199] 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234 235 236
#> [217] 237 238 239 240 241 242 243 244 245 246 247 249 250 251 253 254 255 256
#> [235] 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275
#> [253] 276 277 278 279 280 281 282 283 284 285 287 288 289 290 291 292 293 296
#> [271] 297 298 299 300 301 302 303 304 305 306 308 309 311 312 313 315 316 317
#> [289] 318 319 320 321 322 323 325 326 327 328 329 331 332 333 334 335 336 337
#> [307] 338 339 340 341 343 344 345 346 348 349 350 352 353 354 355 356 357 358
#> [325] 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373 375 376 377
#> [343] 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 394 395 396
#> [361] 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414
#> [379] 415 416 417 418 419 420 422 423 424 426 427 428 429 430 431 432 433 434
#> [397] 435 436 437 438 440 441 442 443 444 445 446 447 448 449 450 451 452 453
#> [415] 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 470 471 472
#> [433] 473 474 475 476 477 478 479 480 481 482 483 484 486 487 488 489 490 492
#> [451] 493 494 495 496 497 498 499 500 501 502 503 504 505 506 508 509 510 511
#> [469] 512 513 514 515 516 518 519 520 521 522 523 524 525 526 527 528 529 530
#> [487] 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548
#> [505] 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566
#> [523] 567 568 569 570 571 573 574 575 576 578 579 580 581 582 583 584 585 586
#> [541] 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604
#> [559] 605 606 607 608 610 611 612 613 614 615 616 617 618 619 620 621 622 623
#> [577] 625 627 628 630 631 633 634 635 636 637 638 639 640 641 642 643 644 646
#> [595] 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662 663 664 665
#> [613] 666 667 668 669 670 671 672 673 674 675 676 677 679 680 681 682 683 684
#> [631] 685 686 687 688 689 690 691 692 693 694 695 696 697 698 700 701 702 703
#> [649] 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721
#> [667] 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739
#> [685] 740 741 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   5   6   7   9  10  11  12  13  14  16  17  18  19  20  21
#>  [19]  22  23  25  27  28  29  30  31  32  33  34  35  36  37  38  39  40  41
#>  [37]  42  43  44  45  46  47  49  50  51  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  65  66  67  68  69  71  72  73  74  75  76  77  78  79  80
#>  [73]  81  82  83  84  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 100 101 103 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119
#> [109] 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138
#> [127] 139 140 141 142 143 144 145 146 147 148 149 152 154 155 156 157 159 160
#> [145] 161 162 163 164 165 167 168 169 170 171 173 174 175 176 177 178 179 180
#> [163] 181 182 183 184 185 186 187 188 189 190 191 193 194 195 196 197 198 199
#> [181] 200 201 202 203 204 205 208 209 210 211 212 213 215 216 217 218 219 220
#> [199] 221 222 223 224 225 226 228 229 230 231 232 233 234 235 236 237 239 240
#> [217] 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258
#> [235] 259 260 262 263 264 266 267 268 269 270 273 274 275 276 277 278 279 281
#> [253] 282 283 284 285 286 288 289 290 291 292 293 294 295 296 297 298 299 300
#> [271] 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318
#> [289] 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336
#> [307] 337 338 339 340 341 342 343 344 345 346 347 348 349 350 351 352 353 355
#> [325] 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373
#> [343] 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391
#> [361] 392 393 394 395 396 397 398 399 400 401 403 404 405 406 407 408 409 410
#> [379] 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426 427 430
#> [397] 431 432 433 434 436 437 438 439 440 441 442 443 444 445 446 447 448 449
#> [415] 450 452 453 454 455 456 457 458 459 460 462 463 464 465 466 467 468 469
#> [433] 471 472 473 474 475 477 478 479 480 481 482 483 484 485 486 487 489 490
#> [451] 491 492 493 494 495 496 497 498 500 501 502 503 504 505 506 507 508 509
#> [469] 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527
#> [487] 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545
#> [505] 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 563 564 565
#> [523] 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [541] 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601
#> [559] 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [577] 620 621 622 624 625 626 627 628 629 630 631 632 634 635 636 637 638 639
#> [595] 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657
#> [613] 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675
#> [631] 676 677 678 679 680 681 682 684 685 686 687 688 689 690 691 692 693 694
#> [649] 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712
#> [667] 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 731
#> [685] 732 733 734 735 736 737 738 739 740 741 742 743 744 745 747 748 749 750
#> [703] 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91
#>  [91]  92  93  94  95  96  97  98  99 100 101 102 104 105 106 107 108 109 110
#> [109] 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 127 128 129
#> [127] 130 131 132 133 134 135 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166
#> [163] 167 168 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186
#> [181] 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204
#> [199] 205 206 207 208 209 210 211 212 213 214 215 217 219 220 221 223 224 225
#> [217] 226 227 228 229 230 231 232 233 235 236 237 238 239 240 241 242 243 244
#> [235] 246 247 248 249 250 251 252 253 254 256 257 258 259 260 261 262 263 264
#> [253] 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280 281 282
#> [271] 283 284 285 286 287 288 289 291 292 293 294 295 296 297 298 299 300 301
#> [289] 302 304 305 306 307 308 310 311 312 313 314 315 316 317 318 319 320 321
#> [307] 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339
#> [325] 340 341 342 343 345 346 347 348 349 350 351 352 353 354 355 358 359 360
#> [343] 361 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379
#> [361] 380 381 382 383 384 385 386 387 388 389 390 391 392 393 395 396 397 398
#> [379] 399 400 401 402 403 404 405 406 407 408 409 410 411 412 414 415 416 417
#> [397] 418 419 420 421 422 423 424 425 427 428 429 430 431 432 433 434 435 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454
#> [433] 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472
#> [451] 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488 489 490
#> [469] 491 492 493 494 495 496 497 498 499 500 501 502 504 505 506 507 508 510
#> [487] 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528
#> [505] 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 545 546 547
#> [523] 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565
#> [541] 566 567 568 569 570 571 572 574 575 576 577 578 579 580 581 582 583 584
#> [559] 586 587 588 589 590 591 592 593 595 596 597 598 599 600 601 602 603 604
#> [577] 605 606 607 608 609 610 611 612 614 615 616 617 618 619 620 621 622 623
#> [595] 624 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 642
#> [613] 643 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661
#> [631] 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679
#> [649] 680 681 682 683 684 685 686 687 688 689 691 692 693 694 695 696 697 698
#> [667] 699 700 701 702 703 704 705 709 710 711 712 713 714 715 716 717 718 719
#> [685] 720 721 722 723 724 725 726 728 729 730 731 732 733 734 735 736 737 738
#> [703] 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#>