(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 133 134 135 136 137 138 139 140 141 142 143 144 145
#> [145] 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163
#> [163] 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181
#> [181] 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198 199
#> [199] 200 201 202 203 204 205 206 207 208 209 210 212 213 214 215 216 217 218
#> [217] 219 220 221 222 223 224 226 227 228 229 230 231 232 233 234 235 236 237
#> [235] 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255
#> [253] 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273
#> [271] 274 275 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290 291
#> [289] 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309
#> [307] 310 311 312 313 314 315 316 317 318 319 320 321 323 324 325 326 327 328
#> [325] 329 330 331 332 333 334 335 337 338 339 340 341 342 343 344 345 346 347
#> [343] 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365
#> [361] 367 369 370 371 372 373 374 375 376 377 378 379 380 381 382 383 384 385
#> [379] 386 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403
#> [397] 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421
#> [415] 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439
#> [433] 440 441 442 443 444 445 446 447 448 449 450 451 453 454 455 456 457 458
#> [451] 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 476 477
#> [469] 478 479 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494 495
#> [487] 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513
#> [505] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531
#> [523] 532 533 534 535 536 537 538 539 540 541 543 544 545 546 547 548 549 550
#> [541] 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [559] 569 570 571 572 573 574 575 576 577 578 579 580 582 583 584 585 586 587
#> [577] 588 589 590 591 592 593 594 595 596 597 598 600 601 602 603 604 605 606
#> [595] 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624
#> [613] 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642
#> [631] 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660
#> [649] 661 662 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679
#> [667] 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696 697
#> [685] 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715
#> [703] 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733
#> [721] 734 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
rrcv$test_set(1)
#> [1] 366 368 452 581
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#> [1] 366 368 452 581
#> 
#> [[1]]$test[[2]]
#>  [1]  49  90 122 183 345 350 403 449 463 474 483 715 743
#> 
#> [[1]]$test[[3]]
#> [1] 488 563 651 681 683
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 133 134 135 136 137 138 139 140 141 142 143 144 145
#> [145] 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163
#> [163] 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181
#> [181] 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198 199
#> [199] 200 201 202 203 204 205 206 207 208 209 210 212 213 214 215 216 217 218
#> [217] 219 220 221 222 223 224 226 227 228 229 230 231 232 233 234 235 236 237
#> [235] 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255
#> [253] 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273
#> [271] 274 275 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290 291
#> [289] 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309
#> [307] 310 311 312 313 314 315 316 317 318 319 320 321 323 324 325 326 327 328
#> [325] 329 330 331 332 333 334 335 337 338 339 340 341 342 343 344 345 346 347
#> [343] 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365
#> [361] 367 369 370 371 372 373 374 375 376 377 378 379 380 381 382 383 384 385
#> [379] 386 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403
#> [397] 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421
#> [415] 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439
#> [433] 440 441 442 443 444 445 446 447 448 449 450 451 453 454 455 456 457 458
#> [451] 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 476 477
#> [469] 478 479 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494 495
#> [487] 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513
#> [505] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531
#> [523] 532 533 534 535 536 537 538 539 540 541 543 544 545 546 547 548 549 550
#> [541] 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [559] 569 570 571 572 573 574 575 576 577 578 579 580 582 583 584 585 586 587
#> [577] 588 589 590 591 592 593 594 595 596 597 598 600 601 602 603 604 605 606
#> [595] 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624
#> [613] 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642
#> [631] 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660
#> [649] 661 662 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679
#> [667] 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696 697
#> [685] 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715
#> [703] 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733
#> [721] 734 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  91  92  93  94
#>  [91]  95  96  97  98  99 100 101 102 104 105 106 107 108 109 111 112 113 114
#> [109] 115 116 117 118 119 120 121 123 124 125 126 127 128 129 130 131 132 133
#> [127] 134 135 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152
#> [145] 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170
#> [163] 171 172 173 174 175 176 177 178 179 180 181 182 184 185 186 187 188 189
#> [181] 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206 207
#> [199] 208 209 210 211 212 213 214 215 216 217 218 219 220 221 223 224 225 226
#> [217] 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 244
#> [235] 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 260 261 262
#> [253] 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280
#> [271] 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298
#> [289] 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316
#> [307] 317 318 319 320 321 322 323 324 326 327 328 329 330 331 332 333 334 335
#> [325] 336 337 338 339 340 341 342 343 344 346 347 348 349 351 352 353 354 355
#> [343] 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373
#> [361] 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391
#> [379] 392 393 394 395 396 397 398 399 401 402 405 406 407 408 409 410 411 412
#> [397] 414 415 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431
#> [415] 432 433 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 450
#> [433] 451 452 453 454 455 456 457 458 459 460 461 462 464 465 466 467 468 469
#> [451] 470 471 472 473 475 476 477 478 479 480 481 482 484 485 487 488 489 491
#> [469] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509
#> [487] 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527
#> [505] 528 529 530 531 532 533 534 535 536 538 539 540 541 542 543 544 545 547
#> [523] 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 565 566
#> [541] 567 568 569 570 571 572 574 576 577 578 579 580 581 582 583 584 585 586
#> [559] 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604
#> [577] 605 606 607 608 609 610 611 612 614 615 616 617 618 619 621 622 623 624
#> [595] 625 626 627 628 629 630 631 632 633 634 635 636 637 638 640 641 642 644
#> [613] 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662
#> [631] 663 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679 680 681
#> [649] 682 683 684 685 686 687 688 689 690 691 692 693 694 695 697 698 699 700
#> [667] 701 702 703 704 705 706 707 708 709 710 711 712 713 714 716 717 718 719
#> [685] 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737
#> [703] 738 739 740 741 742 744 745 746 747 748 749 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   9  10  11  12  13  14  16  17  18  19  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  59  60  61  62  63  64  65  66  67  68  69  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  92  93  94
#>  [91]  95  96  98  99 100 101 103 104 105 106 107 108 109 110 111 112 113 114
#> [109] 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 134 135 136 137 138 139 141 142 143 144 145 146 147 148 149 150 151
#> [145] 152 153 154 155 156 157 158 159 160 161 162 163 164 165 167 168 169 170
#> [163] 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188
#> [181] 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206
#> [199] 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224
#> [217] 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242
#> [235] 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 260
#> [253] 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279
#> [271] 280 281 282 283 284 285 286 288 289 290 291 292 293 294 295 296 297 298
#> [289] 299 300 301 302 303 304 306 307 308 309 310 311 312 313 314 315 316 317
#> [307] 318 319 320 321 322 323 324 325 327 328 329 330 331 332 333 334 335 336
#> [325] 337 338 339 340 341 342 343 344 345 346 347 348 349 350 351 352 353 355
#> [343] 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373
#> [361] 374 375 376 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391
#> [379] 392 393 394 395 396 397 398 399 400 401 403 404 405 406 407 408 409 410
#> [397] 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426 427 428
#> [415] 429 430 431 432 433 434 436 437 438 439 440 441 442 443 444 445 446 447
#> [433] 448 449 450 452 453 454 455 456 457 458 459 460 462 463 464 465 466 467
#> [451] 468 469 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485
#> [469] 486 487 489 490 491 492 493 494 495 496 497 498 500 501 502 503 504 505
#> [487] 506 507 508 509 510 511 513 514 515 516 517 518 519 520 521 522 523 524
#> [505] 525 526 527 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542
#> [523] 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560
#> [541] 562 564 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580
#> [559] 581 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599
#> [577] 600 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617
#> [595] 618 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636
#> [613] 637 638 639 640 641 642 643 644 645 646 647 648 649 650 652 653 654 655
#> [631] 656 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673
#> [649] 674 675 676 677 678 679 680 682 684 685 686 687 688 689 690 691 692 693
#> [667] 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711
#> [685] 712 713 714 715 716 717 718 719 721 722 723 724 725 726 727 728 729 730
#> [703] 731 732 733 734 735 736 737 738 739 740 741 742 743 744 745 746 748 749
#> [721] 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#> [1] 249 516 534 588 650 674 748
#> 
#> [[2]]$test[[2]]
#>  [1]  14  21  54  65  66 105 113 128 130 132 142 145 163 171 181 182 191 202 209
#> [20] 211 213 225 229 253 262 270 308 322 327 336 458 465 475 478 493 498 527 701
#> [39] 714
#> 
#> [[2]]$test[[3]]
#> [1] 117 178 393 545 624 742
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144
#> [145] 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162
#> [163] 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180
#> [181] 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198
#> [199] 199 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216
#> [217] 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234
#> [235] 235 236 237 238 239 240 241 242 243 244 245 246 247 248 250 251 252 253
#> [253] 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271
#> [271] 272 273 274 275 276 277 278 279 280 281 282 283 284 285 286 287 288 289
#> [289] 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307
#> [307] 308 309 310 311 312 313 314 315 316 317 318 319 320 321 322 323 324 325
#> [325] 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343
#> [343] 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361
#> [361] 362 363 364 365 366 367 368 369 370 371 372 373 374 376 377 378 379 380
#> [379] 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398
#> [397] 399 400 401 402 403 404 405 406 407 409 410 411 412 413 414 415 416 418
#> [415] 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436
#> [433] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454
#> [451] 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472
#> [469] 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488 489 490
#> [487] 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508
#> [505] 509 510 511 512 513 514 515 517 518 519 520 521 522 523 524 525 526 527
#> [523] 528 529 530 531 532 533 535 536 537 538 539 540 541 542 543 544 545 546
#> [541] 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564
#> [559] 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582
#> [577] 583 584 585 586 587 589 590 591 592 593 594 595 596 597 598 599 600 601
#> [595] 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [613] 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637
#> [631] 638 639 640 641 642 643 644 645 646 647 648 649 651 652 653 654 655 656
#> [649] 657 658 659 660 661 662 663 664 666 667 668 669 670 671 672 673 675 677
#> [667] 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695
#> [685] 696 697 698 699 700 701 703 704 705 706 707 708 709 710 711 712 713 714
#> [703] 715 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733
#> [721] 734 735 736 737 738 739 740 741 742 743 744 745 746 747 749 750 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  13  15  16  17  18  19  20
#>  [19]  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  41  42  43  44  45  46  47  48  49  50  51  52  53  55  56  57  58
#>  [55]  59  60  61  62  63  64  67  68  69  70  71  72  73  74  75  76  77  80
#>  [73]  81  82  83  84  85  86  87  89  90  91  92  93  95  96  97  98  99 100
#>  [91] 101 102 103 104 106 107 108 110 111 112 114 115 116 117 118 119 120 121
#> [109] 122 123 124 125 126 127 129 131 133 134 135 136 138 139 140 141 143 144
#> [127] 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 164
#> [145] 165 166 167 168 169 170 172 173 174 175 176 177 178 179 180 183 184 185
#> [163] 186 187 188 189 190 192 193 194 195 196 197 198 199 200 201 203 204 205
#> [181] 206 207 208 210 212 214 215 216 217 218 219 220 221 222 223 224 226 227
#> [199] 228 230 231 232 233 234 235 236 237 238 239 240 241 242 243 244 245 246
#> [217] 247 248 249 250 251 252 254 255 256 257 258 259 260 261 263 264 265 266
#> [235] 267 268 269 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285
#> [253] 286 287 288 289 290 291 292 293 294 295 296 297 298 299 300 301 302 303
#> [271] 304 305 306 307 309 310 311 312 313 314 315 316 317 318 319 321 323 324
#> [289] 325 326 328 329 330 332 333 334 335 337 339 340 341 342 343 344 345 346
#> [307] 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364
#> [325] 365 367 368 369 370 371 372 373 374 375 376 377 379 380 381 382 383 384
#> [343] 385 386 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401 402
#> [361] 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420
#> [379] 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438
#> [397] 439 440 441 442 443 444 445 446 447 449 450 451 452 453 454 455 456 457
#> [415] 459 460 461 462 463 464 466 467 468 469 470 471 472 473 474 476 477 479
#> [433] 480 481 482 483 484 485 486 487 488 489 490 491 492 494 495 496 497 499
#> [451] 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515 516 517
#> [469] 518 519 520 521 522 523 524 525 526 528 529 530 531 532 533 534 535 536
#> [487] 537 538 539 540 541 542 543 544 545 546 547 548 549 550 551 552 553 555
#> [505] 556 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573
#> [523] 574 575 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591
#> [541] 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607 608 609
#> [559] 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 625 626 627
#> [577] 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642 643 644 645
#> [595] 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662 663
#> [613] 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679 680 681
#> [631] 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696 697 698 699
#> [649] 700 702 704 705 706 707 708 709 710 711 712 713 715 716 717 718 719 720
#> [667] 721 722 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739
#> [685] 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  34  35  37  38  39
#>  [37]  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  81  82  83  85  86  87  88  89  90  91  92  93  94
#>  [91]  95  96  97  98  99 101 102 103 104 105 106 107 108 109 110 113 114 115
#> [109] 116 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134
#> [127] 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152
#> [145] 153 154 155 156 157 158 159 160 161 162 163 166 169 170 171 172 173 174
#> [163] 176 177 179 180 181 182 183 184 185 186 187 188 189 190 191 192 193 194
#> [181] 195 196 197 198 199 200 201 202 203 204 205 206 207 208 209 211 212 213
#> [199] 214 215 216 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231
#> [217] 232 233 234 235 236 237 238 239 240 241 242 243 244 245 246 247 249 250
#> [235] 251 252 253 254 255 256 258 259 260 261 262 263 264 265 266 267 268 269
#> [253] 270 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285 286 287
#> [271] 288 289 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305
#> [289] 306 307 308 309 310 311 312 313 314 315 316 317 318 319 320 321 322 323
#> [307] 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342
#> [325] 343 344 345 346 347 348 349 350 352 353 354 355 356 357 358 359 360 361
#> [343] 362 365 366 367 368 369 370 371 372 373 375 376 377 378 379 380 381 383
#> [361] 384 385 386 387 388 389 390 392 394 395 396 397 398 399 400 401 402 403
#> [379] 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421
#> [397] 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439
#> [415] 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457
#> [433] 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475
#> [451] 476 477 478 479 480 482 483 484 485 486 487 488 489 490 491 492 493 494
#> [469] 495 496 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513
#> [487] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531
#> [505] 532 533 534 535 536 537 538 539 540 541 542 543 544 546 547 548 549 550
#> [523] 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [541] 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586
#> [559] 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604
#> [577] 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622
#> [595] 623 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 641 642
#> [613] 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660
#> [631] 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678
#> [649] 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 696 697
#> [667] 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715
#> [685] 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732 733
#> [703] 734 735 736 737 738 739 740 741 743 744 745 746 747 748 749 750 751
#> 
#> 
#>