(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   3   4   5   6   7   8  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  31  32  33  34  35  36  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  75
#>  [73]  76  77  78  79  80  82  83  84  85  86  87  88  89  90  91  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113
#> [109] 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 161 162 163 164 165 166 167 168 169
#> [163] 170 171 172 174 175 176 177 178 180 181 182 183 184 185 186 187 188 190
#> [181] 191 192 193 194 195 196 197 198 199 200 201 202 203 204 206 207 208 209
#> [199] 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227
#> [217] 229 230 231 232 233 234 235 236 238 239 240 241 242 243 244 245 246 247
#> [235] 248 249 250 251 252 253 255 256 257 258 259 260 261 262 263 264 265 266
#> [253] 267 268 269 270 271 272 273 274 275 277 278 279 280 281 282 283 284 285
#> [271] 286 287 288 289 290 294 295 296 297 298 299 300 301 302 303 305 306 307
#> [289] 308 309 310 311 312 313 314 316 317 318 319 320 321 322 323 324 325 326
#> [307] 327 328 329 330 331 332 334 336 337 338 341 342 343 344 345 346 347 348
#> [325] 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366
#> [343] 367 368 369 370 371 372 373 374 375 376 377 378 379 381 382 383 384 385
#> [361] 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403 404
#> [379] 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421 422
#> [397] 423 424 425 426 427 428 429 430 431 432 433 434 435 438 439 440 441 442
#> [415] 444 445 446 447 448 449 450 451 452 453 455 456 458 459 460 461 462 463
#> [433] 464 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481
#> [451] 482 483 484 485 486 487 488 489 490 491 492 493 494 495 496 497 498 499
#> [469] 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515 516 517
#> [487] 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532 533 534 535
#> [505] 536 537 538 539 540 541 542 543 544 545 546 547 548 549 550 552 553 554
#> [523] 555 556 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572
#> [541] 573 574 575 576 577 578 579 580 581 582 583 584 585 586 588 589 590 591
#> [559] 592 594 595 596 597 598 599 600 601 602 603 604 605 607 608 609 610 611
#> [577] 612 613 614 615 617 618 619 620 621 623 624 625 626 627 628 629 630 631
#> [595] 632 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648 649
#> [613] 650 651 652 653 654 655 656 657 658 659 660 661 662 663 664 665 666 667
#> [631] 668 669 670 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685
#> [649] 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703
#> [667] 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721
#> [685] 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740
#> [703] 741 742 743 744 745 746 747 748 749 750 751
rrcv$test_set(1)
#>  [1]   9  30  92 173 179 189 205 228 237 254 276 291 293 304 315 333 335 339 340
#> [20] 380 386 437 443 457 722
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>  [1]   9  30  92 173 179 189 205 228 237 254 276 291 293 304 315 333 335 339 340
#> [20] 380 386 437 443 457 722
#> 
#> [[1]]$test[[2]]
#> [1] 117 178 363 364 393 497 545 742
#> 
#> [[1]]$test[[3]]
#>  [1]  58 103 136 222 245 303 356 413 544 573 594 690 706
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  31  32  33  34  35  36  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  75
#>  [73]  76  77  78  79  80  82  83  84  85  86  87  88  89  90  91  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113
#> [109] 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 161 162 163 164 165 166 167 168 169
#> [163] 170 171 172 174 175 176 177 178 180 181 182 183 184 185 186 187 188 190
#> [181] 191 192 193 194 195 196 197 198 199 200 201 202 203 204 206 207 208 209
#> [199] 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227
#> [217] 229 230 231 232 233 234 235 236 238 239 240 241 242 243 244 245 246 247
#> [235] 248 249 250 251 252 253 255 256 257 258 259 260 261 262 263 264 265 266
#> [253] 267 268 269 270 271 272 273 274 275 277 278 279 280 281 282 283 284 285
#> [271] 286 287 288 289 290 294 295 296 297 298 299 300 301 302 303 305 306 307
#> [289] 308 309 310 311 312 313 314 316 317 318 319 320 321 322 323 324 325 326
#> [307] 327 328 329 330 331 332 334 336 337 338 341 342 343 344 345 346 347 348
#> [325] 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366
#> [343] 367 368 369 370 371 372 373 374 375 376 377 378 379 381 382 383 384 385
#> [361] 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403 404
#> [379] 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421 422
#> [397] 423 424 425 426 427 428 429 430 431 432 433 434 435 438 439 440 441 442
#> [415] 444 445 446 447 448 449 450 451 452 453 455 456 458 459 460 461 462 463
#> [433] 464 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481
#> [451] 482 483 484 485 486 487 488 489 490 491 492 493 494 495 496 497 498 499
#> [469] 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515 516 517
#> [487] 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532 533 534 535
#> [505] 536 537 538 539 540 541 542 543 544 545 546 547 548 549 550 552 553 554
#> [523] 555 556 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572
#> [541] 573 574 575 576 577 578 579 580 581 582 583 584 585 586 588 589 590 591
#> [559] 592 594 595 596 597 598 599 600 601 602 603 604 605 607 608 609 610 611
#> [577] 612 613 614 615 617 618 619 620 621 623 624 625 626 627 628 629 630 631
#> [595] 632 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648 649
#> [613] 650 651 652 653 654 655 656 657 658 659 660 661 662 663 664 665 666 667
#> [631] 668 669 670 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685
#> [649] 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703
#> [667] 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721
#> [685] 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740
#> [703] 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  34  35  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74
#>  [73]  75  76  77  78  79  80  81  82  83  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 113
#> [109] 114 115 116 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150
#> [145] 151 152 153 154 155 156 157 158 159 160 161 162 163 164 166 167 168 169
#> [163] 170 171 172 173 174 175 176 177 179 180 181 182 183 184 185 186 187 188
#> [181] 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205 206
#> [199] 207 208 209 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225
#> [217] 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243
#> [235] 244 245 246 247 248 249 250 251 252 253 254 255 256 258 259 260 261 262
#> [253] 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280
#> [271] 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298
#> [289] 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316
#> [307] 317 318 319 320 321 322 323 325 326 327 328 329 330 331 332 333 334 335
#> [325] 336 337 338 339 340 341 342 343 344 345 346 347 348 349 350 352 353 354
#> [343] 355 356 357 358 359 360 361 362 365 366 367 368 369 370 371 372 373 375
#> [361] 376 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 394
#> [379] 395 396 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412
#> [397] 413 414 415 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430
#> [415] 431 432 433 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448
#> [433] 449 450 451 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466
#> [451] 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484
#> [469] 485 486 487 488 489 490 491 492 493 494 495 496 498 499 500 501 502 503
#> [487] 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [505] 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 537 538 539
#> [523] 540 541 542 543 544 546 547 548 549 550 551 552 553 554 555 556 557 558
#> [541] 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575 576
#> [559] 577 578 579 580 581 582 583 585 586 587 588 589 590 591 592 593 594 595
#> [577] 596 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 612 613
#> [595] 614 615 616 617 618 619 620 621 622 623 625 626 627 628 629 630 631 632
#> [613] 633 634 635 636 637 638 639 641 642 643 644 645 646 647 648 649 650 651
#> [631] 652 653 654 655 656 657 658 659 660 661 662 663 664 665 666 667 668 669
#> [649] 670 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685 686 687
#> [667] 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705
#> [685] 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721 722 723
#> [703] 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740 741
#> [721] 743 744 745 746 747 748 749 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  91  92
#>  [91]  93  94  95  96  97  98  99 100 101 102 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 118 119 120 121 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167
#> [163] 168 169 171 172 173 174 175 176 177 178 179 180 181 182 184 185 186 187
#> [181] 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204 205
#> [199] 206 207 208 209 210 211 212 213 214 215 216 217 219 220 221 223 224 225
#> [217] 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243
#> [235] 244 246 247 248 249 250 251 252 253 254 255 256 257 258 259 260 261 262
#> [253] 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280
#> [271] 281 282 283 284 285 286 287 288 289 291 292 293 294 295 296 297 298 299
#> [289] 300 301 302 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318
#> [307] 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336
#> [325] 337 338 339 340 341 342 343 344 346 347 348 349 350 351 352 353 354 355
#> [343] 357 358 359 360 361 363 364 365 366 367 368 369 370 371 372 373 374 375
#> [361] 376 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393
#> [379] 394 395 396 397 398 399 400 401 402 404 405 406 407 408 409 410 411 412
#> [397] 414 415 416 417 418 419 420 421 422 423 424 425 427 428 429 430 431 432
#> [415] 433 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 450 451
#> [433] 452 453 454 455 456 457 458 459 460 461 462 464 465 466 467 468 469 470
#> [451] 471 472 473 475 476 477 478 479 480 481 482 484 485 487 488 489 491 492
#> [469] 493 494 495 496 497 498 499 500 501 502 504 505 506 507 508 510 511 512
#> [487] 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530
#> [505] 531 532 533 534 535 536 537 538 539 540 541 542 543 545 546 547 548 549
#> [523] 550 551 552 553 554 555 556 557 558 559 560 561 562 563 565 566 567 568
#> [541] 569 570 571 572 574 575 576 577 578 579 580 581 582 583 584 586 587 588
#> [559] 589 590 591 592 593 595 596 597 598 599 600 601 602 603 604 605 606 607
#> [577] 608 609 610 611 612 614 615 616 617 618 619 620 621 622 623 624 625 626
#> [595] 627 628 629 630 631 632 633 634 635 636 637 638 639 640 642 643 645 646
#> [613] 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662 663 664
#> [631] 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679 680 681 682
#> [649] 683 684 685 686 687 688 689 691 692 693 694 695 696 697 698 699 700 701
#> [667] 702 703 704 705 707 709 710 711 712 713 714 715 716 717 718 719 720 721
#> [685] 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739
#> [703] 740 741 742 744 745 746 747 748 749 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#> [1]  71 185 247 278 398 438 574 725 738
#> 
#> [[2]]$test[[2]]
#>  [1]  27  67  82  98 108 141 154 180 233 235 275 296 328 353 389 392 401 415 456
#> [20] 600
#> 
#> [[2]]$test[[3]]
#>  [1] 143 188 204 285 318 381 409 477 505 567 591 740
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  72  74
#>  [73]  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92
#>  [91]  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110
#> [109] 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128
#> [127] 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146
#> [145] 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164
#> [163] 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182
#> [181] 183 184 186 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201
#> [199] 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 218 219 220
#> [217] 221 222 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238
#> [235] 239 240 241 242 243 244 245 246 248 249 250 251 252 253 254 255 256 257
#> [253] 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275
#> [271] 276 277 279 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294
#> [289] 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312
#> [307] 313 314 315 316 317 318 319 320 321 322 323 324 325 326 327 328 329 330
#> [325] 331 332 333 334 335 336 337 338 339 340 341 342 343 344 345 346 347 348
#> [343] 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366
#> [361] 367 368 369 370 371 372 373 374 375 376 377 378 379 380 381 382 383 384
#> [379] 385 386 387 388 389 390 391 392 393 394 395 396 397 399 400 401 402 403
#> [397] 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420 421
#> [415] 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 439 440
#> [433] 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 458
#> [451] 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475 476
#> [469] 477 478 479 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494
#> [487] 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512
#> [505] 514 515 516 517 518 519 521 522 523 524 525 526 527 528 529 530 531 532
#> [523] 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549 550
#> [541] 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [559] 569 570 571 572 573 575 576 577 578 579 580 581 582 583 584 585 586 587
#> [577] 588 590 591 593 594 595 596 597 598 599 600 601 602 603 604 606 607 608
#> [595] 609 611 613 614 616 617 618 619 620 621 622 623 624 625 626 627 628 629
#> [613] 630 631 632 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647
#> [631] 648 649 650 651 652 653 654 655 657 658 660 661 662 663 664 665 666 667
#> [649] 668 669 670 671 672 673 674 675 676 677 678 680 681 682 683 684 685 686
#> [667] 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704
#> [685] 705 706 708 709 710 711 712 714 715 716 717 718 719 720 721 722 723 724
#> [703] 726 727 728 729 730 731 732 733 734 735 736 737 739 740 741 742 743 744
#> [721] 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  21  22  23  24  25  26  28  29  30  32  33  34  35  36  37  38  39
#>  [37]  42  43  44  45  46  47  48  49  50  51  52  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  68  70  71  74  75  76  77  78  79  80  81  83
#>  [73]  84  85  86  87  88  89  90  91  92  93  94  95  96  97  99 100 101 102
#>  [91] 103 104 105 106 107 109 110 111 112 113 115 116 117 118 119 120 121 122
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> [127] 142 143 144 145 146 147 148 149 150 151 152 153 155 156 157 158 159 160
#> [145] 161 162 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178
#> [163] 179 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197
#> [181] 198 199 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215
#> [199] 216 217 218 219 220 221 222 223 224 225 227 228 229 230 231 232 234 236
#> [217] 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254
#> [235] 255 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272
#> [253] 273 274 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290 291
#> [271] 292 293 294 295 297 300 302 303 304 305 306 307 308 309 310 311 312 313
#> [289] 314 315 316 317 318 319 320 321 322 323 324 325 327 329 330 331 332 333
#> [307] 335 336 337 338 339 340 341 342 344 345 346 347 348 349 350 351 352 354
#> [325] 355 356 357 358 359 361 362 363 364 365 366 368 369 370 371 372 373 374
#> [343] 375 376 377 378 379 380 381 382 383 384 385 386 387 388 390 391 393 394
#> [361] 395 396 397 398 399 400 402 403 404 405 406 407 408 409 410 411 412 413
#> [379] 414 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432
#> [397] 433 434 435 436 437 438 439 440 442 443 444 445 446 447 448 449 450 451
#> [415] 452 453 454 455 457 458 459 460 461 463 464 465 466 467 468 469 470 471
#> [433] 472 473 474 475 476 477 478 479 480 481 482 483 485 486 487 488 489 490
#> [451] 491 492 493 494 495 496 497 498 499 501 502 503 504 505 506 507 508 509
#> [469] 510 511 512 513 514 515 516 517 518 520 521 522 523 524 525 526 527 528
#> [487] 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546
#> [505] 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564
#> [523] 565 566 567 568 569 570 571 572 573 574 575 577 578 579 580 581 583 584
#> [541] 585 586 587 588 589 590 591 592 593 594 596 597 598 599 601 602 603 604
#> [559] 605 606 607 609 611 612 613 614 615 616 617 618 619 620 621 622 623 624
#> [577] 625 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642
#> [595] 643 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660
#> [613] 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678
#> [631] 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696
#> [649] 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714
#> [667] 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731 732
#> [685] 733 734 735 736 737 738 739 740 741 742 743 744 745 746 748 749 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 124 125 126 127
#> [127] 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 144 145 146
#> [145] 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164
#> [163] 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181 182
#> [181] 183 184 185 186 187 189 190 191 192 193 194 195 196 197 198 199 200 201
#> [199] 202 203 205 206 207 208 209 210 211 212 213 214 216 217 218 219 220 221
#> [217] 222 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239
#> [235] 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257
#> [253] 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275
#> [271] 276 277 278 280 281 282 283 284 286 287 288 289 290 291 292 293 294 295
#> [289] 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313
#> [307] 314 315 316 317 319 320 321 322 323 324 325 326 327 328 329 330 331 332
#> [325] 333 334 335 336 337 338 339 340 341 342 343 344 345 346 347 348 349 350
#> [343] 351 352 353 354 355 356 357 358 359 360 361 362 363 364 366 367 368 369
#> [361] 370 372 373 374 375 376 377 378 379 380 382 383 384 385 386 387 388 389
#> [379] 390 391 392 393 394 395 396 397 398 399 400 401 402 403 404 405 406 407
#> [397] 408 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425 426
#> [415] 427 428 429 430 432 433 434 435 436 437 438 439 440 441 442 443 444 445
#> [433] 447 448 449 450 451 452 453 454 455 456 457 458 459 460 461 462 463 464
#> [451] 465 466 467 468 469 470 471 472 473 474 475 476 478 479 480 481 482 483
#> [469] 484 485 486 487 488 489 490 491 492 493 494 495 497 498 499 500 501 502
#> [487] 503 504 506 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [505] 522 523 524 525 526 527 528 529 530 531 532 534 535 536 537 539 540 541
#> [523] 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557 558 560
#> [541] 561 562 563 564 565 566 568 569 571 572 573 574 575 576 577 578 579 580
#> [559] 581 582 583 584 585 586 587 588 589 590 592 593 594 595 596 597 598 599
#> [577] 600 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617
#> [595] 618 619 620 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636
#> [613] 637 638 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654
#> [631] 655 656 657 659 660 661 662 663 664 665 666 668 669 670 671 672 673 674
#> [649] 675 676 677 678 679 681 682 683 684 685 686 687 688 689 690 691 692 693
#> [667] 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711
#> [685] 712 713 714 715 716 717 718 719 720 721 722 723 724 725 727 728 729 730
#> [703] 731 732 733 734 735 736 737 738 739 741 742 743 744 745 746 747 748 749
#> [721] 751
#> 
#> 
#>