(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  81  82  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166
#> [163] 167 168 169 170 171 172 173 174 175 177 178 179 180 181 182 183 184 185
#> [181] 186 187 188 189 190 191 192 193 194 195 196 198 199 200 201 202 203 204
#> [199] 205 206 207 208 209 210 211 212 213 214 215 216 217 219 220 221 222 223
#> [217] 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241
#> [235] 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259
#> [253] 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277
#> [271] 278 279 280 281 282 283 284 285 286 287 288 289 291 292 293 294 295 296
#> [289] 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314
#> [307] 315 316 317 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332
#> [325] 333 334 335 336 337 338 339 340 342 343 344 345 346 347 348 349 350 351
#> [343] 353 354 355 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371
#> [361] 372 373 374 375 377 378 379 380 381 382 383 384 385 386 387 388 389 390
#> [379] 391 392 393 394 395 396 397 398 399 400 401 402 403 404 405 406 407 408
#> [397] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 425 426 427
#> [415] 428 429 430 431 432 433 434 435 436 437 438 439 440 441 442 443 444 445
#> [433] 446 447 448 449 451 452 453 454 455 456 457 458 459 460 461 462 463 464
#> [451] 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481 482
#> [469] 483 484 485 487 488 489 490 491 492 493 494 495 496 497 498 499 500 501
#> [487] 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [505] 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 537 538 539
#> [523] 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557
#> [541] 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575
#> [559] 576 577 578 579 580 581 582 583 584 586 587 588 589 590 591 592 593 594
#> [577] 595 596 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 613
#> [595] 614 616 617 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632
#> [613] 633 635 636 637 638 639 640 641 642 643 645 646 647 648 649 650 651 653
#> [631] 654 655 656 657 658 659 660 661 663 664 665 666 667 668 669 670 671 672
#> [649] 673 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690
#> [667] 691 692 693 694 695 696 697 699 700 701 702 703 704 705 706 707 708 709
#> [685] 710 711 712 713 714 715 716 717 718 719 720 721 722 723 725 726 727 728
#> [703] 729 730 731 732 733 734 735 736 737 738 739 740 741 742 743 744 745 746
#> [721] 747 748 749 750 751
rrcv$test_set(1)
#> [1] 502 634 652 724
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#> [1] 502 634 652 724
#> 
#> [[1]]$test[[2]]
#>  [1]   8  15 166 261 280 287 354 435 451 461 488 499 561 683
#> 
#> [[1]]$test[[3]]
#>  [1]  74 119 159 160 187 194 244 292 323 355 390 436 454 506 593 625 717
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  81  82  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166
#> [163] 167 168 169 170 171 172 173 174 175 177 178 179 180 181 182 183 184 185
#> [181] 186 187 188 189 190 191 192 193 194 195 196 198 199 200 201 202 203 204
#> [199] 205 206 207 208 209 210 211 212 213 214 215 216 217 219 220 221 222 223
#> [217] 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 240 241
#> [235] 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259
#> [253] 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277
#> [271] 278 279 280 281 282 283 284 285 286 287 288 289 291 292 293 294 295 296
#> [289] 297 298 299 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314
#> [307] 315 316 317 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332
#> [325] 333 334 335 336 337 338 339 340 342 343 344 345 346 347 348 349 350 351
#> [343] 353 354 355 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371
#> [361] 372 373 374 375 377 378 379 380 381 382 383 384 385 386 387 388 389 390
#> [379] 391 392 393 394 395 396 397 398 399 400 401 402 403 404 405 406 407 408
#> [397] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 425 426 427
#> [415] 428 429 430 431 432 433 434 435 436 437 438 439 440 441 442 443 444 445
#> [433] 446 447 448 449 451 452 453 454 455 456 457 458 459 460 461 462 463 464
#> [451] 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481 482
#> [469] 483 484 485 487 488 489 490 491 492 493 494 495 496 497 498 499 500 501
#> [487] 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521
#> [505] 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 537 538 539
#> [523] 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557
#> [541] 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575
#> [559] 576 577 578 579 580 581 582 583 584 586 587 588 589 590 591 592 593 594
#> [577] 595 596 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 613
#> [595] 614 616 617 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632
#> [613] 633 635 636 637 638 639 640 641 642 643 645 646 647 648 649 650 651 653
#> [631] 654 655 656 657 658 659 660 661 663 664 665 666 667 668 669 670 671 672
#> [649] 673 674 675 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690
#> [667] 691 692 693 694 695 696 697 699 700 701 702 703 704 705 706 707 708 709
#> [685] 710 711 712 713 714 715 716 717 718 719 720 721 722 723 725 726 727 728
#> [703] 729 730 731 732 733 734 735 736 737 738 739 740 741 742 743 744 745 746
#> [721] 747 748 749 750 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   9  10  11  12  13  14  16  17  18  19  20
#>  [19]  21  22  23  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  71  72  73  74  75  76
#>  [73]  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94
#>  [91]  95  96  98  99 100 101 103 105 106 107 108 109 110 111 112 113 114 115
#> [109] 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133
#> [127] 134 135 136 137 138 139 141 142 143 144 145 146 147 148 149 150 151 152
#> [145] 154 155 156 157 158 159 160 161 162 163 164 165 167 168 169 170 171 172
#> [163] 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 190
#> [181] 191 193 194 195 196 197 198 199 200 201 202 203 204 205 206 208 209 210
#> [199] 211 212 213 215 216 217 218 219 220 221 222 223 224 225 226 228 229 230
#> [217] 231 232 233 234 235 236 237 238 239 240 241 242 243 244 245 246 247 248
#> [235] 249 250 251 252 253 254 255 256 257 258 259 260 262 263 264 265 266 267
#> [253] 268 269 270 271 272 273 274 275 276 277 278 279 281 282 283 284 285 286
#> [271] 288 289 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 306
#> [289] 307 308 309 310 311 312 313 314 315 316 317 318 319 320 321 322 323 324
#> [307] 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342
#> [325] 343 344 345 346 347 348 349 350 351 352 353 355 356 357 358 359 360 361
#> [343] 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379
#> [361] 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396 397
#> [379] 398 399 400 401 403 404 405 406 407 408 409 410 411 412 413 414 415 416
#> [397] 417 418 419 420 421 422 423 424 425 426 427 429 430 431 432 433 434 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 452 453 454 455
#> [433] 456 457 458 459 460 462 463 464 465 466 467 468 469 470 471 472 473 474
#> [451] 475 476 477 478 479 480 481 482 483 484 485 486 487 489 490 491 492 493
#> [469] 494 495 496 497 498 500 501 502 503 504 505 506 507 508 509 510 511 513
#> [487] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531
#> [505] 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549
#> [523] 550 551 552 553 554 555 556 557 559 560 564 565 566 567 568 569 570 571
#> [541] 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586 587 588 589
#> [559] 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607
#> [577] 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 625
#> [595] 626 627 628 629 630 631 632 634 635 636 637 638 639 640 641 642 643 644
#> [613] 645 646 647 648 649 650 652 653 654 655 656 657 658 659 660 661 662 663
#> [631] 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679 680 681
#> [649] 682 684 685 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700
#> [667] 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718
#> [685] 719 720 721 722 723 724 725 726 727 728 729 731 732 733 734 735 736 737
#> [703] 738 739 740 741 742 743 744 745 747 748 749 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  75  76  77  78  79  80  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 118 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 156 157 158 161 162 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 177 178 180 181 182 183 184 185 186 188
#> [181] 190 191 192 193 195 196 197 198 199 200 201 202 203 204 205 206 207 209
#> [199] 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227
#> [217] 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 245 246
#> [235] 247 248 249 250 251 252 253 255 256 257 258 259 260 261 262 263 264 265
#> [253] 266 267 268 269 270 271 272 273 274 275 277 278 279 280 281 282 283 284
#> [271] 285 286 287 288 289 290 293 294 295 296 297 298 299 300 301 302 303 305
#> [289] 306 307 308 309 310 311 312 313 314 316 317 318 319 320 321 322 324 325
#> [307] 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343
#> [325] 344 345 346 347 348 349 350 351 352 353 354 356 357 358 359 360 361 362
#> [343] 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379 381
#> [361] 382 383 384 385 386 387 388 389 391 392 393 394 395 396 397 398 399 400
#> [379] 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418
#> [397] 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 437
#> [415] 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 455 456
#> [433] 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475
#> [451] 476 477 478 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494
#> [469] 495 496 497 498 499 500 501 502 503 504 505 507 508 509 510 511 512 513
#> [487] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531
#> [505] 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549
#> [523] 550 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [541] 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586
#> [559] 587 588 589 590 591 592 594 595 596 598 599 600 602 603 604 605 606 607
#> [577] 608 609 610 611 612 613 614 615 617 618 619 620 621 622 623 624 626 627
#> [595] 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642 643 644 645
#> [613] 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662 663
#> [631] 664 665 666 667 668 669 670 671 673 674 675 676 677 678 679 680 681 682
#> [649] 683 684 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701
#> [667] 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 718 719 720
#> [685] 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738
#> [703] 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#> [1] 365 371 381 431 496 538 621 658
#> 
#> [[2]]$test[[2]]
#> [1]  66 105 527 703
#> 
#> [[2]]$test[[3]]
#> [1] 123 162 223 288 446 472 532
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 144 145
#> [145] 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163
#> [163] 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181
#> [181] 182 183 184 185 186 187 189 190 191 192 193 194 195 196 197 198 199 200
#> [199] 201 202 203 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219
#> [217] 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237
#> [235] 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255
#> [253] 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273
#> [271] 274 275 276 277 278 279 280 281 282 283 284 286 287 288 289 290 291 292
#> [289] 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308 309 310
#> [307] 311 312 313 314 315 316 317 319 320 321 322 323 324 325 326 327 328 329
#> [325] 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344 345 346 347
#> [343] 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 366
#> [361] 367 368 369 370 372 373 374 375 376 377 378 379 380 382 383 384 385 386
#> [379] 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403 404
#> [397] 405 406 407 408 410 411 412 413 414 415 416 417 418 419 420 421 422 423
#> [415] 424 425 426 427 428 429 430 432 433 434 435 436 437 438 439 440 441 442
#> [433] 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 458 459 460
#> [451] 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475 476 478 479
#> [469] 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494 495 497 498
#> [487] 499 500 501 502 503 504 506 507 508 509 510 511 512 513 514 515 516 517
#> [505] 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532 534 536 537
#> [523] 539 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554 555 556
#> [541] 557 558 559 560 561 562 563 564 565 568 569 570 571 572 573 574 575 576
#> [559] 577 578 579 580 581 582 583 584 585 586 587 588 589 590 592 593 594 595
#> [577] 596 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 612 613
#> [595] 614 615 616 617 618 619 620 622 623 624 625 626 627 628 629 630 631 632
#> [613] 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648 649 650
#> [631] 651 652 653 654 655 656 657 659 660 661 662 663 664 665 666 668 669 670
#> [649] 671 672 673 674 675 676 677 678 679 680 681 682 683 685 686 687 688 689
#> [667] 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705 706 707
#> [685] 708 709 710 711 712 713 714 715 716 717 718 719 720 721 722 723 724 725
#> [703] 726 727 728 729 730 731 732 733 735 736 737 738 739 741 742 743 744 745
#> [721] 746 747 748 749 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  13  15  16  17  18  19  20
#>  [19]  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  41  42  43  44  45  46  47  48  49  50  51  52  53  55  56  57  58
#>  [55]  59  60  61  62  63  64  67  68  69  70  71  72  73  74  75  76  77  78
#>  [73]  79  80  81  82  83  84  85  87  88  89  90  91  92  93  95  96  97  98
#>  [91]  99 100 101 102 103 104 106 107 108 109 110 111 112 113 114 115 116 117
#> [109] 118 119 120 121 122 123 124 125 126 127 128 129 131 133 134 135 136 137
#> [127] 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153 154 155
#> [145] 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 172 173 174
#> [163] 175 176 177 178 179 180 182 183 184 185 186 187 188 189 190 192 193 194
#> [181] 196 197 198 199 200 201 203 204 205 206 207 208 209 210 212 213 214 215
#> [199] 216 217 218 219 220 221 222 223 224 226 227 228 229 230 231 232 233 234
#> [217] 235 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252
#> [235] 253 254 255 256 257 258 259 260 261 263 264 265 266 267 268 269 270 271
#> [253] 272 273 274 275 276 278 279 280 281 283 284 285 286 287 288 289 290 291
#> [271] 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 309 310
#> [289] 311 312 313 314 315 316 317 318 319 320 321 323 324 325 326 327 328 329
#> [307] 330 331 332 333 334 335 337 338 339 340 341 342 343 344 345 346 347 348
#> [325] 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366
#> [343] 367 368 369 370 371 372 373 374 375 376 377 378 379 380 381 382 383 384
#> [361] 385 386 387 388 389 390 391 392 393 394 396 397 398 399 400 401 402 403
#> [379] 404 405 406 407 408 409 410 411 412 413 414 415 417 418 419 420 421 422
#> [397] 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439 440
#> [415] 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 459
#> [433] 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475 476 477
#> [451] 479 480 481 482 483 484 485 486 487 488 489 490 491 492 494 495 496 497
#> [469] 499 500 501 502 503 504 505 506 507 508 509 510 512 513 514 515 516 517
#> [487] 518 519 520 521 522 523 524 525 526 528 529 530 531 532 533 534 535 536
#> [505] 537 538 539 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554
#> [523] 555 556 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572
#> [541] 573 574 575 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590
#> [559] 591 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607 608
#> [577] 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 625 626
#> [595] 627 628 629 630 632 633 634 635 636 637 638 639 640 641 642 643 644 645
#> [613] 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662 663
#> [631] 664 665 666 667 668 669 670 672 673 674 675 676 677 678 679 680 681 682
#> [649] 683 684 685 686 687 688 689 690 692 693 694 695 696 697 698 699 700 702
#> [667] 704 705 706 707 708 709 710 711 712 713 715 716 717 718 719 720 721 722
#> [685] 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740
#> [703] 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91
#>  [91]  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109
#> [109] 110 111 112 113 114 115 116 117 118 119 120 121 122 124 125 126 127 128
#> [127] 129 130 131 132 133 134 135 136 137 138 139 141 142 144 145 146 148 149
#> [145] 150 151 152 153 154 155 156 157 158 159 160 161 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186
#> [181] 187 188 189 190 191 192 193 194 195 197 198 199 201 202 203 205 206 207
#> [199] 208 209 210 211 212 213 214 216 217 218 219 220 221 222 224 225 226 227
#> [217] 228 229 230 231 232 233 234 235 236 237 238 239 241 243 244 245 247 248
#> [235] 249 250 251 252 253 254 255 256 257 258 259 260 261 262 263 264 265 266
#> [253] 267 268 269 270 271 272 273 274 275 276 277 278 280 281 282 283 284 285
#> [271] 286 287 289 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304
#> [289] 305 306 307 308 309 310 311 312 313 314 315 316 317 319 320 321 322 323
#> [307] 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341
#> [325] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [343] 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377
#> [361] 378 379 380 381 382 383 385 386 387 388 389 390 391 392 393 394 395 396
#> [379] 397 398 399 400 401 402 403 404 405 406 407 408 410 411 412 413 414 415
#> [397] 416 417 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434
#> [415] 435 436 437 438 439 440 441 442 443 444 445 447 448 449 450 451 452 453
#> [433] 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471
#> [451] 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488 489 490
#> [469] 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508
#> [487] 509 510 511 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527
#> [505] 528 529 530 531 534 535 536 537 538 539 540 541 542 543 544 545 546 547
#> [523] 548 549 550 551 552 553 555 556 557 559 560 561 562 563 564 565 566 568
#> [541] 569 571 572 573 574 575 576 577 578 579 581 582 583 584 585 586 587 588
#> [559] 589 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606
#> [577] 608 609 610 611 612 613 614 615 616 618 619 620 621 622 623 624 625 626
#> [595] 627 628 629 630 631 632 633 634 635 636 637 639 640 641 642 643 644 645
#> [613] 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662 663
#> [631] 664 665 666 667 668 669 670 671 672 673 674 675 676 677 678 679 680 681
#> [649] 682 683 684 685 686 687 688 689 690 691 692 693 694 695 696 697 698 699
#> [667] 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 717
#> [685] 718 719 720 721 722 724 725 727 728 729 731 732 733 734 735 736 737 738
#> [703] 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#>