(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  75  76  77  78  79  80  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 118 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 156 157 158 161 162 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 177 178 180 181 182 183 184 185 186 188
#> [181] 190 191 192 193 195 196 197 198 199 200 201 202 203 204 205 206 207 209
#> [199] 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227
#> [217] 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 245 246
#> [235] 247 248 249 250 251 252 253 255 256 257 258 259 260 261 262 263 264 265
#> [253] 266 267 268 269 270 271 272 273 274 275 277 278 279 280 281 282 283 284
#> [271] 285 286 287 288 289 290 293 294 295 296 297 298 299 300 301 302 303 305
#> [289] 306 307 308 309 310 311 312 313 314 316 317 318 319 320 321 322 324 325
#> [307] 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343
#> [325] 344 345 346 347 348 349 350 351 352 353 354 356 357 358 359 360 361 362
#> [343] 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379 381
#> [361] 382 383 384 385 386 387 388 389 391 392 393 394 395 396 397 398 399 400
#> [379] 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418
#> [397] 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 437
#> [415] 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 455 456
#> [433] 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475
#> [451] 476 477 478 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494
#> [469] 495 496 497 498 499 500 501 502 503 504 505 507 508 509 510 511 512 513
#> [487] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531
#> [505] 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549
#> [523] 550 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [541] 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586
#> [559] 587 588 589 590 591 592 594 595 596 598 599 600 602 603 604 605 606 607
#> [577] 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 626
#> [595] 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642 643 644
#> [613] 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662
#> [631] 663 664 665 666 667 668 669 670 671 673 674 675 676 677 678 679 680 681
#> [649] 682 683 684 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700
#> [667] 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 718 719
#> [685] 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737
#> [703] 738 739 740 741 742 743 744 745 746 747 748 749 750 751
rrcv$test_set(1)
#>  [1]  74  81 119 159 160 187 194 244 355 390 436 506 593 625 672 717
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>  [1]  74  81 119 159 160 187 194 244 355 390 436 506 593 625 672 717
#> 
#> [[1]]$test[[2]]
#>  [1]   9  92 173 189 205 228 237 254 276 293 333 335 339 340 386 437 443 587 722
#> 
#> [[1]]$test[[3]]
#>  [1]  22  80  83 101 176 197 341 352 450 556 568 614 648
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  75  76  77  78  79  80  82  83  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111
#> [109] 112 113 114 115 116 117 118 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148
#> [145] 149 150 151 152 153 154 155 156 157 158 161 162 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 177 178 180 181 182 183 184 185 186 188
#> [181] 190 191 192 193 195 196 197 198 199 200 201 202 203 204 205 206 207 209
#> [199] 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225 226 227
#> [217] 228 229 230 231 232 233 234 235 236 237 238 239 240 241 242 243 245 246
#> [235] 247 248 249 250 251 252 253 255 256 257 258 259 260 261 262 263 264 265
#> [253] 266 267 268 269 270 271 272 273 274 275 277 278 279 280 281 282 283 284
#> [271] 285 286 287 288 289 290 293 294 295 296 297 298 299 300 301 302 303 305
#> [289] 306 307 308 309 310 311 312 313 314 316 317 318 319 320 321 322 324 325
#> [307] 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343
#> [325] 344 345 346 347 348 349 350 351 352 353 354 356 357 358 359 360 361 362
#> [343] 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379 381
#> [361] 382 383 384 385 386 387 388 389 391 392 393 394 395 396 397 398 399 400
#> [379] 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418
#> [397] 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 437
#> [415] 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 455 456
#> [433] 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475
#> [451] 476 477 478 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494
#> [469] 495 496 497 498 499 500 501 502 503 504 505 507 508 509 510 511 512 513
#> [487] 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529 530 531
#> [505] 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549
#> [523] 550 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568
#> [541] 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586
#> [559] 587 588 589 590 591 592 594 595 596 598 599 600 602 603 604 605 606 607
#> [577] 608 609 610 611 612 613 614 615 616 617 618 619 620 621 622 623 624 626
#> [595] 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642 643 644
#> [613] 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662
#> [631] 663 664 665 666 667 668 669 670 671 673 674 675 676 677 678 679 680 681
#> [649] 682 683 684 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700
#> [667] 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 718 719
#> [685] 720 721 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737
#> [703] 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  31  32  33  34  35  36  37  38
#>  [37]  39  40  41  42  44  45  46  47  48  49  50  51  52  53  54  55  56  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75
#>  [73]  76  77  78  79  80  82  83  84  85  86  87  88  89  90  91  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113
#> [109] 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167
#> [163] 168 169 170 171 172 174 175 176 177 178 180 181 182 183 184 185 186 187
#> [181] 188 190 191 192 193 194 195 196 197 198 199 200 201 202 203 204 206 207
#> [199] 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224 225
#> [217] 226 227 229 230 231 232 233 234 235 236 238 239 240 241 242 243 244 245
#> [235] 246 247 248 249 250 251 252 253 255 256 257 258 259 260 261 262 263 264
#> [253] 265 266 267 268 269 270 271 272 273 274 275 277 278 279 280 281 282 283
#> [271] 285 286 287 288 289 290 292 294 295 296 297 298 299 300 301 302 303 305
#> [289] 306 307 308 309 310 311 312 313 314 316 317 318 319 320 321 322 323 324
#> [307] 325 326 327 328 329 330 331 332 334 336 337 338 341 342 343 344 345 346
#> [325] 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362 363 364
#> [343] 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379 381 382 383
#> [361] 384 385 387 388 389 390 391 392 393 394 395 396 397 398 399 400 401 402
#> [379] 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 420
#> [397] 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435 438 439 440
#> [415] 441 442 444 445 446 447 448 449 450 451 452 453 454 455 456 458 459 460
#> [433] 461 462 463 464 465 466 467 468 469 470 471 472 473 474 475 476 477 478
#> [451] 479 480 481 482 483 484 485 486 487 488 489 490 491 492 493 494 496 497
#> [469] 498 499 500 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515
#> [487] 516 517 518 519 520 522 523 524 525 526 527 528 529 530 531 532 533 534
#> [505] 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549 550 552 553
#> [523] 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571
#> [541] 572 573 574 575 576 577 578 579 580 581 582 583 584 585 586 588 589 590
#> [559] 591 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607 608
#> [577] 609 610 611 612 613 614 615 617 618 619 620 621 623 624 625 626 627 628
#> [595] 629 630 631 632 633 634 635 636 637 638 639 640 641 642 643 644 645 646
#> [613] 647 648 649 650 651 652 653 654 655 656 657 658 660 661 662 663 664 665
#> [631] 666 667 668 669 670 671 672 673 674 676 677 678 679 680 681 682 683 684
#> [649] 685 686 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702
#> [667] 703 704 705 706 707 708 709 710 711 712 714 715 716 717 718 719 720 721
#> [685] 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740
#> [703] 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  81  82  84  85  86  87  88  89  90  91  92  93
#>  [91]  94  95  96  97  98  99 100 102 103 104 105 106 107 108 109 110 111 112
#> [109] 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130
#> [127] 131 132 133 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149
#> [145] 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167
#> [163] 168 169 170 171 172 173 174 175 178 179 180 181 182 183 184 185 186 187
#> [181] 188 189 190 191 192 193 194 195 196 198 199 200 201 202 203 204 205 206
#> [199] 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223 224
#> [217] 225 226 227 228 229 230 231 232 233 234 235 236 237 238 240 241 242 243
#> [235] 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 260 261
#> [253] 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279
#> [271] 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 298
#> [289] 299 300 301 302 303 304 305 306 307 308 309 310 312 313 314 315 316 317
#> [307] 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335
#> [325] 336 337 338 339 340 342 343 344 345 346 347 348 349 350 351 353 354 355
#> [343] 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373
#> [361] 374 375 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392
#> [379] 393 394 395 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412
#> [397] 413 414 415 416 417 418 419 420 421 422 423 425 426 427 428 429 430 431
#> [415] 432 433 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 451
#> [433] 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469
#> [451] 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487
#> [469] 488 489 490 491 492 493 494 495 496 497 498 499 500 501 503 504 505 506
#> [487] 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524
#> [505] 525 526 527 528 529 530 531 532 533 534 535 537 538 539 540 541 542 543
#> [523] 544 545 546 547 548 549 550 551 552 553 554 555 557 558 559 560 561 562
#> [541] 563 564 565 566 567 569 570 571 572 573 574 575 576 577 578 579 580 581
#> [559] 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599
#> [577] 600 601 602 603 604 605 606 607 608 609 610 611 612 613 616 617 618 619
#> [595] 620 621 622 623 624 625 626 627 628 629 630 631 632 633 635 636 637 638
#> [613] 639 640 641 642 643 644 645 646 647 649 650 651 653 654 655 656 657 658
#> [631] 659 660 661 662 663 664 665 667 668 669 670 671 672 673 674 675 676 677
#> [649] 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694 695
#> [667] 696 697 699 700 701 702 703 704 705 706 707 708 709 710 711 712 714 715
#> [685] 716 717 718 719 720 721 722 723 725 726 727 728 729 730 731 732 733 734
#> [703] 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#> [1] 117 178 393 545 624 742
#> 
#> [[2]]$test[[2]]
#>  [1]   1   6  23  33  34  84 100 112 146 164 167 168 175 193 210 248 257 294 307
#> [20] 310 314 324 330 342 347 351 374 425 485 491 572 577 626 632 699
#> 
#> [[2]]$test[[3]]
#>  [1]  23  34  51 107 146 149 193 232 252 286 307 314 330 347 379 432 439 485 491
#> [20] 507 577 626 632 645 647 660 678
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  34  35  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58
#>  [55]  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76
#>  [73]  77  78  79  80  81  82  83  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  98  99 101 102 103 104 105 106 107 108 109 110 113 114 115 116
#> [109] 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135
#> [127] 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153
#> [145] 154 155 156 157 158 159 160 161 162 163 166 169 170 171 172 173 174 176
#> [163] 177 179 180 181 182 183 185 186 187 188 189 190 191 192 193 194 195 196
#> [181] 197 198 199 200 201 202 203 204 205 206 207 208 209 211 212 213 214 215
#> [199] 216 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233
#> [217] 234 235 236 237 238 239 240 241 242 243 244 245 246 247 249 250 251 252
#> [235] 253 254 255 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272
#> [253] 273 274 275 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290
#> [271] 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308
#> [289] 309 311 312 313 314 315 316 317 318 319 320 321 322 323 325 326 327 328
#> [307] 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344 345 346
#> [325] 347 348 349 350 352 353 354 355 356 357 358 359 360 361 362 365 366 367
#> [343] 368 369 370 371 372 373 375 376 377 378 379 380 381 383 384 385 386 387
#> [361] 388 389 390 392 394 395 396 397 398 399 400 401 402 403 404 405 406 408
#> [379] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 426 427
#> [397] 428 429 430 431 432 433 434 435 436 437 438 439 440 441 442 443 444 445
#> [415] 446 447 448 449 450 451 452 453 454 455 456 457 458 459 460 461 462 463
#> [433] 464 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 482
#> [451] 483 484 485 486 487 488 489 490 491 492 493 494 495 496 498 499 500 501
#> [469] 502 503 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518 519
#> [487] 520 521 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536 537
#> [505] 538 539 540 541 542 543 544 546 547 548 549 550 551 552 553 554 555 556
#> [523] 557 558 559 560 561 562 563 564 565 566 567 568 569 570 571 572 573 574
#> [541] 575 576 577 578 579 580 581 582 583 584 585 586 587 588 589 590 591 592
#> [559] 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607 608 609 610
#> [577] 611 612 613 614 615 616 617 618 619 620 621 622 623 625 626 627 628 629
#> [595] 630 631 632 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647
#> [613] 648 649 650 651 652 653 654 655 656 657 658 659 660 661 662 663 664 665
#> [631] 666 667 668 669 670 671 672 673 674 675 676 677 678 679 680 681 682 683
#> [649] 684 685 686 687 688 689 690 691 692 693 694 696 697 698 700 701 702 703
#> [667] 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721
#> [685] 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739
#> [703] 740 741 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[2]]
#>   [1]   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  35  36  37  38  39  40  41
#>  [37]  42  43  44  45  46  47  48  49  50  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78
#>  [73]  79  80  81  82  83  85  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 101 102 103 104 105 106 107 108 109 110 111 113 114 115 116 117
#> [109] 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135
#> [127] 136 137 138 139 140 141 142 143 144 145 147 150 151 152 153 154 155 156
#> [145] 157 158 159 160 161 162 163 166 169 170 171 172 173 174 176 177 178 179
#> [163] 180 181 182 183 184 185 186 187 188 189 190 191 192 194 195 196 197 198
#> [181] 199 200 201 202 203 204 205 206 207 208 209 211 212 213 214 215 216 217
#> [199] 218 219 220 221 222 223 224 225 226 227 228 229 230 231 233 234 235 236
#> [217] 237 238 239 240 241 242 243 244 245 246 247 249 250 251 253 254 255 256
#> [235] 259 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276
#> [253] 277 278 279 280 281 282 283 284 285 287 288 289 290 291 292 293 296 297
#> [271] 298 299 300 301 302 303 304 305 306 308 309 311 312 313 315 316 317 318
#> [289] 319 320 321 322 323 325 326 327 328 329 331 332 333 334 335 336 337 338
#> [307] 339 340 341 343 344 345 346 348 349 350 352 353 354 355 356 357 358 359
#> [325] 360 361 362 363 364 365 366 367 368 369 370 371 372 373 375 376 377 378
#> [343] 380 381 382 383 384 385 386 388 389 390 391 392 393 394 395 396 397 398
#> [361] 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416
#> [379] 417 418 419 420 422 423 424 426 427 428 429 430 431 433 434 435 436 437
#> [397] 438 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456
#> [415] 457 458 459 460 461 462 463 464 465 466 467 468 470 471 472 473 474 475
#> [433] 476 477 478 479 480 481 482 483 484 486 487 488 489 490 492 493 494 495
#> [451] 496 497 498 499 500 501 502 503 504 505 506 508 509 510 511 512 513 514
#> [469] 515 516 518 519 520 521 522 523 524 525 526 527 528 529 530 531 532 533
#> [487] 534 535 536 537 538 539 540 541 542 543 544 545 546 547 548 549 550 551
#> [505] 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566 567 568 569
#> [523] 570 571 573 574 575 576 578 579 580 581 582 583 584 585 586 587 588 589
#> [541] 590 591 592 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607
#> [559] 608 610 611 612 613 614 615 616 617 618 619 620 621 622 623 625 627 628
#> [577] 630 631 633 634 635 636 637 638 639 640 641 642 643 644 646 648 649 650
#> [595] 651 652 653 654 655 656 657 658 659 661 662 663 664 665 666 667 668 669
#> [613] 670 671 672 673 674 675 676 677 679 680 681 682 683 684 685 686 687 688
#> [631] 689 690 691 692 693 694 695 696 697 698 700 701 702 703 704 705 706 707
#> [649] 708 709 710 711 712 713 714 715 716 717 718 719 720 721 722 723 724 725
#> [667] 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740 741 743 744
#> [685] 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   2   3   4   5   7   8   9  10  11  12  13  14  15  17  18  19  20  21
#>  [19]  22  24  25  26  27  28  29  30  31  32  35  36  37  38  39  40  41  42
#>  [37]  43  44  45  46  47  48  49  50  52  53  54  55  57  58  59  60  61  62
#>  [55]  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80
#>  [73]  81  82  83  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 101 102 103 104 105 106 108 109 110 111 112 113 114 115 116 117 118 119
#> [109] 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137
#> [127] 138 139 140 141 142 143 144 145 147 150 151 152 153 154 155 156 157 158
#> [145] 159 160 161 162 163 165 166 169 170 171 172 173 176 177 178 179 180 181
#> [163] 182 183 184 185 186 187 188 189 190 191 192 194 195 196 197 198 199 200
#> [181] 201 202 203 204 205 206 207 208 209 211 213 214 215 216 217 218 219 220
#> [199] 221 222 223 224 225 226 227 228 229 230 231 233 234 235 236 237 238 239
#> [217] 240 241 242 243 244 245 246 247 249 250 251 253 254 255 256 257 259 260
#> [235] 261 262 263 265 266 267 268 269 270 271 272 274 275 276 277 278 279 280
#> [253] 281 282 283 284 285 287 288 289 290 291 292 293 296 297 298 299 300 301
#> [271] 303 304 305 306 308 309 311 313 315 317 318 319 320 321 322 323 325 326
#> [289] 327 328 329 331 332 333 334 335 336 337 338 339 340 341 343 344 345 346
#> [307] 348 349 350 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366
#> [325] 367 368 370 371 372 373 375 376 377 378 380 381 382 383 384 385 386 388
#> [343] 389 390 391 392 393 394 395 396 397 398 399 400 401 402 403 404 405 406
#> [361] 407 408 409 410 411 412 413 414 415 416 417 418 419 422 423 424 426 427
#> [379] 428 429 430 431 433 434 435 436 437 438 440 441 442 443 444 446 448 449
#> [397] 450 451 452 453 454 455 456 457 458 460 461 462 463 464 465 466 467 468
#> [415] 470 471 472 473 474 475 476 477 478 479 480 481 483 484 486 487 488 489
#> [433] 490 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506 508 509
#> [451] 510 511 512 513 514 516 518 519 520 521 522 523 524 525 526 527 528 529
#> [469] 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547
#> [487] 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565
#> [505] 566 567 568 569 570 571 573 574 575 576 578 579 580 581 582 583 584 585
#> [523] 586 587 588 589 591 592 593 594 595 596 597 598 599 600 601 602 603 604
#> [541] 605 606 607 608 610 611 612 613 614 615 616 617 618 619 620 621 622 623
#> [559] 624 625 627 628 630 631 633 634 635 636 637 638 639 640 641 642 643 644
#> [577] 646 648 649 650 651 652 653 654 655 656 657 658 659 661 662 663 664 665
#> [595] 666 667 668 669 670 671 672 673 674 675 676 677 679 680 681 682 683 684
#> [613] 685 686 687 688 689 690 691 692 693 694 695 696 697 698 700 701 702 703
#> [631] 704 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721
#> [649] 722 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739
#> [667] 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> 
#>