(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

(sperrorest) Repeated spatial "disc" resampling

Parameters

repeats (integer(1))
Number of repeats.

References

Brenning A (2012). “Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest.” In 2012 IEEE International Geoscience and Remote Sensing Symposium. doi:10.1109/igarss.2012.6352393 .

Super class

mlr3::Resampling -> ResamplingRepeatedSpCVDisc

Active bindings

iters

integer(1)
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedSpCVDisc$new()

ResamplingRepeatedSpCVDisc$folds()

ResamplingRepeatedSpCVDisc$repeats()

ResamplingRepeatedSpCVDisc$instantiate()

ResamplingRepeatedSpCVDisc$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Create a "Spatial 'Disc' resampling" resampling instance.

For a list of available arguments, please see sperrorest::partition_disc.

Usage

ResamplingRepeatedSpCVDisc$new(id = "repeated_spcv_disc")

Arguments

id

character(1)
Identifier for the resampling strategy.

Method folds()

Translates iteration numbers to fold number.

Usage

ResamplingRepeatedSpCVDisc$folds(iters)

Arguments

iters

integer()
Iteration number.

Method repeats()

Translates iteration numbers to repetition number.

Usage

ResamplingRepeatedSpCVDisc$repeats(iters)

Arguments

iters

integer()
Iteration number.

Method instantiate()

Materializes fixed training and test splits for a given task.

Usage

ResamplingRepeatedSpCVDisc$instantiate(task)

Arguments

task

Task
A task to instantiate.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedSpCVDisc$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rrcv = rsmp("repeated_spcv_disc",
  folds = 3L, repeats = 2,
  radius = 200L, buffer = 200L)
rrcv$instantiate(task)

# Individual sets:
rrcv$iters
#> [1] 6
rrcv$folds(1:6)
#> [1] 1 2 3 1 2 3
rrcv$repeats(1:6)
#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)
#>   [1]   1   2   3   4   5   6   7   8   9  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  40  41  42  43  44  46  47  48  49  50  51  52  53  54  55  56  58
#>  [55]  59  60  62  63  64  65  66  67  69  70  71  72  73  74  75  76  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 100 101 102 103 104 105 107 108 109 110 111 112 113 114 116 117
#> [109] 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135
#> [127] 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153
#> [145] 154 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172
#> [163] 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 191
#> [181] 192 193 194 195 196 197 198 199 200 202 203 204 205 206 207 208 209 210
#> [199] 211 212 213 214 215 216 217 218 220 221 222 223 224 225 226 227 228 229
#> [217] 230 231 232 233 234 235 236 237 238 239 240 241 242 244 245 246 247 248
#> [235] 249 250 251 252 253 254 255 256 257 258 259 260 261 262 264 265 266 267
#> [253] 268 269 270 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285
#> [271] 286 287 288 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304
#> [289] 305 306 307 308 309 310 311 312 313 314 315 316 318 319 320 321 322 323
#> [307] 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341
#> [325] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [343] 360 361 362 363 364 365 366 367 368 369 370 371 373 374 375 376 377 378
#> [361] 379 380 381 382 383 384 386 387 388 389 390 391 392 393 394 395 396 397
#> [379] 398 399 400 401 402 403 404 406 407 408 409 410 411 412 413 415 416 417
#> [397] 418 419 420 421 422 423 424 425 426 428 429 430 431 432 433 434 435 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 454 455
#> [433] 456 457 458 459 460 461 462 463 465 466 467 468 469 470 471 472 473 474
#> [451] 475 476 477 478 479 480 481 482 483 484 485 486 488 489 490 491 492 493
#> [469] 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511
#> [487] 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529
#> [505] 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547
#> [523] 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565
#> [541] 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [559] 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601
#> [577] 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [595] 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637
#> [613] 638 639 640 641 643 644 645 646 647 648 650 651 652 653 654 655 656 657
#> [631] 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 674 675 676
#> [649] 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694
#> [667] 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 711 712 713
#> [685] 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731
#> [703] 732 733 734 735 737 738 739 740 741 742 743 744 745 746 747 748 749 750
#> [721] 751
rrcv$test_set(1)
#>  [1]  10  39  57  61  68  77 106 115 155 190 201 219 243 263 289 317 372 405 414
#> [20] 427 487 642 673
intersect(rrcv$train_set(1), rrcv$test_set(1))
#> integer(0)

# Internal storage:
rrcv$instance # table
#> [[1]]
#> [[1]]$test
#> [[1]]$test[[1]]
#>  [1]  10  39  57  61  68  77 106 115 155 190 201 219 243 263 289 317 372 405 414
#> [20] 427 487 642 673
#> 
#> [[1]]$test[[2]]
#> [1]  86  91 416 619 720
#> 
#> [[1]]$test[[3]]
#> [1] 170 218 290 356 503 585 644
#> 
#> 
#> [[1]]$train
#> [[1]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  40  41  42  43  44  46  47  48  49  50  51  52  53  54  55  56  58
#>  [55]  59  60  62  63  64  65  66  67  69  70  71  72  73  74  75  76  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 100 101 102 103 104 105 107 108 109 110 111 112 113 114 116 117
#> [109] 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135
#> [127] 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153
#> [145] 154 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172
#> [163] 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 191
#> [181] 192 193 194 195 196 197 198 199 200 202 203 204 205 206 207 208 209 210
#> [199] 211 212 213 214 215 216 217 218 220 221 222 223 224 225 226 227 228 229
#> [217] 230 231 232 233 234 235 236 237 238 239 240 241 242 244 245 246 247 248
#> [235] 249 250 251 252 253 254 255 256 257 258 259 260 261 262 264 265 266 267
#> [253] 268 269 270 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285
#> [271] 286 287 288 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304
#> [289] 305 306 307 308 309 310 311 312 313 314 315 316 318 319 320 321 322 323
#> [307] 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341
#> [325] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [343] 360 361 362 363 364 365 366 367 368 369 370 371 373 374 375 376 377 378
#> [361] 379 380 381 382 383 384 386 387 388 389 390 391 392 393 394 395 396 397
#> [379] 398 399 400 401 402 403 404 406 407 408 409 410 411 412 413 415 416 417
#> [397] 418 419 420 421 422 423 424 425 426 428 429 430 431 432 433 434 435 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 454 455
#> [433] 456 457 458 459 460 461 462 463 465 466 467 468 469 470 471 472 473 474
#> [451] 475 476 477 478 479 480 481 482 483 484 485 486 488 489 490 491 492 493
#> [469] 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510 511
#> [487] 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528 529
#> [505] 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547
#> [523] 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564 565
#> [541] 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [559] 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601
#> [577] 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [595] 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637
#> [613] 638 639 640 641 643 644 645 646 647 648 650 651 652 653 654 655 656 657
#> [631] 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 674 675 676
#> [649] 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693 694
#> [667] 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 711 712 713
#> [685] 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729 730 731
#> [703] 732 733 734 735 737 738 739 740 741 742 743 744 745 746 747 748 749 750
#> [721] 751
#> 
#> [[1]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  13  14  15  16  18  19  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74
#>  [73]  75  76  77  78  79  80  81  82  83  84  85  87  88  89  90  92  93  94
#>  [91]  95  96  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113
#> [109] 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128 129 130 131
#> [127] 132 133 134 135 136 137 138 139 141 142 143 144 145 146 147 148 149 150
#> [145] 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168
#> [163] 169 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186
#> [181] 187 188 189 190 191 192 193 194 195 197 198 199 200 201 202 203 204 205
#> [199] 206 207 208 209 210 211 212 213 214 215 216 217 218 219 220 221 222 223
#> [217] 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 241 243
#> [235] 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 260 261
#> [253] 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 278 279 280
#> [271] 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298
#> [289] 299 300 301 302 303 304 306 307 308 309 310 311 312 313 314 315 316 317
#> [307] 318 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335
#> [325] 336 337 338 339 340 341 342 343 344 345 346 347 348 349 350 351 352 353
#> [343] 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371
#> [361] 372 373 374 375 376 377 378 379 380 381 382 383 385 386 387 388 389 390
#> [379] 391 392 393 394 395 396 397 398 399 400 401 402 403 404 405 406 407 408
#> [397] 409 410 411 412 413 414 415 417 418 419 420 421 422 423 424 425 426 427
#> [415] 428 429 430 431 432 433 434 435 436 437 438 439 440 441 442 443 444 445
#> [433] 446 447 448 449 450 451 452 453 454 455 456 457 458 459 460 461 462 463
#> [451] 464 465 466 467 468 469 470 471 472 473 474 475 476 477 478 479 480 481
#> [469] 482 483 484 485 486 487 488 489 490 491 492 493 494 495 496 497 498 499
#> [487] 500 501 502 503 504 505 506 507 508 509 510 511 513 514 515 516 517 518
#> [505] 519 520 521 522 523 524 525 526 527 528 529 530 531 532 533 534 535 536
#> [523] 537 538 539 540 541 542 543 544 545 546 547 548 549 550 551 552 553 554
#> [541] 555 556 557 558 559 560 561 562 564 565 566 567 568 569 570 571 572 573
#> [559] 574 575 576 577 578 579 581 582 583 584 585 586 587 588 589 590 591 592
#> [577] 593 594 595 596 597 598 599 600 601 602 603 604 605 606 607 608 609 610
#> [595] 611 612 613 614 615 616 617 618 620 621 622 623 624 625 626 627 628 629
#> [613] 630 632 633 634 635 636 637 639 640 641 642 643 644 645 646 647 648 649
#> [631] 650 652 653 654 655 656 657 658 659 660 661 662 663 664 665 666 667 668
#> [649] 669 670 671 672 673 674 675 676 677 678 679 680 682 683 684 685 686 687
#> [667] 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704 705
#> [685] 706 707 708 709 710 711 712 713 714 715 716 717 718 719 721 722 723 724
#> [703] 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740 741 742
#> [721] 743 744 745 746 748 749 750 751
#> 
#> [[1]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91
#>  [91]  92  93  94  95  96  97  98  99 100 101 102 104 105 106 107 108 109 110
#> [109] 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128
#> [127] 129 130 131 132 133 134 135 137 138 139 140 141 142 143 144 145 146 147
#> [145] 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165
#> [163] 166 167 168 169 171 172 173 174 175 176 177 178 179 180 181 182 183 184
#> [181] 185 186 187 188 189 190 191 192 193 194 195 196 197 198 199 200 201 202
#> [199] 203 204 205 206 207 208 209 210 211 212 213 214 215 217 219 220 221 223
#> [217] 224 225 226 227 228 229 230 231 232 233 235 236 237 238 239 240 241 242
#> [235] 243 244 246 247 248 249 250 251 252 253 254 256 257 258 259 260 261 262
#> [253] 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 280
#> [271] 281 282 283 284 285 286 287 288 289 291 292 293 294 295 296 297 298 299
#> [289] 300 301 302 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318
#> [307] 319 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336
#> [325] 337 338 339 340 341 342 343 345 346 347 348 349 350 351 352 353 354 355
#> [343] 357 358 359 360 361 363 364 365 366 367 368 369 370 371 372 373 374 375
#> [361] 376 377 378 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393
#> [379] 394 395 396 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411
#> [397] 412 414 415 416 417 418 419 420 421 422 423 424 425 427 428 429 430 431
#> [415] 432 433 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449
#> [433] 450 451 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467
#> [451] 468 469 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485
#> [469] 487 488 489 490 491 492 493 494 495 496 497 498 499 500 501 504 505 506
#> [487] 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524
#> [505] 525 526 527 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542
#> [523] 543 545 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561
#> [541] 562 563 564 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579
#> [559] 580 581 582 583 584 586 587 588 589 590 591 592 593 595 596 597 598 599
#> [577] 600 601 602 603 604 605 606 607 608 609 610 611 613 614 616 617 618 619
#> [595] 620 621 622 623 624 625 626 627 628 629 630 631 632 633 635 636 637 638
#> [613] 639 640 641 642 643 645 646 647 648 649 650 651 652 653 654 655 656 657
#> [631] 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674 675
#> [649] 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693
#> [667] 694 695 696 697 698 699 700 701 702 703 704 705 708 709 710 711 712 713
#> [685] 714 715 716 717 718 719 720 721 722 723 725 726 727 728 729 730 731 732
#> [703] 733 734 735 736 737 738 739 740 741 742 743 744 745 746 747 748 749 750
#> [721] 751
#> 
#> 
#> 
#> [[2]]
#> [[2]]$test
#> [[2]]$test[[1]]
#>  [1]  74  81 159 160 179 189 276 315 390 436 551
#> 
#> [[2]]$test[[2]]
#> [1] 535 566 734
#> 
#> [[2]]$test[[3]]
#>  [1]  10  39  57  61  68  77 106 115 155 190 201 219 243 263 289 317 372 385 405
#> [20] 414 427 487 642 673
#> 
#> 
#> [[2]]$train
#> [[2]]$train[[1]]
#>   [1]   1   2   3   4   5   6   7   8  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  31  32  33  34  35  36  37  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56
#>  [55]  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73  75
#>  [73]  76  77  78  79  80  82  83  84  85  86  87  88  89  90  91  93  94  95
#>  [91]  96  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113
#> [109] 114 115 116 117 118 120 121 122 123 124 125 126 127 128 129 130 131 132
#> [127] 133 134 135 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150
#> [145] 151 152 153 154 155 156 157 158 161 162 163 164 165 166 167 168 169 170
#> [163] 171 172 174 175 176 177 178 180 181 182 183 184 185 188 190 191 192 193
#> [181] 195 196 197 198 199 200 201 202 203 204 206 207 209 210 211 212 213 214
#> [199] 215 216 217 218 219 220 221 222 223 224 225 226 227 229 230 231 232 233
#> [217] 234 235 236 238 239 240 241 242 243 245 246 247 248 249 250 251 252 253
#> [235] 255 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271 272
#> [253] 273 274 275 277 278 279 280 281 282 283 284 285 286 287 288 289 290 294
#> [271] 295 296 297 298 299 300 301 302 303 305 306 307 308 309 310 311 312 313
#> [289] 314 316 317 318 319 320 321 322 324 325 326 327 328 329 330 331 332 334
#> [307] 336 337 338 341 342 343 344 345 346 347 348 349 350 351 352 353 354 356
#> [325] 357 358 359 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374
#> [343] 375 376 377 378 379 381 382 383 384 385 387 388 389 391 392 393 394 395
#> [361] 396 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413
#> [379] 414 415 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431
#> [397] 432 433 434 435 438 439 440 441 442 444 445 446 447 448 449 450 451 452
#> [415] 453 455 456 458 459 460 461 462 463 464 465 466 467 468 469 470 471 472
#> [433] 473 474 475 476 477 478 480 481 482 483 484 485 486 487 488 489 490 491
#> [451] 492 493 494 495 496 497 498 499 500 501 502 503 504 505 507 508 509 510
#> [469] 511 512 513 514 515 516 517 518 519 520 521 523 524 525 526 527 528 529
#> [487] 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546 547
#> [505] 548 549 550 552 553 554 555 556 557 558 559 560 561 562 563 564 565 566
#> [523] 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583 584
#> [541] 585 586 588 589 590 591 592 594 595 596 598 599 600 601 602 603 604 605
#> [559] 606 607 608 609 610 611 612 613 614 615 617 618 619 620 621 622 623 624
#> [577] 626 627 628 629 630 631 632 633 634 635 636 637 638 639 640 641 642 643
#> [595] 644 645 646 647 648 649 650 651 652 653 654 655 656 657 658 659 660 661
#> [613] 662 663 664 665 666 667 668 669 670 671 673 674 675 676 677 678 679 680
#> [631] 681 682 683 684 686 687 688 689 690 691 692 693 694 695 696 697 698 699
#> [649] 700 701 702 703 704 705 706 707 708 709 710 711 712 713 714 715 716 718
#> [667] 719 720 721 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737
#> [685] 738 739 740 741 742 743 744 745 746 747 748 749 750 751
#> 
#> [[2]]$train[[2]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  45  46  47  48  49  50  51  52  53  54  55
#>  [55]  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72  73
#>  [73]  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91
#>  [91]  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108 109
#> [109] 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127
#> [127] 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145
#> [145] 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163
#> [163] 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181
#> [181] 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198 199
#> [199] 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217
#> [217] 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234 235
#> [235] 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253
#> [253] 254 255 256 257 258 260 261 262 263 264 265 266 267 268 269 270 271 272
#> [271] 273 274 275 276 277 278 279 280 281 282 283 284 285 286 287 288 289 290
#> [289] 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308
#> [307] 309 310 311 312 313 314 315 316 317 318 319 320 321 322 323 324 325 326
#> [325] 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344
#> [343] 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362
#> [361] 363 364 366 367 368 369 370 371 372 373 374 375 376 377 378 379 380 381
#> [379] 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398 399
#> [397] 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417
#> [415] 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434 435
#> [433] 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453
#> [451] 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470 471
#> [469] 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488 489
#> [487] 490 491 492 493 495 496 497 498 499 500 501 502 503 504 505 506 507 508
#> [505] 509 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526
#> [523] 527 528 529 530 531 532 533 534 536 537 539 540 541 542 543 544 545 546
#> [541] 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564
#> [559] 565 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582 583
#> [577] 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600 601
#> [595] 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618 619
#> [613] 620 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636 637 638
#> [631] 639 640 641 642 643 644 645 646 647 648 649 650 651 652 653 654 655 656
#> [649] 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 673 674
#> [667] 675 676 677 678 679 680 681 682 683 685 686 687 688 689 690 691 692 693
#> [685] 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711
#> [703] 712 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729
#> [721] 730 731 732 733 735 736 737 738 739 740 741 742 743 744 745 746 747 748
#> [739] 749 750 751
#> 
#> [[2]]$train[[3]]
#>   [1]   1   2   3   4   5   6   7   8   9  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  40  41  42  43  44  46  47  48  49  50  51  52  53  54  55  56  58
#>  [55]  59  60  62  63  64  65  66  67  69  70  71  72  73  74  75  76  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97
#>  [91]  98  99 100 101 102 103 104 105 107 108 109 110 111 112 113 114 116 117
#> [109] 118 119 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135
#> [127] 136 137 138 139 140 141 142 143 144 145 146 147 148 149 150 151 152 153
#> [145] 154 156 157 158 159 160 161 162 163 164 165 166 167 168 169 170 171 172
#> [163] 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188 189 191
#> [181] 192 193 194 195 196 197 198 199 200 202 203 204 205 206 207 208 209 210
#> [199] 211 212 213 214 215 216 217 218 220 221 222 223 224 225 226 227 228 229
#> [217] 230 231 232 233 234 235 236 237 238 239 240 241 242 244 245 246 247 248
#> [235] 249 250 251 252 253 254 255 256 257 258 259 260 261 262 264 265 266 267
#> [253] 268 269 270 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285
#> [271] 286 287 288 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304
#> [289] 305 306 307 308 309 310 311 312 313 314 315 316 318 319 320 321 322 323
#> [307] 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341
#> [325] 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359
#> [343] 360 361 362 363 364 365 366 367 368 369 370 371 373 374 375 376 377 378
#> [361] 379 380 381 382 383 384 386 387 388 389 390 391 392 393 394 395 396 397
#> [379] 398 399 400 401 402 403 404 406 407 408 409 410 411 412 413 415 416 417
#> [397] 418 419 420 421 422 423 424 425 426 428 429 430 431 432 433 434 435 436
#> [415] 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453 454
#> [433] 455 456 457 458 459 460 461 462 463 465 466 467 468 469 470 471 472 473
#> [451] 474 475 476 477 478 479 480 481 482 483 484 485 486 488 489 490 491 492
#> [469] 493 494 495 496 497 498 499 500 501 502 503 504 505 506 507 508 509 510
#> [487] 511 512 513 514 515 516 517 518 519 520 521 522 523 524 525 526 527 528
#> [505] 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546
#> [523] 547 548 549 550 551 552 553 554 555 556 557 558 559 560 561 562 563 564
#> [541] 565 566 567 568 569 570 571 572 573 574 575 576 577 578 579 580 581 582
#> [559] 583 584 585 586 587 588 589 590 591 592 593 594 595 596 597 598 599 600
#> [577] 601 602 603 604 605 606 607 608 609 610 611 612 613 614 615 616 617 618
#> [595] 619 620 621 622 623 624 625 626 627 628 629 630 631 632 633 634 635 636
#> [613] 637 638 639 640 641 643 644 645 646 647 648 650 651 652 653 654 655 656
#> [631] 657 658 659 660 661 662 663 664 665 666 667 668 669 670 671 672 674 675
#> [649] 676 677 678 679 680 681 682 683 684 685 686 687 688 689 690 691 692 693
#> [667] 694 695 696 697 698 699 700 701 702 703 704 705 706 707 708 709 710 711
#> [685] 712 713 714 715 716 717 718 719 720 721 722 723 724 725 726 727 728 729
#> [703] 730 731 732 733 734 735 737 738 739 740 741 742 743 744 745 746 747 748
#> [721] 749 750 751
#> 
#> 
#>